% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_access_control_proxy.R
\docType{data}
\name{account_access_control_proxy}
\alias{account_access_control_proxy}
\title{These APIs manage access rules on resources in an account. Currently, only
grant rules are supported. A grant rule specifies a role assigned to a set of
principals. A list of rules attached to a resource is called a rule set. A
workspace must belong to an account for these APIs to work.}
\format{
An object of class \code{list} of length 3.
}
\usage{
account_access_control_proxy
}
\description{
These APIs manage access rules on resources in an account. Currently, only
grant rules are supported. A grant rule specifies a role assigned to a set of
principals. A list of rules attached to a resource is called a rule set. A
workspace must belong to an account for these APIs to work.
}
\section{Operations}{

\tabular{ll}{
\link[=account_access_control_proxy_get_assignable_roles_for_resource]{get_assignable_roles_for_resource} \tab Get assignable roles for a resource.\cr
\link[=account_access_control_proxy_get_rule_set]{get_rule_set} \tab Get a rule set.\cr
\link[=account_access_control_proxy_update_rule_set]{update_rule_set} \tab Update a rule set.\cr
}
}

\keyword{datasets}
