% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_access_control_proxy.R
\name{account_access_control_proxy_get_assignable_roles_for_resource}
\alias{account_access_control_proxy_get_assignable_roles_for_resource}
\title{Get assignable roles for a resource.}
\usage{
account_access_control_proxy_get_assignable_roles_for_resource(resource)
}
\arguments{
\item{resource}{Required. The resource name for which assignable roles will be listed.}
}
\description{
Gets all the roles that can be granted on an account-level resource. A role
is grantable if the rule set on the resource can contain an access rule of
the role.
}
\keyword{internal}
