% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_access_control_proxy.R
\name{account_access_control_proxy_get_rule_set}
\alias{account_access_control_proxy_get_rule_set}
\title{Get a rule set.}
\usage{
account_access_control_proxy_get_rule_set(name, etag)
}
\arguments{
\item{name}{Required. The ruleset name associated with the request.}

\item{etag}{Required. Etag used for versioning.}
}
\description{
Get a rule set by its name. A rule set is always attached to a resource and
contains a list of access rules on the said resource. Currently only a
default rule set for each resource is supported.
}
\keyword{internal}
