% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_access_control_proxy.R
\name{account_access_control_proxy_update_rule_set}
\alias{account_access_control_proxy_update_rule_set}
\title{Update a rule set.}
\usage{
account_access_control_proxy_update_rule_set(name, rule_set)
}
\arguments{
\item{name}{Required. Name of the rule set.}

\item{rule_set}{Required.}
}
\description{
Replace the rules of a rule set. First, use a GET rule set request to read
the current version of the rule set before modifying it. This pattern helps
prevent conflicts between concurrent updates.
}
\keyword{internal}
