% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance_profiles.R
\name{add_instance_profile}
\alias{add_instance_profile}
\title{Register an instance profile.}
\usage{
add_instance_profile(
  client,
  instance_profile_arn,
  iam_role_arn = NULL,
  is_meta_instance_profile = NULL,
  skip_validation = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{instance_profile_arn}{Required. The AWS ARN of the instance profile to register with Databricks.}

\item{iam_role_arn}{The AWS IAM role ARN of the role associated with the instance profile.}

\item{is_meta_instance_profile}{Boolean flag indicating whether the instance profile should only be used in credential passthrough scenarios.}

\item{skip_validation}{By default, Databricks validates that it has sufficient permissions to launch instances with the instance profile.}
}
\description{
In the UI, you can select the instance profile when launching clusters. This
API is only available to admin users.
}
