% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\docType{data}
\name{alerts}
\alias{alerts}
\title{The alerts API can be used to perform CRUD operations on alerts. An alert is
a Databricks SQL object that periodically runs a query, evaluates a condition
of its result, and notifies one or more users and/or notification
destinations if the condition was met. Alerts can be scheduled using the
\code{sql_task} type of the Jobs API, e.g. :method:jobs/create.}
\format{
An object of class \code{list} of length 5.
}
\usage{
alerts
}
\description{
The alerts API can be used to perform CRUD operations on alerts. An alert is
a Databricks SQL object that periodically runs a query, evaluates a condition
of its result, and notifies one or more users and/or notification
destinations if the condition was met. Alerts can be scheduled using the
\code{sql_task} type of the Jobs API, e.g. :method:jobs/create.
}
\section{Operations}{

\tabular{ll}{
\link[=alerts_create]{create} \tab Create an alert.\cr
\link[=alerts_delete]{delete} \tab Delete an alert.\cr
\link[=alerts_get]{get} \tab Get an alert.\cr
\link[=alerts_list]{list} \tab Get alerts.\cr
\link[=alerts_update]{update} \tab Update an alert.\cr
}
}

\keyword{datasets}
