% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{alerts_create}
\alias{alerts_create}
\title{Create an alert.}
\usage{
alerts_create(name, options, query_id, parent = NULL, rearm = NULL)
}
\arguments{
\item{name}{Required. Name of the alert.}

\item{options}{Required. Alert configuration options.}

\item{query_id}{Required. ID of the query evaluated by the alert.}

\item{parent}{The identifier of the workspace folder containing the alert.}

\item{rearm}{Number of seconds after being triggered before the alert rearms itself and can be triggered again.}
}
\description{
Creates an alert. An alert is a Databricks SQL object that periodically runs
a query, evaluates a condition of its result, and notifies users or
notification destinations if the condition was met.
}
\keyword{internal}
