% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{alerts_create_schedule}
\alias{alerts_create_schedule}
\title{\link{DEPRECATED} Create a refresh schedule.}
\usage{
alerts_create_schedule(cron, alert_id, data_source_id = NULL)
}
\arguments{
\item{cron}{Required. Cron string representing the refresh schedule.}

\item{alert_id}{Required.}

\item{data_source_id}{ID of the SQL warehouse to refresh with.}
}
\description{
Creates a new refresh schedule for an alert.
}
\details{
\strong{Note:} The structure of refresh schedules is subject to change.

\strong{Note:} This API is deprecated: Use :method:jobs/create to create a job
with the alert.
}
\keyword{internal}
