% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{alerts_delete}
\alias{alerts_delete}
\title{Delete an alert.}
\usage{
alerts_delete(alert_id, ...)
}
\arguments{
\item{alert_id}{Required.}
}
\description{
Deletes an alert. Deleted alerts are no longer accessible and cannot be
restored. \strong{Note:} Unlike queries and dashboards, alerts cannot be moved to
the trash.
}
\keyword{internal}
