% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{alerts_delete_schedule}
\alias{alerts_delete_schedule}
\title{\link{DEPRECATED} Delete a refresh schedule.}
\usage{
alerts_delete_schedule(alert_id, schedule_id)
}
\arguments{
\item{alert_id}{Required.}

\item{schedule_id}{Required.}
}
\description{
Deletes an alert's refresh schedule. The refresh schedule specifies when to
refresh and evaluate the associated query result.
}
\details{
\strong{Note:} This API is deprecated: Use :method:jobs/delete to delete a job for
the alert.
}
\keyword{internal}
