% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{alerts_get_subscriptions}
\alias{alerts_get_subscriptions}
\title{\link{DEPRECATED} Get an alert's subscriptions.}
\usage{
alerts_get_subscriptions(alert_id, ...)
}
\arguments{
\item{alert_id}{\link{required}}
}
\description{
Get the subscriptions for an alert. An alert subscription represents exactly
one recipient being notified whenever the alert is triggered. The alert
recipient is specified by either the \code{user} field or the \code{destination} field.
The \code{user} field is ignored if \code{destination} is non-\code{null}.
}
\details{
\strong{Note:} This API is deprecated: Use :method:jobs/get to get the
subscriptions associated with a job for an alert.
}
\keyword{internal}
