% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{alerts_list_schedules}
\alias{alerts_list_schedules}
\title{\link{DEPRECATED} Get refresh schedules.}
\usage{
alerts_list_schedules(alert_id, ...)
}
\arguments{
\item{alert_id}{\link{required}}
}
\description{
Gets the refresh schedules for the specified alert. Alerts can have refresh
schedules that specify when to refresh and evaluate the associated query
result.
}
\details{
\strong{Note:} Although refresh schedules are returned in a list, only one refresh
schedule per alert is currently supported. The structure of refresh schedules
is subject to change.

\strong{Note:} This API is deprecated: Use :method:jobs/list to list jobs and
filter by the alert.
}
\keyword{internal}
