% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{alerts_subscribe}
\alias{alerts_subscribe}
\title{\link{DEPRECATED} Subscribe to an alert.}
\usage{
alerts_subscribe(alert_id, destination_id = NULL, user_id = NULL, ...)
}
\arguments{
\item{alert_id}{\link{required} ID of the alert.}

\item{destination_id}{ID of the alert subscriber (if subscribing an alert destination).}

\item{user_id}{ID of the alert subscriber (if subscribing a user).}
}
\description{
\strong{Note:} This API is deprecated: Use :method:jobs/update to subscribe to a
job for an alert.
}
\keyword{internal}
