% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{alerts_unsubscribe}
\alias{alerts_unsubscribe}
\title{\link{DEPRECATED} Unsubscribe to an alert.}
\usage{
alerts_unsubscribe(alert_id, subscription_id, ...)
}
\arguments{
\item{alert_id}{\link{required}}

\item{subscription_id}{\link{required}}
}
\description{
Unsubscribes a user or a destination to an alert.
}
\details{
\strong{Note:} This API is deprecated: Use :method:jobs/update to unsubscribe to a
job for an alert.
}
\keyword{internal}
