% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraries.R
\name{all_cluster_library_statuses}
\alias{all_cluster_library_statuses}
\title{Get all statuses.}
\usage{
all_cluster_library_statuses(client)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}
}
\description{
Get the status of all libraries on all clusters. A status will be available
for all libraries installed on this cluster via the API or the libraries UI
as well as libraries set to be installed on all clusters via the libraries
UI.
}
