% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{appsDeleteApp}
\alias{appsDeleteApp}
\title{Delete an application.}
\usage{
appsDeleteApp(client, name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. The name of an application.}
}
\description{
Delete an application definition
}
