% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{appsGetApps}
\alias{appsGetApps}
\title{List all applications.}
\usage{
appsGetApps(client)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}
}
\description{
List all available applications
}
