% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{appsGetEvents}
\alias{appsGetEvents}
\title{Get deployment events for an application.}
\usage{
appsGetEvents(client, name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. The name of an application.}
}
\description{
Get deployment events for an application
}
