% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command_execution.R
\name{cancel_command_execution}
\alias{cancel_command_execution}
\alias{commandExecutionCancel}
\title{Cancel a command.}
\usage{
cancel_command_execution(
  client,
  cluster_id = NULL,
  command_id = NULL,
  context_id = NULL
)

commandExecutionCancel(
  client,
  cluster_id = NULL,
  command_id = NULL,
  context_id = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{cluster_id}{This field has no description yet.}

\item{command_id}{This field has no description yet.}

\item{context_id}{This field has no description yet.

See the \href{https://docs.databricks.com/api/workspace/commandexecution/cancel}{API documentation} for more
information on the request and response field structure.}
}
\description{
Cancels a currently running command within an execution context.
}
\details{
The command ID is obtained from a prior successful call to \strong{execute}.
}
