% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{cancel_job_run}
\alias{cancel_job_run}
\alias{jobsCancelRun}
\title{Cancel a run.}
\usage{
cancel_job_run(client, run_id)

jobsCancelRun(client, run_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{run_id}{Required. This field is required.

See the \href{https://docs.databricks.com/api/workspace/jobs/cancelrun}{API documentation} for more
information on the request and response field structure.}
}
\description{
Cancels a job run or a task run. The run is canceled asynchronously, so it
may still be running when this request completes.
}
