% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{cancel_job_run_and_wait}
\alias{cancel_job_run_and_wait}
\title{Cancel a run.}
\usage{
cancel_job_run_and_wait(client, run_id, timeout = 20, callback = cli_reporter)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{run_id}{Required. This field is required.}

\item{timeout}{Time to wait for the operation to complete in minutes.}

\item{callback}{Function to report the status of the operation. By default, it reports to console.}
}
\description{
This is a long-running operation, which blocks until Jobs on Databricks reach
TERMINATED or SKIPPED state with the timeout of 20 minutes, that you can change via \code{timeout} parameter.
By default, the state of Databricks Jobs is reported to console. You can change this behavior
by changing the \code{callback} parameter.
}
\details{
Cancels a job run or a task run. The run is canceled asynchronously, so it
may still be running when this request completes.
}
