% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakehouse_monitors.R
\name{cancel_lakehouse_monitor_refresh}
\alias{cancel_lakehouse_monitor_refresh}
\alias{lakehouseMonitorsCancelRefresh}
\title{Cancel refresh.}
\usage{
cancel_lakehouse_monitor_refresh(client, full_name, refresh_id)

lakehouseMonitorsCancelRefresh(client, full_name, refresh_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{full_name}{Required. Full name of the table.}

\item{refresh_id}{Required. ID of the refresh.

See the \href{https://docs.databricks.com/api/workspace/lakehousemonitors/cancelrefresh}{API documentation} for more
information on the request and response field structure.}
}
\description{
Cancel an active monitor refresh for the given refresh ID.
}
\details{
The caller must either: 1. be an owner of the table's parent catalog 2. have
\strong{USE_CATALOG} on the table's parent catalog and be an owner of the table's
parent schema 3. have the following permissions: - \strong{USE_CATALOG} on the
table's parent catalog - \strong{USE_SCHEMA} on the table's parent schema - be an
owner of the table

Additionally, the call must be made from the workspace where the monitor was
created.
}
