% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogs.R
\docType{data}
\name{catalogs}
\alias{catalogs}
\title{A catalog is the first layer of Unity Catalog’s three-level namespace.
It’s used to organize your data assets. Users can see all catalogs on which
they have been assigned the USE_CATALOG data permission.}
\format{
An object of class \code{list} of length 5.
}
\usage{
catalogs
}
\description{
In Unity Catalog, admins and data stewards manage users and their access to
data centrally across all of the workspaces in a Databricks account. Users in
different workspaces can share access to the same data, depending on
privileges granted centrally in Unity Catalog.
}
\section{Operations}{

\tabular{ll}{
\link[=catalogs_create]{create} \tab Create a catalog.\cr
\link[=catalogs_delete]{delete} \tab Delete a catalog.\cr
\link[=catalogs_get]{get} \tab Get a catalog.\cr
\link[=catalogs_list]{list} \tab List catalogs.\cr
\link[=catalogs_update]{update} \tab Update a catalog.\cr
}
}

\keyword{datasets}
