% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogs.R
\name{catalogs_create}
\alias{catalogs_create}
\title{Create a catalog.}
\usage{
catalogs_create(
  name,
  comment = NULL,
  properties = NULL,
  provider_name = NULL,
  share_name = NULL,
  storage_root = NULL
)
}
\arguments{
\item{name}{Required. Name of catalog.}

\item{comment}{User-provided free-form text description.}

\item{properties}{A map of key-value properties attached to the securable.}

\item{provider_name}{The name of delta sharing provider.}

\item{share_name}{The name of the share under the share provider.}

\item{storage_root}{Storage root URL for managed tables within catalog.}
}
\description{
Creates a new catalog instance in the parent metastore if the caller is a
metastore admin or has the \strong{CREATE_CATALOG} privilege.
}
\keyword{internal}
