% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogs.R
\name{catalogs_delete}
\alias{catalogs_delete}
\title{Delete a catalog.}
\usage{
catalogs_delete(name, force = NULL, ...)
}
\arguments{
\item{name}{\link{required} The name of the catalog.}

\item{force}{Force deletion even if the catalog is not empty.}
}
\description{
Deletes the catalog that matches the supplied name. The caller must be a
metastore admin or the owner of the catalog.
}
\keyword{internal}
