% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogs.R
\name{catalogs_get}
\alias{catalogs_get}
\title{Get a catalog.}
\usage{
catalogs_get(name, ...)
}
\arguments{
\item{name}{\link{required} The name of the catalog.}
}
\description{
Gets the specified catalog in a metastore. The caller must be a metastore
admin, the owner of the catalog, or a user that has the \strong{USE_CATALOG}
privilege set for their account.
}
\keyword{internal}
