% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogs.R
\name{catalogs_list}
\alias{catalogs_list}
\title{List catalogs.}
\usage{
catalogs_list(...)
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Gets an array of catalogs in the metastore. If the caller is the metastore
admin, all catalogs will be retrieved. Otherwise, only catalogs owned by the
caller (or for which the caller has the \strong{USE_CATALOG} privilege) will be
retrieved. There is no guarantee of a specific ordering of the elements in
the array.#'
}
\keyword{internal}
