% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_rooms.R
\name{cleanRoomsList}
\alias{cleanRoomsList}
\title{List clean rooms.}
\usage{
cleanRoomsList(client, max_results = NULL, page_token = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{max_results}{Maximum number of clean rooms to return.}

\item{page_token}{Opaque pagination token to go to next page based on previous query.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Gets an array of data object clean rooms from the metastore. The caller must
be a metastore admin or the owner of the clean room. There is no guarantee of
a specific ordering of the elements in the array.
}
