% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraries.R
\name{cluster_library_status}
\alias{cluster_library_status}
\alias{librariesClusterStatus}
\title{Get status.}
\usage{
cluster_library_status(client, cluster_id)

librariesClusterStatus(client, cluster_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{cluster_id}{Required. Unique identifier of the cluster whose status should be retrieved.}
}
\value{
\code{data.frame} with all of the response pages.

See the \href{https://docs.databricks.com/api/workspace/libraries/clusterstatus}{API documentation} for more
information on the request and response field structure.
}
\description{
Get the status of libraries on a cluster. A status will be available for all
libraries installed on this cluster via the API or the libraries UI as well
as libraries set to be installed on all clusters via the libraries UI. The
order of returned libraries will be as follows.
}
\details{
\enumerate{
\item Libraries set to be installed on this cluster will be returned first.
Within this group, the final order will be order in which the libraries were
added to the cluster.
\item Libraries set to be installed on all clusters are returned next. Within
this group there is no order guarantee.
\item Libraries that were previously requested on this cluster or on all
clusters, but now marked for removal. Within this group there is no order
guarantee.
}
}
