% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_policies.R
\docType{data}
\name{cluster_policies}
\alias{cluster_policies}
\title{Cluster policy limits the ability to configure clusters based on a set of
rules. The policy rules limit the attributes or attribute values available
for cluster creation. Cluster policies have ACLs that limit their use to
specific users and groups.}
\format{
An object of class \code{list} of length 5.
}
\usage{
cluster_policies
}
\description{
Cluster policies let you limit users to create clusters with prescribed
settings, simplify the user interface and enable more users to create their
own clusters (by fixing and hiding some values), control cost by limiting per
cluster maximum cost (by setting limits on attributes whose values contribute
to hourly price).
}
\details{
Cluster policy permissions limit which policies a user can select in the
Policy drop-down when the user creates a cluster: - A user who has cluster
create permission can select the Unrestricted policy and create
fully-configurable clusters. - A user who has both cluster create permission
and access to cluster policies can select the Unrestricted policy and
policies they have access to. - A user that has access to only cluster
policies, can select the policies they have access to.

If no policies have been created in the workspace, the Policy drop-down does
not display.

Only admin users can create, edit, and delete policies. Admin users also have
access to all policies.
}
\section{Operations}{

\tabular{ll}{
\link[=cluster_policies_create]{create} \tab Create a new policy.\cr
\link[=cluster_policies_delete]{delete} \tab Delete a cluster policy.\cr
\link[=cluster_policies_edit]{edit} \tab Update a cluster policy.\cr
\link[=cluster_policies_get]{get} \tab Get entity.\cr
\link[=cluster_policies_list]{list} \tab Get a cluster policy.\cr
}
}

\keyword{datasets}
