% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_policies.R
\name{cluster_policies_create}
\alias{cluster_policies_create}
\title{Create a new policy.}
\usage{
cluster_policies_create(
  name,
  definition = NULL,
  description = NULL,
  max_clusters_per_user = NULL,
  policy_family_definition_overrides = NULL,
  policy_family_id = NULL,
  ...
)
}
\arguments{
\item{name}{Required. Cluster Policy name requested by the user.}

\item{definition}{Policy definition document expressed in Databricks Cluster Policy Definition Language.}

\item{description}{Additional human-readable description of the cluster policy.}

\item{max_clusters_per_user}{Max number of clusters per user that can be active using this policy.}

\item{policy_family_definition_overrides}{Policy definition JSON document expressed in Databricks Policy Definition Language.}

\item{policy_family_id}{ID of the policy family.}
}
\description{
Creates a new policy with prescribed settings.
}
\keyword{internal}
