% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_policies.R
\name{cluster_policies_delete}
\alias{cluster_policies_delete}
\title{Delete a cluster policy.}
\usage{
cluster_policies_delete(policy_id)
}
\arguments{
\item{policy_id}{Required. The ID of the policy to delete.}
}
\description{
Delete a policy for a cluster. Clusters governed by this policy can still
run, but cannot be edited.
}
\keyword{internal}
