% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_policies.R
\name{cluster_policies_get}
\alias{cluster_policies_get}
\title{Get entity.}
\usage{
cluster_policies_get(policy_id, ...)
}
\arguments{
\item{policy_id}{Required. Canonical unique identifier for the cluster policy.}
}
\description{
Get a cluster policy entity. Creation and editing is available to admins
only.
}
\keyword{internal}
