% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_policies.R
\name{cluster_policies_list}
\alias{cluster_policies_list}
\title{Get a cluster policy.}
\usage{
cluster_policies_list(sort_column = NULL, sort_order = NULL)
}
\arguments{
\item{sort_column}{The cluster policy attribute to sort by.}

\item{sort_order}{The order in which the policies get listed.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Returns a list of policies accessible by the requesting user.
}
\keyword{internal}
