% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_policies.R
\name{clusterPoliciesClusterPolicyId}
\alias{clusterPoliciesClusterPolicyId}
\title{Get cluster policy permissions.}
\usage{
clusterPoliciesClusterPolicyId(client, cluster_policy_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{cluster_policy_id}{Required. The cluster policy for which to get or manage permissions.}
}
\description{
Gets the permissions of a cluster policy. Cluster policies can inherit
permissions from their root object.
}
