% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_policies.R
\name{clusterPoliciesDelete}
\alias{clusterPoliciesDelete}
\title{Delete a cluster policy.}
\usage{
clusterPoliciesDelete(client, policy_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{policy_id}{Required. The ID of the policy to delete.}
}
\description{
Delete a policy for a cluster. Clusters governed by this policy can still
run, but cannot be edited.
}
