% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_policies.R
\name{clusterPoliciesGet}
\alias{clusterPoliciesGet}
\title{Get a cluster policy.}
\usage{
clusterPoliciesGet(client, policy_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{policy_id}{Required. Canonical unique identifier for the cluster policy.}
}
\description{
Get a cluster policy entity. Creation and editing is available to admins
only.
}
