% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_policies.R
\name{clusterPoliciesPermissionLevels}
\alias{clusterPoliciesPermissionLevels}
\title{Get cluster policy permission levels.}
\usage{
clusterPoliciesPermissionLevels(client, cluster_policy_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{cluster_policy_id}{Required. The cluster policy for which to get or manage permissions.}
}
\description{
Gets the permission levels that a user can have on an object.
}
