% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_policies.R
\name{clusterPoliciesSetClusterPolicyPermissions}
\alias{clusterPoliciesSetClusterPolicyPermissions}
\title{Set cluster policy permissions.}
\usage{
clusterPoliciesSetClusterPolicyPermissions(
  client,
  cluster_policy_id,
  access_control_list = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{cluster_policy_id}{Required. The cluster policy for which to get or manage permissions.}

\item{access_control_list}{}
}
\description{
Sets permissions on a cluster policy. Cluster policies can inherit
permissions from their root object.
}
