% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\docType{data}
\name{clusters}
\alias{clusters}
\title{The Clusters API allows you to create, start, edit, list, terminate, and
delete clusters.}
\format{
An object of class \code{list} of length 16.
}
\usage{
clusters
}
\description{
Databricks maps cluster node instance types to compute units known as DBUs.
See the instance type pricing page for a list of the supported instance types
and their corresponding DBUs.
}
\details{
A Databricks cluster is a set of computation resources and configurations on
which you run data engineering, data science, and data analytics workloads,
such as production ETL pipelines, streaming analytics, ad-hoc analytics, and
machine learning.

You run these workloads as a set of commands in a notebook or as an automated
job. Databricks makes a distinction between all-purpose clusters and job
clusters. You use all-purpose clusters to analyze data collaboratively using
interactive notebooks. You use job clusters to run fast and robust automated
jobs.

You can create an all-purpose cluster using the UI, CLI, or REST API. You can
manually terminate and restart an all-purpose cluster. Multiple users can
share such clusters to do collaborative interactive analysis.

IMPORTANT: Databricks retains cluster configuration information for up to 200
all-purpose clusters terminated in the last 30 days and up to 30 job clusters
recently terminated by the job scheduler. To keep an all-purpose cluster
configuration even after it has been terminated for more than 30 days, an
administrator can pin a cluster to the cluster list.
}
\section{Operations}{

\tabular{ll}{
\link[=clusters_change_owner]{change_owner} \tab Change cluster owner.\cr
\link[=clusters_create]{create} \tab Create new cluster.\cr
\link[=clusters_delete]{delete} \tab Terminate cluster.\cr
\link[=clusters_edit]{edit} \tab Update cluster configuration.\cr
\link[=clusters_events]{events} \tab List cluster activity events.\cr
\link[=clusters_get]{get} \tab Get cluster info.\cr
\link[=clusters_list]{list} \tab List all clusters.\cr
\link[=clusters_list_node_types]{list_node_types} \tab List node types.\cr
\link[=clusters_list_zones]{list_zones} \tab List availability zones.\cr
\link[=clusters_permanent_delete]{permanent_delete} \tab Permanently delete cluster.\cr
\link[=clusters_pin]{pin} \tab Pin cluster.\cr
\link[=clusters_resize]{resize} \tab Resize cluster.\cr
\link[=clusters_restart]{restart} \tab Restart cluster.\cr
\link[=clusters_spark_versions]{spark_versions} \tab List available Spark versions.\cr
\link[=clusters_start]{start} \tab Start terminated cluster.\cr
\link[=clusters_unpin]{unpin} \tab Unpin cluster.\cr
}
}

\keyword{datasets}
