% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{clusters_change_owner}
\alias{clusters_change_owner}
\title{Change cluster owner.}
\usage{
clusters_change_owner(cluster_id, owner_username)
}
\arguments{
\item{cluster_id}{Required. \if{html}{\out{<needs content added>}}.}

\item{owner_username}{Required. New owner of the cluster_id after this RPC.}
}
\description{
Change the owner of the cluster. You must be an admin to perform this
operation.
}
\keyword{internal}
