% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{clusters_events}
\alias{clusters_events}
\title{List cluster activity events.}
\usage{
clusters_events(
  cluster_id,
  end_time = NULL,
  event_types = NULL,
  limit = NULL,
  offset = NULL,
  order = NULL,
  start_time = NULL,
  ...
)
}
\arguments{
\item{cluster_id}{\link{required} The ID of the cluster to retrieve events about.}

\item{end_time}{The end time in epoch milliseconds.}

\item{event_types}{An optional set of event types to filter on.}

\item{limit}{The maximum number of events to include in a page of events.}

\item{offset}{The offset in the result set.}

\item{order}{The order to list events in; either 'ASC' or 'DESC'.}

\item{start_time}{The start time in epoch milliseconds.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Retrieves a list of events about the activity of a cluster. This API is
paginated. If there are more events to read, the response includes all the
nparameters necessary to request the next page of events.
}
\keyword{internal}
