% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{clusters_get}
\alias{clusters_get}
\title{Get cluster info.}
\usage{
clusters_get(cluster_id, ...)
}
\arguments{
\item{cluster_id}{\link{required} The cluster about which to retrieve information.}
}
\description{
'Retrieves the information for a cluster given its identifier. Clusters can
be described while they are running, or up to 60 days after they are
terminated.
}
\keyword{internal}
