% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{clusters_list_node_types}
\alias{clusters_list_node_types}
\title{List node types.}
\usage{
clusters_list_node_types()
}
\description{
Returns a list of supported Spark node types. These node types can be used to
launch a cluster.#'
}
\keyword{internal}
