% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{clusters_list_zones}
\alias{clusters_list_zones}
\title{List availability zones.}
\usage{
clusters_list_zones()
}
\description{
Returns a list of availability zones where clusters can be created in (For
example, us-west-2a). These zones can be used to launch a cluster.#'
}
\keyword{internal}
