% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{clusters_permanent_delete}
\alias{clusters_permanent_delete}
\title{Permanently delete cluster.}
\usage{
clusters_permanent_delete(cluster_id)
}
\arguments{
\item{cluster_id}{Required. The cluster to be deleted.}
}
\description{
Permanently deletes a Spark cluster. This cluster is terminated and resources
are asynchronously removed.
}
\details{
In addition, users will no longer see permanently deleted clusters in the
cluster list, and API users can no longer perform any action on permanently
deleted clusters.
}
\keyword{internal}
