% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{clusters_pin}
\alias{clusters_pin}
\title{Pin cluster.}
\usage{
clusters_pin(cluster_id, ...)
}
\arguments{
\item{cluster_id}{Required. \if{html}{\out{<needs content added>}}.}
}
\description{
Pinning a cluster ensures that the cluster will always be returned by the
ListClusters API. Pinning a cluster that is already pinned will have no
effect. This API can only be called by workspace admins.
}
\keyword{internal}
