% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{clusters_spark_versions}
\alias{clusters_spark_versions}
\title{List available Spark versions.}
\usage{
clusters_spark_versions(...)
}
\description{
Returns the list of available Spark versions. These versions can be used to
launch a cluster.#'
}
\keyword{internal}
