% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{clusters_unpin}
\alias{clusters_unpin}
\title{Unpin cluster.}
\usage{
clusters_unpin(cluster_id)
}
\arguments{
\item{cluster_id}{Required. \if{html}{\out{<needs content added>}}.}
}
\description{
Unpinning a cluster will allow the cluster to eventually be removed from the
ListClusters API. Unpinning a cluster that is not pinned will have no effect.
This API can only be called by workspace admins.
}
\keyword{internal}
