% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{clustersChangeOwner}
\alias{clustersChangeOwner}
\title{Change cluster owner.}
\usage{
clustersChangeOwner(client, cluster_id, owner_username)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{cluster_id}{Required. \if{html}{\out{<needs content added>}}.}

\item{owner_username}{Required. New owner of the cluster_id after this RPC.}
}
\description{
Change the owner of the cluster. You must be an admin to perform this
operation.
}
