% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{clustersGet}
\alias{clustersGet}
\title{Get cluster info.}
\usage{
clustersGet(client, cluster_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{cluster_id}{Required. The cluster about which to retrieve information.}
}
\description{
Retrieves the information for a cluster given its identifier. Clusters can be
described while they are running, or up to 60 days after they are terminated.
}
