% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{clustersListNodeTypes}
\alias{clustersListNodeTypes}
\title{List node types.}
\usage{
clustersListNodeTypes(client)
}
\description{
Returns a list of supported Spark node types. These node types can be used to
launch a cluster.#'
}
