% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{clustersRestart}
\alias{clustersRestart}
\title{Restart cluster.}
\usage{
clustersRestart(client, cluster_id, restart_user = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{cluster_id}{Required. The cluster to be started.}

\item{restart_user}{\if{html}{\out{<needs content added>}}.}
}
\description{
Restarts a Spark cluster with the supplied ID. If the cluster is not
currently in a \code{RUNNING} state, nothing will happen.
}
