% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{clustersSetPermissions}
\alias{clustersSetPermissions}
\title{Set cluster permissions.}
\usage{
clustersSetPermissions(client, cluster_id, access_control_list = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{cluster_id}{Required. The cluster for which to get or manage permissions.}

\item{access_control_list}{}
}
\description{
Sets permissions on a cluster. Clusters can inherit permissions from their
root object.
}
