% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command_execution.R
\docType{data}
\name{command_execution}
\alias{command_execution}
\title{This API allows execution of Python, Scala, SQL, or R commands on running
Databricks Clusters.}
\format{
An object of class \code{list} of length 6.
}
\usage{
command_execution
}
\description{
This API allows execution of Python, Scala, SQL, or R commands on running
Databricks Clusters.
}
\section{Operations}{

\tabular{ll}{
\link[=command_execution_cancel]{cancel} \tab Cancel a command.\cr
\link[=command_execution_command_status]{command_status} \tab Get command info.\cr
\link[=command_execution_context_status]{context_status} \tab Get status.\cr
\link[=command_execution_create]{create} \tab Create an execution context.\cr
\link[=command_execution_destroy]{destroy} \tab Delete an execution context.\cr
\link[=command_execution_execute]{execute} \tab Run a command.\cr
}
}

\keyword{datasets}
