% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command_execution.R
\name{command_execution_context_status}
\alias{command_execution_context_status}
\title{Get status.}
\usage{
command_execution_context_status(cluster_id, context_id, ...)
}
\arguments{
\item{cluster_id}{Required.}

\item{context_id}{Required.}
}
\description{
Gets the status for an execution context.
}
\keyword{internal}
