% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command_execution.R
\name{command_execution_create}
\alias{command_execution_create}
\title{Create an execution context.}
\usage{
command_execution_create(
  cluster_id = NULL,
  language = NULL,
  timeout = 20,
  callback = cli_reporter
)
}
\arguments{
\item{cluster_id}{Running cluster id.}

\item{language}{}
}
\description{
This is a long-running operation, which blocks until Command Execution on Databricks reach
Running state with the timeout of 20 minutes, that you can change via \code{timeout} parameter.
By default, the state of Databricks Command Execution is reported to console. You can change this behavior
by changing the \code{callback} parameter.
}
\details{
Creates an execution context for running cluster commands.

If successful, this method returns the ID of the new execution context.
}
\keyword{internal}
