% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command_execution.R
\name{command_execution_destroy}
\alias{command_execution_destroy}
\title{Delete an execution context.}
\usage{
command_execution_destroy(cluster_id, context_id, ...)
}
\arguments{
\item{cluster_id}{\link{required}}

\item{context_id}{\link{required}}
}
\description{
Deletes an execution context.
}
\keyword{internal}
