% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command_execution.R
\name{command_execution_execute}
\alias{command_execution_execute}
\title{Run a command.}
\usage{
command_execution_execute(
  cluster_id = NULL,
  command = NULL,
  context_id = NULL,
  language = NULL,
  timeout = 20,
  callback = cli_reporter
)
}
\arguments{
\item{cluster_id}{Running cluster id.}

\item{command}{Executable code.}

\item{context_id}{Running context id.}

\item{language}{}
}
\description{
This is a long-running operation, which blocks until Command Execution on Databricks reach
Finished or Error state with the timeout of 20 minutes, that you can change via \code{timeout} parameter.
By default, the state of Databricks Command Execution is reported to console. You can change this behavior
by changing the \code{callback} parameter.
}
\details{
Runs a cluster command in the given execution context, using the provided
language.

If successful, it returns an ID for tracking the status of the command's
execution.
}
\keyword{internal}
