% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command_execution.R
\name{command_execution_status}
\alias{command_execution_status}
\alias{commandExecutionCommandStatus}
\title{Get command info.}
\usage{
command_execution_status(client, cluster_id, context_id, command_id)

commandExecutionCommandStatus(client, cluster_id, context_id, command_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{cluster_id}{Required. This field has no description yet.}

\item{context_id}{Required. This field has no description yet.}

\item{command_id}{Required. This field has no description yet.}
}
\description{
Gets the status of and, if available, the results from a currently executing
command.
}
\details{
The command ID is obtained from a prior successful call to \strong{execute}.
}
