% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command_execution.R
\name{commandExecutionCancel}
\alias{commandExecutionCancel}
\title{Cancel a command.}
\usage{
commandExecutionCancel(
  client,
  cluster_id = NULL,
  command_id = NULL,
  context_id = NULL,
  timeout = 20,
  callback = cli_reporter
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{cluster_id}{}

\item{command_id}{}

\item{context_id}{}
}
\description{
This is a long-running operation, which blocks until Command Execution on Databricks reach
Cancelled state with the timeout of 20 minutes, that you can change via \code{timeout} parameter.
By default, the state of Databricks Command Execution is reported to console. You can change this behavior
by changing the \code{callback} parameter.
}
\details{
Cancels a currently running command within an execution context.

The command ID is obtained from a prior successful call to \strong{execute}.
}
