% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command_execution.R
\name{commandExecutionContextStatus}
\alias{commandExecutionContextStatus}
\title{Get status.}
\usage{
commandExecutionContextStatus(client, cluster_id, context_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{cluster_id}{Required.}

\item{context_id}{Required.}
}
\description{
Gets the status for an execution context.
}
