% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command_execution.R
\name{commandExecutionDestroy}
\alias{commandExecutionDestroy}
\title{Delete an execution context.}
\usage{
commandExecutionDestroy(client, cluster_id, context_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{cluster_id}{Required. This field has no description yet.}

\item{context_id}{Required. This field has no description yet.}
}
\description{
Deletes an execution context.
}
