% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\docType{data}
\name{connections}
\alias{connections}
\title{Connections allow for creating a connection to an external data source.}
\format{
An object of class \code{list} of length 5.
}
\usage{
connections
}
\description{
A connection is an abstraction of an external data source that can be
connected from Databricks Compute. Creating a connection object is the first
step to managing external data sources within Unity Catalog, with the second
step being creating a data object (catalog, schema, or table) using the
connection. Data objects derived from a connection can be written to or read
from similar to other Unity Catalog data objects based on cloud storage.
Users may create different types of connections with each connection having a
unique set of configuration options to support credential management and
other settings.
}
\section{Operations}{

\tabular{ll}{
\link[=connections_create]{create} \tab Create a connection.\cr
\link[=connections_delete]{delete} \tab Delete a connection.\cr
\link[=connections_get]{get} \tab Get a connection.\cr
\link[=connections_list]{list} \tab List connections.\cr
\link[=connections_update]{update} \tab Update a connection.\cr
}
}

\keyword{datasets}
