% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{connections_delete}
\alias{connections_delete}
\title{Delete a connection.}
\usage{
connections_delete(name_arg)
}
\arguments{
\item{name_arg}{Required. The name of the connection to be deleted.}
}
\description{
Deletes the connection that matches the supplied name.
}
\keyword{internal}
