% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{connections_get}
\alias{connections_get}
\title{Get a connection.}
\usage{
connections_get(name_arg)
}
\arguments{
\item{name_arg}{Required. Name of the connection.}
}
\description{
Gets a connection from it's name.
}
\keyword{internal}
