% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{connections_list}
\alias{connections_list}
\title{List connections.}
\usage{
connections_list()
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
List all connections.#'
}
\keyword{internal}
