% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{connections_update}
\alias{connections_update}
\title{Update a connection.}
\usage{
connections_update(name, options_kvpairs, name_arg)
}
\arguments{
\item{name}{Required. Name of the connection.}

\item{options_kvpairs}{Required. Object properties as map of string key-value pairs.}

\item{name_arg}{Required. Name of the connection.}
}
\description{
Updates the connection that matches the supplied name.
}
\keyword{internal}
