% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command_execution.R
\name{context_command_execution_status}
\alias{context_command_execution_status}
\alias{commandExecutionContextStatus}
\title{Get status.}
\usage{
context_command_execution_status(client, cluster_id, context_id)

commandExecutionContextStatus(client, cluster_id, context_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{cluster_id}{Required. This field has no description yet.}

\item{context_id}{Required. This field has no description yet.

See the \href{https://docs.databricks.com/api/workspace/commandexecution/contextstatus}{API documentation} for more
information on the request and response field structure.}
}
\description{
Gets the status for an execution context.
}
