% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{create_alert}
\alias{create_alert}
\alias{alertsCreate}
\title{Create an alert.}
\usage{
create_alert(client, name, options, query_id, parent = NULL, rearm = NULL)

alertsCreate(client, name, options, query_id, parent = NULL, rearm = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Name of the alert.}

\item{options}{Required. Alert configuration options.}

\item{query_id}{Required. Query ID.}

\item{parent}{The identifier of the workspace folder containing the object.}

\item{rearm}{Number of seconds after being triggered before the alert rearms itself and can be triggered again.

See the \href{https://docs.databricks.com/api/workspace/alerts/create}{API documentation} for more
information on the request and response field structure.}
}
\description{
Creates an alert. An alert is a Databricks SQL object that periodically runs
a query, evaluates a condition of its result, and notifies users or
notification destinations if the condition was met.
}
