% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{create_app}
\alias{create_app}
\alias{appsCreate}
\title{Create and deploy an application.}
\usage{
create_app(client, manifest, resources = NULL)

appsCreate(client, manifest, resources = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{manifest}{Required. Manifest that specifies the application requirements.}

\item{resources}{Information passed at app deployment time to fulfill app dependencies.

See the \href{https://docs.databricks.com/api/workspace/apps/create}{API documentation} for more
information on the request and response field structure.}
}
\description{
Creates and deploys an application.
}
