% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogs.R
\name{create_catalog}
\alias{create_catalog}
\title{Create a catalog.}
\usage{
create_catalog(
  client,
  name,
  comment = NULL,
  connection_name = NULL,
  options = NULL,
  properties = NULL,
  provider_name = NULL,
  share_name = NULL,
  storage_root = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Name of catalog.}

\item{comment}{User-provided free-form text description.}

\item{connection_name}{The name of the connection to an external data source.}

\item{options}{A map of key-value properties attached to the securable.}

\item{properties}{A map of key-value properties attached to the securable.}

\item{provider_name}{The name of delta sharing provider.}

\item{share_name}{The name of the share under the share provider.}

\item{storage_root}{Storage root URL for managed tables within catalog.}
}
\description{
Creates a new catalog instance in the parent metastore if the caller is a
metastore admin or has the \strong{CREATE_CATALOG} privilege.
}
