% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_rooms.R
\name{create_clean_room}
\alias{create_clean_room}
\alias{cleanRoomsCreate}
\title{Create a clean room.}
\usage{
create_clean_room(client, name, remote_detailed_info, comment = NULL)

cleanRoomsCreate(client, name, remote_detailed_info, comment = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Name of the clean room.}

\item{remote_detailed_info}{Required. Central clean room details.}

\item{comment}{User-provided free-form text description.}
}
\description{
Creates a new clean room with specified colaborators. The caller must be a
metastore admin or have the \strong{CREATE_CLEAN_ROOM} privilege on the metastore.
}
