% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{create_connection}
\alias{create_connection}
\title{Create a connection.}
\usage{
create_connection(
  client,
  name,
  connection_type,
  options,
  comment = NULL,
  properties = NULL,
  read_only = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Name of the connection.}

\item{connection_type}{Required. The type of connection.}

\item{options}{Required. A map of key-value properties attached to the securable.}

\item{comment}{User-provided free-form text description.}

\item{properties}{An object containing map of key-value properties attached to the connection.}

\item{read_only}{If the connection is read only.}
}
\description{
Creates a new connection
}
\details{
Creates a new connection to an external data source. It allows users to
specify connection details and configurations for interaction with the
external server.
}
