% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_widgets.R
\name{create_dashboard_widget}
\alias{create_dashboard_widget}
\title{Add widget to a dashboard.}
\usage{
create_dashboard_widget(
  client,
  dashboard_id,
  options,
  width,
  text = NULL,
  visualization_id = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{dashboard_id}{Required. Dashboard ID returned by :method:dashboards/create.}

\item{options}{Required. This field has no description yet.}

\item{width}{Required. Width of a widget.}

\item{text}{If this is a textbox widget, the application displays this text.}

\item{visualization_id}{Query Vizualization ID returned by :method:queryvisualizations/create.}
}
\description{
Add widget to a dashboard.
}
