% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_locations.R
\name{create_external_location}
\alias{create_external_location}
\alias{externalLocationsCreate}
\title{Create an external location.}
\usage{
create_external_location(
  client,
  name,
  url,
  credential_name,
  access_point = NULL,
  comment = NULL,
  encryption_details = NULL,
  read_only = NULL,
  skip_validation = NULL
)

externalLocationsCreate(
  client,
  name,
  url,
  credential_name,
  access_point = NULL,
  comment = NULL,
  encryption_details = NULL,
  read_only = NULL,
  skip_validation = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Name of the external location.}

\item{url}{Required. Path URL of the external location.}

\item{credential_name}{Required. Name of the storage credential used with this location.}

\item{access_point}{The AWS access point to use when accesing s3 for this external location.}

\item{comment}{User-provided free-form text description.}

\item{encryption_details}{Encryption options that apply to clients connecting to cloud storage.}

\item{read_only}{Indicates whether the external location is read-only.}

\item{skip_validation}{Skips validation of the storage credential associated with the external location.}
}
\description{
Creates a new external location entry in the metastore. The caller must be a
metastore admin or have the \strong{CREATE_EXTERNAL_LOCATION} privilege on both
the metastore and the associated storage credential.
}
