% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_function}
\alias{create_function}
\title{Create a function.}
\usage{
create_function(client, function_info)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{function_info}{Required. Partial \strong{FunctionInfo} specifying the function to be created.}
}
\description{
Creates a new function
}
\details{
The user must have the following permissions in order for the function to be
created: - \strong{USE_CATALOG} on the function's parent catalog - \strong{USE_SCHEMA}
and \strong{CREATE_FUNCTION} on the function's parent schema
}
