% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_credentials.R
\name{create_git_credential}
\alias{create_git_credential}
\alias{gitCredentialsCreate}
\title{Create a credential entry.}
\usage{
create_git_credential(
  client,
  git_provider,
  git_username = NULL,
  personal_access_token = NULL
)

gitCredentialsCreate(
  client,
  git_provider,
  git_username = NULL,
  personal_access_token = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{git_provider}{Required. Git provider.}

\item{git_username}{Git username.}

\item{personal_access_token}{The personal access token used to authenticate to the corresponding Git provider.

See the \href{https://docs.databricks.com/api/workspace/gitcredentials/create}{API documentation} for more
information on the request and response field structure.}
}
\description{
Creates a Git credential entry for the user. Only one Git credential per user
is supported, so any attempts to create credentials if an entry already
exists will fail. Use the PATCH endpoint to update existing credentials, or
the DELETE endpoint to delete existing credentials.
}
