% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_init_scripts.R
\name{create_global_init_script}
\alias{create_global_init_script}
\title{Create init script.}
\usage{
create_global_init_script(
  client,
  name,
  script,
  enabled = NULL,
  position = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. The name of the script.}

\item{script}{Required. The Base64-encoded content of the script.}

\item{enabled}{Specifies whether the script is enabled.}

\item{position}{The position of a global init script, where 0 represents the first script to run, 1 is the second script to run, in ascending order.}
}
\description{
Creates a new global init script in this workspace.
}
