% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakehouse_monitors.R
\name{create_lakehouse_monitor}
\alias{create_lakehouse_monitor}
\alias{lakehouseMonitorsCreate}
\title{Create a table monitor.}
\usage{
create_lakehouse_monitor(
  client,
  full_name,
  assets_dir,
  output_schema_name,
  baseline_table_name = NULL,
  custom_metrics = NULL,
  data_classification_config = NULL,
  inference_log = NULL,
  notifications = NULL,
  schedule = NULL,
  skip_builtin_dashboard = NULL,
  slicing_exprs = NULL,
  snapshot = NULL,
  time_series = NULL,
  warehouse_id = NULL
)

lakehouseMonitorsCreate(
  client,
  full_name,
  assets_dir,
  output_schema_name,
  baseline_table_name = NULL,
  custom_metrics = NULL,
  data_classification_config = NULL,
  inference_log = NULL,
  notifications = NULL,
  schedule = NULL,
  skip_builtin_dashboard = NULL,
  slicing_exprs = NULL,
  snapshot = NULL,
  time_series = NULL,
  warehouse_id = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{full_name}{Required. Full name of the table.}

\item{assets_dir}{Required. The directory to store monitoring assets (e.g.}

\item{output_schema_name}{Required. Schema where output metric tables are created.}

\item{baseline_table_name}{Name of the baseline table from which drift metrics are computed from.}

\item{custom_metrics}{Custom metrics to compute on the monitored table.}

\item{data_classification_config}{The data classification config for the monitor.}

\item{inference_log}{Configuration for monitoring inference logs.}

\item{notifications}{The notification settings for the monitor.}

\item{schedule}{The schedule for automatically updating and refreshing metric tables.}

\item{skip_builtin_dashboard}{Whether to skip creating a default dashboard summarizing data quality metrics.}

\item{slicing_exprs}{List of column expressions to slice data with for targeted analysis.}

\item{snapshot}{Configuration for monitoring snapshot tables.}

\item{time_series}{Configuration for monitoring time series tables.}

\item{warehouse_id}{Optional argument to specify the warehouse for dashboard creation.

See the \href{https://docs.databricks.com/api/workspace/lakehousemonitors/create}{API documentation} for more
information on the request and response field structure.}
}
\description{
Creates a new monitor for the specified table.
}
\details{
The caller must either: 1. be an owner of the table's parent catalog, have
\strong{USE_SCHEMA} on the table's parent schema, and have \strong{SELECT} access on
the table 2. have \strong{USE_CATALOG} on the table's parent catalog, be an owner
of the table's parent schema, and have \strong{SELECT} access on the table. 3.
have the following permissions: - \strong{USE_CATALOG} on the table's parent
catalog - \strong{USE_SCHEMA} on the table's parent schema - be an owner of the
table.

Workspace assets, such as the dashboard, will be created in the workspace
where this call was made.
}
