% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeview.R
\name{create_lakeview}
\alias{create_lakeview}
\title{Create dashboard.}
\usage{
create_lakeview(
  client,
  display_name,
  parent_path = NULL,
  serialized_dashboard = NULL,
  warehouse_id = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{display_name}{Required. The display name of the dashboard.}

\item{parent_path}{The workspace path of the folder containing the dashboard.}

\item{serialized_dashboard}{The contents of the dashboard in serialized string form.}

\item{warehouse_id}{The warehouse ID used to run the dashboard.}
}
\description{
Create a draft dashboard.
}
