% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{create_model}
\alias{create_model}
\title{Create a model.}
\usage{
create_model(client, name, description = NULL, tags = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Register models under this name.}

\item{description}{Optional description for registered model.}

\item{tags}{Additional metadata for registered model.}
}
\description{
Creates a new registered model with the name specified in the request body.
}
\details{
Throws \code{RESOURCE_ALREADY_EXISTS} if a registered model with the given name
exists.
}
