% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{create_model_comment}
\alias{create_model_comment}
\alias{modelRegistryCreateComment}
\title{Post a comment.}
\usage{
create_model_comment(client, name, version, comment)

modelRegistryCreateComment(client, name, version, comment)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Name of the model.}

\item{version}{Required. Version of the model.}

\item{comment}{Required. User-provided comment on the action.}
}
\description{
Posts a comment on a model version. A comment can be submitted either by a
user or programmatically to display relevant information about the model. For
example, test results or deployment errors.
}
