% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{create_model_version}
\alias{create_model_version}
\title{Create a model version.}
\usage{
create_model_version(
  client,
  name,
  source,
  description = NULL,
  run_id = NULL,
  run_link = NULL,
  tags = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Register model under this name.}

\item{source}{Required. URI indicating the location of the model artifacts.}

\item{description}{Optional description for model version.}

\item{run_id}{MLflow run ID for correlation, if \code{source} was generated by an experiment run in MLflow tracking server.}

\item{run_link}{MLflow run link - this is the exact link of the run that generated this model version, potentially hosted at another instance of MLflow.}

\item{tags}{Additional metadata for model version.}
}
\description{
Creates a model version.
}
