% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{create_model_webhook}
\alias{create_model_webhook}
\title{Create a webhook.}
\usage{
create_model_webhook(
  client,
  events,
  description = NULL,
  http_url_spec = NULL,
  job_spec = NULL,
  model_name = NULL,
  status = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{events}{Required. Events that can trigger a registry webhook: * \code{MODEL_VERSION_CREATED}: A new model version was created for the associated model.}

\item{description}{User-specified description for the webhook.}

\item{http_url_spec}{This field has no description yet.}

\item{job_spec}{This field has no description yet.}

\item{model_name}{Name of the model whose events would trigger this webhook.}

\item{status}{Enable or disable triggering the webhook, or put the webhook into test mode.}
}
\description{
\strong{NOTE}: This endpoint is in Public Preview.
}
\details{
Creates a registry webhook.
}
