% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_management.R
\name{create_obo_token}
\alias{create_obo_token}
\alias{tokenManagementCreateOboToken}
\title{Create on-behalf token.}
\usage{
create_obo_token(
  client,
  application_id,
  comment = NULL,
  lifetime_seconds = NULL
)

tokenManagementCreateOboToken(
  client,
  application_id,
  comment = NULL,
  lifetime_seconds = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{application_id}{Required. Application ID of the service principal.}

\item{comment}{Comment that describes the purpose of the token.}

\item{lifetime_seconds}{The number of seconds before the token expires.}
}
\description{
Creates a token on behalf of a service principal.
}
