% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/providers.R
\name{create_provider}
\alias{create_provider}
\alias{providersCreate}
\title{Create an auth provider.}
\usage{
create_provider(
  client,
  name,
  authentication_type,
  comment = NULL,
  recipient_profile_str = NULL
)

providersCreate(
  client,
  name,
  authentication_type,
  comment = NULL,
  recipient_profile_str = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. The name of the Provider.}

\item{authentication_type}{Required. The delta sharing authentication type.}

\item{comment}{Description about the provider.}

\item{recipient_profile_str}{This field is required when the \strong{authentication_type} is \strong{TOKEN} or not provided.}
}
\description{
Creates a new authentication provider minimally based on a name and
authentication type. The caller must be an admin on the metastore.
}
