% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{create_query}
\alias{create_query}
\alias{queriesCreate}
\title{Create a new query definition.}
\usage{
create_query(
  client,
  data_source_id = NULL,
  description = NULL,
  name = NULL,
  options = NULL,
  parent = NULL,
  query = NULL,
  run_as_role = NULL
)

queriesCreate(
  client,
  data_source_id = NULL,
  description = NULL,
  name = NULL,
  options = NULL,
  parent = NULL,
  query = NULL,
  run_as_role = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{data_source_id}{Data source ID maps to the ID of the data source used by the resource and is distinct from the warehouse ID.}

\item{description}{General description that conveys additional information about this query such as usage notes.}

\item{name}{The title of this query that appears in list views, widget headings, and on the query page.}

\item{options}{Exclusively used for storing a list parameter definitions.}

\item{parent}{The identifier of the workspace folder containing the object.}

\item{query}{The text of the query to be run.}

\item{run_as_role}{Sets the \strong{Run as} role for the object.

See the \href{https://docs.databricks.com/api/workspace/queries/create}{API documentation} for more
information on the request and response field structure.}
}
\description{
Creates a new query definition. Queries created with this endpoint belong to
the authenticated user making the request.
}
\details{
The \code{data_source_id} field specifies the ID of the SQL warehouse to run this
query against. You can use the Data Sources API to see a complete list of
available SQL warehouses. Or you can copy the \code{data_source_id} from an
existing query.

\strong{Note}: You cannot add a visualization until you create the query.
}
