% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_visualizations.R
\name{create_query_visualization}
\alias{create_query_visualization}
\alias{queryVisualizationsCreate}
\title{Add visualization to a query.}
\usage{
create_query_visualization(
  client,
  query_id,
  type,
  options,
  description = NULL,
  name = NULL
)

queryVisualizationsCreate(
  client,
  query_id,
  type,
  options,
  description = NULL,
  name = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{query_id}{Required. The identifier returned by :method:queries/create.}

\item{type}{Required. The type of visualization: chart, table, pivot table, and so on.}

\item{options}{Required. The options object varies widely from one visualization type to the next and is unsupported.}

\item{description}{A short description of this visualization.}

\item{name}{The name of the visualization that appears on dashboards and the query screen.}
}
\description{
Add visualization to a query.
}
