% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registered_models.R
\name{create_registered_model}
\alias{create_registered_model}
\title{Create a Registered Model.}
\usage{
create_registered_model(
  client,
  catalog_name,
  schema_name,
  name,
  comment = NULL,
  storage_location = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{catalog_name}{Required. The name of the catalog where the schema and the registered model reside.}

\item{schema_name}{Required. The name of the schema where the registered model resides.}

\item{name}{Required. The name of the registered model.}

\item{comment}{The comment attached to the registered model.}

\item{storage_location}{The storage location on the cloud under which model version data files are stored.}
}
\description{
Creates a new registered model in Unity Catalog.
}
\details{
File storage for model versions in the registered model will be located in
the default location which is specified by the parent schema, or the parent
catalog, or the Metastore.

For registered model creation to succeed, the user must satisfy the following
conditions: - The caller must be a metastore admin, or be the owner of the
parent catalog and schema, or have the \strong{USE_CATALOG} privilege on the
parent catalog and the \strong{USE_SCHEMA} privilege on the parent schema. - The
caller must have the \strong{CREATE MODEL} or \strong{CREATE FUNCTION} privilege on the
parent schema.
}
