% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{create_repo}
\alias{create_repo}
\alias{reposCreate}
\title{Create a repo.}
\usage{
create_repo(client, url, provider, path = NULL, sparse_checkout = NULL)

reposCreate(client, url, provider, path = NULL, sparse_checkout = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{url}{Required. URL of the Git repository to be linked.}

\item{provider}{Required. Git provider.}

\item{path}{Desired path for the repo in the workspace.}

\item{sparse_checkout}{If specified, the repo will be created with sparse checkout enabled.}
}
\description{
Creates a repo in the workspace and links it to the remote Git repo
specified. Note that repos created programmatically must be linked to a
remote Git repo, unlike repos created in the browser.
}
