% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service_principals.R
\name{create_service_principal}
\alias{create_service_principal}
\alias{servicePrincipalsCreate}
\title{Create a service principal.}
\usage{
create_service_principal(
  client,
  active = NULL,
  application_id = NULL,
  display_name = NULL,
  entitlements = NULL,
  external_id = NULL,
  groups = NULL,
  id = NULL,
  roles = NULL,
  schemas = NULL
)

servicePrincipalsCreate(
  client,
  active = NULL,
  application_id = NULL,
  display_name = NULL,
  entitlements = NULL,
  external_id = NULL,
  groups = NULL,
  id = NULL,
  roles = NULL,
  schemas = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{active}{If this user is active.}

\item{application_id}{UUID relating to the service principal.}

\item{display_name}{String that represents a concatenation of given and family names.}

\item{entitlements}{Entitlements assigned to the service principal.}

\item{external_id}{This field has no description yet.}

\item{groups}{This field has no description yet.}

\item{id}{Databricks service principal ID.}

\item{roles}{Corresponds to AWS instance profile/arn role.}

\item{schemas}{The schema of the List response.}
}
\description{
Creates a new service principal in the Databricks workspace.
}
