% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_credentials.R
\name{create_storage_credential}
\alias{create_storage_credential}
\alias{storageCredentialsCreate}
\title{Create a storage credential.}
\usage{
create_storage_credential(
  client,
  name,
  aws_iam_role = NULL,
  azure_managed_identity = NULL,
  azure_service_principal = NULL,
  cloudflare_api_token = NULL,
  comment = NULL,
  databricks_gcp_service_account = NULL,
  read_only = NULL,
  skip_validation = NULL
)

storageCredentialsCreate(
  client,
  name,
  aws_iam_role = NULL,
  azure_managed_identity = NULL,
  azure_service_principal = NULL,
  cloudflare_api_token = NULL,
  comment = NULL,
  databricks_gcp_service_account = NULL,
  read_only = NULL,
  skip_validation = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. The credential name.}

\item{aws_iam_role}{The AWS IAM role configuration.}

\item{azure_managed_identity}{The Azure managed identity configuration.}

\item{azure_service_principal}{The Azure service principal configuration.}

\item{cloudflare_api_token}{The Cloudflare API token configuration.}

\item{comment}{Comment associated with the credential.}

\item{databricks_gcp_service_account}{The Databricks managed GCP service account configuration.}

\item{read_only}{Whether the storage credential is only usable for read operations.}

\item{skip_validation}{Supplying true to this argument skips validation of the created credential.}
}
\description{
Creates a new storage credential.
}
