% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_constraints.R
\name{create_table_constraint}
\alias{create_table_constraint}
\alias{tableConstraintsCreate}
\title{Create a table constraint.}
\usage{
create_table_constraint(client, full_name_arg, constraint)

tableConstraintsCreate(client, full_name_arg, constraint)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{full_name_arg}{Required. The full name of the table referenced by the constraint.

See the \href{https://docs.databricks.com/api/workspace/tableconstraints/create}{API documentation} for more
information on the request and response field structure.}

\item{constraint}{Required. A table constraint, as defined by \emph{one} of the following fields being set: \strong{primary_key_constraint}, \strong{foreign_key_constraint}, \strong{named_table_constraint}.}
}
\description{
Creates a new table constraint.
}
\details{
For the table constraint creation to succeed, the user must satisfy both of
these conditions: - the user must have the \strong{USE_CATALOG} privilege on the
table's parent catalog, the \strong{USE_SCHEMA} privilege on the table's parent
schema, and be the owner of the table. - if the new constraint is a
\strong{ForeignKeyConstraint}, the user must have the \strong{USE_CATALOG} privilege on
the referenced parent table's catalog, the \strong{USE_SCHEMA} privilege on the
referenced parent table's schema, and be the owner of the referenced parent
table.
}
