% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csp_enablement.R
\name{cspEnablementGet}
\alias{cspEnablementGet}
\title{Get the compliance security profile setting.}
\usage{
cspEnablementGet(client, etag = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{etag}{etag used for versioning.}
}
\description{
Gets the compliance security profile setting.
}
