% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metastores.R
\name{current_metastore}
\alias{current_metastore}
\alias{metastoresCurrent}
\title{Get metastore assignment for workspace.}
\usage{
current_metastore(client)

metastoresCurrent(client)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()

See the \href{https://docs.databricks.com/api/workspace/metastores/current}{API documentation} for more
information on the request and response field structure.}
}
\description{
Gets the metastore assignment for the workspace being accessed.
}
