% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_user.R
\docType{data}
\name{current_user}
\alias{current_user}
\title{This API allows retrieving information about currently authenticated user or
service principal.}
\format{
An object of class \code{list} of length 1.
}
\usage{
current_user
}
\description{
This API allows retrieving information about currently authenticated user or
service principal.
}
\section{Operations}{

\tabular{ll}{
\link[=current_user_me]{me} \tab Get current user info.\cr
}
}

\keyword{datasets}
