% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_user.R
\name{currentUserMe}
\alias{currentUserMe}
\title{Get current user info.}
\usage{
currentUserMe(client)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}
}
\description{
Get details about the current method caller's identity.
}
