% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboards.R
\docType{data}
\name{dashboards}
\alias{dashboards}
\title{In general, there is little need to modify dashboards using the API. However,
it can be useful to use dashboard objects to look-up a collection of related
query IDs. The API can also be used to duplicate multiple dashboards at once
since you can get a dashboard definition with a GET request and then POST it
to create a new one.}
\format{
An object of class \code{list} of length 5.
}
\usage{
dashboards
}
\description{
\strong{Note}: Programmatic operations on refresh schedules via the Databricks SQL
API are deprecated. Dashboard refresh schedules can be created, updated,
fetched and deleted using Jobs API, e.g. :method:jobs/create.
}
\section{Operations}{

\tabular{ll}{
\link[=dashboards_create]{create} \tab Create a dashboard object.\cr
\link[=dashboards_delete]{delete} \tab Remove a dashboard.\cr
\link[=dashboards_get]{get} \tab Retrieve a definition.\cr
\link[=dashboards_list]{list} \tab Get dashboard objects.\cr
\link[=dashboards_restore]{restore} \tab Restore a dashboard.\cr
}
}

\keyword{datasets}
