% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboards.R
\name{dashboards_create}
\alias{dashboards_create}
\title{Create a dashboard object.}
\usage{
dashboards_create(is_favorite = NULL, name = NULL, parent = NULL, tags = NULL)
}
\arguments{
\item{is_favorite}{Indicates whether this query object should appear in the current user's favorites list.}

\item{name}{The title of this dashboard that appears in list views and at the top of the dashboard page.}

\item{parent}{The identifier of the workspace folder containing the dashboard.}

\item{tags}{}
}
\description{
Create a dashboard object.
}
\keyword{internal}
