% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboards.R
\name{dashboards_delete}
\alias{dashboards_delete}
\title{Remove a dashboard.}
\usage{
dashboards_delete(dashboard_id, ...)
}
\arguments{
\item{dashboard_id}{Required.}
}
\description{
Moves a dashboard to the trash. Trashed dashboards do not appear in list
views or searches, and cannot be shared.
}
\keyword{internal}
