% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboards.R
\name{dashboards_get}
\alias{dashboards_get}
\title{Retrieve a definition.}
\usage{
dashboards_get(dashboard_id, ...)
}
\arguments{
\item{dashboard_id}{\link{required}}
}
\description{
Returns a JSON representation of a dashboard object, including its
visualization and query objects.
}
\keyword{internal}
