% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboards.R
\name{dashboards_list}
\alias{dashboards_list}
\title{Get dashboard objects.}
\usage{
dashboards_list(order = NULL, page = NULL, page_size = NULL, q = NULL)
}
\arguments{
\item{order}{Name of dashboard attribute to order by.}

\item{page}{Page number to retrieve.}

\item{page_size}{Number of dashboards to return per page.}

\item{q}{Full text search term.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Fetch a paginated list of dashboard objects.
}
\keyword{internal}
