% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboards.R
\name{dashboards_restore}
\alias{dashboards_restore}
\title{Restore a dashboard.}
\usage{
dashboards_restore(dashboard_id)
}
\arguments{
\item{dashboard_id}{Required.}
}
\description{
A restored dashboard appears in list views and searches and can be shared.
}
\keyword{internal}
