% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboards.R
\name{dashboardsDelete}
\alias{dashboardsDelete}
\title{Remove a dashboard.}
\usage{
dashboardsDelete(client, dashboard_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{dashboard_id}{Required.}
}
\description{
Moves a dashboard to the trash. Trashed dashboards do not appear in list
views or searches, and cannot be shared.
}
