% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboards.R
\name{dashboardsGet}
\alias{dashboardsGet}
\title{Retrieve a definition.}
\usage{
dashboardsGet(client, dashboard_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{dashboard_id}{Required. This field has no description yet.}
}
\description{
Returns a JSON representation of a dashboard object, including its
visualization and query objects.
}
