% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboards.R
\name{dashboardsRestore}
\alias{dashboardsRestore}
\title{Restore a dashboard.}
\usage{
dashboardsRestore(client, dashboard_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{dashboard_id}{Required. This field has no description yet.}
}
\description{
A restored dashboard appears in list views and searches and can be shared.
}
