% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_widgets.R
\name{dashboardWidgetsUpdate}
\alias{dashboardWidgetsUpdate}
\title{Update existing widget.}
\usage{
dashboardWidgetsUpdate(
  client,
  id,
  dashboard_id,
  options,
  width,
  text = NULL,
  visualization_id = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{id}{Required. Widget ID returned by :method:dashboardwidgets/create.}

\item{dashboard_id}{Required. Dashboard ID returned by :method:dashboards/create.}

\item{options}{Required.}

\item{width}{Required. Width of a widget.}

\item{text}{If this is a textbox widget, the application displays this text.}

\item{visualization_id}{Query Vizualization ID returned by :method:queryvisualizations/create.}
}
\description{
Update existing widget.
}
