% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sources.R
\docType{data}
\name{data_sources}
\alias{data_sources}
\title{This API is provided to assist you in making new query objects. When creating
a query object, you may optionally specify a \code{data_source_id} for the SQL
warehouse against which it will run. If you don't already know the
\code{data_source_id} for your desired SQL warehouse, this API will help you find
it.}
\format{
An object of class \code{list} of length 1.
}
\usage{
data_sources
}
\description{
This API does not support searches. It returns the full list of SQL
warehouses in your workspace. We advise you to use any text editor, REST
client, or \code{grep} to search the response from this API for the name of your
SQL warehouse as it appears in Databricks SQL.
}
\section{Operations}{

\tabular{ll}{
\link[=data_sources_list]{list} \tab Get a list of SQL warehouses.\cr
}
}

\keyword{datasets}
