% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-package.R
\docType{package}
\name{databricks-package}
\alias{databricks}
\alias{databricks-package}
\title{databricks: Databricks SDK for R (Experimental)}
\description{
Call all Databricks REST APIs from R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://databrickslabs.github.io/databricks-sdk-r/}
}

}
\author{
\strong{Maintainer}: Serge Smertin \email{serge.smertin@databricks.com}

Other contributors:
\itemize{
  \item Databricks [copyright holder, funder]
}

}
\keyword{internal}
