% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{DatabricksClient}
\alias{DatabricksClient}
\title{DatabricksClient is a constructor for class that performs any operations with
Databricks REST API and handle a subset of Unified Client Authentication.}
\usage{
DatabricksClient(profile = NULL, host = NULL, token = NULL, config_file = NULL)
}
\arguments{
\item{profile}{configuration profile from ~/.databrickscfg. Defaults to DEFAULT}

\item{host}{URL of Databricks Workspace}

\item{token}{Personal Access Token}

\item{config_file}{path to Databricks CLI configuration file. Defaults to ~/.databrickscfg}
}
\description{
DatabricksClient is a constructor for class that performs any operations with
Databricks REST API and handle a subset of Unified Client Authentication.
}
