% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfs.R
\docType{data}
\name{dbfs}
\alias{dbfs}
\title{DBFS API makes it simple to interact with various data sources without having
to include a users credentials every time to read a file.}
\format{
An object of class \code{list} of length 10.
}
\usage{
dbfs
}
\description{
DBFS API makes it simple to interact with various data sources without having
to include a users credentials every time to read a file.
}
\section{Operations}{

\tabular{ll}{
\link[=dbfs_add_block]{add_block} \tab Append data block.\cr
\link[=dbfs_close]{close} \tab Close the stream.\cr
\link[=dbfs_create]{create} \tab Open a stream.\cr
\link[=dbfs_delete]{delete} \tab Delete a file/directory.\cr
\link[=dbfs_get_status]{get_status} \tab Get the information of a file or directory.\cr
\link[=dbfs_list]{list} \tab List directory contents or file details.\cr
\link[=dbfs_mkdirs]{mkdirs} \tab Create a directory.\cr
\link[=dbfs_move]{move} \tab Move a file.\cr
\link[=dbfs_put]{put} \tab Upload a file.\cr
\link[=dbfs_read]{read} \tab Get the contents of a file.\cr
}
}

\keyword{datasets}
