% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfs.R
\name{dbfs_close}
\alias{dbfs_close}
\title{Close the stream.}
\usage{
dbfs_close(handle, ...)
}
\arguments{
\item{handle}{Required. The handle on an open stream.}
}
\description{
Closes the stream specified by the input handle. If the handle does not
exist, this call throws an exception with \code{RESOURCE_DOES_NOT_EXIST}.
}
\keyword{internal}
