% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfs.R
\name{dbfs_get_status}
\alias{dbfs_get_status}
\title{Get the information of a file or directory.}
\usage{
dbfs_get_status(path, ...)
}
\arguments{
\item{path}{Required. The path of the file or directory.}
}
\description{
Gets the file information for a file or directory. If the file or directory
does not exist, this call throws an exception with \code{RESOURCE_DOES_NOT_EXIST}.
}
\keyword{internal}
