% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfs.R
\name{dbfs_read}
\alias{dbfs_read}
\title{Get the contents of a file.}
\usage{
dbfs_read(path, length = NULL, offset = NULL)
}
\arguments{
\item{path}{Required. The path of the file to read.}

\item{length}{The number of bytes to read starting from the offset.}

\item{offset}{The offset to read from in bytes.}
}
\description{
Returns the contents of a file. If the file does not exist, this call throws
an exception with \code{RESOURCE_DOES_NOT_EXIST}. If the path is a directory, the
read length is negative, or if the offset is negative, this call throws an
exception with \code{INVALID_PARAMETER_VALUE}. If the read length exceeds 1 MB,
this call throws an exception with \code{MAX_READ_SIZE_EXCEEDED}.
}
\details{
If \code{offset + length} exceeds the number of bytes in a file, it reads the
contents until the end of file.',
}
\keyword{internal}
