% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfs.R
\name{dbfsClose}
\alias{dbfsClose}
\title{Close the stream.}
\usage{
dbfsClose(client, handle)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{handle}{Required. The handle on an open stream.}
}
\description{
Closes the stream specified by the input handle. If the handle does not
exist, this call throws an exception with \code{RESOURCE_DOES_NOT_EXIST}.
}
