% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfs.R
\name{dbfsMkdirs}
\alias{dbfsMkdirs}
\title{Create a directory.}
\usage{
dbfsMkdirs(client, path)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{path}{Required. The path of the new directory.}
}
\description{
Creates the given directory and necessary parent directories if they do not
exist. If a file (not a directory) exists at any prefix of the input path,
this call throws an exception with \code{RESOURCE_ALREADY_EXISTS}. \strong{Note}: If
this operation fails, it might have succeeded in creating some of the
necessary parent directories.
}
