% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfs.R
\name{dbfsPut}
\alias{dbfsPut}
\title{Upload a file.}
\usage{
dbfsPut(client, path, contents = NULL, overwrite = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{path}{Required. The path of the new file.}

\item{contents}{This parameter might be absent, and instead a posted file will be used.}

\item{overwrite}{The flag that specifies whether to overwrite existing file/files.}
}
\description{
Uploads a file through the use of multipart form post. It is mainly used for
streaming uploads, but can also be used as a convenient single call for data
upload.
}
\details{
Alternatively you can pass contents as base64 string.

The amount of data that can be passed (when not streaming) using the
\strong{contents} parameter is limited to 1 MB. \code{MAX_BLOCK_SIZE_EXCEEDED} will be
thrown if this limit is exceeded.

If you want to upload large files, use the streaming upload. For details, see
:method:dbfs/create, :method:dbfs/addBlock, :method:dbfs/close.
}
