% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbsql_permissions.R
\docType{data}
\name{dbsql_permissions}
\alias{dbsql_permissions}
\title{The SQL Permissions API is similar to the endpoints of the
:method:permissions/set. However, this exposes only one endpoint, which gets
the Access Control List for a given object. You cannot modify any permissions
using this API.}
\format{
An object of class \code{list} of length 3.
}
\usage{
dbsql_permissions
}
\description{
There are three levels of permission:
}
\details{
\itemize{
\item \code{CAN_VIEW}: Allows read-only access
\item \code{CAN_RUN}: Allows read access and run access (superset of \code{CAN_VIEW})
\item \code{CAN_MANAGE}: Allows all actions: read, run, edit, delete, modify
permissions (superset of \code{CAN_RUN})
}
}
\section{Operations}{

\tabular{ll}{
\link[=dbsql_permissions_get]{get} \tab Get object ACL.\cr
\link[=dbsql_permissions_set]{set} \tab Set object ACL.\cr
\link[=dbsql_permissions_transfer_ownership]{transfer_ownership} \tab Transfer object ownership.\cr
}
}

\keyword{datasets}
