% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbsql_permissions.R
\name{dbsql_permissions_get}
\alias{dbsql_permissions_get}
\title{Get object ACL.}
\usage{
dbsql_permissions_get(object_type, object_id, ...)
}
\arguments{
\item{object_type}{\link{required} The type of object permissions to check.}

\item{object_id}{\link{required} Object ID.}
}
\description{
Gets a JSON representation of the access control list (ACL) for a specified
object.
}
\keyword{internal}
