% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbsql_permissions.R
\name{dbsql_permissions_set}
\alias{dbsql_permissions_set}
\title{Set object ACL.}
\usage{
dbsql_permissions_set(object_type, object_id, access_control_list = NULL)
}
\arguments{
\item{object_type}{Required. The type of object permission to set.}

\item{object_id}{Required. Object ID.}

\item{access_control_list}{}
}
\description{
Sets the access control list (ACL) for a specified object. This operation
will complete rewrite the ACL.
}
\keyword{internal}
