% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbsql_permissions.R
\name{dbsql_permissions_transfer_ownership}
\alias{dbsql_permissions_transfer_ownership}
\title{Transfer object ownership.}
\usage{
dbsql_permissions_transfer_ownership(
  object_type,
  object_id,
  new_owner = NULL,
  ...
)
}
\arguments{
\item{object_type}{Required. The type of object on which to change ownership.}

\item{object_id}{Required. The ID of the object on which to change ownership.}

\item{new_owner}{Email address for the new owner, who must exist in the workspace.}
}
\description{
Transfers ownership of a dashboard, query, or alert to an active user.
Requires an admin API key.
}
\keyword{internal}
