% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbsql_permissions.R
\name{dbsqlPermissionsGet}
\alias{dbsqlPermissionsGet}
\title{Get object ACL.}
\usage{
dbsqlPermissionsGet(client, object_type, object_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{object_type}{Required. The type of object permissions to check.}

\item{object_id}{Required. Object ID.}
}
\description{
Gets a JSON representation of the access control list (ACL) for a specified
object.
}
