% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{delete_app}
\alias{delete_app}
\alias{appsDeleteApp}
\title{Delete an application.}
\usage{
delete_app(client, name)

appsDeleteApp(client, name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. The name of an application.

See the \href{https://docs.databricks.com/api/workspace/apps/deleteapp}{API documentation} for more
information on the request and response field structure.}
}
\description{
Delete an application definition
}
