% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogs.R
\name{delete_catalog}
\alias{delete_catalog}
\alias{catalogsDelete}
\title{Delete a catalog.}
\usage{
delete_catalog(client, name, force = NULL)

catalogsDelete(client, name, force = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. The name of the catalog.

See the \href{https://docs.databricks.com/api/workspace/catalogs/delete}{API documentation} for more
information on the request and response field structure.}

\item{force}{Force deletion even if the catalog is not empty.}
}
\description{
Deletes the catalog that matches the supplied name. The caller must be a
metastore admin or the owner of the catalog.
}
