% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_rooms.R
\name{delete_clean_room}
\alias{delete_clean_room}
\alias{cleanRoomsDelete}
\title{Delete a clean room.}
\usage{
delete_clean_room(client, name)

cleanRoomsDelete(client, name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. The name of the clean room.}
}
\description{
Deletes a data object clean room from the metastore. The caller must be an
owner of the clean room.
}
