% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{delete_cluster}
\alias{delete_cluster}
\alias{clustersDelete}
\title{Terminate cluster.}
\usage{
delete_cluster(client, cluster_id)

clustersDelete(client, cluster_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{cluster_id}{Required. The cluster to be terminated.

See the \href{https://docs.databricks.com/api/workspace/clusters/delete}{API documentation} for more
information on the request and response field structure.}
}
\description{
Terminates the Spark cluster with the specified ID. The cluster is removed
asynchronously. Once the termination has completed, the cluster will be in a
\code{TERMINATED} state. If the cluster is already in a \code{TERMINATING} or
\code{TERMINATED} state, nothing will happen.
}
