% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboards.R
\name{delete_dashboard}
\alias{delete_dashboard}
\alias{dashboardsDelete}
\title{Remove a dashboard.}
\usage{
delete_dashboard(client, dashboard_id)

dashboardsDelete(client, dashboard_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{dashboard_id}{Required. This field has no description yet.

See the \href{https://docs.databricks.com/api/workspace/dashboards/delete}{API documentation} for more
information on the request and response field structure.}
}
\description{
Moves a dashboard to the trash. Trashed dashboards do not appear in list
views or searches, and cannot be shared.
}
