% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_namespace.R
\name{delete_default_namespace}
\alias{delete_default_namespace}
\alias{defaultNamespaceDelete}
\title{Delete the default namespace setting.}
\usage{
delete_default_namespace(client, etag = NULL)

defaultNamespaceDelete(client, etag = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{etag}{etag used for versioning.}
}
\description{
Deletes the default namespace setting for the workspace. A fresh etag needs
to be provided in \code{DELETE} requests (as a query parameter). The etag can be
retrieved by making a \code{GET} request before the \code{DELETE} request. If the
setting is updated/deleted concurrently, \code{DELETE} fails with 409 and the
request must be retried by using the fresh etag in the 409 response.
}
