% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{delete_experiment_run}
\alias{delete_experiment_run}
\title{Delete a run.}
\usage{
delete_experiment_run(client, run_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{run_id}{Required. ID of the run to delete.}
}
\description{
Marks a run for deletion.
}
