% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_locations.R
\name{delete_external_location}
\alias{delete_external_location}
\title{Delete an external location.}
\usage{
delete_external_location(client, name, force = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Name of the external location.}

\item{force}{Force deletion even if there are dependent external tables or mounts.}
}
\description{
Deletes the specified external location from the metastore. The caller must
be the owner of the external location.
}
