% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{delete_file}
\alias{delete_file}
\alias{filesDelete}
\title{Delete a file.}
\usage{
delete_file(client, file_path)

filesDelete(client, file_path)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{file_path}{Required. The absolute path of the file.

See the \href{https://docs.databricks.com/api/workspace/files/delete}{API documentation} for more
information on the request and response field structure.}
}
\description{
Deletes a file. If the request is successful, there is no response body.
}
