% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{delete_function}
\alias{delete_function}
\title{Delete a function.}
\usage{
delete_function(client, name, force = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. The fully-qualified name of the function (of the form \strong{catalog_name}.\strong{schema_name}.\strong{function__name}).}

\item{force}{Force deletion even if the function is notempty.}
}
\description{
Deletes the function that matches the supplied name. For the deletion to
succeed, the user must satisfy one of the following conditions: - Is the
owner of the function's parent catalog - Is the owner of the function's
parent schema and have the \strong{USE_CATALOG} privilege on its parent catalog -
Is the owner of the function itself and have both the \strong{USE_CATALOG}
privilege on its parent catalog and the \strong{USE_SCHEMA} privilege on its
parent schema
}
