% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_credentials.R
\name{delete_git_credential}
\alias{delete_git_credential}
\alias{gitCredentialsDelete}
\title{Delete a credential.}
\usage{
delete_git_credential(client, credential_id)

gitCredentialsDelete(client, credential_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{credential_id}{Required. The ID for the corresponding credential to access.

See the \href{https://docs.databricks.com/api/workspace/gitcredentials/delete}{API documentation} for more
information on the request and response field structure.}
}
\description{
Deletes the specified Git credential.
}
