% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{delete_group}
\alias{delete_group}
\alias{groupsDelete}
\title{Delete a group.}
\usage{
delete_group(client, id)

groupsDelete(client, id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{id}{Required. Unique ID for a group in the Databricks workspace.

See the \href{https://docs.databricks.com/api/workspace/groups/delete}{API documentation} for more
information on the request and response field structure.}
}
\description{
Deletes a group from the Databricks workspace.
}
