% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance_pools.R
\name{delete_instance_pool}
\alias{delete_instance_pool}
\alias{instancePoolsDelete}
\title{Delete an instance pool.}
\usage{
delete_instance_pool(client, instance_pool_id)

instancePoolsDelete(client, instance_pool_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{instance_pool_id}{Required. The instance pool to be terminated.

See the \href{https://docs.databricks.com/api/workspace/instancepools/delete}{API documentation} for more
information on the request and response field structure.}
}
\description{
Deletes the instance pool permanently. The idle instances in the pool are
terminated asynchronously.
}
