% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{delete_job}
\alias{delete_job}
\alias{jobsDelete}
\title{Delete a job.}
\usage{
delete_job(client, job_id)

jobsDelete(client, job_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{job_id}{Required. The canonical identifier of the job to delete.

See the \href{https://docs.databricks.com/api/workspace/jobs/delete}{API documentation} for more
information on the request and response field structure.}
}
\description{
Deletes a job.
}
