% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metastores.R
\name{delete_metastore}
\alias{delete_metastore}
\alias{metastoresDelete}
\title{Delete a metastore.}
\usage{
delete_metastore(client, id, force = NULL)

metastoresDelete(client, id, force = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{id}{Required. Unique ID of the metastore.}

\item{force}{Force deletion even if the metastore is not empty.}
}
\description{
Deletes a metastore. The caller must be a metastore admin.
}
