% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{delete_model_tag}
\alias{delete_model_tag}
\alias{modelRegistryDeleteModelTag}
\title{Delete a model tag.}
\usage{
delete_model_tag(client, name, key)

modelRegistryDeleteModelTag(client, name, key)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Name of the registered model that the tag was logged under.

See the \href{https://docs.databricks.com/api/workspace/modelregistry/deletemodeltag}{API documentation} for more
information on the request and response field structure.}

\item{key}{Required. Name of the tag.}
}
\description{
Deletes the tag for a registered model.
}
