% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R, R/model_versions.R
\name{delete_model_version}
\alias{delete_model_version}
\alias{modelRegistryDeleteModelVersion}
\alias{modelVersionsDelete}
\title{Delete a model version.}
\usage{
delete_model_version(client, full_name, version)

modelRegistryDeleteModelVersion(client, name, version)

delete_model_version(client, full_name, version)

modelVersionsDelete(client, full_name, version)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{full_name}{Required. The three-level (fully qualified) name of the model version.}

\item{version}{Required. The integer version number of the model version.}

\item{name}{Required. Name of the registered model.}
}
\description{
Deletes a model version.

Deletes a model version from the specified registered model. Any aliases
assigned to the model version will also be deleted.
}
\details{
The caller must be a metastore admin or an owner of the parent registered
model. For the latter case, the caller must also be the owner or have the
\strong{USE_CATALOG} privilege on the parent catalog and the \strong{USE_SCHEMA}
privilege on the parent schema.
}
