% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{delete_model_version_tag}
\alias{delete_model_version_tag}
\alias{modelRegistryDeleteModelVersionTag}
\title{Delete a model version tag.}
\usage{
delete_model_version_tag(client, name, version, key)

modelRegistryDeleteModelVersionTag(client, name, version, key)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Name of the registered model that the tag was logged under.}

\item{version}{Required. Model version number that the tag was logged under.

See the \href{https://docs.databricks.com/api/workspace/modelregistry/deletemodelversiontag}{API documentation} for more
information on the request and response field structure.}

\item{key}{Required. Name of the tag.}
}
\description{
Deletes a model version tag.
}
