% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{delete_model_webhook}
\alias{delete_model_webhook}
\alias{modelRegistryDeleteWebhook}
\title{Delete a webhook.}
\usage{
delete_model_webhook(client, id = NULL)

modelRegistryDeleteWebhook(client, id = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{id}{Webhook ID required to delete a registry webhook.}
}
\description{
\strong{NOTE:} This endpoint is in Public Preview.
}
\details{
Deletes a registry webhook.
}
