% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{delete_notebook}
\alias{delete_notebook}
\alias{workspaceDelete}
\title{Delete a workspace object.}
\usage{
delete_notebook(client, path, recursive = NULL)

workspaceDelete(client, path, recursive = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{path}{Required. The absolute path of the notebook or directory.}

\item{recursive}{The flag that specifies whether to delete the object recursively.

See the \href{https://docs.databricks.com/api/workspace/workspace/delete}{API documentation} for more
information on the request and response field structure.}
}
\description{
Deletes an object or a directory (and optionally recursively deletes all
objects in the directory). * If \code{path} does not exist, this call returns an
error \code{RESOURCE_DOES_NOT_EXIST}. * If \code{path} is a non-empty directory and
\code{recursive} is set to \code{false}, this call returns an error
\code{DIRECTORY_NOT_EMPTY}.
}
\details{
Object deletion cannot be undone and deleting a directory recursively is not
atomic.
}
