% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_tables.R
\name{delete_online_table}
\alias{delete_online_table}
\alias{onlineTablesDelete}
\title{Delete an Online Table.}
\usage{
delete_online_table(client, name)

onlineTablesDelete(client, name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Full three-part (catalog, schema, table) name of the table.

See the \href{https://docs.databricks.com/api/workspace/onlinetables/delete}{API documentation} for more
information on the request and response field structure.}
}
\description{
Delete an online table. Warning: This will delete all the data in the online
table. If the source Delta table was deleted or modified since this Online
Table was created, this will lose the data forever!
}
