% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/providers.R
\name{delete_provider}
\alias{delete_provider}
\alias{providersDelete}
\title{Delete a provider.}
\usage{
delete_provider(client, name)

providersDelete(client, name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Name of the provider.

See the \href{https://docs.databricks.com/api/workspace/providers/delete}{API documentation} for more
information on the request and response field structure.}
}
\description{
Deletes an authentication provider, if the caller is a metastore admin or is
the owner of the provider.
}
