% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{delete_query}
\alias{delete_query}
\alias{queriesDelete}
\title{Delete a query.}
\usage{
delete_query(client, query_id)

queriesDelete(client, query_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{query_id}{Required. This field has no description yet.

See the \href{https://docs.databricks.com/api/workspace/queries/delete}{API documentation} for more
information on the request and response field structure.}
}
\description{
Moves a query to the trash. Trashed queries immediately disappear from
searches and list views, and they cannot be used for alerts. The trash is
deleted after 30 days.
}
