% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registered_models.R
\name{delete_registered_model_alias}
\alias{delete_registered_model_alias}
\alias{registeredModelsDeleteAlias}
\title{Delete a Registered Model Alias.}
\usage{
delete_registered_model_alias(client, full_name, alias)

registeredModelsDeleteAlias(client, full_name, alias)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{full_name}{Required. The three-level (fully qualified) name of the registered model.

See the \href{https://docs.databricks.com/api/workspace/registeredmodels/deletealias}{API documentation} for more
information on the request and response field structure.}

\item{alias}{Required. The name of the alias.}
}
\description{
Deletes a registered model alias.
}
\details{
The caller must be a metastore admin or an owner of the registered model. For
the latter case, the caller must also be the owner or have the
\strong{USE_CATALOG} privilege on the parent catalog and the \strong{USE_SCHEMA}
privilege on the parent schema.
}
