% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\name{delete_schema}
\alias{delete_schema}
\title{Delete a schema.}
\usage{
delete_schema(client, full_name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{full_name}{Required. Full name of the schema.}
}
\description{
Deletes the specified schema from the parent catalog. The caller must be the
owner of the schema or an owner of the parent catalog.
}
