% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{delete_secret}
\alias{delete_secret}
\alias{secretsDeleteSecret}
\title{Delete a secret.}
\usage{
delete_secret(client, scope, key)

secretsDeleteSecret(client, scope, key)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{scope}{Required. The name of the scope that contains the secret to delete.}

\item{key}{Required. Name of the secret to delete.}
}
\description{
Deletes the secret stored in this secret scope. You must have \code{WRITE} or
\code{MANAGE} permission on the secret scope.
}
\details{
Throws \code{RESOURCE_DOES_NOT_EXIST} if no such secret scope or secret exists.
Throws \code{PERMISSION_DENIED} if the user does not have permission to make this
API call.
}
