% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{delete_secret_acl}
\alias{delete_secret_acl}
\alias{secretsDeleteAcl}
\title{Delete an ACL.}
\usage{
delete_secret_acl(client, scope, principal)

secretsDeleteAcl(client, scope, principal)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{scope}{Required. The name of the scope to remove permissions from.}

\item{principal}{Required. The principal to remove an existing ACL from.}
}
\description{
Deletes the given ACL on the given scope.
}
\details{
Users must have the \code{MANAGE} permission to invoke this API. Throws
\code{RESOURCE_DOES_NOT_EXIST} if no such secret scope, principal, or ACL exists.
Throws \code{PERMISSION_DENIED} if the user does not have permission to make this
API call.
}
