% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{delete_secret_scope}
\alias{delete_secret_scope}
\alias{secretsDeleteScope}
\title{Delete a secret scope.}
\usage{
delete_secret_scope(client, scope)

secretsDeleteScope(client, scope)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{scope}{Required. Name of the scope to delete.}
}
\description{
Deletes a secret scope.
}
\details{
Throws \code{RESOURCE_DOES_NOT_EXIST} if the scope does not exist. Throws
\code{PERMISSION_DENIED} if the user does not have permission to make this API
call.
}
