% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service_principals.R
\name{delete_service_principal}
\alias{delete_service_principal}
\alias{servicePrincipalsDelete}
\title{Delete a service principal.}
\usage{
delete_service_principal(client, id)

servicePrincipalsDelete(client, id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{id}{Required. Unique ID for a service principal in the Databricks workspace.

See the \href{https://docs.databricks.com/api/workspace/serviceprincipals/delete}{API documentation} for more
information on the request and response field structure.}
}
\description{
Delete a single service principal in the Databricks workspace.
}
