% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{delete_table}
\alias{delete_table}
\alias{tablesDelete}
\title{Delete a table.}
\usage{
delete_table(client, full_name)

tablesDelete(client, full_name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{full_name}{Required. Full name of the table.}
}
\description{
Deletes a table from the specified parent catalog and schema. The caller must
be the owner of the parent catalog, have the \strong{USE_CATALOG} privilege on the
parent catalog and be the owner of the parent schema, or be the owner of the
table and have the \strong{USE_CATALOG} privilege on the parent catalog and the
\strong{USE_SCHEMA} privilege on the parent schema.
}
