% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{delete_token}
\alias{delete_token}
\alias{tokensDelete}
\title{Revoke token.}
\usage{
delete_token(client, token_id)

tokensDelete(client, token_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{token_id}{Required. The ID of the token to be revoked.}
}
\description{
Revokes an access token.
}
\details{
If a token with the specified ID is not valid, this call returns an error
\strong{RESOURCE_DOES_NOT_EXIST}.
}
