% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_search_indexes.R
\name{delete_vector_search_index}
\alias{delete_vector_search_index}
\alias{vectorSearchIndexesDeleteIndex}
\title{Delete an index.}
\usage{
delete_vector_search_index(client, index_name)

vectorSearchIndexesDeleteIndex(client, index_name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{index_name}{Required. Name of the index.

See the \href{https://docs.databricks.com/api/workspace/vectorsearchindexes/deleteindex}{API documentation} for more
information on the request and response field structure.}
}
\description{
Delete an index.
}
