% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_search_indexes.R
\name{delete_vector_search_index_data}
\alias{delete_vector_search_index_data}
\alias{vectorSearchIndexesDeleteDataVectorIndex}
\title{Delete data from index.}
\usage{
delete_vector_search_index_data(client, index_name, primary_keys)

vectorSearchIndexesDeleteDataVectorIndex(client, index_name, primary_keys)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{index_name}{Required. Name of the vector index where data is to be deleted.}

\item{primary_keys}{Required. List of primary keys for the data to be deleted.

See the \href{https://docs.databricks.com/api/workspace/vectorsearchindexes/deletedatavectorindex}{API documentation} for more
information on the request and response field structure.}
}
\description{
Handles the deletion of data from a specified vector index.
}
