% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volumes.R
\name{delete_volume}
\alias{delete_volume}
\title{Delete a Volume.}
\usage{
delete_volume(client, name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. The three-level (fully qualified) name of the volume.}
}
\description{
Deletes a volume from the specified parent catalog and schema.
}
\details{
The caller must be a metastore admin or an owner of the volume. For the
latter case, the caller must also be the owner or have the \strong{USE_CATALOG}
privilege on the parent catalog and the \strong{USE_SCHEMA} privilege on the
parent schema.
}
