% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouses.R
\name{delete_warehouse}
\alias{delete_warehouse}
\alias{warehousesDelete}
\title{Delete a warehouse.}
\usage{
delete_warehouse(client, id)

warehousesDelete(client, id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{id}{Required. Required.

See the \href{https://docs.databricks.com/api/workspace/warehouses/delete}{API documentation} for more
information on the request and response field structure.}
}
\description{
Deletes a SQL warehouse.
}
