% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_schemas.R
\name{disable_system_schema}
\alias{disable_system_schema}
\alias{systemSchemasDisable}
\title{Disable a system schema.}
\usage{
disable_system_schema(client, metastore_id, schema_name)

systemSchemasDisable(client, metastore_id, schema_name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{metastore_id}{Required. The metastore ID under which the system schema lives.}

\item{schema_name}{Required. Full name of the system schema.

See the \href{https://docs.databricks.com/api/workspace/systemschemas/disable}{API documentation} for more
information on the request and response field structure.}
}
\description{
Disables the system schema and removes it from the system catalog. The caller
must be an account admin or a metastore admin.
}
