% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance_pools.R
\name{edit_instance_pool}
\alias{edit_instance_pool}
\alias{instancePoolsEdit}
\title{Edit an existing instance pool.}
\usage{
edit_instance_pool(
  client,
  instance_pool_id,
  instance_pool_name,
  node_type_id,
  custom_tags = NULL,
  idle_instance_autotermination_minutes = NULL,
  max_capacity = NULL,
  min_idle_instances = NULL
)

instancePoolsEdit(
  client,
  instance_pool_id,
  instance_pool_name,
  node_type_id,
  custom_tags = NULL,
  idle_instance_autotermination_minutes = NULL,
  max_capacity = NULL,
  min_idle_instances = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{instance_pool_id}{Required. Instance pool ID.}

\item{instance_pool_name}{Required. Pool name requested by the user.}

\item{node_type_id}{Required. This field encodes, through a single value, the resources available to each of the Spark nodes in this cluster.

See the \href{https://docs.databricks.com/api/workspace/instancepools/edit}{API documentation} for more
information on the request and response field structure.}

\item{custom_tags}{Additional tags for pool resources.}

\item{idle_instance_autotermination_minutes}{Automatically terminates the extra instances in the pool cache after they are inactive for this time in minutes if min_idle_instances requirement is already met.}

\item{max_capacity}{Maximum number of outstanding instances to keep in the pool, including both instances used by clusters and idle instances.}

\item{min_idle_instances}{Minimum number of idle instances to keep in the instance pool.}
}
\description{
Modifies the configuration of an existing instance pool.
}
