% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_schemas.R
\name{enable_system_schema}
\alias{enable_system_schema}
\alias{systemSchemasEnable}
\title{Enable a system schema.}
\usage{
enable_system_schema(client, metastore_id, schema_name)

systemSchemasEnable(client, metastore_id, schema_name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{metastore_id}{Required. The metastore ID under which the system schema lives.}

\item{schema_name}{Required. Full name of the system schema.

See the \href{https://docs.databricks.com/api/workspace/systemschemas/enable}{API documentation} for more
information on the request and response field structure.}
}
\description{
Enables the system schema and adds it to the system catalog. The caller must
be an account admin or a metastore admin.
}
