% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{events_cluster}
\alias{events_cluster}
\alias{clustersEvents}
\title{List cluster activity events.}
\usage{
events_cluster(
  client,
  cluster_id,
  end_time = NULL,
  event_types = NULL,
  limit = NULL,
  offset = NULL,
  order = NULL,
  start_time = NULL
)

clustersEvents(
  client,
  cluster_id,
  end_time = NULL,
  event_types = NULL,
  limit = NULL,
  offset = NULL,
  order = NULL,
  start_time = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{cluster_id}{Required. The ID of the cluster to retrieve events about.}

\item{end_time}{The end time in epoch milliseconds.}

\item{event_types}{An optional set of event types to filter on.}

\item{limit}{The maximum number of events to include in a page of events.}

\item{offset}{The offset in the result set.}

\item{order}{The order to list events in; either 'ASC' or 'DESC'.}

\item{start_time}{The start time in epoch milliseconds.}
}
\value{
\code{data.frame} with all of the response pages.

See the \href{https://docs.databricks.com/api/workspace/clusters/events}{API documentation} for more
information on the request and response field structure.
}
\description{
Retrieves a list of events about the activity of a cluster. This API is
paginated. If there are more events to read, the response includes all the
nparameters necessary to request the next page of events.
}
