% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command_execution.R
\name{execute_command}
\alias{execute_command}
\alias{commandExecutionExecute}
\title{Run a command.}
\usage{
execute_command(
  client,
  cluster_id = NULL,
  command = NULL,
  context_id = NULL,
  language = NULL
)

commandExecutionExecute(
  client,
  cluster_id = NULL,
  command = NULL,
  context_id = NULL,
  language = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{cluster_id}{Running cluster id.}

\item{command}{Executable code.}

\item{context_id}{Running context id.}

\item{language}{This field has no description yet.}
}
\description{
Runs a cluster command in the given execution context, using the provided
language.
}
\details{
If successful, it returns an ID for tracking the status of the command's
execution.
}
