% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{exists_table}
\alias{exists_table}
\alias{tablesExists}
\title{Get boolean reflecting if table exists.}
\usage{
exists_table(client, full_name)

tablesExists(client, full_name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{full_name}{Required. Full name of the table.}
}
\description{
Gets if a table exists in the metastore for a specific catalog and schema.
The caller must satisfy one of the following requirements: * Be a metastore
admin * Be the owner of the parent catalog * Be the owner of the parent
schema and have the USE_CATALOG privilege on the parent catalog * Have the
\strong{USE_CATALOG} privilege on the parent catalog and the \strong{USE_SCHEMA}
privilege on the parent schema, and either be the table owner or have the
SELECT privilege on the table. * Have BROWSE privilege on the parent catalog
\itemize{
\item Have BROWSE privilege on the parent schema.
}
}
