% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\docType{data}
\name{experiments}
\alias{experiments}
\title{Experiments are the primary unit of organization in MLflow; all MLflow runs
belong to an experiment. Each experiment lets you visualize, search, and
compare runs, as well as download run artifacts or metadata for analysis in
other tools. Experiments are maintained in a Databricks hosted MLflow
tracking server.}
\format{
An object of class \code{list} of length 24.
}
\usage{
experiments
}
\description{
Experiments are located in the workspace file tree. You manage experiments
using the same tools you use to manage other workspace objects such as
folders, notebooks, and libraries.
}
\section{Operations}{

\tabular{ll}{
\link[=experiments_create_experiment]{create_experiment} \tab Create experiment.\cr
\link[=experiments_create_run]{create_run} \tab Create a run.\cr
\link[=experiments_delete_experiment]{delete_experiment} \tab Delete an experiment.\cr
\link[=experiments_delete_run]{delete_run} \tab Delete a run.\cr
\link[=experiments_delete_tag]{delete_tag} \tab Delete a tag.\cr
\link[=experiments_get_by_name]{get_by_name} \tab Get metadata.\cr
\link[=experiments_get_experiment]{get_experiment} \tab Get an experiment.\cr
\link[=experiments_get_history]{get_history} \tab Get history of a given metric within a run.\cr
\link[=experiments_get_run]{get_run} \tab Get a run.\cr
\link[=experiments_list_artifacts]{list_artifacts} \tab Get all artifacts.\cr
\link[=experiments_list_experiments]{list_experiments} \tab List experiments.\cr
\link[=experiments_log_batch]{log_batch} \tab Log a batch.\cr
\link[=experiments_log_inputs]{log_inputs} \tab Log inputs to a run.\cr
\link[=experiments_log_metric]{log_metric} \tab Log a metric.\cr
\link[=experiments_log_model]{log_model} \tab Log a model.\cr
\link[=experiments_log_param]{log_param} \tab Log a param.\cr
\link[=experiments_restore_experiment]{restore_experiment} \tab Restores an experiment.\cr
\link[=experiments_restore_run]{restore_run} \tab Restore a run.\cr
\link[=experiments_search_experiments]{search_experiments} \tab Search experiments.\cr
\link[=experiments_search_runs]{search_runs} \tab Search for runs.\cr
\link[=experiments_set_experiment_tag]{set_experiment_tag} \tab Set a tag.\cr
\link[=experiments_set_tag]{set_tag} \tab Set a tag.\cr
\link[=experiments_update_experiment]{update_experiment} \tab Update an experiment.\cr
\link[=experiments_update_run]{update_run} \tab Update a run.\cr
}
}

\keyword{datasets}
