% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{experiments_delete}
\alias{experiments_delete}
\title{Delete an experiment.}
\usage{
experiments_delete(experiment_id)
}
\arguments{
\item{experiment_id}{Required. ID of the associated experiment.}
}
\description{
Marks an experiment and associated metadata, runs, metrics, params, and tags
for deletion. If the experiment uses FileStore, artifacts associated with
experiment are also deleted.
}
\keyword{internal}
