% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{experiments_delete_run}
\alias{experiments_delete_run}
\title{Delete a run.}
\usage{
experiments_delete_run(run_id)
}
\arguments{
\item{run_id}{Required. ID of the run to delete.}
}
\description{
Marks a run for deletion.
}
\keyword{internal}
