% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{experiments_get}
\alias{experiments_get}
\title{Get an experiment.}
\usage{
experiments_get(experiment_id, ...)
}
\arguments{
\item{experiment_id}{Required. ID of the associated experiment.}
}
\description{
Gets metadata for an experiment. This method works on deleted experiments.
}
\keyword{internal}
