% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{experiments_get_by_name}
\alias{experiments_get_by_name}
\title{Get metadata.}
\usage{
experiments_get_by_name(experiment_name)
}
\arguments{
\item{experiment_name}{Required. Name of the associated experiment.}
}
\description{
Gets metadata for an experiment.
}
\details{
This endpoint will return deleted experiments, but prefers the active
experiment if an active and deleted experiment share the same name. If
multiple deleted experiments share the same name, the API will return one of
them.

Throws \code{RESOURCE_DOES_NOT_EXIST} if no experiment with the specified name
exists.
}
\keyword{internal}
