% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{experiments_get_history}
\alias{experiments_get_history}
\title{Get history of a given metric within a run.}
\usage{
experiments_get_history(
  metric_key,
  max_results = NULL,
  page_token = NULL,
  run_id = NULL,
  run_uuid = NULL
)
}
\arguments{
\item{metric_key}{Required. Name of the metric.}

\item{max_results}{Maximum number of Metric records to return per paginated request.}

\item{page_token}{Token indicating the page of metric histories to fetch.}

\item{run_id}{ID of the run from which to fetch metric values.}

\item{run_uuid}{\link{Deprecated, use run_id instead} ID of the run from which to fetch metric values.}
}
\description{
Gets a list of all values for the specified metric for a given run.
}
\keyword{internal}
