% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{experiments_list}
\alias{experiments_list}
\title{List experiments.}
\usage{
experiments_list(max_results = NULL, page_token = NULL, view_type = NULL, ...)
}
\arguments{
\item{max_results}{Maximum number of experiments desired.}

\item{page_token}{Token indicating the page of experiments to fetch.}

\item{view_type}{Qualifier for type of experiments to be returned.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Gets a list of all experiments.
}
\keyword{internal}
