% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{experiments_log_inputs}
\alias{experiments_log_inputs}
\title{Log inputs to a run.}
\usage{
experiments_log_inputs(datasets = NULL, run_id = NULL)
}
\arguments{
\item{datasets}{Dataset inputs.}

\item{run_id}{ID of the run to log under.}
}
\description{
\strong{NOTE:} Experimental: This API may change or be removed in a future release
without warning.
}
\keyword{internal}
