% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{experiments_log_model}
\alias{experiments_log_model}
\title{Log a model.}
\usage{
experiments_log_model(model_json = NULL, run_id = NULL)
}
\arguments{
\item{model_json}{MLmodel file in json format.}

\item{run_id}{ID of the run to log under.}
}
\description{
\strong{NOTE:} Experimental: This API may change or be removed in a future release
without warning.
}
\keyword{internal}
