% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{experiments_restore_experiment}
\alias{experiments_restore_experiment}
\title{Restores an experiment.}
\usage{
experiments_restore_experiment(experiment_id)
}
\arguments{
\item{experiment_id}{Required. ID of the associated experiment.}
}
\description{
Restore an experiment marked for deletion. This also restores associated
metadata, runs, metrics, params, and tags. If experiment uses FileStore,
underlying artifacts associated with experiment are also restored.
}
\details{
Throws \code{RESOURCE_DOES_NOT_EXIST} if experiment was never created or was
permanently deleted.
}
\keyword{internal}
