% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{experiments_restore_run}
\alias{experiments_restore_run}
\title{Restore a run.}
\usage{
experiments_restore_run(run_id)
}
\arguments{
\item{run_id}{Required. ID of the run to restore.}
}
\description{
Restores a deleted run.
}
\keyword{internal}
