% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{experiments_set_experiment_tag}
\alias{experiments_set_experiment_tag}
\title{Set a tag.}
\usage{
experiments_set_experiment_tag(experiment_id, key, value)
}
\arguments{
\item{experiment_id}{Required. ID of the experiment under which to log the tag.}

\item{key}{Required. Name of the tag.}

\item{value}{Required. String value of the tag being logged.}
}
\description{
Sets a tag on an experiment. Experiment tags are metadata that can be
updated.
}
\keyword{internal}
