% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{experiments_update}
\alias{experiments_update}
\title{Update an experiment.}
\usage{
experiments_update(experiment_id, new_name = NULL, ...)
}
\arguments{
\item{experiment_id}{\link{required} ID of the associated experiment.}

\item{new_name}{If provided, the experiment's name is changed to the new name.}
}
\description{
Updates experiment metadata.
}
\keyword{internal}
