% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{experiments_update_experiment}
\alias{experiments_update_experiment}
\title{Update an experiment.}
\usage{
experiments_update_experiment(experiment_id, new_name = NULL)
}
\arguments{
\item{experiment_id}{Required. ID of the associated experiment.}

\item{new_name}{If provided, the experiment's name is changed to the new name.}
}
\description{
Updates experiment metadata.
}
\keyword{internal}
