% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{experiments_update_run}
\alias{experiments_update_run}
\title{Update a run.}
\usage{
experiments_update_run(
  end_time = NULL,
  run_id = NULL,
  run_uuid = NULL,
  status = NULL
)
}
\arguments{
\item{end_time}{Unix timestamp in milliseconds of when the run ended.}

\item{run_id}{ID of the run to update.}

\item{run_uuid}{\link{Deprecated, use run_id instead} ID of the run to update.}

\item{status}{Updated status of the run.}
}
\description{
Updates run metadata.
}
\keyword{internal}
