% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{experimentsGetExperiment}
\alias{experimentsGetExperiment}
\title{Get an experiment.}
\usage{
experimentsGetExperiment(client, experiment_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{experiment_id}{Required. ID of the associated experiment.}
}
\description{
Gets metadata for an experiment. This method works on deleted experiments.
}
