% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{experimentsGetExperimentPermissions}
\alias{experimentsGetExperimentPermissions}
\title{Get experiment permissions.}
\usage{
experimentsGetExperimentPermissions(client, experiment_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{experiment_id}{Required. The experiment for which to get or manage permissions.}
}
\description{
Gets the permissions of an experiment. Experiments can inherit permissions
from their root object.
}
