% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{experimentsGetHistory}
\alias{experimentsGetHistory}
\title{Get history of a given metric within a run.}
\usage{
experimentsGetHistory(
  client,
  metric_key,
  max_results = NULL,
  page_token = NULL,
  run_id = NULL,
  run_uuid = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{metric_key}{Required. Name of the metric.}

\item{max_results}{Maximum number of Metric records to return per paginated request.}

\item{page_token}{Token indicating the page of metric histories to fetch.}

\item{run_id}{ID of the run from which to fetch metric values.}

\item{run_uuid}{Deprecated, use run_id instead. ID of the run from which to fetch metric values.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Gets a list of all values for the specified metric for a given run.
}
