% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{experimentsGetPermissionLevels}
\alias{experimentsGetPermissionLevels}
\title{Get experiment permission levels.}
\usage{
experimentsGetPermissionLevels(client, experiment_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{experiment_id}{Required. The experiment for which to get or manage permissions.}
}
\description{
Gets the permission levels that a user can have on an object.
}
