% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{experimentsListArtifacts}
\alias{experimentsListArtifacts}
\title{Get all artifacts.}
\usage{
experimentsListArtifacts(
  client,
  page_token = NULL,
  path = NULL,
  run_id = NULL,
  run_uuid = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{page_token}{Token indicating the page of artifact results to fetch.}

\item{path}{Filter artifacts matching this path (a relative path from the root artifact directory).}

\item{run_id}{ID of the run whose artifacts to list.}

\item{run_uuid}{\link{Deprecated, use run_id instead} ID of the run whose artifacts to list.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
List artifacts for a run. Takes an optional \code{artifact_path} prefix. If it is
specified, the response contains only artifacts with the specified prefix.',
}
