% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{experimentsListExperiments}
\alias{experimentsListExperiments}
\title{List experiments.}
\usage{
experimentsListExperiments(
  client,
  max_results = NULL,
  page_token = NULL,
  view_type = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{max_results}{Maximum number of experiments desired.}

\item{page_token}{Token indicating the page of experiments to fetch.}

\item{view_type}{Qualifier for type of experiments to be returned.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Gets a list of all experiments.
}
