% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{experimentsRestoreExperiment}
\alias{experimentsRestoreExperiment}
\title{Restores an experiment.}
\usage{
experimentsRestoreExperiment(client, experiment_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{experiment_id}{Required. ID of the associated experiment.}
}
\description{
Restore an experiment marked for deletion. This also restores associated
metadata, runs, metrics, params, and tags. If experiment uses FileStore,
underlying artifacts associated with experiment are also restored.
}
\details{
Throws \code{RESOURCE_DOES_NOT_EXIST} if experiment was never created or was
permanently deleted.
}
