% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{experimentsRestoreRuns}
\alias{experimentsRestoreRuns}
\title{Restore runs by deletion time.}
\usage{
experimentsRestoreRuns(
  client,
  experiment_id,
  min_timestamp_millis,
  max_runs = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{experiment_id}{Required. The ID of the experiment containing the runs to restore.}

\item{min_timestamp_millis}{Required. The minimum deletion timestamp in milliseconds since the UNIX epoch for restoring runs.}

\item{max_runs}{An optional positive integer indicating the maximum number of runs to restore.}
}
\description{
Bulk restore runs in an experiment that were deleted no earlier than the
specified timestamp. Restores at most max_runs per request. To call this API
from a Databricks Notebook in Python, you can use the client code snippet on
https://learn.microsoft.com/en-us/azure/databricks/mlflow/runs#bulk-restore.
}
