% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{experimentsSetExperimentPermissions}
\alias{experimentsSetExperimentPermissions}
\title{Set experiment permissions.}
\usage{
experimentsSetExperimentPermissions(
  client,
  experiment_id,
  access_control_list = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{experiment_id}{Required. The experiment for which to get or manage permissions.}

\item{access_control_list}{}
}
\description{
Sets permissions on an experiment. Experiments can inherit permissions from
their root object.
}
