% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{experimentsSetTag}
\alias{experimentsSetTag}
\title{Set a tag.}
\usage{
experimentsSetTag(client, key, value, run_id = NULL, run_uuid = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{key}{Required. Name of the tag.}

\item{value}{Required. String value of the tag being logged.}

\item{run_id}{ID of the run under which to log the tag.}

\item{run_uuid}{\link{Deprecated, use run_id instead} ID of the run under which to log the tag.}
}
\description{
Sets a tag on a run. Tags are run metadata that can be updated during a run
and after a run completes.
}
