% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{export_job_run}
\alias{export_job_run}
\alias{jobsExportRun}
\title{Export and retrieve a job run.}
\usage{
export_job_run(client, run_id, views_to_export = NULL)

jobsExportRun(client, run_id, views_to_export = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{run_id}{Required. The canonical identifier for the run.}

\item{views_to_export}{Which views to export (CODE, DASHBOARDS, or ALL).}
}
\description{
Export and retrieve the job run task.
}
