% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serving_endpoints.R
\name{export_serving_endpoint_metrics}
\alias{export_serving_endpoint_metrics}
\alias{servingEndpointsExportMetrics}
\title{Get metrics of a serving endpoint.}
\usage{
export_serving_endpoint_metrics(client, name)

servingEndpointsExportMetrics(client, name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. The name of the serving endpoint to retrieve metrics for.

See the \href{https://docs.databricks.com/api/workspace/servingendpoints/exportmetrics}{API documentation} for more
information on the request and response field structure.}
}
\description{
Retrieves the metrics associated with the provided serving endpoint in either
Prometheus or OpenMetrics exposition format.
}
