% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_locations.R
\docType{data}
\name{external_locations}
\alias{external_locations}
\title{An external location is an object that combines a cloud storage path with a
storage credential that authorizes access to the cloud storage path. Each
external location is subject to Unity Catalog access-control policies that
control which users and groups can access the credential. If a user does not
have access to an external location in Unity Catalog, the request fails and
Unity Catalog does not attempt to authenticate to your cloud tenant on the
user’s behalf.}
\format{
An object of class \code{list} of length 5.
}
\usage{
external_locations
}
\description{
Databricks recommends using external locations rather than using storage
credentials directly.
}
\details{
To create external locations, you must be a metastore admin or a user with
the \strong{CREATE_EXTERNAL_LOCATION} privilege.
}
\section{Operations}{

\tabular{ll}{
\link[=external_locations_create]{create} \tab Create an external location.\cr
\link[=external_locations_delete]{delete} \tab Delete an external location.\cr
\link[=external_locations_get]{get} \tab Get an external location.\cr
\link[=external_locations_list]{list} \tab List external locations.\cr
\link[=external_locations_update]{update} \tab Update an external location.\cr
}
}

\keyword{datasets}
