% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_locations.R
\name{external_locations_create}
\alias{external_locations_create}
\title{Create an external location.}
\usage{
external_locations_create(
  name,
  url,
  credential_name,
  comment = NULL,
  read_only = NULL,
  skip_validation = NULL,
  ...
)
}
\arguments{
\item{name}{Required. Name of the external location.}

\item{url}{Required. Path URL of the external location.}

\item{credential_name}{Required. Name of the storage credential used with this location.}

\item{comment}{User-provided free-form text description.}

\item{read_only}{Indicates whether the external location is read-only.}

\item{skip_validation}{Skips validation of the storage credential associated with the external location.}
}
\description{
Creates a new external location entry in the metastore. The caller must be a
metastore admin or have the \strong{CREATE_EXTERNAL_LOCATION} privilege on both
the metastore and the associated storage credential.
}
\keyword{internal}
