% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_locations.R
\name{external_locations_delete}
\alias{external_locations_delete}
\title{Delete an external location.}
\usage{
external_locations_delete(name, force = NULL, ...)
}
\arguments{
\item{name}{Required. Name of the external location.}

\item{force}{Force deletion even if there are dependent external tables or mounts.}
}
\description{
Deletes the specified external location from the metastore. The caller must
be the owner of the external location.
}
\keyword{internal}
