% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_locations.R
\name{external_locations_get}
\alias{external_locations_get}
\title{Get an external location.}
\usage{
external_locations_get(name, ...)
}
\arguments{
\item{name}{\link{required} Name of the external location.}
}
\description{
Gets an external location from the metastore. The caller must be either a
metastore admin, the owner of the external location, or a user that has some
privilege on the external location.
}
\keyword{internal}
