% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_locations.R
\name{external_locations_list}
\alias{external_locations_list}
\title{List external locations.}
\usage{
external_locations_list()
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Gets an array of external locations (\strong{ExternalLocationInfo} objects) from
the metastore. The caller must be a metastore admin, the owner of the
external location, or a user that has some privilege on the external
location. There is no guarantee of a specific ordering of the elements in the
array.#'
}
\keyword{internal}
