% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_locations.R
\name{externalLocationsUpdate}
\alias{externalLocationsUpdate}
\title{Update an external location.}
\usage{
externalLocationsUpdate(
  client,
  name,
  comment = NULL,
  credential_name = NULL,
  force = NULL,
  owner = NULL,
  read_only = NULL,
  url = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Name of the external location.}

\item{comment}{User-provided free-form text description.}

\item{credential_name}{Name of the storage credential used with this location.}

\item{force}{Force update even if changing url invalidates dependent external tables or mounts.}

\item{owner}{The owner of the external location.}

\item{read_only}{Indicates whether the external location is read-only.}

\item{url}{Path URL of the external location.}
}
\description{
Updates an external location in the metastore. The caller must be the owner
of the external location, or be a metastore admin. In the second case, the
admin can only update the name of the external location.
}
