% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{filesCreateDirectory}
\alias{filesCreateDirectory}
\title{Create a directory.}
\usage{
filesCreateDirectory(client, directory_path)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{directory_path}{Required. The absolute path of a directory.}
}
\description{
Creates an empty directory. If necessary, also creates any parent directories
of the new, empty directory (like the shell command \code{mkdir -p}). If called on
an existing directory, returns a success response; this method is idempotent
(it will succeed if the directory already exists).
}
