% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{filesDelete}
\alias{filesDelete}
\title{Delete a file.}
\usage{
filesDelete(client, file_path)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{file_path}{Required. The absolute path of the file.}
}
\description{
Deletes a file. If the request is successful, there is no response body.
}
