% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{filesGetMetadata}
\alias{filesGetMetadata}
\title{Get file metadata.}
\usage{
filesGetMetadata(client, file_path)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{file_path}{Required. The absolute path of the file.}
}
\description{
Get the metadata of a file. The response HTTP headers contain the metadata.
There is no response body.
}
