% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{functions}
\alias{functions}
\title{Functions implement User-Defined Functions (UDFs) in Unity Catalog.}
\format{
An object of class \code{list} of length 5.
}
\usage{
functions
}
\description{
The function implementation can be any SQL expression or Query, and it can be
invoked wherever a table reference is allowed in a query. In Unity Catalog, a
function resides at the same level as a table, so it can be referenced with
the form \strong{catalog_name}.\strong{schema_name}.\strong{function_name}.
}
\section{Operations}{

\tabular{ll}{
\link[=functions_create]{create} \tab Create a function.\cr
\link[=functions_delete]{delete} \tab Delete a function.\cr
\link[=functions_get]{get} \tab Get a function.\cr
\link[=functions_list]{list} \tab List functions.\cr
\link[=functions_update]{update} \tab Update a function.\cr
}
}

\keyword{datasets}
