% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{functions_create}
\alias{functions_create}
\title{Create a function.}
\usage{
functions_create(
  name,
  catalog_name,
  schema_name,
  input_params,
  data_type,
  full_data_type,
  return_params,
  routine_body,
  routine_definition,
  routine_dependencies,
  parameter_style,
  is_deterministic,
  sql_data_access,
  is_null_call,
  security_type,
  specific_name,
  comment = NULL,
  external_language = NULL,
  external_name = NULL,
  properties = NULL,
  sql_path = NULL,
  ...
)
}
\arguments{
\item{name}{\link{required} Name of function, relative to parent schema.}

\item{catalog_name}{\link{required} Name of parent catalog.}

\item{schema_name}{\link{required} Name of parent schema relative to its parent catalog.}

\item{input_params}{\link{required} The array of \strong{FunctionParameterInfo} definitions of the function's parameters.}

\item{data_type}{\link{required} Scalar function return data type.}

\item{full_data_type}{\link{required} Pretty printed function data type.}

\item{return_params}{\link{required} Table function return parameters.}

\item{routine_body}{\link{required} Function language.}

\item{routine_definition}{\link{required} Function body.}

\item{routine_dependencies}{\link{required} Function dependencies.}

\item{parameter_style}{\link{required} Function parameter style.}

\item{is_deterministic}{\link{required} Whether the function is deterministic.}

\item{sql_data_access}{\link{required} Function SQL data access.}

\item{is_null_call}{\link{required} Function null call.}

\item{security_type}{\link{required} Function security type.}

\item{specific_name}{\link{required} Specific name of the function; Reserved for future use.}

\item{comment}{User-provided free-form text description.}

\item{external_language}{External function language.}

\item{external_name}{External function name.}

\item{properties}{A map of key-value properties attached to the securable.}

\item{sql_path}{List of schemes whose objects can be referenced without qualification.}
}
\description{
Creates a new function
}
\details{
The user must have the following permissions in order for the function to be
created: - \strong{USE_CATALOG} on the function's parent catalog - \strong{USE_SCHEMA}
and \strong{CREATE_FUNCTION} on the function's parent schema
}
\keyword{internal}
