% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{functions_get}
\alias{functions_get}
\title{Get a function.}
\usage{
functions_get(name, ...)
}
\arguments{
\item{name}{\link{required} The fully-qualified name of the function (of the form \strong{catalog_name}.\strong{schema_name}.\strong{function__name}).}
}
\description{
Gets a function from within a parent catalog and schema. For the fetch to
succeed, the user must satisfy one of the following requirements: - Is a
metastore admin - Is an owner of the function's parent catalog - Have the
\strong{USE_CATALOG} privilege on the function's parent catalog and be the owner
of the function - Have the \strong{USE_CATALOG} privilege on the function's parent
catalog, the \strong{USE_SCHEMA} privilege on the function's parent schema, and
the \strong{EXECUTE} privilege on the function itself
}
\keyword{internal}
