% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{functionsList}
\alias{functionsList}
\title{List functions.}
\usage{
functionsList(client, catalog_name, schema_name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{catalog_name}{Required. Name of parent catalog for functions of interest.}

\item{schema_name}{Required. Parent schema of functions.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
List functions within the specified parent catalog and schema. If the user is
a metastore admin, all functions are returned in the output list. Otherwise,
the user must have the \strong{USE_CATALOG} privilege on the catalog and the
\strong{USE_SCHEMA} privilege on the schema, and the output list contains only
functions for which either the user has the \strong{EXECUTE} privilege or the user
is the owner. There is no guarantee of a specific ordering of the elements in
the array.
}
