% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{get_alert}
\alias{get_alert}
\alias{alertsGet}
\title{Get an alert.}
\usage{
get_alert(client, alert_id)

alertsGet(client, alert_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{alert_id}{Required. This field has no description yet.

See the \href{https://docs.databricks.com/api/workspace/alerts/get}{API documentation} for more
information on the request and response field structure.}
}
\description{
Gets an alert.
}
