% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{get_app}
\alias{get_app}
\alias{appsGetApp}
\title{Get definition for an application.}
\usage{
get_app(client, name)

appsGetApp(client, name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. The name of an application.

See the \href{https://docs.databricks.com/api/workspace/apps/getapp}{API documentation} for more
information on the request and response field structure.}
}
\description{
Get an application definition
}
