% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact_allowlists.R
\name{get_artifact_allowlist}
\alias{get_artifact_allowlist}
\title{Get an artifact allowlist.}
\usage{
get_artifact_allowlist(client, artifact_type)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{artifact_type}{Required. The artifact type of the allowlist.}
}
\description{
Get the artifact allowlist of a certain artifact type. The caller must be a
metastore admin or have the \strong{MANAGE ALLOWLIST} privilege on the metastore.
}
