% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogs.R
\name{get_catalog}
\alias{get_catalog}
\alias{catalogsGet}
\title{Get a catalog.}
\usage{
get_catalog(client, name, include_browse = NULL)

catalogsGet(client, name, include_browse = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. The name of the catalog.

See the \href{https://docs.databricks.com/api/workspace/catalogs/get}{API documentation} for more
information on the request and response field structure.}

\item{include_browse}{Whether to include catalogs in the response for which the principal can only access selective metadata for.}
}
\description{
Gets the specified catalog in a metastore. The caller must be a metastore
admin, the owner of the catalog, or a user that has the \strong{USE_CATALOG}
privilege set for their account.
}
