% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_rooms.R
\name{get_clean_room}
\alias{get_clean_room}
\alias{cleanRoomsGet}
\title{Get a clean room.}
\usage{
get_clean_room(client, name, include_remote_details = NULL)

cleanRoomsGet(client, name, include_remote_details = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. The name of the clean room.}

\item{include_remote_details}{Whether to include remote details (central) on the clean room.}
}
\description{
Gets a data object clean room from the metastore. The caller must be a
metastore admin or the owner of the clean room.
}
