% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{get_cluster_permission_levels}
\alias{get_cluster_permission_levels}
\alias{clustersGetPermissionLevels}
\title{Get cluster permission levels.}
\usage{
get_cluster_permission_levels(client, cluster_id)

clustersGetPermissionLevels(client, cluster_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{cluster_id}{Required. The cluster for which to get or manage permissions.}
}
\description{
Gets the permission levels that a user can have on an object.
}
