% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_policies.R
\name{get_cluster_policy}
\alias{get_cluster_policy}
\alias{clusterPoliciesGet}
\title{Get a cluster policy.}
\usage{
get_cluster_policy(client, policy_id)

clusterPoliciesGet(client, policy_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{policy_id}{Required. Canonical unique identifier for the cluster policy.

See the \href{https://docs.databricks.com/api/workspace/clusterpolicies/get}{API documentation} for more
information on the request and response field structure.}
}
\description{
Get a cluster policy entity. Creation and editing is available to admins
only.
}
