% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_families.R
\name{get_cluster_policy_family}
\alias{get_cluster_policy_family}
\alias{policyFamiliesGet}
\title{Get policy family information.}
\usage{
get_cluster_policy_family(client, policy_family_id)

policyFamiliesGet(client, policy_family_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{policy_family_id}{Required. This field has no description yet.

See the \href{https://docs.databricks.com/api/workspace/policyfamilies/get}{API documentation} for more
information on the request and response field structure.}
}
\description{
Retrieve the information for an policy family based on its identifier.
}
