% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{get_connection}
\alias{get_connection}
\alias{connectionsGet}
\title{Get a connection.}
\usage{
get_connection(client, name)

connectionsGet(client, name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Name of the connection.

See the \href{https://docs.databricks.com/api/workspace/connections/get}{API documentation} for more
information on the request and response field structure.}
}
\description{
Gets a connection from it's name.
}
