% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboards.R
\name{get_dashboard}
\alias{get_dashboard}
\alias{dashboardsGet}
\title{Retrieve a definition.}
\usage{
get_dashboard(client, dashboard_id)

dashboardsGet(client, dashboard_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{dashboard_id}{Required. This field has no description yet.

See the \href{https://docs.databricks.com/api/workspace/dashboards/get}{API documentation} for more
information on the request and response field structure.}
}
\description{
Returns a JSON representation of a dashboard object, including its
visualization and query objects.
}
