% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfs.R
\name{get_dbfs_status}
\alias{get_dbfs_status}
\alias{dbfsGetStatus}
\title{Get the information of a file or directory.}
\usage{
get_dbfs_status(client, path)

dbfsGetStatus(client, path)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{path}{Required. The path of the file or directory.

See the \href{https://docs.databricks.com/api/workspace/dbfs/getstatus}{API documentation} for more
information on the request and response field structure.}
}
\description{
Gets the file information for a file or directory. If the file or directory
does not exist, this call throws an exception with \code{RESOURCE_DOES_NOT_EXIST}.
}
