% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbsql_permissions.R
\name{get_dbsql_permission}
\alias{get_dbsql_permission}
\alias{dbsqlPermissionsGet}
\title{Get object ACL.}
\usage{
get_dbsql_permission(client, object_type, object_id)

dbsqlPermissionsGet(client, object_type, object_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{object_type}{Required. The type of object permissions to check.

See the \href{https://docs.databricks.com/api/workspace/dbsqlpermissions/get}{API documentation} for more
information on the request and response field structure.}

\item{object_id}{Required. Object ID.}
}
\description{
Gets a JSON representation of the access control list (ACL) for a specified
object.
}
