% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esm_enablement.R
\name{get_esm_enablement}
\alias{get_esm_enablement}
\title{Get the enhanced security monitoring setting.}
\usage{
get_esm_enablement(client, etag = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{etag}{etag used for versioning.}
}
\description{
Gets the enhanced security monitoring setting.
}
