% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{get_experiment_by_name}
\alias{get_experiment_by_name}
\alias{experimentsGetByName}
\title{Get metadata.}
\usage{
get_experiment_by_name(client, experiment_name)

experimentsGetByName(client, experiment_name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{experiment_name}{Required. Name of the associated experiment.

See the \href{https://docs.databricks.com/api/workspace/experiments/getbyname}{API documentation} for more
information on the request and response field structure.}
}
\description{
Gets metadata for an experiment.
}
\details{
This endpoint will return deleted experiments, but prefers the active
experiment if an active and deleted experiment share the same name. If
multiple deleted experiments share the same name, the API will return one of
them.

Throws \code{RESOURCE_DOES_NOT_EXIST} if no experiment with the specified name
exists.
}
