% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{get_experiment_permission_levels}
\alias{get_experiment_permission_levels}
\alias{experimentsGetPermissionLevels}
\title{Get experiment permission levels.}
\usage{
get_experiment_permission_levels(client, experiment_id)

experimentsGetPermissionLevels(client, experiment_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{experiment_id}{Required. The experiment for which to get or manage permissions.

See the \href{https://docs.databricks.com/api/workspace/experiments/getpermissionlevels}{API documentation} for more
information on the request and response field structure.}
}
\description{
Gets the permission levels that a user can have on an object.
}
