% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{get_experiment_permissions}
\alias{get_experiment_permissions}
\alias{experimentsGetPermissions}
\title{Get experiment permissions.}
\usage{
get_experiment_permissions(client, experiment_id)

experimentsGetPermissions(client, experiment_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{experiment_id}{Required. The experiment for which to get or manage permissions.}
}
\description{
Gets the permissions of an experiment. Experiments can inherit permissions
from their root object.
}
