% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{get_experiment_run}
\alias{get_experiment_run}
\alias{experimentsGetRun}
\title{Get a run.}
\usage{
get_experiment_run(client, run_id, run_uuid = NULL)

experimentsGetRun(client, run_id, run_uuid = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{run_id}{Required. ID of the run to fetch.}

\item{run_uuid}{Deprecated, use run_id instead. ID of the run to fetch.

See the \href{https://docs.databricks.com/api/workspace/experiments/getrun}{API documentation} for more
information on the request and response field structure.}
}
\description{
Gets the metadata, metrics, params, and tags for a run. In the case where
multiple metrics with the same key are logged for a run, return only the
value with the latest timestamp.
}
\details{
If there are multiple values with the latest timestamp, return the maximum of
these values.
}
