% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_locations.R
\name{get_external_location}
\alias{get_external_location}
\title{Get an external location.}
\usage{
get_external_location(client, name, include_browse = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Name of the external location.}

\item{include_browse}{Whether to include external locations in the response for which the principal can only access selective metadata for.}
}
\description{
Gets an external location from the metastore. The caller must be either a
metastore admin, the owner of the external location, or a user that has some
privilege on the external location.
}
