% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{get_file_directory_metadata}
\alias{get_file_directory_metadata}
\title{Get directory metadata.}
\usage{
get_file_directory_metadata(client, directory_path)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{directory_path}{Required. The absolute path of a directory.}
}
\description{
Get the metadata of a directory. The response HTTP headers contain the
metadata. There is no response body.
}
\details{
This method is useful to check if a directory exists and the caller has
access to it.

If you wish to ensure the directory exists, you can instead use \code{PUT}, which
will create the directory if it does not exist, and is idempotent (it will
succeed if the directory already exists).
}
