% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{get_file_metadata}
\alias{get_file_metadata}
\alias{filesGetMetadata}
\title{Get file metadata.}
\usage{
get_file_metadata(client, file_path)

filesGetMetadata(client, file_path)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{file_path}{Required. The absolute path of the file.

See the \href{https://docs.databricks.com/api/workspace/files/getmetadata}{API documentation} for more
information on the request and response field structure.}
}
\description{
Get the metadata of a file. The response HTTP headers contain the metadata.
There is no response body.
}
