% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_init_scripts.R
\name{get_global_init_script}
\alias{get_global_init_script}
\alias{globalInitScriptsGet}
\title{Get an init script.}
\usage{
get_global_init_script(client, script_id)

globalInitScriptsGet(client, script_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{script_id}{Required. The ID of the global init script.}
}
\description{
Gets all the details of a script, including its Base64-encoded contents.
}
