% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grants.R
\name{get_grant_effective}
\alias{get_grant_effective}
\alias{grantsGetEffective}
\title{Get effective permissions.}
\usage{
get_grant_effective(client, securable_type, full_name, principal = NULL)

grantsGetEffective(client, securable_type, full_name, principal = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{securable_type}{Required. Type of securable.

See the \href{https://docs.databricks.com/api/workspace/grants/geteffective}{API documentation} for more
information on the request and response field structure.}

\item{full_name}{Required. Full name of securable.}

\item{principal}{If provided, only the effective permissions for the specified principal (user or group) are returned.}
}
\description{
Gets the effective permissions for a securable.
}
