% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance_pools.R
\name{get_instance_pool}
\alias{get_instance_pool}
\title{Get instance pool information.}
\usage{
get_instance_pool(client, instance_pool_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{instance_pool_id}{Required. The canonical unique identifier for the instance pool.}
}
\description{
Retrieve the information for an instance pool based on its identifier.
}
