% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance_pools.R
\name{get_instance_pool_permissions}
\alias{get_instance_pool_permissions}
\title{Get instance pool permissions.}
\usage{
get_instance_pool_permissions(client, instance_pool_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{instance_pool_id}{Required. The instance pool for which to get or manage permissions.}
}
\description{
Gets the permissions of an instance pool. Instance pools can inherit
permissions from their root object.
}
