% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{get_job}
\alias{get_job}
\alias{jobsGet}
\title{Get a single job.}
\usage{
get_job(client, job_id)

jobsGet(client, job_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{job_id}{Required. The canonical identifier of the job to retrieve information about.

See the \href{https://docs.databricks.com/api/workspace/jobs/get}{API documentation} for more
information on the request and response field structure.}
}
\description{
Retrieves the details for a single job.
}
