% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{get_job_permissions}
\alias{get_job_permissions}
\alias{jobsGetPermissions}
\title{Get job permissions.}
\usage{
get_job_permissions(client, job_id)

jobsGetPermissions(client, job_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{job_id}{Required. The job for which to get or manage permissions.

See the \href{https://docs.databricks.com/api/workspace/jobs/getpermissions}{API documentation} for more
information on the request and response field structure.}
}
\description{
Gets the permissions of a job. Jobs can inherit permissions from their root
object.
}
