% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{get_job_run}
\alias{get_job_run}
\alias{jobsGetRun}
\title{Get a single job run.}
\usage{
get_job_run(
  client,
  run_id,
  include_history = NULL,
  include_resolved_values = NULL
)

jobsGetRun(
  client,
  run_id,
  include_history = NULL,
  include_resolved_values = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{run_id}{Required. The canonical identifier of the run for which to retrieve the metadata.

See the \href{https://docs.databricks.com/api/workspace/jobs/getrun}{API documentation} for more
information on the request and response field structure.}

\item{include_history}{Whether to include the repair history in the response.}

\item{include_resolved_values}{Whether to include resolved parameter values in the response.}
}
\description{
Retrieve the metadata of a run.
}
