% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakehouse_monitors.R
\name{get_lakehouse_monitor}
\alias{get_lakehouse_monitor}
\alias{lakehouseMonitorsGet}
\title{Get a table monitor.}
\usage{
get_lakehouse_monitor(client, full_name)

lakehouseMonitorsGet(client, full_name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{full_name}{Required. Full name of the table.}
}
\description{
Gets a monitor for the specified table.
}
\details{
The caller must either: 1. be an owner of the table's parent catalog 2. have
\strong{USE_CATALOG} on the table's parent catalog and be an owner of the table's
parent schema. 3. have the following permissions: - \strong{USE_CATALOG} on the
table's parent catalog - \strong{USE_SCHEMA} on the table's parent schema -
\strong{SELECT} privilege on the table.

The returned information includes configuration values, as well as
information on assets created by the monitor. Some information (e.g.,
dashboard) may be filtered out if the caller is in a different workspace than
where the monitor was created.
}
