% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakehouse_monitors.R
\name{get_lakehouse_monitor_refresh}
\alias{get_lakehouse_monitor_refresh}
\alias{lakehouseMonitorsGetRefresh}
\title{Get refresh.}
\usage{
get_lakehouse_monitor_refresh(client, full_name, refresh_id)

lakehouseMonitorsGetRefresh(client, full_name, refresh_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{full_name}{Required. Full name of the table.}

\item{refresh_id}{Required. ID of the refresh.}
}
\description{
Gets info about a specific monitor refresh using the given refresh ID.
}
\details{
The caller must either: 1. be an owner of the table's parent catalog 2. have
\strong{USE_CATALOG} on the table's parent catalog and be an owner of the table's
parent schema 3. have the following permissions: - \strong{USE_CATALOG} on the
table's parent catalog - \strong{USE_SCHEMA} on the table's parent schema -
\strong{SELECT} privilege on the table.

Additionally, the call must be made from the workspace where the monitor was
created.
}
