% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeview.R
\name{get_lakeview}
\alias{get_lakeview}
\alias{lakeviewGet}
\title{Get dashboard.}
\usage{
get_lakeview(client, dashboard_id)

lakeviewGet(client, dashboard_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{dashboard_id}{Required. UUID identifying the dashboard.

See the \href{https://docs.databricks.com/api/workspace/lakeview/get}{API documentation} for more
information on the request and response field structure.}
}
\description{
Get a draft dashboard.
}
