% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metastores.R
\name{get_metastore}
\alias{get_metastore}
\alias{metastoresGet}
\title{Get a metastore.}
\usage{
get_metastore(client, id)

metastoresGet(client, id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{id}{Required. Unique ID of the metastore.

See the \href{https://docs.databricks.com/api/workspace/metastores/get}{API documentation} for more
information on the request and response field structure.}
}
\description{
Gets a metastore that matches the supplied ID. The caller must be a metastore
admin to retrieve this info.
}
