% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{get_model}
\alias{get_model}
\title{Get model.}
\usage{
get_model(client, name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Registered model unique name identifier.}
}
\description{
Get the details of a model. This is a Databricks workspace version of the
MLflow endpoint that also returns the model's Databricks workspace ID and
the permission level of the requesting user on the model.
}
\details{
MLflow endpoint: https://www.mlflow.org/docs/latest/rest-api.html#get-registeredmodel
}
