% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_versions.R
\name{get_model_version_by_alias}
\alias{get_model_version_by_alias}
\alias{modelVersionsGetByAlias}
\title{Get Model Version By Alias.}
\usage{
get_model_version_by_alias(client, full_name, alias)

modelVersionsGetByAlias(client, full_name, alias)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{full_name}{Required. The three-level (fully qualified) name of the registered model.

See the \href{https://docs.databricks.com/api/workspace/modelversions/getbyalias}{API documentation} for more
information on the request and response field structure.}

\item{alias}{Required. The name of the alias.}
}
\description{
Get a model version by alias.
}
\details{
The caller must be a metastore admin or an owner of (or have the \strong{EXECUTE}
privilege on) the registered model. For the latter case, the caller must also
be the owner or have the \strong{USE_CATALOG} privilege on the parent catalog and
the \strong{USE_SCHEMA} privilege on the parent schema.
}
