% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{get_model_version_download_uri}
\alias{get_model_version_download_uri}
\title{Get a model version URI.}
\usage{
get_model_version_download_uri(client, name, version)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Name of the registered model.}

\item{version}{Required. Model version number.}
}
\description{
Gets a URI to download the model version.
}
