% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{get_notebook_permission_levels}
\alias{get_notebook_permission_levels}
\alias{workspaceGetPermissionLevels}
\title{Get workspace object permission levels.}
\usage{
get_notebook_permission_levels(
  client,
  workspace_object_type,
  workspace_object_id
)

workspaceGetPermissionLevels(
  client,
  workspace_object_type,
  workspace_object_id
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{workspace_object_type}{Required. The workspace object type for which to get or manage permissions.

See the \href{https://docs.databricks.com/api/workspace/workspace/getpermissionlevels}{API documentation} for more
information on the request and response field structure.}

\item{workspace_object_id}{Required. The workspace object for which to get or manage permissions.}
}
\description{
Gets the permission levels that a user can have on an object.
}
