% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{get_notebook_permissions}
\alias{get_notebook_permissions}
\alias{workspaceGetPermissions}
\title{Get workspace object permissions.}
\usage{
get_notebook_permissions(client, workspace_object_type, workspace_object_id)

workspaceGetPermissions(client, workspace_object_type, workspace_object_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{workspace_object_type}{Required. The workspace object type for which to get or manage permissions.

See the \href{https://docs.databricks.com/api/workspace/workspace/getpermissions}{API documentation} for more
information on the request and response field structure.}

\item{workspace_object_id}{Required. The workspace object for which to get or manage permissions.}
}
\description{
Gets the permissions of a workspace object. Workspace objects can inherit
permissions from their parent objects or root object.
}
