% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_tables.R
\name{get_online_table}
\alias{get_online_table}
\alias{onlineTablesGet}
\title{Get an Online Table.}
\usage{
get_online_table(client, name)

onlineTablesGet(client, name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Full three-part (catalog, schema, table) name of the table.

See the \href{https://docs.databricks.com/api/workspace/onlinetables/get}{API documentation} for more
information on the request and response field structure.}
}
\description{
Get information about an existing online table and its status.
}
