% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{get_pipeline}
\alias{get_pipeline}
\alias{pipelinesGet}
\title{Get a pipeline.}
\usage{
get_pipeline(client, pipeline_id)

pipelinesGet(client, pipeline_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{pipeline_id}{Required. This field has no description yet.

See the \href{https://docs.databricks.com/api/workspace/pipelines/get}{API documentation} for more
information on the request and response field structure.}
}
\description{
Get a pipeline.
}
