% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{get_pipeline_permission_levels}
\alias{get_pipeline_permission_levels}
\alias{pipelinesGetPermissionLevels}
\title{Get pipeline permission levels.}
\usage{
get_pipeline_permission_levels(client, pipeline_id)

pipelinesGetPermissionLevels(client, pipeline_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{pipeline_id}{Required. The pipeline for which to get or manage permissions.

See the \href{https://docs.databricks.com/api/workspace/pipelines/getpermissionlevels}{API documentation} for more
information on the request and response field structure.}
}
\description{
Gets the permission levels that a user can have on an object.
}
