% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{get_pipeline_permissions}
\alias{get_pipeline_permissions}
\alias{pipelinesGetPermissions}
\title{Get pipeline permissions.}
\usage{
get_pipeline_permissions(client, pipeline_id)

pipelinesGetPermissions(client, pipeline_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{pipeline_id}{Required. The pipeline for which to get or manage permissions.

See the \href{https://docs.databricks.com/api/workspace/pipelines/getpermissions}{API documentation} for more
information on the request and response field structure.}
}
\description{
Gets the permissions of a pipeline. Pipelines can inherit permissions from
their root object.
}
