% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/providers.R
\name{get_provider}
\alias{get_provider}
\alias{providersGet}
\title{Get a provider.}
\usage{
get_provider(client, name)

providersGet(client, name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Name of the provider.

See the \href{https://docs.databricks.com/api/workspace/providers/get}{API documentation} for more
information on the request and response field structure.}
}
\description{
Gets a specific authentication provider. The caller must supply the name of
the provider, and must either be a metastore admin or the owner of the
provider.
}
