% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_query}
\alias{get_query}
\alias{queriesGet}
\title{Get a query definition.}
\usage{
get_query(client, query_id)

queriesGet(client, query_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{query_id}{Required. This field has no description yet.

See the \href{https://docs.databricks.com/api/workspace/queries/get}{API documentation} for more
information on the request and response field structure.}
}
\description{
Retrieve a query object definition along with contextual permissions
information about the currently authenticated user.
}
