% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registered_models.R
\name{get_registered_model}
\alias{get_registered_model}
\alias{registeredModelsGet}
\title{Get a Registered Model.}
\usage{
get_registered_model(client, full_name, include_browse = NULL)

registeredModelsGet(client, full_name, include_browse = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{full_name}{Required. The three-level (fully qualified) name of the registered model.}

\item{include_browse}{Whether to include registered models in the response for which the principal can only access selective metadata for.

See the \href{https://docs.databricks.com/api/workspace/registeredmodels/get}{API documentation} for more
information on the request and response field structure.}
}
\description{
Get a registered model.
}
\details{
The caller must be a metastore admin or an owner of (or have the \strong{EXECUTE}
privilege on) the registered model. For the latter case, the caller must also
be the owner or have the \strong{USE_CATALOG} privilege on the parent catalog and
the \strong{USE_SCHEMA} privilege on the parent schema.
}
