% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{get_repo}
\alias{get_repo}
\alias{reposGet}
\title{Get a repo.}
\usage{
get_repo(client, repo_id)

reposGet(client, repo_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{repo_id}{Required. The ID for the corresponding repo to access.

See the \href{https://docs.databricks.com/api/workspace/repos/get}{API documentation} for more
information on the request and response field structure.}
}
\description{
Returns the repo with the given repo ID.
}
