% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrict_workspace_admins.R
\name{get_restrict_workspace_admin}
\alias{get_restrict_workspace_admin}
\title{Get the restrict workspace admins setting.}
\usage{
get_restrict_workspace_admin(client, etag = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{etag}{etag used for versioning.}
}
\description{
Gets the restrict workspace admins setting.
}
