% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serving_endpoints.R
\name{get_serving_endpoint}
\alias{get_serving_endpoint}
\alias{servingEndpointsGet}
\title{Get a single serving endpoint.}
\usage{
get_serving_endpoint(client, name)

servingEndpointsGet(client, name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. The name of the serving endpoint.

See the \href{https://docs.databricks.com/api/workspace/servingendpoints/get}{API documentation} for more
information on the request and response field structure.}
}
\description{
Retrieves the details for a single serving endpoint.
}
