% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shares.R
\name{get_share}
\alias{get_share}
\alias{sharesGet}
\title{Get a share.}
\usage{
get_share(client, name, include_shared_data = NULL)

sharesGet(client, name, include_shared_data = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. The name of the share.

See the \href{https://docs.databricks.com/api/workspace/shares/get}{API documentation} for more
information on the request and response field structure.}

\item{include_shared_data}{Query for data to include in the share.}
}
\description{
Gets a data object share from the metastore. The caller must be a metastore
admin or the owner of the share.
}
