% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement_execution.R
\name{get_statement_execution}
\alias{get_statement_execution}
\title{Get status, manifest, and result first chunk.}
\usage{
get_statement_execution(client, statement_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{statement_id}{Required. The statement ID is returned upon successfully submitting a SQL statement, and is a required reference for all subsequent calls.}
}
\description{
This request can be used to poll for the statement's status. When the
\code{status.state} field is \code{SUCCEEDED} it will also return the result manifest
and the first chunk of the result data. When the statement is in the terminal
states \code{CANCELED}, \code{CLOSED} or \code{FAILED}, it returns HTTP 200 with the state
set. After at least 12 hours in terminal state, the statement is removed from
the warehouse and further calls will receive an HTTP 404 response.
}
\details{
\strong{NOTE} This call currently might take up to 5 seconds to get the latest
status and result.
}
