% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement_execution.R
\name{get_statement_execution_result_chunk_n}
\alias{get_statement_execution_result_chunk_n}
\title{Get result chunk by index.}
\usage{
get_statement_execution_result_chunk_n(client, statement_id, chunk_index)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{statement_id}{Required. The statement ID is returned upon successfully submitting a SQL statement, and is a required reference for all subsequent calls.}

\item{chunk_index}{Required. This field has no description yet.}
}
\description{
After the statement execution has \code{SUCCEEDED}, this request can be used to
fetch any chunk by index. Whereas the first chunk with \code{chunk_index=0} is
typically fetched with :method:statementexecution/executeStatement or
:method:statementexecution/getStatement, this request can be used to fetch
subsequent chunks. The response structure is identical to the nested \code{result}
element described in the :method:statementexecution/getStatement request, and
similarly includes the \code{next_chunk_index} and \code{next_chunk_internal_link}
fields for simple iteration through the result set.
}
