% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_credentials.R
\name{get_storage_credential}
\alias{get_storage_credential}
\title{Get a credential.}
\usage{
get_storage_credential(client, name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Name of the storage credential.}
}
\description{
Gets a storage credential from the metastore. The caller must be a metastore
admin, the owner of the storage credential, or have some permission on the
storage credential.
}
