% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{get_table}
\alias{get_table}
\alias{tablesGet}
\title{Get a table.}
\usage{
get_table(
  client,
  full_name,
  include_browse = NULL,
  include_delta_metadata = NULL
)

tablesGet(
  client,
  full_name,
  include_browse = NULL,
  include_delta_metadata = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{full_name}{Required. Full name of the table.}

\item{include_browse}{Whether to include tables in the response for which the principal can only access selective metadata for.}

\item{include_delta_metadata}{Whether delta metadata should be included in the response.

See the \href{https://docs.databricks.com/api/workspace/tables/get}{API documentation} for more
information on the request and response field structure.}
}
\description{
Gets a table from the metastore for a specific catalog and schema. The caller
must satisfy one of the following requirements: * Be a metastore admin * Be
the owner of the parent catalog * Be the owner of the parent schema and have
the USE_CATALOG privilege on the parent catalog * Have the \strong{USE_CATALOG}
privilege on the parent catalog and the \strong{USE_SCHEMA} privilege on the
parent schema, and either be the table owner or have the SELECT privilege on
the table.
}
