% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_management.R
\name{get_token_management}
\alias{get_token_management}
\alias{tokenManagementGet}
\title{Get token info.}
\usage{
get_token_management(client, token_id)

tokenManagementGet(client, token_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{token_id}{Required. The ID of the token to get.

See the \href{https://docs.databricks.com/api/workspace/tokenmanagement/get}{API documentation} for more
information on the request and response field structure.}
}
\description{
Gets information about a token, specified by its ID.
}
