% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{get_user}
\alias{get_user}
\title{Get user details.}
\usage{
get_user(
  client,
  id,
  attributes = NULL,
  count = NULL,
  excluded_attributes = NULL,
  filter = NULL,
  sort_by = NULL,
  sort_order = NULL,
  start_index = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{id}{Required. Unique ID for a user in the Databricks workspace.}

\item{attributes}{Comma-separated list of attributes to return in response.}

\item{count}{Desired number of results per page.}

\item{excluded_attributes}{Comma-separated list of attributes to exclude in response.}

\item{filter}{Query by which the results have to be filtered.}

\item{sort_by}{Attribute to sort the results.}

\item{sort_order}{The order to sort the results.}

\item{start_index}{Specifies the index of the first result.}
}
\description{
Gets information for a specific user in Databricks workspace.
}
