% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{get_user_permissions}
\alias{get_user_permissions}
\title{Get password permissions.}
\usage{
get_user_permissions(client)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}
}
\description{
Gets the permissions of all passwords. Passwords can inherit permissions from
their root object.
}
