% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_search_endpoints.R
\name{get_vector_search_endpoint}
\alias{get_vector_search_endpoint}
\alias{vectorSearchEndpointsGetEndpoint}
\title{Get an endpoint.}
\usage{
get_vector_search_endpoint(client, endpoint_name)

vectorSearchEndpointsGetEndpoint(client, endpoint_name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{endpoint_name}{Required. Name of the endpoint.

See the \href{https://docs.databricks.com/api/workspace/vectorsearchendpoints/getendpoint}{API documentation} for more
information on the request and response field structure.}
}
\description{
Get an endpoint.
}
