% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouses.R
\name{get_warehouse}
\alias{get_warehouse}
\alias{warehousesGet}
\title{Get warehouse info.}
\usage{
get_warehouse(client, id)

warehousesGet(client, id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{id}{Required. Required.

See the \href{https://docs.databricks.com/api/workspace/warehouses/get}{API documentation} for more
information on the request and response field structure.}
}
\description{
Gets the information for a single SQL warehouse.
}
