% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouses.R
\name{get_warehouse_permissions}
\alias{get_warehouse_permissions}
\alias{warehousesGetPermissions}
\title{Get SQL warehouse permissions.}
\usage{
get_warehouse_permissions(client, warehouse_id)

warehousesGetPermissions(client, warehouse_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{warehouse_id}{Required. The SQL warehouse for which to get or manage permissions.

See the \href{https://docs.databricks.com/api/workspace/warehouses/getpermissions}{API documentation} for more
information on the request and response field structure.}
}
\description{
Gets the permissions of a SQL warehouse. SQL warehouses can inherit
permissions from their root object.
}
