% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouses.R
\name{get_warehouse_workspace_config}
\alias{get_warehouse_workspace_config}
\title{Get the workspace configuration.}
\usage{
get_warehouse_workspace_config(client)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}
}
\description{
Gets the workspace level configuration that is shared by all SQL warehouses
in a workspace.
}
