% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace_bindings.R
\name{get_workspace_binding}
\alias{get_workspace_binding}
\alias{workspaceBindingsGet}
\title{Get catalog workspace bindings.}
\usage{
get_workspace_binding(client, name)

workspaceBindingsGet(client, name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. The name of the catalog.

See the \href{https://docs.databricks.com/api/workspace/workspacebindings/get}{API documentation} for more
information on the request and response field structure.}
}
\description{
Gets workspace bindings of the catalog. The caller must be a metastore admin
or an owner of the catalog.
}
