% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace_bindings.R
\name{get_workspace_binding_bindings}
\alias{get_workspace_binding_bindings}
\alias{workspaceBindingsGetBindings}
\title{Get securable workspace bindings.}
\usage{
get_workspace_binding_bindings(client, securable_type, securable_name)

workspaceBindingsGetBindings(client, securable_type, securable_name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{securable_type}{Required. The type of the securable.

See the \href{https://docs.databricks.com/api/workspace/workspacebindings/getbindings}{API documentation} for more
information on the request and response field structure.}

\item{securable_name}{Required. The name of the securable.}
}
\description{
Gets workspace bindings of the securable. The caller must be a metastore
admin or an owner of the securable.
}
