% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace_conf.R
\name{get_workspace_conf_status}
\alias{get_workspace_conf_status}
\alias{workspaceConfGetStatus}
\title{Check configuration status.}
\usage{
get_workspace_conf_status(client, keys)

workspaceConfGetStatus(client, keys)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{keys}{Required. This field has no description yet.

See the \href{https://docs.databricks.com/api/workspace/workspaceconf/getstatus}{API documentation} for more
information on the request and response field structure.}
}
\description{
Gets the configuration status for a workspace.
}
