% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_credentials.R
\docType{data}
\name{git_credentials}
\alias{git_credentials}
\title{Registers personal access token for Databricks to do operations on behalf of
the user.}
\format{
An object of class \code{list} of length 5.
}
\usage{
git_credentials
}
\description{
See \link{more info}.
}
\details{

}
\section{Operations}{

\tabular{ll}{
\link[=git_credentials_create]{create} \tab Create a credential entry.\cr
\link[=git_credentials_delete]{delete} \tab Delete a credential.\cr
\link[=git_credentials_get]{get} \tab Get a credential entry.\cr
\link[=git_credentials_list]{list} \tab Get Git credentials.\cr
\link[=git_credentials_update]{update} \tab Update a credential.\cr
}
}

\keyword{datasets}
