% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_credentials.R
\name{git_credentials_create}
\alias{git_credentials_create}
\title{Create a credential entry.}
\usage{
git_credentials_create(
  git_provider,
  git_username = NULL,
  personal_access_token = NULL,
  ...
)
}
\arguments{
\item{git_provider}{Required. Git provider.}

\item{git_username}{Git username.}

\item{personal_access_token}{The personal access token used to authenticate to the corresponding Git provider.}
}
\description{
Creates a Git credential entry for the user. Only one Git credential per user
is supported, so any attempts to create credentials if an entry already
exists will fail. Use the PATCH endpoint to update existing credentials, or
the DELETE endpoint to delete existing credentials.
}
\keyword{internal}
