% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_credentials.R
\name{git_credentials_delete}
\alias{git_credentials_delete}
\title{Delete a credential.}
\usage{
git_credentials_delete(credential_id, ...)
}
\arguments{
\item{credential_id}{Required. The ID for the corresponding credential to access.}
}
\description{
Deletes the specified Git credential.
}
\keyword{internal}
