% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_credentials.R
\name{git_credentials_get}
\alias{git_credentials_get}
\title{Get a credential entry.}
\usage{
git_credentials_get(credential_id, ...)
}
\arguments{
\item{credential_id}{Required. The ID for the corresponding credential to access.}
}
\description{
Gets the Git credential with the specified credential ID.
}
\keyword{internal}
