% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_credentials.R
\name{git_credentials_list}
\alias{git_credentials_list}
\title{Get Git credentials.}
\usage{
git_credentials_list()
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Lists the calling user's Git credentials. One credential per user is
supported.#'
}
\keyword{internal}
