% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_credentials.R
\name{git_credentials_update}
\alias{git_credentials_update}
\title{Update a credential.}
\usage{
git_credentials_update(
  credential_id,
  git_provider = NULL,
  git_username = NULL,
  personal_access_token = NULL,
  ...
)
}
\arguments{
\item{credential_id}{\link{required} The ID for the corresponding credential to access.}

\item{git_provider}{Git provider.}

\item{git_username}{Git username.}

\item{personal_access_token}{The personal access token used to authenticate to the corresponding Git provider.}
}
\description{
Updates the specified Git credential.
}
\keyword{internal}
