% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_credentials.R
\name{gitCredentialsGet}
\alias{gitCredentialsGet}
\title{Get a credential entry.}
\usage{
gitCredentialsGet(client, credential_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{credential_id}{Required. The ID for the corresponding credential to access.}
}
\description{
Gets the Git credential with the specified credential ID.
}
