% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_credentials.R
\name{gitCredentialsList}
\alias{gitCredentialsList}
\title{Get Git credentials.}
\usage{
gitCredentialsList(client)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Lists the calling user's Git credentials. One credential per user is
supported.
}
