% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_credentials.R
\name{gitCredentialsUpdate}
\alias{gitCredentialsUpdate}
\title{Update a credential.}
\usage{
gitCredentialsUpdate(
  client,
  credential_id,
  git_provider = NULL,
  git_username = NULL,
  personal_access_token = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{credential_id}{Required. The ID for the corresponding credential to access.}

\item{git_provider}{Git provider.}

\item{git_username}{Git username.}

\item{personal_access_token}{The personal access token used to authenticate to the corresponding Git provider.}
}
\description{
Updates the specified Git credential.
}
