% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_init_scripts.R
\docType{data}
\name{global_init_scripts}
\alias{global_init_scripts}
\title{The Global Init Scripts API enables Workspace administrators to configure
global initialization scripts for their workspace. These scripts run on every
node in every cluster in the workspace.}
\format{
An object of class \code{list} of length 5.
}
\usage{
global_init_scripts
}
\description{
\strong{Important:} Existing clusters must be restarted to pick up any changes
made to global init scripts. Global init scripts are run in order. If the
init script returns with a bad exit code, the Apache Spark container fails to
launch and init scripts with later position are skipped. If enough containers
fail, the entire cluster fails with a \code{GLOBAL_INIT_SCRIPT_FAILURE} error
code.
}
\section{Operations}{

\tabular{ll}{
\link[=global_init_scripts_create]{create} \tab Create init script.\cr
\link[=global_init_scripts_delete]{delete} \tab Delete init script.\cr
\link[=global_init_scripts_get]{get} \tab Get an init script.\cr
\link[=global_init_scripts_list]{list} \tab Get init scripts.\cr
\link[=global_init_scripts_update]{update} \tab Update init script.\cr
}
}

\keyword{datasets}
