% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_init_scripts.R
\name{global_init_scripts_delete}
\alias{global_init_scripts_delete}
\title{Delete init script.}
\usage{
global_init_scripts_delete(script_id, ...)
}
\arguments{
\item{script_id}{\link{required} The ID of the global init script.}
}
\description{
Deletes a global init script.
}
\keyword{internal}
