% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_init_scripts.R
\name{global_init_scripts_get}
\alias{global_init_scripts_get}
\title{Get an init script.}
\usage{
global_init_scripts_get(script_id, ...)
}
\arguments{
\item{script_id}{\link{required} The ID of the global init script.}
}
\description{
Gets all the details of a script, including its Base64-encoded contents.
}
\keyword{internal}
