% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_init_scripts.R
\name{global_init_scripts_list}
\alias{global_init_scripts_list}
\title{Get init scripts.}
\usage{
global_init_scripts_list()
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Get a list of all global init scripts for this workspace. This returns all
properties for each script but \strong{not} the script contents. To retrieve the
contents of a script, use the \href{#operation/get-script}{get a global init script} operation.#'
}
\keyword{internal}
