% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_init_scripts.R
\name{globalInitScriptsDelete}
\alias{globalInitScriptsDelete}
\title{Delete init script.}
\usage{
globalInitScriptsDelete(client, script_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{script_id}{Required. The ID of the global init script.}
}
\description{
Deletes a global init script.
}
