% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_init_scripts.R
\name{globalInitScriptsList}
\alias{globalInitScriptsList}
\title{Get init scripts.}
\usage{
globalInitScriptsList(client)
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Get a list of all global init scripts for this workspace. This returns all
properties for each script but \strong{not} the script contents. To retrieve the
contents of a script, use the \href{#operation/get-script}{get a global init script} operation.#'
}
