% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grants.R
\docType{data}
\name{grants}
\alias{grants}
\title{In Unity Catalog, data is secure by default. Initially, users have no access
to data in a metastore. Access can be granted by either a metastore admin,
the owner of an object, or the owner of the catalog or schema that contains
the object. Securable objects in Unity Catalog are hierarchical and
privileges are inherited downward.}
\format{
An object of class \code{list} of length 3.
}
\usage{
grants
}
\description{
Securable objects in Unity Catalog are hierarchical and privileges are
inherited downward. This means that granting a privilege on the catalog
automatically grants the privilege to all current and future objects within
the catalog. Similarly, privileges granted on a schema are inherited by all
current and future objects within that schema.
}
\section{Operations}{

\tabular{ll}{
\link[=grants_get]{get} \tab Get permissions.\cr
\link[=grants_get_effective]{get_effective} \tab Get effective permissions.\cr
\link[=grants_update]{update} \tab Update permissions.\cr
}
}

\keyword{datasets}
