% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grants.R
\name{grants_get}
\alias{grants_get}
\title{Get permissions.}
\usage{
grants_get(securable_type, full_name, principal = NULL, ...)
}
\arguments{
\item{securable_type}{Required. Type of securable.}

\item{full_name}{Required. Full name of securable.}

\item{principal}{If provided, only the permissions for the specified principal (user or group) are returned.}
}
\description{
Gets the permissions for a securable.
}
\keyword{internal}
