% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grants.R
\name{grants_get_effective}
\alias{grants_get_effective}
\title{Get effective permissions.}
\usage{
grants_get_effective(securable_type, full_name, principal = NULL)
}
\arguments{
\item{securable_type}{Required. Type of securable.}

\item{full_name}{Required. Full name of securable.}

\item{principal}{If provided, only the effective permissions for the specified principal (user or group) are returned.}
}
\description{
Gets the effective permissions for a securable.
}
\keyword{internal}
