% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grants.R
\name{grants_update}
\alias{grants_update}
\title{Update permissions.}
\usage{
grants_update(securable_type, full_name, changes = NULL, ...)
}
\arguments{
\item{securable_type}{\link{required} Type of securable.}

\item{full_name}{\link{required} Full name of securable.}

\item{changes}{Array of permissions change objects.}
}
\description{
Updates the permissions for a securable.
}
\keyword{internal}
