% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grants.R
\name{grantsGet}
\alias{grantsGet}
\title{Get permissions.}
\usage{
grantsGet(client, securable_type, full_name, principal = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{securable_type}{Required. Type of securable.}

\item{full_name}{Required. Full name of securable.}

\item{principal}{If provided, only the permissions for the specified principal (user or group) are returned.}
}
\description{
Gets the permissions for a securable.
}
