% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\docType{data}
\name{groups}
\alias{groups}
\title{Groups simplify identity management, making it easier to assign access to
Databricks Workspace, data, and other securable objects.}
\format{
An object of class \code{list} of length 6.
}
\usage{
groups
}
\description{
It is best practice to assign access to workspaces and access-control
policies in Unity Catalog to groups, instead of to users individually. All
Databricks Workspace identities can be assigned as members of groups, and
members inherit permissions that are assigned to their group.
}
\section{Operations}{

\tabular{ll}{
\link[=groups_create]{create} \tab Create a new group.\cr
\link[=groups_delete]{delete} \tab Delete a group.\cr
\link[=groups_get]{get} \tab Get group details.\cr
\link[=groups_list]{list} \tab List group details.\cr
\link[=groups_patch]{patch} \tab Update group details.\cr
\link[=groups_update]{update} \tab Replace a group.\cr
}
}

\keyword{datasets}
