% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{groups_create}
\alias{groups_create}
\title{Create a new group.}
\usage{
groups_create(
  id,
  display_name = NULL,
  entitlements = NULL,
  external_id = NULL,
  groups = NULL,
  members = NULL,
  roles = NULL
)
}
\arguments{
\item{id}{Databricks group ID.}

\item{display_name}{String that represents a human-readable group name.}

\item{entitlements}{}

\item{external_id}{}

\item{groups}{}

\item{members}{}

\item{roles}{}
}
\description{
Creates a group in the Databricks Workspace with a unique name, using the
supplied group details.
}
\keyword{internal}
