% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{groups_delete}
\alias{groups_delete}
\title{Delete a group.}
\usage{
groups_delete(id)
}
\arguments{
\item{id}{Required. Unique ID for a group in the Databricks Workspace.}
}
\description{
Deletes a group from the Databricks Workspace.
}
\keyword{internal}
