% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{groups_get}
\alias{groups_get}
\title{Get group details.}
\usage{
groups_get(id)
}
\arguments{
\item{id}{Required. Unique ID for a group in the Databricks Workspace.}
}
\description{
Gets the information for a specific group in the Databricks Workspace.
}
\keyword{internal}
