% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{groups_patch}
\alias{groups_patch}
\title{Update group details.}
\usage{
groups_patch(id, operations = NULL, ...)
}
\arguments{
\item{id}{\link{required} Unique ID for a group in the Databricks Account.}

\item{operations}{}
}
\description{
Partially updates the details of a group.
}
\keyword{internal}
