% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{groups_update}
\alias{groups_update}
\title{Replace a group.}
\usage{
groups_update(
  id,
  display_name = NULL,
  entitlements = NULL,
  external_id = NULL,
  groups = NULL,
  members = NULL,
  roles = NULL
)
}
\arguments{
\item{id}{Databricks group ID.}

\item{display_name}{String that represents a human-readable group name.}

\item{entitlements}{}

\item{external_id}{}

\item{groups}{}

\item{members}{}

\item{roles}{}
}
\description{
Updates the details of a group by replacing the entire group entity.
}
\keyword{internal}
