% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{groupsCreate}
\alias{groupsCreate}
\title{Create a new group.}
\usage{
groupsCreate(
  client,
  display_name = NULL,
  entitlements = NULL,
  external_id = NULL,
  groups = NULL,
  id = NULL,
  members = NULL,
  meta = NULL,
  roles = NULL,
  schemas = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{display_name}{String that represents a human-readable group name.}

\item{entitlements}{Entitlements assigned to the group.}

\item{external_id}{This field has no description yet.}

\item{groups}{This field has no description yet.}

\item{id}{Databricks group ID.}

\item{members}{This field has no description yet.}

\item{meta}{Container for the group identifier.}

\item{roles}{Corresponds to AWS instance profile/arn role.}

\item{schemas}{The schema of the group.}
}
\description{
Creates a group in the Databricks workspace with a unique name, using the
supplied group details.
}
