% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{groupsDelete}
\alias{groupsDelete}
\title{Delete a group.}
\usage{
groupsDelete(client, id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{id}{Required. Unique ID for a group in the Databricks workspace.}
}
\description{
Deletes a group from the Databricks workspace.
}
