% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{groupsGet}
\alias{groupsGet}
\title{Get group details.}
\usage{
groupsGet(client, id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{id}{Required. Unique ID for a group in the Databricks workspace.}
}
\description{
Gets the information for a specific group in the Databricks workspace.
}
