% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraries.R
\name{install_cluster_library}
\alias{install_cluster_library}
\title{Add a library.}
\usage{
install_cluster_library(client, cluster_id, libraries)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{cluster_id}{Required. Unique identifier for the cluster on which to install these libraries.}

\item{libraries}{Required. The libraries to install.}
}
\description{
Add libraries to be installed on a cluster. The installation is asynchronous;
it happens in the background after the completion of this request.
}
\details{
\strong{Note}: The actual set of libraries to be installed on a cluster is the
union of the libraries specified via this method and the libraries set to be
installed on all clusters via the libraries UI.
}
