% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance_pools.R
\docType{data}
\name{instance_pools}
\alias{instance_pools}
\title{Instance Pools API are used to create, edit, delete and list instance pools
by using ready-to-use cloud instances which reduces a cluster start and
auto-scaling times.}
\format{
An object of class \code{list} of length 5.
}
\usage{
instance_pools
}
\description{
Databricks pools reduce cluster start and auto-scaling times by maintaining a
set of idle, ready-to-use instances. When a cluster is attached to a pool,
cluster nodes are created using the pool’s idle instances. If the pool has
no idle instances, the pool expands by allocating a new instance from the
instance provider in order to accommodate the cluster’s request. When a
cluster releases an instance, it returns to the pool and is free for another
cluster to use. Only clusters attached to a pool can use that pool’s idle
instances.
}
\details{
You can specify a different pool for the driver node and worker nodes, or use
the same pool for both.

Databricks does not charge DBUs while instances are idle in the pool.
Instance provider billing does apply. See pricing.
}
\section{Operations}{

\tabular{ll}{
\link[=instance_pools_create]{create} \tab Create a new instance pool.\cr
\link[=instance_pools_delete]{delete} \tab Delete an instance pool.\cr
\link[=instance_pools_edit]{edit} \tab Edit an existing instance pool.\cr
\link[=instance_pools_get]{get} \tab Get instance pool information.\cr
\link[=instance_pools_list]{list} \tab List instance pool info.\cr
}
}

\keyword{datasets}
