% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance_pools.R
\name{instance_pools_create}
\alias{instance_pools_create}
\title{Create a new instance pool.}
\usage{
instance_pools_create(
  instance_pool_name,
  node_type_id,
  aws_attributes = NULL,
  azure_attributes = NULL,
  custom_tags = NULL,
  disk_spec = NULL,
  enable_elastic_disk = NULL,
  idle_instance_autotermination_minutes = NULL,
  instance_pool_fleet_attributes = NULL,
  max_capacity = NULL,
  min_idle_instances = NULL,
  preloaded_docker_images = NULL,
  preloaded_spark_versions = NULL
)
}
\arguments{
\item{instance_pool_name}{Required. Pool name requested by the user.}

\item{node_type_id}{Required. This field encodes, through a single value, the resources available to each of the Spark nodes in this cluster.}

\item{aws_attributes}{Attributes related to pool running on Amazon Web Services.}

\item{azure_attributes}{Attributes related to pool running on Azure.}

\item{custom_tags}{Additional tags for pool resources.}

\item{disk_spec}{Defines the specification of the disks that will be attached to all spark containers.}

\item{enable_elastic_disk}{Autoscaling Local Storage: when enabled, this instances in this pool will dynamically acquire additional disk space when its Spark workers are running low on disk space.}

\item{idle_instance_autotermination_minutes}{Automatically terminates the extra instances in the pool cache after they are inactive for this time in minutes if min_idle_instances requirement is already met.}

\item{instance_pool_fleet_attributes}{The fleet related setting to power the instance pool.}

\item{max_capacity}{Maximum number of outstanding instances to keep in the pool, including both instances used by clusters and idle instances.}

\item{min_idle_instances}{Minimum number of idle instances to keep in the instance pool.}

\item{preloaded_docker_images}{Custom Docker Image BYOC.}

\item{preloaded_spark_versions}{A list of preloaded Spark image versions for the pool.}
}
\description{
Creates a new instance pool using idle and ready-to-use cloud instances.
}
\keyword{internal}
