% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance_pools.R
\name{instance_pools_delete}
\alias{instance_pools_delete}
\title{Delete an instance pool.}
\usage{
instance_pools_delete(instance_pool_id)
}
\arguments{
\item{instance_pool_id}{Required. The instance pool to be terminated.}
}
\description{
Deletes the instance pool permanently. The idle instances in the pool are
terminated asynchronously.
}
\keyword{internal}
