% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance_pools.R
\name{instance_pools_get}
\alias{instance_pools_get}
\title{Get instance pool information.}
\usage{
instance_pools_get(instance_pool_id)
}
\arguments{
\item{instance_pool_id}{Required. The canonical unique identifier for the instance pool.}
}
\description{
Retrieve the information for an instance pool based on its identifier.
}
\keyword{internal}
