% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance_pools.R
\name{instance_pools_list}
\alias{instance_pools_list}
\title{List instance pool info.}
\usage{
instance_pools_list(...)
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Gets a list of instance pools with their statistics.#'
}
\keyword{internal}
