% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance_profiles.R
\docType{data}
\name{instance_profiles}
\alias{instance_profiles}
\title{The Instance Profiles API allows admins to add, list, and remove instance
profiles that users can launch clusters with. Regular users can list the
instance profiles available to them. See \link{Secure access to S3 buckets} using
instance profiles for more information.}
\format{
An object of class \code{list} of length 4.
}
\usage{
instance_profiles
}
\description{

}
\section{Operations}{

\tabular{ll}{
\link[=instance_profiles_add]{add} \tab Register an instance profile.\cr
\link[=instance_profiles_edit]{edit} \tab Edit an instance profile.\cr
\link[=instance_profiles_list]{list} \tab List available instance profiles.\cr
\link[=instance_profiles_remove]{remove} \tab Remove the instance profile.\cr
}
}

\keyword{datasets}
