% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance_profiles.R
\name{instance_profiles_add}
\alias{instance_profiles_add}
\title{Register an instance profile.}
\usage{
instance_profiles_add(
  instance_profile_arn,
  iam_role_arn = NULL,
  is_meta_instance_profile = NULL,
  skip_validation = NULL,
  ...
)
}
\arguments{
\item{instance_profile_arn}{\link{required} The AWS ARN of the instance profile to register with Databricks.}

\item{iam_role_arn}{The AWS IAM role ARN of the role associated with the instance profile.}

\item{is_meta_instance_profile}{By default, Databricks validates that it has sufficient permissions to launch instances with the instance profile.}

\item{skip_validation}{By default, Databricks validates that it has sufficient permissions to launch instances with the instance profile.}
}
\description{
In the UI, you can select the instance profile when launching clusters. This
API is only available to admin users.
}
\keyword{internal}
