% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance_profiles.R
\name{instance_profiles_remove}
\alias{instance_profiles_remove}
\title{Remove the instance profile.}
\usage{
instance_profiles_remove(instance_profile_arn, ...)
}
\arguments{
\item{instance_profile_arn}{Required. The ARN of the instance profile to remove.}
}
\description{
Remove the instance profile with the provided ARN. Existing clusters with
this instance profile will continue to function.
}
\details{
This API is only accessible to admin users.
}
\keyword{internal}
