% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance_pools.R
\name{instancePoolsList}
\alias{instancePoolsList}
\title{List instance pool info.}
\usage{
instancePoolsList(client)
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Gets a list of instance pools with their statistics.#'
}
