% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance_profiles.R
\name{instanceProfilesEdit}
\alias{instanceProfilesEdit}
\title{Edit an instance profile.}
\usage{
instanceProfilesEdit(
  client,
  instance_profile_arn,
  iam_role_arn = NULL,
  is_meta_instance_profile = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{instance_profile_arn}{Required. The AWS ARN of the instance profile to register with Databricks.}

\item{iam_role_arn}{The AWS IAM role ARN of the role associated with the instance profile.}

\item{is_meta_instance_profile}{Boolean flag indicating whether the instance profile should only be used in credential passthrough scenarios.}
}
\description{
The only supported field to change is the optional IAM role ARN associated
with the instance profile. It is required to specify the IAM role ARN if both
of the following are true:
}
\details{
\itemize{
\item Your role name and instance profile name do not match. The name is the part
after the last slash in each ARN. * You want to use the instance profile with
\href{https://docs.databricks.com/sql/admin/serverless.html}{Databricks SQL Serverless}.
}

To understand where these fields are in the AWS console, see \href{https://docs.databricks.com/sql/admin/serverless.html}{Enable serverless SQL warehouses}.

This API is only available to admin users.
}
