% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance_profiles.R
\name{instanceProfilesList}
\alias{instanceProfilesList}
\title{List available instance profiles.}
\usage{
instanceProfilesList(client)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
List the instance profiles that the calling user can use to launch a cluster.
}
\details{
This API is available to all users.
}
