% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance_profiles.R
\name{instanceProfilesRemove}
\alias{instanceProfilesRemove}
\title{Remove the instance profile.}
\usage{
instanceProfilesRemove(client, instance_profile_arn)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{instance_profile_arn}{Required. The ARN of the instance profile to remove.}
}
\description{
Remove the instance profile with the provided ARN. Existing clusters with
this instance profile will continue to function.
}
\details{
This API is only accessible to admin users.
}
