% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip_access_lists.R
\docType{data}
\name{ip_access_lists}
\alias{ip_access_lists}
\title{IP Access List enables admins to configure IP access lists.}
\format{
An object of class \code{list} of length 6.
}
\usage{
ip_access_lists
}
\description{
IP access lists affect web application access and REST API access to this
workspace only. If the feature is disabled for a workspace, all access is
allowed for this workspace. There is support for allow lists (inclusion) and
block lists (exclusion).
}
\details{
When a connection is attempted: 1. \strong{First, all block lists are checked.} If
the connection IP address matches any block list, the connection is rejected.
2. \strong{If the connection was not rejected by block lists}, the IP address is
compared with the allow lists.

If there is at least one allow list for the workspace, the connection is
allowed only if the IP address matches an allow list. If there are no allow
lists for the workspace, all IP addresses are allowed.

For all allow lists and block lists combined, the workspace supports a
maximum of 1000 IP/CIDR values, where one CIDR counts as a single value.

After changes to the IP access list feature, it can take a few minutes for
changes to take effect.
}
\section{Operations}{

\tabular{ll}{
\link[=ip_access_lists_create]{create} \tab Create access list.\cr
\link[=ip_access_lists_delete]{delete} \tab Delete access list.\cr
\link[=ip_access_lists_get]{get} \tab Get access list.\cr
\link[=ip_access_lists_list]{list} \tab Get access lists.\cr
\link[=ip_access_lists_replace]{replace} \tab Replace access list.\cr
\link[=ip_access_lists_update]{update} \tab Update access list.\cr
}
}

\keyword{datasets}
