% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip_access_lists.R
\name{ip_access_lists_create}
\alias{ip_access_lists_create}
\title{Create access list.}
\usage{
ip_access_lists_create(label, list_type, ip_addresses, ...)
}
\arguments{
\item{label}{\link{required} Label for the IP access list.}

\item{list_type}{\link{required} This describes an enum.}

\item{ip_addresses}{\link{required} Array of IP addresses or CIDR values to be added to the IP access list.}
}
\description{
Creates an IP access list for this workspace.
}
\details{
A list can be an allow list or a block list. See the top of this file for a
description of how the server treats allow lists and block lists at runtime.

When creating or updating an IP access list:
\itemize{
\item For all allow lists and block lists combined, the API supports a maximum of
1000 IP/CIDR values, where one CIDR counts as a single value. Attempts to
exceed that number return error 400 with \code{error_code} value \code{QUOTA_EXCEEDED}.
\item If the new list would block the calling user's current IP, error 400 is
returned with \code{error_code} value \code{INVALID_STATE}.
}

It can take a few minutes for the changes to take effect. \strong{Note}: Your new
IP access list has no effect until you enable the feature. See
:method:workspaceconf/setStatus
}
\keyword{internal}
