% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip_access_lists.R
\name{ip_access_lists_delete}
\alias{ip_access_lists_delete}
\title{Delete access list.}
\usage{
ip_access_lists_delete(ip_access_list_id, ...)
}
\arguments{
\item{ip_access_list_id}{Required. The ID for the corresponding IP access list to modify.}
}
\description{
Deletes an IP access list, specified by its list ID.
}
\keyword{internal}
