% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip_access_lists.R
\name{ip_access_lists_list}
\alias{ip_access_lists_list}
\title{Get access lists.}
\usage{
ip_access_lists_list()
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Gets all IP access lists for the specified workspace.#'
}
\keyword{internal}
