% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip_access_lists.R
\name{ipAccessListsUpdate}
\alias{ipAccessListsUpdate}
\title{Update access list.}
\usage{
ipAccessListsUpdate(
  client,
  ip_access_list_id,
  enabled = NULL,
  ip_addresses = NULL,
  label = NULL,
  list_type = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{ip_access_list_id}{Required. The ID for the corresponding IP access list.}

\item{enabled}{Specifies whether this IP access list is enabled.}

\item{ip_addresses}{}

\item{label}{Label for the IP access list.}

\item{list_type}{Type of IP access list.}
}
\description{
Updates an existing IP access list, specified by its ID.
}
\details{
A list can include allow lists and block lists. See the top of this file for
a description of how the server treats allow lists and block lists at run
time.

When updating an IP access list:
\itemize{
\item For all allow lists and block lists combined, the API supports a maximum of
1000 IP/CIDR values, where one CIDR counts as a single value. Attempts to
exceed that number return error 400 with \code{error_code} value \code{QUOTA_EXCEEDED}.
\item If the updated list would block the calling user's current IP, error 400 is
returned with \code{error_code} value \code{INVALID_STATE}.
}

It can take a few minutes for the changes to take effect. Note that your
resulting IP access list has no effect until you enable the feature. See
:method:workspaceconf/setStatus.
}
