% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\docType{data}
\name{jobs}
\alias{jobs}
\title{The Jobs API allows you to create, edit, and delete jobs.}
\format{
An object of class \code{list} of length 16.
}
\usage{
jobs
}
\description{
You can use a Databricks job to run a data processing or data analysis task
in a Databricks cluster with scalable resources. Your job can consist of a
single task or can be a large, multi-task workflow with complex dependencies.
Databricks manages the task orchestration, cluster management, monitoring,
and error reporting for all of your jobs. You can run your jobs immediately
or periodically through an easy-to-use scheduling system. You can implement
job tasks using notebooks, JARS, Delta Live Tables pipelines, or Python,
Scala, Spark submit, and Java applications.
}
\details{
You should never hard code secrets or store them in plain text. Use the
\href{https://docs.databricks.com/dev-tools/cli/secrets-cli.html}{Secrets CLI} to manage secrets in the \href{https://docs.databricks.com/dev-tools/cli/index.html}{Databricks CLI}. Use the \href{https://docs.databricks.com/dev-tools/databricks-utils.html#dbutils-secrets}{Secrets utility} to reference secrets in notebooks and jobs.
}
\section{Operations}{

\tabular{ll}{
\link[=jobs_cancel_all_runs]{cancel_all_runs} \tab Cancel all runs of a job.\cr
\link[=jobs_cancel_run]{cancel_run} \tab Cancel a job run.\cr
\link[=jobs_create]{create} \tab Create a new job.\cr
\link[=jobs_delete]{delete} \tab Delete a job.\cr
\link[=jobs_delete_run]{delete_run} \tab Delete a job run.\cr
\link[=jobs_export_run]{export_run} \tab Export and retrieve a job run.\cr
\link[=jobs_get]{get} \tab Get a single job.\cr
\link[=jobs_get_run]{get_run} \tab Get a single job run.\cr
\link[=jobs_get_run_output]{get_run_output} \tab Get the output for a single run.\cr
\link[=jobs_list]{list} \tab List all jobs.\cr
\link[=jobs_list_runs]{list_runs} \tab List runs for a job.\cr
\link[=jobs_repair_run]{repair_run} \tab Repair a job run.\cr
\link[=jobs_reset]{reset} \tab Overwrites all settings for a job.\cr
\link[=jobs_run_now]{run_now} \tab Trigger a new job run.\cr
\link[=jobs_submit]{submit} \tab Create and trigger a one-time run.\cr
\link[=jobs_update]{update} \tab Partially update a job.\cr
}
}

\keyword{datasets}
