% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobs_cancel_all_runs}
\alias{jobs_cancel_all_runs}
\title{Cancel all runs of a job.}
\usage{
jobs_cancel_all_runs(job_id)
}
\arguments{
\item{job_id}{Required. The canonical identifier of the job to cancel all runs of.}
}
\description{
Cancels all active runs of a job. The runs are canceled asynchronously, so it
doesn't prevent new runs from being started.
}
\keyword{internal}
