% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobs_cancel_run}
\alias{jobs_cancel_run}
\title{Cancel a job run.}
\usage{
jobs_cancel_run(run_id, timeout = 20, callback = cli_reporter)
}
\arguments{
\item{run_id}{Required. This field is required.}
}
\description{
This is a long-running operation, which blocks until Jobs on Databricks reach
TERMINATED or SKIPPED state with the timeout of 20 minutes, that you can change via \code{timeout} parameter.
By default, the state of Databricks Jobs is reported to console. You can change this behavior
by changing the \code{callback} parameter.
}
\details{
Cancels a job run. The run is canceled asynchronously, so it may still be
running when this request completes.
}
\keyword{internal}
