% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobs_create}
\alias{jobs_create}
\title{Create a new job.}
\usage{
jobs_create(
  access_control_list = NULL,
  continuous = NULL,
  email_notifications = NULL,
  format = NULL,
  git_source = NULL,
  job_clusters = NULL,
  max_concurrent_runs = NULL,
  name = NULL,
  notification_settings = NULL,
  run_as = NULL,
  schedule = NULL,
  tags = NULL,
  tasks = NULL,
  timeout_seconds = NULL,
  trigger = NULL,
  webhook_notifications = NULL
)
}
\arguments{
\item{access_control_list}{List of permissions to set on the job.}

\item{continuous}{An optional continuous property for this job.}

\item{email_notifications}{An optional set of email addresses that is notified when runs of this job begin or complete as well as when this job is deleted.}

\item{format}{Used to tell what is the format of the job.}

\item{git_source}{An optional specification for a remote repository containing the notebooks used by this job's notebook tasks.}

\item{job_clusters}{A list of job cluster specifications that can be shared and reused by tasks of this job.}

\item{max_concurrent_runs}{An optional maximum allowed number of concurrent runs of the job.}

\item{name}{An optional name for the job.}

\item{notification_settings}{Optional notification settings that are used when sending notifications to each of the \code{email_notifications} and \code{webhook_notifications} for this job.}

\item{run_as}{Write-only setting, available only in Create/Update/Reset and Submit calls.}

\item{schedule}{An optional periodic schedule for this job.}

\item{tags}{A map of tags associated with the job.}

\item{tasks}{A list of task specifications to be executed by this job.}

\item{timeout_seconds}{An optional timeout applied to each run of this job.}

\item{trigger}{Trigger settings for the job.}

\item{webhook_notifications}{A collection of system notification IDs to notify when the run begins or completes.}
}
\description{
Create a new job.
}
\keyword{internal}
