% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobs_delete}
\alias{jobs_delete}
\title{Delete a job.}
\usage{
jobs_delete(job_id, ...)
}
\arguments{
\item{job_id}{Required. The canonical identifier of the job to delete.}
}
\description{
Deletes a job.
}
\keyword{internal}
