% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobs_delete_run}
\alias{jobs_delete_run}
\title{Delete a job run.}
\usage{
jobs_delete_run(run_id, ...)
}
\arguments{
\item{run_id}{Required. The canonical identifier of the run for which to retrieve the metadata.}
}
\description{
Deletes a non-active run. Returns an error if the run is active.
}
\keyword{internal}
