% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobs_export_run}
\alias{jobs_export_run}
\title{Export and retrieve a job run.}
\usage{
jobs_export_run(run_id, views_to_export = NULL, ...)
}
\arguments{
\item{run_id}{Required. The canonical identifier for the run.}

\item{views_to_export}{Which views to export (CODE, DASHBOARDS, or ALL).}
}
\description{
Export and retrieve the job run task.
}
\keyword{internal}
