% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobs_get}
\alias{jobs_get}
\title{Get a single job.}
\usage{
jobs_get(job_id)
}
\arguments{
\item{job_id}{Required. The canonical identifier of the job to retrieve information about.}
}
\description{
Retrieves the details for a single job.
}
\keyword{internal}
