% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobs_get_run_output}
\alias{jobs_get_run_output}
\title{Get the output for a single run.}
\usage{
jobs_get_run_output(run_id, ...)
}
\arguments{
\item{run_id}{\link{required} The canonical identifier for the run.}
}
\description{
Retrieve the output and metadata of a single task run. When a notebook task
returns a value through the \code{dbutils.notebook.exit()} call, you can use this
endpoint to retrieve that value. Databricks restricts this API to returning
the first 5 MB of the output. To return a larger result, you can store job
results in a cloud storage service.
}
\details{
This endpoint validates that the \strong{run_id} parameter is valid and returns an
HTTP status code 400 if the \strong{run_id} parameter is invalid. Runs are
automatically removed after 60 days. If you to want to reference them beyond
60 days, you must save old run results before they expire.
}
\keyword{internal}
