% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobs_list}
\alias{jobs_list}
\title{List all jobs.}
\usage{
jobs_list(expand_tasks = NULL, limit = NULL, name = NULL, offset = NULL, ...)
}
\arguments{
\item{expand_tasks}{Whether to include task and cluster details in the response.}

\item{limit}{The number of jobs to return.}

\item{name}{A filter on the list based on the exact (case insensitive) job name.}

\item{offset}{The offset of the first job to return, relative to the most recently created job.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Retrieves a list of jobs.
}
\keyword{internal}
