% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobs_list_runs}
\alias{jobs_list_runs}
\title{List runs for a job.}
\usage{
jobs_list_runs(
  active_only = NULL,
  completed_only = NULL,
  expand_tasks = NULL,
  job_id = NULL,
  limit = NULL,
  offset = NULL,
  run_type = NULL,
  start_time_from = NULL,
  start_time_to = NULL
)
}
\arguments{
\item{active_only}{If active_only is \code{true}, only active runs are included in the results; otherwise, lists both active and completed runs.}

\item{completed_only}{If completed_only is \code{true}, only completed runs are included in the results; otherwise, lists both active and completed runs.}

\item{expand_tasks}{Whether to include task and cluster details in the response.}

\item{job_id}{The job for which to list runs.}

\item{limit}{The number of runs to return.}

\item{offset}{The offset of the first run to return, relative to the most recent run.}

\item{run_type}{The type of runs to return.}

\item{start_time_from}{Show runs that started \emph{at or after} this value.}

\item{start_time_to}{Show runs that started \emph{at or before} this value.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
List runs in descending order by start time.
}
\keyword{internal}
