% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobs_reset}
\alias{jobs_reset}
\title{Overwrites all settings for a job.}
\usage{
jobs_reset(job_id, new_settings, ...)
}
\arguments{
\item{job_id}{Required. The canonical identifier of the job to reset.}

\item{new_settings}{Required. The new settings of the job.}
}
\description{
Overwrites all the settings for a specific job. Use the Update endpoint to
update job settings partially.
}
\keyword{internal}
