% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobs_update}
\alias{jobs_update}
\title{Partially updates a job.}
\usage{
jobs_update(job_id, fields_to_remove = NULL, new_settings = NULL, ...)
}
\arguments{
\item{job_id}{\link{required} The canonical identifier of the job to update.}

\item{fields_to_remove}{Remove top-level fields in the job settings.}

\item{new_settings}{The new settings for the job.}
}
\description{
Add, update, or remove specific settings of an existing job. Use the ResetJob
to overwrite all job settings.
}
\keyword{internal}
