% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobsCancelAllRuns}
\alias{jobsCancelAllRuns}
\title{Cancel all runs of a job.}
\usage{
jobsCancelAllRuns(client, job_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{job_id}{Required. The canonical identifier of the job to cancel all runs of.}
}
\description{
Cancels all active runs of a job. The runs are canceled asynchronously, so it
doesn't prevent new runs from being started.
}
