% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobsDelete}
\alias{jobsDelete}
\title{Delete a job.}
\usage{
jobsDelete(client, job_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{job_id}{Required. The canonical identifier of the job to delete.}
}
\description{
Deletes a job.
}
