% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobsDeleteRun}
\alias{jobsDeleteRun}
\title{Delete a job run.}
\usage{
jobsDeleteRun(client, run_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{run_id}{Required. The canonical identifier of the run for which to retrieve the metadata.}
}
\description{
Deletes a non-active run. Returns an error if the run is active.
}
