% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobsGet}
\alias{jobsGet}
\title{Get a single job.}
\usage{
jobsGet(client, job_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{job_id}{Required. The canonical identifier of the job to retrieve information about.}
}
\description{
Retrieves the details for a single job.
}
