% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobsGetJobPermissionLevels}
\alias{jobsGetJobPermissionLevels}
\title{Get job permission levels.}
\usage{
jobsGetJobPermissionLevels(client, job_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{job_id}{Required. The job for which to get or manage permissions.}
}
\description{
Gets the permission levels that a user can have on an object.
}
