% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobsGetPermissions}
\alias{jobsGetPermissions}
\title{Get job permissions.}
\usage{
jobsGetPermissions(client, job_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{job_id}{Required. The job for which to get or manage permissions.}
}
\description{
Gets the permissions of a job. Jobs can inherit permissions from their root
object.
}
