% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobsGetRun}
\alias{jobsGetRun}
\title{Get a single job run.}
\usage{
jobsGetRun(
  client,
  run_id,
  include_history = NULL,
  timeout = 20,
  callback = cli_reporter
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{run_id}{Required. The canonical identifier of the run for which to retrieve the metadata.}

\item{include_history}{Whether to include the repair history in the response.}
}
\description{
This is a long-running operation, which blocks until Jobs on Databricks reach
TERMINATED or SKIPPED state with the timeout of 20 minutes, that you can change via \code{timeout} parameter.
By default, the state of Databricks Jobs is reported to console. You can change this behavior
by changing the \code{callback} parameter.
}
\details{
Retrieve the metadata of a run.
}
