% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobsList}
\alias{jobsList}
\title{List all jobs.}
\usage{
jobsList(
  client,
  expand_tasks = NULL,
  limit = NULL,
  name = NULL,
  offset = NULL,
  page_token = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{expand_tasks}{Whether to include task and cluster details in the response.}

\item{limit}{The number of jobs to return.}

\item{name}{A filter on the list based on the exact (case insensitive) job name.}

\item{offset}{The offset of the first job to return, relative to the most recently created job.}

\item{page_token}{Use \code{next_page_token} or \code{prev_page_token} returned from the previous request to list the next or previous page of jobs respectively.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Retrieves a list of jobs.
}
