% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobsSubmit}
\alias{jobsSubmit}
\title{Create and trigger a one-time run.}
\usage{
jobsSubmit(
  client,
  access_control_list = NULL,
  email_notifications = NULL,
  git_source = NULL,
  health = NULL,
  idempotency_token = NULL,
  notification_settings = NULL,
  run_name = NULL,
  tasks = NULL,
  timeout_seconds = NULL,
  webhook_notifications = NULL,
  timeout = 20,
  callback = cli_reporter
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{access_control_list}{List of permissions to set on the job.}

\item{email_notifications}{An optional set of email addresses notified when the run begins or completes.}

\item{git_source}{An optional specification for a remote repository containing the notebooks used by this job's notebook tasks.}

\item{health}{An optional set of health rules that can be defined for this job.}

\item{idempotency_token}{An optional token that can be used to guarantee the idempotency of job run requests.}

\item{notification_settings}{Optional notification settings that are used when sending notifications to each of the \code{webhook_notifications} for this run.}

\item{run_name}{An optional name for the run.}

\item{tasks}{}

\item{timeout_seconds}{An optional timeout applied to each run of this job.}

\item{webhook_notifications}{A collection of system notification IDs to notify when the run begins or completes.}
}
\description{
This is a long-running operation, which blocks until Jobs on Databricks reach
TERMINATED or SKIPPED state with the timeout of 20 minutes, that you can change via \code{timeout} parameter.
By default, the state of Databricks Jobs is reported to console. You can change this behavior
by changing the \code{callback} parameter.
}
\details{
Submit a one-time run. This endpoint allows you to submit a workload directly
without creating a job. Runs submitted using this endpoint don’t display in
the UI. Use the \code{jobs/runs/get} API to check the run state after the job is
submitted.
}
