% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakehouse_monitors.R
\name{lakehouseMonitorsDelete}
\alias{lakehouseMonitorsDelete}
\title{Delete a table monitor.}
\usage{
lakehouseMonitorsDelete(client, full_name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{full_name}{Required. Full name of the table.}
}
\description{
Deletes a monitor for the specified table.
}
\details{
The caller must either: 1. be an owner of the table's parent catalog 2. have
\strong{USE_CATALOG} on the table's parent catalog and be an owner of the table's
parent schema 3. have the following permissions: - \strong{USE_CATALOG} on the
table's parent catalog - \strong{USE_SCHEMA} on the table's parent schema - be an
owner of the table.

Additionally, the call must be made from the workspace where the monitor was
created.

Note that the metric tables and dashboard will not be deleted as part of this
call; those assets must be manually cleaned up (if desired).
}
