% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakehouse_monitors.R
\name{lakehouseMonitorsRunRefresh}
\alias{lakehouseMonitorsRunRefresh}
\title{Queue a metric refresh for a monitor.}
\usage{
lakehouseMonitorsRunRefresh(client, full_name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{full_name}{Required. Full name of the table.}
}
\description{
Queues a metric refresh on the monitor for the specified table. The refresh
will execute in the background.
}
\details{
The caller must either: 1. be an owner of the table's parent catalog 2. have
\strong{USE_CATALOG} on the table's parent catalog and be an owner of the table's
parent schema 3. have the following permissions: - \strong{USE_CATALOG} on the
table's parent catalog - \strong{USE_SCHEMA} on the table's parent schema - be an
owner of the table

Additionally, the call must be made from the workspace where the monitor was
created.
}
