% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakehouse_monitors.R
\name{lakehouseMonitorsUpdate}
\alias{lakehouseMonitorsUpdate}
\title{Update a table monitor.}
\usage{
lakehouseMonitorsUpdate(
  client,
  full_name,
  output_schema_name,
  baseline_table_name = NULL,
  custom_metrics = NULL,
  data_classification_config = NULL,
  inference_log = NULL,
  notifications = NULL,
  schedule = NULL,
  slicing_exprs = NULL,
  snapshot = NULL,
  time_series = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{full_name}{Required. Full name of the table.}

\item{output_schema_name}{Required. Schema where output metric tables are created.}

\item{baseline_table_name}{Name of the baseline table from which drift metrics are computed from.}

\item{custom_metrics}{Custom metrics to compute on the monitored table.}

\item{data_classification_config}{The data classification config for the monitor.}

\item{inference_log}{Configuration for monitoring inference logs.}

\item{notifications}{The notification settings for the monitor.}

\item{schedule}{The schedule for automatically updating and refreshing metric tables.}

\item{slicing_exprs}{List of column expressions to slice data with for targeted analysis.}

\item{snapshot}{Configuration for monitoring snapshot tables.}

\item{time_series}{Configuration for monitoring time series tables.}
}
\description{
Updates a monitor for the specified table.
}
\details{
The caller must either: 1. be an owner of the table's parent catalog 2. have
\strong{USE_CATALOG} on the table's parent catalog and be an owner of the table's
parent schema 3. have the following permissions: - \strong{USE_CATALOG} on the
table's parent catalog - \strong{USE_SCHEMA} on the table's parent schema - be an
owner of the table.

Additionally, the call must be made from the workspace where the monitor was
created, and the caller must be the original creator of the monitor.

Certain configuration fields, such as output asset identifiers, cannot be
updated.
}
