% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeview.R
\name{lakeviewGet}
\alias{lakeviewGet}
\title{Get dashboard.}
\usage{
lakeviewGet(client, dashboard_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{dashboard_id}{Required. UUID identifying the dashboard.}
}
\description{
Get a draft dashboard.
}
