% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraries.R
\docType{data}
\name{libraries}
\alias{libraries}
\title{The Libraries API allows you to install and uninstall libraries and get the
status of libraries on a cluster.}
\format{
An object of class \code{list} of length 4.
}
\usage{
libraries
}
\description{
To make third-party or custom code available to notebooks and jobs running on
your clusters, you can install a library. Libraries can be written in Python,
Java, Scala, and R. You can upload Java, Scala, and Python libraries and
point to external packages in PyPI, Maven, and CRAN repositories.
}
\details{
Cluster libraries can be used by all notebooks running on a cluster. You can
install a cluster library directly from a public repository such as PyPI or
Maven, using a previously installed workspace library, or using an init
script.

When you install a library on a cluster, a notebook already attached to that
cluster will not immediately see the new library. You must first detach and
then reattach the notebook to the cluster.

When you uninstall a library from a cluster, the library is removed only when
you restart the cluster. Until you restart the cluster, the status of the
uninstalled library appears as Uninstall pending restart.
}
\section{Operations}{

\tabular{ll}{
\link[=libraries_all_cluster_statuses]{all_cluster_statuses} \tab Get all statuses.\cr
\link[=libraries_cluster_status]{cluster_status} \tab Get status.\cr
\link[=libraries_install]{install} \tab Add a library.\cr
\link[=libraries_uninstall]{uninstall} \tab Uninstall libraries.\cr
}
}

\keyword{datasets}
