% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraries.R
\name{libraries_uninstall}
\alias{libraries_uninstall}
\title{Uninstall libraries.}
\usage{
libraries_uninstall(cluster_id, libraries, ...)
}
\arguments{
\item{cluster_id}{Required. Unique identifier for the cluster on which to uninstall these libraries.}

\item{libraries}{Required. The libraries to uninstall.}
}
\description{
Set libraries to be uninstalled on a cluster. The libraries won't be
uninstalled until the cluster is restarted. Uninstalling libraries that are
not installed on the cluster will have no impact but is not an error.
}
\keyword{internal}
