% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{list_alerts}
\alias{list_alerts}
\alias{alertsList}
\title{Get alerts.}
\usage{
list_alerts(client)

alertsList(client)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()

See the \href{https://docs.databricks.com/api/workspace/alerts/list}{API documentation} for more
information on the request and response field structure.}
}
\description{
Gets a list of alerts.
}
