% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_policies.R
\name{list_cluster_policies}
\alias{list_cluster_policies}
\alias{clusterPoliciesList}
\title{List cluster policies.}
\usage{
list_cluster_policies(client, sort_column = NULL, sort_order = NULL)

clusterPoliciesList(client, sort_column = NULL, sort_order = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{sort_column}{The cluster policy attribute to sort by.}

\item{sort_order}{The order in which the policies get listed.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Returns a list of policies accessible by the requesting user.
}
