% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_families.R
\name{list_cluster_policy_families}
\alias{list_cluster_policy_families}
\alias{policyFamiliesList}
\title{List policy families.}
\usage{
list_cluster_policy_families(client, max_results = NULL, page_token = NULL)

policyFamiliesList(client, max_results = NULL, page_token = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{max_results}{The max number of policy families to return.}

\item{page_token}{A token that can be used to get the next page of results.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Retrieve a list of policy families. This API is paginated.
}
