% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{list_cluster_zones}
\alias{list_cluster_zones}
\alias{clustersListZones}
\title{List availability zones.}
\usage{
list_cluster_zones(client)

clustersListZones(client)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()

See the \href{https://docs.databricks.com/api/workspace/clusters/listzones}{API documentation} for more
information on the request and response field structure.}
}
\description{
Returns a list of availability zones where clusters can be created in (For
example, us-west-2a). These zones can be used to launch a cluster.
}
