% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{list_clusters}
\alias{list_clusters}
\alias{clustersList}
\title{List all clusters.}
\usage{
list_clusters(client, can_use_client = NULL)

clustersList(client, can_use_client = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{can_use_client}{Filter clusters based on what type of client it can be used for.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Return information about all pinned clusters, active clusters, up to 200 of
the most recently terminated all-purpose clusters in the past 30 days, and up
to 30 of the most recently terminated job clusters in the past 30 days.
}
\details{
For example, if there is 1 pinned cluster, 4 active clusters, 45 terminated
all-purpose clusters in the past 30 days, and 50 terminated job clusters in
the past 30 days, then this API returns the 1 pinned cluster, 4 active
clusters, all 45 terminated all-purpose clusters, and the 30 most recently
terminated job clusters.
}
