% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sources.R
\name{list_data_sources}
\alias{list_data_sources}
\alias{dataSourcesList}
\title{Get a list of SQL warehouses.}
\usage{
list_data_sources(client)

dataSourcesList(client)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()

See the \href{https://docs.databricks.com/api/workspace/datasources/list}{API documentation} for more
information on the request and response field structure.}
}
\description{
Retrieves a full list of SQL warehouses available in this workspace. All
fields that appear in this API response are enumerated for clarity. However,
you need only a SQL warehouse's \code{id} to create new queries against it.
}
