% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfs.R
\name{list_dbfs}
\alias{list_dbfs}
\title{List directory contents or file details.}
\usage{
list_dbfs(client, path)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{path}{Required. The path of the file or directory.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
List the contents of a directory, or details of the file. If the file or
directory does not exist, this call throws an exception with
\code{RESOURCE_DOES_NOT_EXIST}.
}
\details{
When calling list on a large directory, the list operation will time out
after approximately 60 seconds. We strongly recommend using list only on
directories containing less than 10K files and discourage using the DBFS REST
API for operations that list more than 10K files. Instead, we recommend that
you perform such operations in the context of a cluster, using the \href{/dev-tools/databricks-utils.html#dbutils-fs}{File system utility (dbutils.fs)},
which provides the same functionality without timing out.
}
