% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{list_experiment_experiments}
\alias{list_experiment_experiments}
\alias{experimentsListExperiments}
\title{List experiments.}
\usage{
list_experiment_experiments(
  client,
  max_results = NULL,
  page_token = NULL,
  view_type = NULL
)

experimentsListExperiments(
  client,
  max_results = NULL,
  page_token = NULL,
  view_type = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{max_results}{Maximum number of experiments desired.}

\item{page_token}{Token indicating the page of experiments to fetch.}

\item{view_type}{Qualifier for type of experiments to be returned.}
}
\value{
\code{data.frame} with all of the response pages.

See the \href{https://docs.databricks.com/api/workspace/experiments/listexperiments}{API documentation} for more
information on the request and response field structure.
}
\description{
Gets a list of all experiments.
}
