% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{list_file_directory_contents}
\alias{list_file_directory_contents}
\alias{filesListDirectoryContents}
\title{List directory contents.}
\usage{
list_file_directory_contents(
  client,
  directory_path,
  page_size = NULL,
  page_token = NULL
)

filesListDirectoryContents(
  client,
  directory_path,
  page_size = NULL,
  page_token = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{directory_path}{Required. The absolute path of a directory.}

\item{page_size}{The maximum number of directory entries to return.}

\item{page_token}{An opaque page token which was the \code{next_page_token} in the response of the previous request to list the contents of this directory.}
}
\value{
\code{data.frame} with all of the response pages.

See the \href{https://docs.databricks.com/api/workspace/files/listdirectorycontents}{API documentation} for more
information on the request and response field structure.
}
\description{
Returns the contents of a directory. If there is no directory at the
specified path, the API returns a HTTP 404 error.
}
