% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{list_functions}
\alias{list_functions}
\alias{functionsList}
\title{List functions.}
\usage{
list_functions(
  client,
  catalog_name,
  schema_name,
  include_browse = NULL,
  max_results = NULL,
  page_token = NULL
)

functionsList(
  client,
  catalog_name,
  schema_name,
  include_browse = NULL,
  max_results = NULL,
  page_token = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{catalog_name}{Required. Name of parent catalog for functions of interest.}

\item{schema_name}{Required. Parent schema of functions.}

\item{include_browse}{Whether to include functions in the response for which the principal can only access selective metadata for.}

\item{max_results}{Maximum number of functions to return.}

\item{page_token}{Opaque pagination token to go to next page based on previous query.}
}
\value{
\code{data.frame} with all of the response pages.

See the \href{https://docs.databricks.com/api/workspace/functions/list}{API documentation} for more
information on the request and response field structure.
}
\description{
List functions within the specified parent catalog and schema. If the user is
a metastore admin, all functions are returned in the output list. Otherwise,
the user must have the \strong{USE_CATALOG} privilege on the catalog and the
\strong{USE_SCHEMA} privilege on the schema, and the output list contains only
functions for which either the user has the \strong{EXECUTE} privilege or the user
is the owner. There is no guarantee of a specific ordering of the elements in
the array.
}
