% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance_pools.R
\name{list_instance_pools}
\alias{list_instance_pools}
\alias{instancePoolsList}
\title{List instance pool info.}
\usage{
list_instance_pools(client)

instancePoolsList(client)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}
}
\value{
\code{data.frame} with all of the response pages.

See the \href{https://docs.databricks.com/api/workspace/instancepools/list}{API documentation} for more
information on the request and response field structure.
}
\description{
Gets a list of instance pools with their statistics.
}
