% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance_profiles.R
\name{list_instance_profiles}
\alias{list_instance_profiles}
\alias{instanceProfilesList}
\title{List available instance profiles.}
\usage{
list_instance_profiles(client)

instanceProfilesList(client)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}
}
\value{
\code{data.frame} with all of the response pages.

See the \href{https://docs.databricks.com/api/workspace/instanceprofiles/list}{API documentation} for more
information on the request and response field structure.
}
\description{
List the instance profiles that the calling user can use to launch a cluster.
}
\details{
This API is available to all users.
}
