% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakehouse_monitors.R
\name{list_lakehouse_monitor_refreshes}
\alias{list_lakehouse_monitor_refreshes}
\title{List refreshes.}
\usage{
list_lakehouse_monitor_refreshes(client, full_name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{full_name}{Required. Full name of the table.}
}
\description{
Gets an array containing the history of the most recent refreshes (up to 25)
for this table.
}
\details{
The caller must either: 1. be an owner of the table's parent catalog 2. have
\strong{USE_CATALOG} on the table's parent catalog and be an owner of the table's
parent schema 3. have the following permissions: - \strong{USE_CATALOG} on the
table's parent catalog - \strong{USE_SCHEMA} on the table's parent schema -
\strong{SELECT} privilege on the table.

Additionally, the call must be made from the workspace where the monitor was
created.
}
