% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{list_model_models}
\alias{list_model_models}
\title{List models.}
\usage{
list_model_models(client, max_results = NULL, page_token = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{max_results}{Maximum number of registered models desired.}

\item{page_token}{Pagination token to go to the next page based on a previous query.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Lists all available registered models, up to the limit specified in
\strong{max_results}.
}
