% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_versions.R
\name{list_model_versions}
\alias{list_model_versions}
\alias{modelVersionsList}
\title{List Model Versions.}
\usage{
list_model_versions(
  client,
  full_name,
  include_browse = NULL,
  max_results = NULL,
  page_token = NULL
)

modelVersionsList(
  client,
  full_name,
  include_browse = NULL,
  max_results = NULL,
  page_token = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{full_name}{Required. The full three-level name of the registered model under which to list model versions.}

\item{include_browse}{Whether to include model versions in the response for which the principal can only access selective metadata for.}

\item{max_results}{Maximum number of model versions to return.}

\item{page_token}{Opaque pagination token to go to next page based on previous query.}
}
\value{
\code{data.frame} with all of the response pages.

See the \href{https://docs.databricks.com/api/workspace/modelversions/list}{API documentation} for more
information on the request and response field structure.
}
\description{
List model versions. You can list model versions under a particular schema,
or list all model versions in the current metastore.
}
\details{
The returned models are filtered based on the privileges of the calling user.
For example, the metastore admin is able to list all the model versions. A
regular user needs to be the owner or have the \strong{EXECUTE} privilege on the
parent registered model to recieve the model versions in the response. For
the latter case, the caller must also be the owner or have the
\strong{USE_CATALOG} privilege on the parent catalog and the \strong{USE_SCHEMA}
privilege on the parent schema.

There is no guarantee of a specific ordering of the elements in the response.
}
