% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{list_pipeline_updates}
\alias{list_pipeline_updates}
\alias{pipelinesListUpdates}
\title{List pipeline updates.}
\usage{
list_pipeline_updates(
  client,
  pipeline_id,
  max_results = NULL,
  page_token = NULL,
  until_update_id = NULL
)

pipelinesListUpdates(
  client,
  pipeline_id,
  max_results = NULL,
  page_token = NULL,
  until_update_id = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{pipeline_id}{Required. The pipeline to return updates for.}

\item{max_results}{Max number of entries to return in a single page.}

\item{page_token}{Page token returned by previous call.}

\item{until_update_id}{If present, returns updates until and including this update_id.

See the \href{https://docs.databricks.com/api/workspace/pipelines/listupdates}{API documentation} for more
information on the request and response field structure.}
}
\description{
List updates for an active pipeline.
}
