% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{list_queries}
\alias{list_queries}
\alias{queriesList}
\title{Get a list of queries.}
\usage{
list_queries(client, order = NULL, page = NULL, page_size = NULL, q = NULL)

queriesList(client, order = NULL, page = NULL, page_size = NULL, q = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{order}{Name of query attribute to order by.}

\item{page}{Page number to retrieve.}

\item{page_size}{Number of queries to return per page.}

\item{q}{Full text search term.}
}
\value{
\code{data.frame} with all of the response pages.

See the \href{https://docs.databricks.com/api/workspace/queries/list}{API documentation} for more
information on the request and response field structure.
}
\description{
Gets a list of queries. Optionally, this list can be filtered by a search
term.
}
\details{
\subsection{**Warning: Calling this API concurrently 10 or more times could result in}{

throttling, service degradation, or a temporary ban.**
}
}
