% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registered_models.R
\name{list_registered_models}
\alias{list_registered_models}
\alias{registeredModelsList}
\title{List Registered Models.}
\usage{
list_registered_models(
  client,
  catalog_name = NULL,
  include_browse = NULL,
  max_results = NULL,
  page_token = NULL,
  schema_name = NULL
)

registeredModelsList(
  client,
  catalog_name = NULL,
  include_browse = NULL,
  max_results = NULL,
  page_token = NULL,
  schema_name = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{catalog_name}{The identifier of the catalog under which to list registered models.}

\item{include_browse}{Whether to include registered models in the response for which the principal can only access selective metadata for.}

\item{max_results}{Max number of registered models to return.}

\item{page_token}{Opaque token to send for the next page of results (pagination).}

\item{schema_name}{The identifier of the schema under which to list registered models.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
List registered models. You can list registered models under a particular
schema, or list all registered models in the current metastore.
}
\details{
The returned models are filtered based on the privileges of the calling user.
For example, the metastore admin is able to list all the registered models. A
regular user needs to be the owner or have the \strong{EXECUTE} privilege on the
registered model to recieve the registered models in the response. For the
latter case, the caller must also be the owner or have the \strong{USE_CATALOG}
privilege on the parent catalog and the \strong{USE_SCHEMA} privilege on the
parent schema.

There is no guarantee of a specific ordering of the elements in the response.
}
