% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\name{list_schemas}
\alias{list_schemas}
\title{List schemas.}
\usage{
list_schemas(
  client,
  catalog_name,
  include_browse = NULL,
  max_results = NULL,
  page_token = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{catalog_name}{Required. Parent catalog for schemas of interest.}

\item{include_browse}{Whether to include schemas in the response for which the principal can only access selective metadata for.}

\item{max_results}{Maximum number of schemas to return.}

\item{page_token}{Opaque pagination token to go to next page based on previous query.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Gets an array of schemas for a catalog in the metastore. If the caller is the
metastore admin or the owner of the parent catalog, all schemas for the
catalog will be retrieved. Otherwise, only schemas owned by the caller (or
for which the caller has the \strong{USE_SCHEMA} privilege) will be retrieved.
There is no guarantee of a specific ordering of the elements in the array.
}
