% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serving_endpoints.R
\name{list_serving_endpoints}
\alias{list_serving_endpoints}
\alias{servingEndpointsList}
\title{Get all serving endpoints.}
\usage{
list_serving_endpoints(client)

servingEndpointsList(client)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}
}
\value{
\code{data.frame} with all of the response pages.

See the \href{https://docs.databricks.com/api/workspace/servingendpoints/list}{API documentation} for more
information on the request and response field structure.
}
\description{
Get all serving endpoints.
}
