% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shares.R
\name{list_shares}
\alias{list_shares}
\alias{sharesList}
\title{List shares.}
\usage{
list_shares(client)

sharesList(client)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}
}
\value{
\code{data.frame} with all of the response pages.

See the \href{https://docs.databricks.com/api/workspace/shares/list}{API documentation} for more
information on the request and response field structure.
}
\description{
Gets an array of data object shares from the metastore. The caller must be a
metastore admin or the owner of the share. There is no guarantee of a
specific ordering of the elements in the array.
}
