% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_credentials.R
\name{list_storage_credentials}
\alias{list_storage_credentials}
\alias{storageCredentialsList}
\title{List credentials.}
\usage{
list_storage_credentials(client, max_results = NULL, page_token = NULL)

storageCredentialsList(client, max_results = NULL, page_token = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{max_results}{Maximum number of storage credentials to return.}

\item{page_token}{Opaque pagination token to go to next page based on previous query.}
}
\value{
\code{data.frame} with all of the response pages.

See the \href{https://docs.databricks.com/api/workspace/storagecredentials/list}{API documentation} for more
information on the request and response field structure.
}
\description{
Gets an array of storage credentials (as \strong{StorageCredentialInfo} objects).
The array is limited to only those storage credentials the caller has
permission to access. If the caller is a metastore admin, retrieval of
credentials is unrestricted. There is no guarantee of a specific ordering of
the elements in the array.
}
