% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_schemas.R
\name{list_system_schemas}
\alias{list_system_schemas}
\title{List system schemas.}
\usage{
list_system_schemas(client, metastore_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{metastore_id}{Required. The ID for the metastore in which the system schema resides.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Gets an array of system schemas for a metastore. The caller must be an
account admin or a metastore admin.
}
