% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_search_endpoints.R
\name{list_vector_search_endpoint_endpoints}
\alias{list_vector_search_endpoint_endpoints}
\alias{vectorSearchEndpointsListEndpoints}
\title{List all endpoints.}
\usage{
list_vector_search_endpoint_endpoints(client, page_token = NULL)

vectorSearchEndpointsListEndpoints(client, page_token = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{page_token}{Token for pagination.}
}
\value{
\code{data.frame} with all of the response pages.

See the \href{https://docs.databricks.com/api/workspace/vectorsearchendpoints/listendpoints}{API documentation} for more
information on the request and response field structure.
}
\description{
List all endpoints.
}
