% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volumes.R
\name{list_volumes}
\alias{list_volumes}
\title{List Volumes.}
\usage{
list_volumes(
  client,
  catalog_name,
  schema_name,
  include_browse = NULL,
  max_results = NULL,
  page_token = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{catalog_name}{Required. The identifier of the catalog.}

\item{schema_name}{Required. The identifier of the schema.}

\item{include_browse}{Whether to include volumes in the response for which the principal can only access selective metadata for.}

\item{max_results}{Maximum number of volumes to return (page length).}

\item{page_token}{Opaque token returned by a previous request.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Gets an array of volumes for the current metastore under the parent catalog
and schema.
}
\details{
The returned volumes are filtered based on the privileges of the calling
user. For example, the metastore admin is able to list all the volumes. A
regular user needs to be the owner or have the \strong{READ VOLUME} privilege on
the volume to recieve the volumes in the response. For the latter case, the
caller must also be the owner or have the \strong{USE_CATALOG} privilege on the
parent catalog and the \strong{USE_SCHEMA} privilege on the parent schema.

There is no guarantee of a specific ordering of the elements in the array.
}
