% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouses.R
\name{list_warehouses}
\alias{list_warehouses}
\alias{warehousesList}
\title{List warehouses.}
\usage{
list_warehouses(client, run_as_user_id = NULL)

warehousesList(client, run_as_user_id = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{run_as_user_id}{Service Principal which will be used to fetch the list of warehouses.}
}
\value{
\code{data.frame} with all of the response pages.

See the \href{https://docs.databricks.com/api/workspace/warehouses/list}{API documentation} for more
information on the request and response field structure.
}
\description{
Lists all SQL warehouses that a user has manager permissions on.
}
