% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{log_experiment_inputs}
\alias{log_experiment_inputs}
\title{Log inputs to a run.}
\usage{
log_experiment_inputs(client, datasets = NULL, run_id = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{datasets}{Dataset inputs.}

\item{run_id}{ID of the run to log under.}
}
\description{
\strong{NOTE:} Experimental: This API may change or be removed in a future release
without warning.
}
