% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{log_experiment_metric}
\alias{log_experiment_metric}
\alias{experimentsLogMetric}
\title{Log a metric.}
\usage{
log_experiment_metric(
  client,
  key,
  value,
  timestamp,
  run_id = NULL,
  run_uuid = NULL,
  step = NULL
)

experimentsLogMetric(
  client,
  key,
  value,
  timestamp,
  run_id = NULL,
  run_uuid = NULL,
  step = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{key}{Required. Name of the metric.}

\item{value}{Required. Double value of the metric being logged.

See the \href{https://docs.databricks.com/api/workspace/experiments/logmetric}{API documentation} for more
information on the request and response field structure.}

\item{timestamp}{Required. Unix timestamp in milliseconds at the time metric was logged.}

\item{run_id}{ID of the run under which to log the metric.}

\item{run_uuid}{Deprecated, use run_id instead. ID of the run under which to log the metric.}

\item{step}{Step at which to log the metric.}
}
\description{
Logs a metric for a run. A metric is a key-value pair (string key, float
value) with an associated timestamp. Examples include the various metrics
that represent ML model accuracy. A metric can be logged multiple times.
}
