% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{log_experiment_model}
\alias{log_experiment_model}
\alias{experimentsLogModel}
\title{Log a model.}
\usage{
log_experiment_model(client, model_json = NULL, run_id = NULL)

experimentsLogModel(client, model_json = NULL, run_id = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{model_json}{MLmodel file in json format.}

\item{run_id}{ID of the run to log under.}
}
\description{
\strong{NOTE:} Experimental: This API may change or be removed in a future release
without warning.
}
