% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{log_experiment_param}
\alias{log_experiment_param}
\title{Log a param.}
\usage{
log_experiment_param(client, key, value, run_id = NULL, run_uuid = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{key}{Required. Name of the param.}

\item{value}{Required. String value of the param being logged.}

\item{run_id}{ID of the run under which to log the param.}

\item{run_uuid}{Deprecated, use run_id instead. ID of the run under which to log the param.}
}
\description{
Logs a param used for a run. A param is a key-value pair (string key, string
value). Examples include hyperparameters used for ML model training and
constant dates and values used in an ETL pipeline. A param can be logged only
once for a run.
}
