% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serving_endpoints.R
\name{logs_serving_endpoint}
\alias{logs_serving_endpoint}
\alias{servingEndpointsLogs}
\title{Get the latest logs for a served model.}
\usage{
logs_serving_endpoint(client, name, served_model_name)

servingEndpointsLogs(client, name, served_model_name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. The name of the serving endpoint that the served model belongs to.}

\item{served_model_name}{Required. The name of the served model that logs will be retrieved for.}
}
\description{
Retrieves the service logs associated with the provided served model.
}
