% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_lflow_artifacts.R
\name{m_lflow_artifacts_list}
\alias{m_lflow_artifacts_list}
\title{Get all artifacts.}
\usage{
m_lflow_artifacts_list(
  page_token = NULL,
  path = NULL,
  run_id = NULL,
  run_uuid = NULL,
  ...
)
}
\arguments{
\item{page_token}{Token indicating the page of artifact results to fetch.}

\item{path}{Filter artifacts matching this path (a relative path from the root artifact directory).}

\item{run_id}{ID of the run whose artifacts to list.}

\item{run_uuid}{\link{Deprecated, use run_id instead} ID of the run whose artifacts to list.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
List artifacts for a run. Takes an optional \code{artifact_path} prefix. If it is
specified, the response contains only artifacts with the specified prefix.',
}
\keyword{internal}
