% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_lflow_databricks.R
\docType{data}
\name{m_lflow_databricks}
\alias{m_lflow_databricks}
\title{These endpoints are modified versions of the MLflow API that accept
additional input parameters or return additional information.}
\format{
An object of class \code{list} of length 2.
}
\usage{
m_lflow_databricks
}
\description{
These endpoints are modified versions of the MLflow API that accept
additional input parameters or return additional information.
}
\section{Operations}{

\tabular{ll}{
\link[=m_lflow_databricks_get]{get} \tab Get model.\cr
\link[=m_lflow_databricks_transition_stage]{transition_stage} \tab Transition a stage.\cr
}
}

\keyword{datasets}
