% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_lflow_databricks.R
\name{m_lflow_databricks_get}
\alias{m_lflow_databricks_get}
\title{Get model.}
\usage{
m_lflow_databricks_get(name, ...)
}
\arguments{
\item{name}{\link{required} Name of the model.}
}
\description{
Get the details of a model. This is a Databricks Workspace version of the
\link{MLflow endpoint} that also returns the model's Databricks Workspace ID and
the permission level of the requesting user on the model.
}
\details{

}
\keyword{internal}
