% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_lflow_runs.R
\name{m_lflow_runs_delete}
\alias{m_lflow_runs_delete}
\title{Delete a run.}
\usage{
m_lflow_runs_delete(run_id, ...)
}
\arguments{
\item{run_id}{\link{required} ID of the run to delete.}
}
\description{
Marks a run for deletion.
}
\keyword{internal}
