% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_lflow_runs.R
\name{m_lflow_runs_delete_tag}
\alias{m_lflow_runs_delete_tag}
\title{Delete a tag.}
\usage{
m_lflow_runs_delete_tag(run_id, key)
}
\arguments{
\item{run_id}{Required. ID of the run that the tag was logged under.}

\item{key}{Required. Name of the tag.}
}
\description{
Deletes a tag on a run. Tags are run metadata that can be updated during a
run and after a run completes.
}
\keyword{internal}
