% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_lflow_runs.R
\name{m_lflow_runs_get}
\alias{m_lflow_runs_get}
\title{Get a run.}
\usage{
m_lflow_runs_get(run_id, run_uuid = NULL, ...)
}
\arguments{
\item{run_id}{Required. ID of the run to fetch.}

\item{run_uuid}{\link{Deprecated, use run_id instead} ID of the run to fetch.}
}
\description{
'Gets the metadata, metrics, params, and tags for a run. In the case where
multiple metrics with the same key are logged for a run, return only the
value with the latest timestamp.
}
\details{
If there are multiple values with the latest timestamp, return the maximum of
these values.
}
\keyword{internal}
