% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_lflow_runs.R
\name{m_lflow_runs_log_metric}
\alias{m_lflow_runs_log_metric}
\title{Log a metric.}
\usage{
m_lflow_runs_log_metric(
  key,
  value,
  timestamp,
  run_id = NULL,
  run_uuid = NULL,
  step = NULL,
  ...
)
}
\arguments{
\item{key}{\link{required} Name of the metric.}

\item{value}{\link{required} Double value of the metric being logged.}

\item{timestamp}{\link{required} Unix timestamp in milliseconds at the time metric was logged.}

\item{run_id}{ID of the run under which to log the metric.}

\item{run_uuid}{\link{Deprecated, use run_id instead} ID of the run under which to log the metric.}

\item{step}{Step at which to log the metric.}
}
\description{
Logs a metric for a run. A metric is a key-value pair (string key, float
value) with an associated timestamp. Examples include the various metrics
that represent ML model accuracy. A metric can be logged multiple times.
}
\keyword{internal}
