% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_lflow_runs.R
\name{m_lflow_runs_log_parameter}
\alias{m_lflow_runs_log_parameter}
\title{Log a param.}
\usage{
m_lflow_runs_log_parameter(key, value, run_id = NULL, run_uuid = NULL, ...)
}
\arguments{
\item{key}{\link{required} Name of the param.}

\item{value}{\link{required} String value of the param being logged.}

\item{run_id}{ID of the run under which to log the param.}

\item{run_uuid}{\link{Deprecated, use run_id instead} ID of the run under which to log the param.}
}
\description{
Logs a param used for a run. A param is a key-value pair (string key, string
value). Examples include hyperparameters used for ML model training and
constant dates and values used in an ETL pipeline. A param can be logged only
once for a run.
}
\keyword{internal}
