% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_lflow_runs.R
\name{m_lflow_runs_restore}
\alias{m_lflow_runs_restore}
\title{Restore a run.}
\usage{
m_lflow_runs_restore(run_id)
}
\arguments{
\item{run_id}{Required. ID of the run to restore.}
}
\description{
Restores a deleted run.
}
\keyword{internal}
