% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_lflow_runs.R
\name{m_lflow_runs_update}
\alias{m_lflow_runs_update}
\title{Update a run.}
\usage{
m_lflow_runs_update(
  end_time = NULL,
  run_id = NULL,
  run_uuid = NULL,
  status = NULL
)
}
\arguments{
\item{end_time}{Unix timestamp in milliseconds of when the run ended.}

\item{run_id}{ID of the run to update.}

\item{run_uuid}{\link{Deprecated, use run_id instead} ID of the run to update.}

\item{status}{Updated status of the run.}
}
\description{
Updates run metadata.
}
\keyword{internal}
