% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_user.R
\name{me}
\alias{me}
\title{Get current user info.}
\usage{
me(client)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}
}
\description{
Get details about the current method caller's identity.
}
