% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metastores.R
\name{metastores_create}
\alias{metastores_create}
\title{Create a metastore.}
\usage{
metastores_create(name, storage_root, region = NULL)
}
\arguments{
\item{name}{Required. The user-specified name of the metastore.}

\item{storage_root}{Required. The storage root URL for metastore.}

\item{region}{Cloud region which the metastore serves (e.g., \code{us-west-2}, \code{westus}).}
}
\description{
Creates a new metastore based on a provided name and storage root path.
}
\keyword{internal}
