% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metastores.R
\name{metastores_current}
\alias{metastores_current}
\title{Get metastore assignment for workspace.}
\usage{
metastores_current()
}
\description{
Gets the metastore assignment for the workspace being accessed.#'
}
\keyword{internal}
