% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metastores.R
\name{metastores_delete}
\alias{metastores_delete}
\title{Delete a metastore.}
\usage{
metastores_delete(id, force = NULL, ...)
}
\arguments{
\item{id}{Required. Unique ID of the metastore.}

\item{force}{Force deletion even if the metastore is not empty.}
}
\description{
Deletes a metastore. The caller must be a metastore admin.
}
\keyword{internal}
