% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metastores.R
\name{metastores_get}
\alias{metastores_get}
\title{Get a metastore.}
\usage{
metastores_get(id, ...)
}
\arguments{
\item{id}{\link{required} Unique ID of the metastore.}
}
\description{
Gets a metastore that matches the supplied ID. The caller must be a metastore
admin to retrieve this info.
}
\keyword{internal}
