% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metastores.R
\name{metastores_list}
\alias{metastores_list}
\title{List metastores.}
\usage{
metastores_list(...)
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Gets an array of the available metastores (as \strong{MetastoreInfo} objects). The
caller must be an admin to retrieve this info. There is no guarantee of a
specific ordering of the elements in the array.#'
}
\keyword{internal}
