% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metastores.R
\name{metastores_maintenance}
\alias{metastores_maintenance}
\title{Enables or disables auto maintenance on the metastore.}
\usage{
metastores_maintenance(metastore_id, enable)
}
\arguments{
\item{metastore_id}{Required. Unique identifier of metastore.}

\item{enable}{Required. Whether to enable auto maintenance on the metastore.}
}
\description{
Enables or disables auto maintenance on the metastore.
}
\keyword{internal}
