% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metastores.R
\name{metastores_summary}
\alias{metastores_summary}
\title{Get a metastore summary.}
\usage{
metastores_summary(...)
}
\description{
Gets information about a metastore. This summary includes the storage
credential, the cloud vendor, the cloud region, and the global metastore ID.#'
}
\keyword{internal}
