% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metastores.R
\name{metastores_update_assignment}
\alias{metastores_update_assignment}
\title{Update an assignment.}
\usage{
metastores_update_assignment(
  workspace_id,
  metastore_id,
  default_catalog_name = NULL,
  ...
)
}
\arguments{
\item{workspace_id}{Required. Workspace ID.}

\item{metastore_id}{The unique ID of the metastore.}

\item{default_catalog_name}{The name of the default catalog for the metastore.}
}
\description{
Updates a metastore assignment. This operation can be used to update
\strong{metastore_id} or \strong{default_catalog_name} for a specified Workspace, if
the Workspace is already assigned a metastore. The caller must be an account
admin to update \strong{metastore_id}; otherwise, the caller can be a Workspace
admin.
}
\keyword{internal}
