% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metastores.R
\name{metastoresAssign}
\alias{metastoresAssign}
\title{Create an assignment.}
\usage{
metastoresAssign(client, metastore_id, default_catalog_name, workspace_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{metastore_id}{Required. The unique ID of the metastore.}

\item{default_catalog_name}{Required. The name of the default catalog in the metastore.}

\item{workspace_id}{Required. A workspace ID.}
}
\description{
Creates a new metastore assignment. If an assignment for the same
\strong{workspace_id} exists, it will be overwritten by the new \strong{metastore_id}
and \strong{default_catalog_name}. The caller must be an account admin.
}
