% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metastores.R
\name{metastoresCurrent}
\alias{metastoresCurrent}
\title{Get metastore assignment for workspace.}
\usage{
metastoresCurrent(client)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}
}
\description{
Gets the metastore assignment for the workspace being accessed.
}
