% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metastores.R
\name{metastoresEnableOptimization}
\alias{metastoresEnableOptimization}
\title{Toggle predictive optimization on the metastore.}
\usage{
metastoresEnableOptimization(client, metastore_id, enable)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{metastore_id}{Required. Unique identifier of metastore.}

\item{enable}{Required. Whether to enable predictive optimization on the metastore.}
}
\description{
Enables or disables predictive optimization on the metastore.
}
