% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metastores.R
\name{metastoresList}
\alias{metastoresList}
\title{List metastores.}
\usage{
metastoresList(client)
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Gets an array of the available metastores (as \strong{MetastoreInfo} objects). The
caller must be an admin to retrieve this info. There is no guarantee of a
specific ordering of the elements in the array.#'
}
