% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metastores.R
\name{metastoresMaintenance}
\alias{metastoresMaintenance}
\title{Enables or disables auto maintenance on the metastore.}
\usage{
metastoresMaintenance(client, metastore_id, enable)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{metastore_id}{Required. Unique identifier of metastore.}

\item{enable}{Required. Whether to enable auto maintenance on the metastore.}
}
\description{
Enables or disables auto maintenance on the metastore.
}
