% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{mkdirs_notebook}
\alias{mkdirs_notebook}
\title{Create a directory.}
\usage{
mkdirs_notebook(client, path)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{path}{Required. The absolute path of the directory.}
}
\description{
Creates the specified directory (and necessary parent directories if they do
not exist). If there is an object (not a directory) at any prefix of the
input path, this call returns an error \code{RESOURCE_ALREADY_EXISTS}.
}
\details{
Note that if this operation fails it may have succeeded in creating some of
the necessary parent directories.
}
