% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\docType{data}
\name{model_registry}
\alias{model_registry}
\title{MLflow Model Registry is a centralized model repository and a UI and set of
APIs that enable you to manage the full lifecycle of MLflow Models.}
\format{
An object of class \code{list} of length 32.
}
\usage{
model_registry
}
\description{
MLflow Model Registry is a centralized model repository and a UI and set of
APIs that enable you to manage the full lifecycle of MLflow Models.
}
\section{Operations}{

\tabular{ll}{
\link[=model_registry_approve_transition_request]{approve_transition_request} \tab Approve transition request.\cr
\link[=model_registry_create_comment]{create_comment} \tab Post a comment.\cr
\link[=model_registry_create_model]{create_model} \tab Create a model.\cr
\link[=model_registry_create_model_version]{create_model_version} \tab Create a model version.\cr
\link[=model_registry_create_transition_request]{create_transition_request} \tab Make a transition request.\cr
\link[=model_registry_create_webhook]{create_webhook} \tab Create a webhook.\cr
\link[=model_registry_delete_comment]{delete_comment} \tab Delete a comment.\cr
\link[=model_registry_delete_model]{delete_model} \tab Delete a model.\cr
\link[=model_registry_delete_model_tag]{delete_model_tag} \tab Delete a model tag.\cr
\link[=model_registry_delete_model_version]{delete_model_version} \tab Delete a model version.\cr
\link[=model_registry_delete_model_version_tag]{delete_model_version_tag} \tab Delete a model version tag.\cr
\link[=model_registry_delete_transition_request]{delete_transition_request} \tab Delete a transition request.\cr
\link[=model_registry_delete_webhook]{delete_webhook} \tab Delete a webhook.\cr
\link[=model_registry_get_latest_versions]{get_latest_versions} \tab Get the latest version.\cr
\link[=model_registry_get_model]{get_model} \tab Get model.\cr
\link[=model_registry_get_model_version]{get_model_version} \tab Get a model version.\cr
\link[=model_registry_get_model_version_download_uri]{get_model_version_download_uri} \tab Get a model version URI.\cr
\link[=model_registry_list_models]{list_models} \tab List models.\cr
\link[=model_registry_list_transition_requests]{list_transition_requests} \tab List transition requests.\cr
\link[=model_registry_list_webhooks]{list_webhooks} \tab List registry webhooks.\cr
\link[=model_registry_reject_transition_request]{reject_transition_request} \tab Reject a transition request.\cr
\link[=model_registry_rename_model]{rename_model} \tab Rename a model.\cr
\link[=model_registry_search_model_versions]{search_model_versions} \tab Searches model versions.\cr
\link[=model_registry_search_models]{search_models} \tab Search models.\cr
\link[=model_registry_set_model_tag]{set_model_tag} \tab Set a tag.\cr
\link[=model_registry_set_model_version_tag]{set_model_version_tag} \tab Set a version tag.\cr
\link[=model_registry_test_registry_webhook]{test_registry_webhook} \tab Test a webhook.\cr
\link[=model_registry_transition_stage]{transition_stage} \tab Transition a stage.\cr
\link[=model_registry_update_comment]{update_comment} \tab Update a comment.\cr
\link[=model_registry_update_model]{update_model} \tab Update model.\cr
\link[=model_registry_update_model_version]{update_model_version} \tab Update model version.\cr
\link[=model_registry_update_webhook]{update_webhook} \tab Update a webhook.\cr
}
}

\keyword{datasets}
