% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{model_registry_create_comment}
\alias{model_registry_create_comment}
\title{Post a comment.}
\usage{
model_registry_create_comment(name, version, comment)
}
\arguments{
\item{name}{Required. Name of the model.}

\item{version}{Required. Version of the model.}

\item{comment}{Required. User-provided comment on the action.}
}
\description{
Posts a comment on a model version. A comment can be submitted either by a
user or programmatically to display relevant information about the model. For
example, test results or deployment errors.
}
\keyword{internal}
