% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{model_registry_create_webhook}
\alias{model_registry_create_webhook}
\title{Create a webhook.}
\usage{
model_registry_create_webhook(
  events,
  description = NULL,
  http_url_spec = NULL,
  job_spec = NULL,
  model_name = NULL,
  status = NULL
)
}
\arguments{
\item{events}{Required. Events that can trigger a registry webhook: * \code{MODEL_VERSION_CREATED}: A new model version was created for the associated model.}

\item{description}{User-specified description for the webhook.}

\item{http_url_spec}{}

\item{job_spec}{}

\item{model_name}{Name of the model whose events would trigger this webhook.}

\item{status}{This describes an enum.}
}
\description{
\strong{NOTE}: This endpoint is in Public Preview.
}
\details{
Creates a registry webhook.
}
\keyword{internal}
