% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{model_registry_delete_comment}
\alias{model_registry_delete_comment}
\title{Delete a comment.}
\usage{
model_registry_delete_comment(id)
}
\arguments{
\item{id}{Required.}
}
\description{
Deletes a comment on a model version.
}
\keyword{internal}
