% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{model_registry_delete_model}
\alias{model_registry_delete_model}
\title{Delete a model.}
\usage{
model_registry_delete_model(name)
}
\arguments{
\item{name}{Required. Registered model unique name identifier.}
}
\description{
Deletes a registered model.
}
\keyword{internal}
