% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{model_registry_delete_model_tag}
\alias{model_registry_delete_model_tag}
\title{Delete a model tag.}
\usage{
model_registry_delete_model_tag(name, key)
}
\arguments{
\item{name}{Required. Name of the registered model that the tag was logged under.}

\item{key}{Required. Name of the tag.}
}
\description{
Deletes the tag for a registered model.
}
\keyword{internal}
