% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{model_registry_delete_webhook}
\alias{model_registry_delete_webhook}
\title{Delete a webhook.}
\usage{
model_registry_delete_webhook(id = NULL)
}
\arguments{
\item{id}{Webhook ID required to delete a registry webhook.}
}
\description{
\strong{NOTE:} This endpoint is in Public Preview.
}
\details{
Deletes a registry webhook.
}
\keyword{internal}
