% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{model_registry_get_latest_versions}
\alias{model_registry_get_latest_versions}
\title{Get the latest version.}
\usage{
model_registry_get_latest_versions(name, stages = NULL)
}
\arguments{
\item{name}{Required. Registered model unique name identifier.}

\item{stages}{List of stages.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Gets the latest version of a registered model.
}
\keyword{internal}
