% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{model_registry_get_model}
\alias{model_registry_get_model}
\title{Get model.}
\usage{
model_registry_get_model(name)
}
\arguments{
\item{name}{Required. Registered model unique name identifier.}
}
\description{
Get the details of a model. This is a Databricks workspace version of the
\link{MLflow endpoint} that also returns the model's Databricks workspace ID and
the permission level of the requesting user on the model.
}
\details{

}
\keyword{internal}
