% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{model_registry_get_model_version}
\alias{model_registry_get_model_version}
\title{Get a model version.}
\usage{
model_registry_get_model_version(name, version)
}
\arguments{
\item{name}{Required. Name of the registered model.}

\item{version}{Required. Model version number.}
}
\description{
Get a model version.
}
\keyword{internal}
