% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{model_registry_list_transition_requests}
\alias{model_registry_list_transition_requests}
\title{List transition requests.}
\usage{
model_registry_list_transition_requests(name, version)
}
\arguments{
\item{name}{Required. Name of the model.}

\item{version}{Required. Version of the model.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Gets a list of all open stage transition requests for the model version.
}
\keyword{internal}
