% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{model_registry_rename_model}
\alias{model_registry_rename_model}
\title{Rename a model.}
\usage{
model_registry_rename_model(name, new_name = NULL)
}
\arguments{
\item{name}{Required. Registered model unique name identifier.}

\item{new_name}{If provided, updates the name for this \code{registered_model}.}
}
\description{
Renames a registered model.
}
\keyword{internal}
