% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{model_registry_search_models}
\alias{model_registry_search_models}
\title{Search models.}
\usage{
model_registry_search_models(
  filter = NULL,
  max_results = NULL,
  order_by = NULL,
  page_token = NULL
)
}
\arguments{
\item{filter}{String filter condition, like 'name LIKE 'my-model-name''.}

\item{max_results}{Maximum number of models desired.}

\item{order_by}{List of columns for ordering search results, which can include model name and last updated timestamp with an optional 'DESC' or 'ASC' annotation, where 'ASC' is the default.}

\item{page_token}{Pagination token to go to the next page based on a previous search query.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Search for registered models based on the specified \strong{filter}.
}
\keyword{internal}
