% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{model_registry_set_model_tag}
\alias{model_registry_set_model_tag}
\title{Set a tag.}
\usage{
model_registry_set_model_tag(name, key, value)
}
\arguments{
\item{name}{Required. Unique name of the model.}

\item{key}{Required. Name of the tag.}

\item{value}{Required. String value of the tag being logged.}
}
\description{
Sets a tag on a registered model.
}
\keyword{internal}
