% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{model_registry_test_registry_webhook}
\alias{model_registry_test_registry_webhook}
\title{Test a webhook.}
\usage{
model_registry_test_registry_webhook(id, event = NULL)
}
\arguments{
\item{id}{Required. Webhook ID.}

\item{event}{If \code{event} is specified, the test trigger uses the specified event.}
}
\description{
\strong{NOTE:} This endpoint is in Public Preview.
}
\details{
Tests a registry webhook.
}
\keyword{internal}
