% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{model_registry_update_comment}
\alias{model_registry_update_comment}
\title{Update a comment.}
\usage{
model_registry_update_comment(id, comment)
}
\arguments{
\item{id}{Required. Unique identifier of an activity.}

\item{comment}{Required. User-provided comment on the action.}
}
\description{
Post an edit to a comment on a model version.
}
\keyword{internal}
