% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{model_registry_update_model_version}
\alias{model_registry_update_model_version}
\title{Update model version.}
\usage{
model_registry_update_model_version(name, version, description = NULL)
}
\arguments{
\item{name}{Required. Name of the registered model.}

\item{version}{Required. Model version number.}

\item{description}{If provided, updates the description for this \code{registered_model}.}
}
\description{
Updates the model version.
}
\keyword{internal}
