% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_version_comments.R
\name{model_version_comments_create}
\alias{model_version_comments_create}
\title{Post a comment.}
\usage{
model_version_comments_create(name, version, comment, ...)
}
\arguments{
\item{name}{\link{required} Name of the model.}

\item{version}{\link{required} Version of the model.}

\item{comment}{\link{required} User-provided comment on the action.}
}
\description{
Posts a comment on a model version. A comment can be submitted either by a
user or programmatically to display relevant information about the model. For
example, test results or deployment errors.
}
\keyword{internal}
