% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_version_comments.R
\name{model_version_comments_delete}
\alias{model_version_comments_delete}
\title{Delete a comment.}
\usage{
model_version_comments_delete(id)
}
\arguments{
\item{id}{Required.}
}
\description{
Deletes a comment on a model version.
}
\keyword{internal}
