% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_version_comments.R
\name{model_version_comments_update}
\alias{model_version_comments_update}
\title{Update a comment.}
\usage{
model_version_comments_update(id, comment, ...)
}
\arguments{
\item{id}{Required. Unique identifier of an activity.}

\item{comment}{Required. User-provided comment on the action.}
}
\description{
Post an edit to a comment on a model version.
}
\keyword{internal}
