% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_versions.R
\name{model_versions_create}
\alias{model_versions_create}
\title{Create a model version.}
\usage{
model_versions_create(
  name,
  source,
  description = NULL,
  run_id = NULL,
  run_link = NULL,
  tags = NULL,
  ...
)
}
\arguments{
\item{name}{\link{required} Register model under this name.}

\item{source}{\link{required} URI indicating the location of the model artifacts.}

\item{description}{Optional description for model version.}

\item{run_id}{MLflow run ID for correlation, if \code{source} was generated by an experiment run in MLflow tracking server.}

\item{run_link}{MLflow run link - this is the exact link of the run that generated this model version, potentially hosted at another instance of MLflow.}

\item{tags}{Additional metadata for model version.}
}
\description{
Creates a model version.
}
\keyword{internal}
