% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_versions.R
\name{model_versions_delete_tag}
\alias{model_versions_delete_tag}
\title{Delete a model version tag.}
\usage{
model_versions_delete_tag(name, version, key, ...)
}
\arguments{
\item{name}{\link{required} Name of the registered model that the tag was logged under.}

\item{version}{\link{required} Model version number that the tag was logged under.}

\item{key}{\link{required} Name of the tag.}
}
\description{
Deletes a model version tag.
}
\keyword{internal}
