% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_versions.R
\name{model_versions_get}
\alias{model_versions_get}
\title{Get a model version.}
\usage{
model_versions_get(name, version, ...)
}
\arguments{
\item{name}{Required. Name of the registered model.}

\item{version}{Required. Model version number.}
}
\description{
Get a model version.
}
\keyword{internal}
