% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_versions.R
\name{model_versions_get_download_uri}
\alias{model_versions_get_download_uri}
\title{Get a model version URI.}
\usage{
model_versions_get_download_uri(name, version, ...)
}
\arguments{
\item{name}{\link{required} Name of the registered model.}

\item{version}{\link{required} Model version number.}
}
\description{
Gets a URI to download the model version.
}
\keyword{internal}
