% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_versions.R
\name{model_versions_set_tag}
\alias{model_versions_set_tag}
\title{Set a version tag.}
\usage{
model_versions_set_tag(name, version, key, value, ...)
}
\arguments{
\item{name}{\link{required} Unique name of the model.}

\item{version}{\link{required} Model version number.}

\item{key}{\link{required} Name of the tag.}

\item{value}{\link{required} String value of the tag being logged.}
}
\description{
Sets a model version tag.
}
\keyword{internal}
