% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{modelRegistryDeleteComment}
\alias{modelRegistryDeleteComment}
\title{Delete a comment.}
\usage{
modelRegistryDeleteComment(client, id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{id}{Required.}
}
\description{
Deletes a comment on a model version.
}
