% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{modelRegistryDeleteModel}
\alias{modelRegistryDeleteModel}
\title{Delete a model.}
\usage{
modelRegistryDeleteModel(client, name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Registered model unique name identifier.}
}
\description{
Deletes a registered model.
}
