% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{modelRegistryDeleteModelVersion}
\alias{modelRegistryDeleteModelVersion}
\title{Delete a model version.}
\usage{
modelRegistryDeleteModelVersion(client, name, version)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Name of the registered model.}

\item{version}{Required. Model version number.}
}
\description{
Deletes a model version.
}
