% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{modelRegistryDeleteTransitionRequest}
\alias{modelRegistryDeleteTransitionRequest}
\title{Delete a transition request.}
\usage{
modelRegistryDeleteTransitionRequest(
  client,
  name,
  version,
  stage,
  creator,
  comment = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Name of the model.}

\item{version}{Required. Version of the model.}

\item{stage}{Required. Target stage of the transition request.}

\item{creator}{Required. Username of the user who created this request.}

\item{comment}{User-provided comment on the action.}
}
\description{
Cancels a model version stage transition request.
}
