% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{modelRegistryGetModelVersion}
\alias{modelRegistryGetModelVersion}
\title{Get a model version.}
\usage{
modelRegistryGetModelVersion(client, name, version)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Name of the registered model.}

\item{version}{Required. Model version number.}
}
\description{
Get a model version.
}
