% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{modelRegistryGetModelVersionDownloadUri}
\alias{modelRegistryGetModelVersionDownloadUri}
\title{Get a model version URI.}
\usage{
modelRegistryGetModelVersionDownloadUri(client, name, version)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Name of the registered model.}

\item{version}{Required. Model version number.}
}
\description{
Gets a URI to download the model version.
}
