% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{modelRegistryGetPermissionLevels}
\alias{modelRegistryGetPermissionLevels}
\title{Get registered model permission levels.}
\usage{
modelRegistryGetPermissionLevels(client, registered_model_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{registered_model_id}{Required. The registered model for which to get or manage permissions.}
}
\description{
Gets the permission levels that a user can have on an object.
}
