% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{modelRegistryGetRegisteredModelPermissions}
\alias{modelRegistryGetRegisteredModelPermissions}
\title{Get registered model permissions.}
\usage{
modelRegistryGetRegisteredModelPermissions(client, registered_model_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{registered_model_id}{Required. The registered model for which to get or manage permissions.}
}
\description{
Gets the permissions of a registered model. Registered models can inherit
permissions from their root object.
}
