% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{modelRegistrySetModelTag}
\alias{modelRegistrySetModelTag}
\title{Set a tag.}
\usage{
modelRegistrySetModelTag(client, name, key, value)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Unique name of the model.}

\item{key}{Required. Name of the tag.}

\item{value}{Required. String value of the tag being logged.}
}
\description{
Sets a tag on a registered model.
}
