% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{modelRegistryTestRegistryWebhook}
\alias{modelRegistryTestRegistryWebhook}
\title{Test a webhook.}
\usage{
modelRegistryTestRegistryWebhook(client, id, event = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{id}{Required. Webhook ID.}

\item{event}{If \code{event} is specified, the test trigger uses the specified event.}
}
\description{
\strong{NOTE:} This endpoint is in Public Preview.
}
\details{
Tests a registry webhook.
}
