% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{modelRegistryUpdateModel}
\alias{modelRegistryUpdateModel}
\title{Update model.}
\usage{
modelRegistryUpdateModel(client, name, description = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Registered model unique name identifier.}

\item{description}{If provided, updates the description for this \code{registered_model}.}
}
\description{
Updates a registered model.
}
