% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{modelRegistryUpdateModelVersion}
\alias{modelRegistryUpdateModelVersion}
\title{Update model version.}
\usage{
modelRegistryUpdateModelVersion(client, name, version, description = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Name of the registered model.}

\item{version}{Required. Model version number.}

\item{description}{If provided, updates the description for this \code{registered_model}.}
}
\description{
Updates the model version.
}
