% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_versions.R
\name{modelVersionsGet}
\alias{modelVersionsGet}
\title{Get a Model Version.}
\usage{
modelVersionsGet(client, full_name, version, include_browse = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{full_name}{Required. The three-level (fully qualified) name of the model version.}

\item{version}{Required. The integer version number of the model version.}

\item{include_browse}{Whether to include model versions in the response for which the principal can only access selective metadata for.}
}
\description{
Get a model version.
}
\details{
The caller must be a metastore admin or an owner of (or have the \strong{EXECUTE}
privilege on) the parent registered model. For the latter case, the caller
must also be the owner or have the \strong{USE_CATALOG} privilege on the parent
catalog and the \strong{USE_SCHEMA} privilege on the parent schema.
}
