% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfs.R
\name{move_dbfs}
\alias{move_dbfs}
\title{Move a file.}
\usage{
move_dbfs(client, source_path, destination_path)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{source_path}{Required. The source path of the file or directory.}

\item{destination_path}{Required. The destination path of the file or directory.}
}
\description{
Moves a file from one location to another location within DBFS. If the source
file does not exist, this call throws an exception with
\code{RESOURCE_DOES_NOT_EXIST}. If a file already exists in the destination path,
this call throws an exception with \code{RESOURCE_ALREADY_EXISTS}. If the given
source path is a directory, this call always recursively moves all files.
}
