% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_tables.R
\name{onlineTablesDelete}
\alias{onlineTablesDelete}
\title{Delete an Online Table.}
\usage{
onlineTablesDelete(client, name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Full three-part (catalog, schema, table) name of the table.}
}
\description{
Delete an online table. Warning: This will delete all the data in the online
table. If the source Delta table was deleted or modified since this Online
Table was created, this will lose the data forever!
}
