% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_tables.R
\name{onlineTablesGet}
\alias{onlineTablesGet}
\title{Get an Online Table.}
\usage{
onlineTablesGet(client, name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Full three-part (catalog, schema, table) name of the table.}
}
\description{
Get information about an existing online table and its status.
}
