% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{patch_group}
\alias{patch_group}
\alias{groupsPatch}
\title{Update group details.}
\usage{
patch_group(client, id, operations = NULL, schemas = NULL)

groupsPatch(client, id, operations = NULL, schemas = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{id}{Required. Unique ID for a group in the Databricks workspace.}

\item{operations}{This field has no description yet.}

\item{schemas}{The schema of the patch request.

See the \href{https://docs.databricks.com/api/workspace/groups/patch}{API documentation} for more
information on the request and response field structure.}
}
\description{
Partially updates the details of a group.
}
