% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service_principals.R
\name{patch_service_principal}
\alias{patch_service_principal}
\title{Update service principal details.}
\usage{
patch_service_principal(client, id, operations = NULL, schemas = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{id}{Required. Unique ID for a service principal in the Databricks workspace.}

\item{operations}{This field has no description yet.}

\item{schemas}{The schema of the patch request.}
}
\description{
Partially updates the details of a single service principal in the Databricks
workspace.
}
