% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serving_endpoints.R
\name{patch_serving_endpoint}
\alias{patch_serving_endpoint}
\title{Update tags of a serving endpoint.}
\usage{
patch_serving_endpoint(client, name, add_tags = NULL, delete_tags = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. The name of the serving endpoint who's tags to patch.}

\item{add_tags}{List of endpoint tags to add.}

\item{delete_tags}{List of tag keys to delete.}
}
\description{
Used to batch add and delete tags from a serving endpoint with a single API
call.
}
