% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{patch_user}
\alias{patch_user}
\alias{usersPatch}
\title{Update user details.}
\usage{
patch_user(client, id, operations = NULL, schemas = NULL)

usersPatch(client, id, operations = NULL, schemas = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{id}{Required. Unique ID for a user in the Databricks workspace.}

\item{operations}{This field has no description yet.}

\item{schemas}{The schema of the patch request.

See the \href{https://docs.databricks.com/api/workspace/users/patch}{API documentation} for more
information on the request and response field structure.}
}
\description{
Partially updates a user resource by applying the supplied operations on
specific user attributes.
}
