% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{permanent_cluster_delete}
\alias{permanent_cluster_delete}
\alias{clustersPermanentDelete}
\title{Permanently delete cluster.}
\usage{
permanent_cluster_delete(client, cluster_id)

clustersPermanentDelete(client, cluster_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{cluster_id}{Required. The cluster to be deleted.

See the \href{https://docs.databricks.com/api/workspace/clusters/permanentdelete}{API documentation} for more
information on the request and response field structure.}
}
\description{
Permanently deletes a Spark cluster. This cluster is terminated and resources
are asynchronously removed.
}
\details{
In addition, users will no longer see permanently deleted clusters in the
cluster list, and API users can no longer perform any action on permanently
deleted clusters.
}
