% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\docType{data}
\name{permissions}
\alias{permissions}
\title{Permissions API are used to create read, write, edit, update and manage
access for various users on different objects and endpoints.}
\format{
An object of class \code{list} of length 4.
}
\usage{
permissions
}
\description{
Permissions API are used to create read, write, edit, update and manage
access for various users on different objects and endpoints.
}
\section{Operations}{

\tabular{ll}{
\link[=permissions_get]{get} \tab Get object permissions.\cr
\link[=permissions_get_permission_levels]{get_permission_levels} \tab Get permission levels.\cr
\link[=permissions_set]{set} \tab Set permissions.\cr
\link[=permissions_update]{update} \tab Update permission.\cr
}
}

\keyword{datasets}
