% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{permissions_get_permission_levels}
\alias{permissions_get_permission_levels}
\title{Get permission levels.}
\usage{
permissions_get_permission_levels(request_object_type, request_object_id, ...)
}
\arguments{
\item{request_object_type}{\link{required} \if{html}{\out{<needs content>}}.}

\item{request_object_id}{\link{required} \if{html}{\out{<needs content>}}.}
}
\description{
Gets the permission levels that a user can have on an object.
}
\keyword{internal}
