% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{permissions_set}
\alias{permissions_set}
\title{Set permissions.}
\usage{
permissions_set(
  request_object_type,
  request_object_id,
  access_control_list = NULL,
  ...
)
}
\arguments{
\item{request_object_type}{\link{required} \if{html}{\out{<needs content>}}.}

\item{request_object_id}{\link{required}}

\item{access_control_list}{}
}
\description{
Sets permissions on object. Objects can inherit permissions from their parent
objects and root objects.
}
\keyword{internal}
