% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{permissionsGet}
\alias{permissionsGet}
\title{Get object permissions.}
\usage{
permissionsGet(client, request_object_type, request_object_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{request_object_type}{Required. \if{html}{\out{<needs content>}}.}

\item{request_object_id}{Required.}
}
\description{
Gets the permission of an object. Objects can inherit permissions from their
parent objects or root objects.
}
