% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{permissionsGetPermissionLevels}
\alias{permissionsGetPermissionLevels}
\title{Get object permission levels.}
\usage{
permissionsGetPermissionLevels(client, request_object_type, request_object_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{request_object_type}{Required. \if{html}{\out{<needs content>}}.}

\item{request_object_id}{Required. \if{html}{\out{<needs content>}}.}
}
\description{
Gets the permission levels that a user can have on an object.
}
