% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{permissionsSet}
\alias{permissionsSet}
\title{Set object permissions.}
\usage{
permissionsSet(
  client,
  request_object_type,
  request_object_id,
  access_control_list = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{request_object_type}{Required. \if{html}{\out{<needs content>}}.}

\item{request_object_id}{Required.}

\item{access_control_list}{}
}
\description{
Sets permissions on an object. Objects can inherit permissions from their
parent objects or root object.
}
