% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{permissionsUpdate}
\alias{permissionsUpdate}
\title{Update permission.}
\usage{
permissionsUpdate(
  client,
  request_object_type,
  request_object_id,
  access_control_list = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{request_object_type}{Required. \if{html}{\out{<needs content>}}.}

\item{request_object_id}{Required.}

\item{access_control_list}{}
}
\description{
Updates the permissions on an object.
}
