% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\docType{data}
\name{pipelines}
\alias{pipelines}
\title{The Delta Live Tables API allows you to create, edit, delete, start, and view
details about pipelines.}
\format{
An object of class \code{list} of length 11.
}
\usage{
pipelines
}
\description{
Delta Live Tables is a framework for building reliable, maintainable, and
testable data processing pipelines. You define the transformations to perform
on your data, and Delta Live Tables manages task orchestration, cluster
management, monitoring, data quality, and error handling.
}
\details{
Instead of defining your data pipelines using a series of separate Apache
Spark tasks, Delta Live Tables manages how your data is transformed based on
a target schema you define for each processing step. You can also enforce
data quality with Delta Live Tables expectations. Expectations allow you to
define expected data quality and specify how to handle records that fail
those expectations.
}
\section{Operations}{

\tabular{ll}{
\link[=pipelines_create]{create} \tab Create a pipeline.\cr
\link[=pipelines_delete]{delete} \tab Delete a pipeline.\cr
\link[=pipelines_get]{get} \tab Get a pipeline.\cr
\link[=pipelines_get_update]{get_update} \tab Get a pipeline update.\cr
\link[=pipelines_list_pipeline_events]{list_pipeline_events} \tab List pipeline events.\cr
\link[=pipelines_list_pipelines]{list_pipelines} \tab List pipelines.\cr
\link[=pipelines_list_updates]{list_updates} \tab List pipeline updates.\cr
\link[=pipelines_reset]{reset} \tab Reset a pipeline.\cr
\link[=pipelines_start_update]{start_update} \tab Queue a pipeline update.\cr
\link[=pipelines_stop]{stop} \tab Stop a pipeline.\cr
\link[=pipelines_update]{update} \tab Edit a pipeline.\cr
}
}

\keyword{datasets}
