% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{pipelines_create}
\alias{pipelines_create}
\title{Create a pipeline.}
\usage{
pipelines_create(
  allow_duplicate_names = NULL,
  catalog = NULL,
  channel = NULL,
  clusters = NULL,
  configuration = NULL,
  continuous = NULL,
  development = NULL,
  dry_run = NULL,
  edition = NULL,
  filters = NULL,
  id = NULL,
  libraries = NULL,
  name = NULL,
  photon = NULL,
  storage = NULL,
  target = NULL,
  trigger = NULL,
  ...
)
}
\arguments{
\item{allow_duplicate_names}{If false, deployment will fail if name conflicts with that of another pipeline.}

\item{catalog}{Catalog in UC to add tables to.}

\item{channel}{DLT Release Channel that specifies which version to use.}

\item{clusters}{Cluster settings for this pipeline deployment.}

\item{configuration}{String-String configuration for this pipeline execution.}

\item{continuous}{Whether the pipeline is continuous or triggered.}

\item{development}{Whether the pipeline is in Development mode.}

\item{dry_run}{}

\item{edition}{Pipeline product edition.}

\item{filters}{Filters on which Pipeline packages to include in the deployed graph.}

\item{id}{Unique identifier for this pipeline.}

\item{libraries}{Libraries or code needed by this deployment.}

\item{name}{Friendly identifier for this pipeline.}

\item{photon}{Whether Photon is enabled for this pipeline.}

\item{storage}{DBFS root directory for storing checkpoints and tables.}

\item{target}{Target schema (database) to add tables in this pipeline to.}

\item{trigger}{Which pipeline trigger to use.}
}
\description{
Creates a new data processing pipeline based on the requested configuration.
If successful, this method returns the ID of the new pipeline.
}
\keyword{internal}
