% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{pipelines_delete}
\alias{pipelines_delete}
\title{Delete a pipeline.}
\usage{
pipelines_delete(pipeline_id)
}
\arguments{
\item{pipeline_id}{Required.}
}
\description{
Deletes a pipeline.
}
\keyword{internal}
