% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{pipelines_get}
\alias{pipelines_get}
\title{Get a pipeline.}
\usage{
pipelines_get(pipeline_id, ...)
}
\arguments{
\item{pipeline_id}{\link{required}}
}
\description{
Get a pipeline.
}
\keyword{internal}
