% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{pipelines_get_update}
\alias{pipelines_get_update}
\title{Get a pipeline update.}
\usage{
pipelines_get_update(pipeline_id, update_id, ...)
}
\arguments{
\item{pipeline_id}{Required. The ID of the pipeline.}

\item{update_id}{Required. The ID of the update.}
}
\description{
Gets an update from an active pipeline.
}
\keyword{internal}
