% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{pipelines_list_pipeline_events}
\alias{pipelines_list_pipeline_events}
\title{List pipeline events.}
\usage{
pipelines_list_pipeline_events(
  pipeline_id,
  filter = NULL,
  max_results = NULL,
  order_by = NULL,
  page_token = NULL
)
}
\arguments{
\item{pipeline_id}{Required.}

\item{filter}{Criteria to select a subset of results, expressed using a SQL-like syntax.}

\item{max_results}{Max number of entries to return in a single page.}

\item{order_by}{A string indicating a sort order by timestamp for the results, for example, \link{'timestamp asc'}.}

\item{page_token}{Page token returned by previous call.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Retrieves events for a pipeline.
}
\keyword{internal}
