% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{pipelines_reset}
\alias{pipelines_reset}
\title{Reset a pipeline.}
\usage{
pipelines_reset(pipeline_id, timeout = 20, callback = cli_reporter, ...)
}
\arguments{
\item{pipeline_id}{Required.}
}
\description{
This is a long-running operation, which blocks until Pipelines on Databricks reach
RUNNING state with the timeout of 20 minutes, that you can change via \code{timeout} parameter.
By default, the state of Databricks Pipelines is reported to console. You can change this behavior
by changing the \code{callback} parameter.
}
\details{
Resets a pipeline.
}
\keyword{internal}
