% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{pipelines_start_update}
\alias{pipelines_start_update}
\title{Queue a pipeline update.}
\usage{
pipelines_start_update(
  pipeline_id,
  cause = NULL,
  full_refresh = NULL,
  full_refresh_selection = NULL,
  refresh_selection = NULL,
  ...
)
}
\arguments{
\item{pipeline_id}{\link{required}}

\item{cause}{}

\item{full_refresh}{If true, this update will reset all tables before running.}

\item{full_refresh_selection}{A list of tables to update with fullRefresh.}

\item{refresh_selection}{A list of tables to update without fullRefresh.}
}
\description{
Starts or queues a pipeline update.
}
\keyword{internal}
