% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{pipelinesDelete}
\alias{pipelinesDelete}
\title{Delete a pipeline.}
\usage{
pipelinesDelete(client, pipeline_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{pipeline_id}{Required. This field has no description yet.}
}
\description{
Deletes a pipeline.
}
