% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{pipelinesGet}
\alias{pipelinesGet}
\title{Get a pipeline.}
\usage{
pipelinesGet(client, pipeline_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{pipeline_id}{Required. This field has no description yet.}
}
\description{
Get a pipeline.
}
