% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{pipelinesGetPipelinePermissionLevels}
\alias{pipelinesGetPipelinePermissionLevels}
\title{Get pipeline permission levels.}
\usage{
pipelinesGetPipelinePermissionLevels(client, pipeline_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{pipeline_id}{Required. The pipeline for which to get or manage permissions.}
}
\description{
Gets the permission levels that a user can have on an object.
}
