% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{pipelinesListPipelines}
\alias{pipelinesListPipelines}
\title{List pipelines.}
\usage{
pipelinesListPipelines(
  client,
  filter = NULL,
  max_results = NULL,
  order_by = NULL,
  page_token = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{filter}{Select a subset of results based on the specified criteria.}

\item{max_results}{The maximum number of entries to return in a single page.}

\item{order_by}{A list of strings specifying the order of results.}

\item{page_token}{Page token returned by previous call.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Lists pipelines defined in the Delta Live Tables system.
}
