% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{pipelinesListUpdates}
\alias{pipelinesListUpdates}
\title{List pipeline updates.}
\usage{
pipelinesListUpdates(
  client,
  pipeline_id,
  max_results = NULL,
  page_token = NULL,
  until_update_id = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{pipeline_id}{Required. The pipeline to return updates for.}

\item{max_results}{Max number of entries to return in a single page.}

\item{page_token}{Page token returned by previous call.}

\item{until_update_id}{If present, returns updates until and including this update_id.}
}
\description{
List updates for an active pipeline.
}
