% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{pipelinesSetPermissions}
\alias{pipelinesSetPermissions}
\title{Set pipeline permissions.}
\usage{
pipelinesSetPermissions(client, pipeline_id, access_control_list = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{pipeline_id}{Required. The pipeline for which to get or manage permissions.}

\item{access_control_list}{This field has no description yet.}
}
\description{
Sets permissions on a pipeline. Pipelines can inherit permissions from their
root object.
}
