% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{pipelinesStartUpdate}
\alias{pipelinesStartUpdate}
\title{Queue a pipeline update.}
\usage{
pipelinesStartUpdate(
  client,
  pipeline_id,
  cause = NULL,
  full_refresh = NULL,
  full_refresh_selection = NULL,
  refresh_selection = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{pipeline_id}{Required.}

\item{cause}{}

\item{full_refresh}{If true, this update will reset all tables before running.}

\item{full_refresh_selection}{A list of tables to update with fullRefresh.}

\item{refresh_selection}{A list of tables to update without fullRefresh.}
}
\description{
Starts or queues a pipeline update.
}
