% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{pipelinesUpdate}
\alias{pipelinesUpdate}
\title{Edit a pipeline.}
\usage{
pipelinesUpdate(
  client,
  pipeline_id,
  allow_duplicate_names = NULL,
  catalog = NULL,
  channel = NULL,
  clusters = NULL,
  configuration = NULL,
  continuous = NULL,
  development = NULL,
  edition = NULL,
  expected_last_modified = NULL,
  filters = NULL,
  id = NULL,
  libraries = NULL,
  name = NULL,
  photon = NULL,
  serverless = NULL,
  storage = NULL,
  target = NULL,
  trigger = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{pipeline_id}{Unique identifier for this pipeline.}

\item{allow_duplicate_names}{If false, deployment will fail if name has changed and conflicts the name of another pipeline.}

\item{catalog}{A catalog in Unity Catalog to publish data from this pipeline to.}

\item{channel}{DLT Release Channel that specifies which version to use.}

\item{clusters}{Cluster settings for this pipeline deployment.}

\item{configuration}{String-String configuration for this pipeline execution.}

\item{continuous}{Whether the pipeline is continuous or triggered.}

\item{development}{Whether the pipeline is in Development mode.}

\item{edition}{Pipeline product edition.}

\item{expected_last_modified}{If present, the last-modified time of the pipeline settings before the edit.}

\item{filters}{Filters on which Pipeline packages to include in the deployed graph.}

\item{id}{Unique identifier for this pipeline.}

\item{libraries}{Libraries or code needed by this deployment.}

\item{name}{Friendly identifier for this pipeline.}

\item{photon}{Whether Photon is enabled for this pipeline.}

\item{serverless}{Whether serverless compute is enabled for this pipeline.}

\item{storage}{DBFS root directory for storing checkpoints and tables.}

\item{target}{Target schema (database) to add tables in this pipeline to.}

\item{trigger}{Which pipeline trigger to use.}
}
\description{
Updates a pipeline with the supplied configuration.
}
