% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_families.R
\docType{data}
\name{policy_families}
\alias{policy_families}
\title{View available policy families. A policy family contains a policy definition
providing best practices for configuring clusters for a particular use case.}
\format{
An object of class \code{list} of length 2.
}
\usage{
policy_families
}
\description{
Databricks manages and provides policy families for several common cluster
use cases. You cannot create, edit, or delete policy families.
}
\details{
Policy families cannot be used directly to create clusters. Instead, you
create cluster policies using a policy family. Cluster policies created using
a policy family inherit the policy family's policy definition.
}
\section{Operations}{

\tabular{ll}{
\link[=policy_families_get]{get} \tab \cr
\link[=policy_families_list]{list} \tab \cr
}
}

\keyword{datasets}
