% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_families.R
\name{policyFamiliesGet}
\alias{policyFamiliesGet}
\title{Get policy family information.}
\usage{
policyFamiliesGet(client, policy_family_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{policy_family_id}{Required.}
}
\description{
Retrieve the information for an policy family based on its identifier.
}
