% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/providers.R
\docType{data}
\name{providers}
\alias{providers}
\title{Databricks Providers REST API}
\format{
An object of class \code{list} of length 6.
}
\usage{
providers
}
\description{
Databricks Providers REST API
}
\section{Operations}{

\tabular{ll}{
\link[=providers_create]{create} \tab Create an auth provider.\cr
\link[=providers_delete]{delete} \tab Delete a provider.\cr
\link[=providers_get]{get} \tab Get a provider.\cr
\link[=providers_list]{list} \tab List providers.\cr
\link[=providers_list_shares]{list_shares} \tab List shares by Provider.\cr
\link[=providers_update]{update} \tab Update a provider.\cr
}
}

\keyword{datasets}
