% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/providers.R
\name{providers_delete}
\alias{providers_delete}
\title{Delete a provider.}
\usage{
providers_delete(name, ...)
}
\arguments{
\item{name}{Required. Name of the provider.}
}
\description{
Deletes an authentication provider, if the caller is a metastore admin or is
the owner of the provider.
}
\keyword{internal}
