% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/providers.R
\name{providers_get}
\alias{providers_get}
\title{Get a provider.}
\usage{
providers_get(name)
}
\arguments{
\item{name}{Required. Name of the provider.}
}
\description{
Gets a specific authentication provider. The caller must supply the name of
the provider, and must either be a metastore admin or the owner of the
provider.
}
\keyword{internal}
