% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/providers.R
\name{providers_list_shares}
\alias{providers_list_shares}
\title{List shares by Provider.}
\usage{
providers_list_shares(name, ...)
}
\arguments{
\item{name}{\link{required} Name of the provider in which to list shares.}
}
\description{
Gets an array of a specified provider's shares within the metastore where:
}
\details{
\itemize{
\item the caller is a metastore admin, or * the caller is the owner.
}
}
\keyword{internal}
