% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/providers.R
\name{providers_update}
\alias{providers_update}
\title{Update a provider.}
\usage{
providers_update(
  name,
  comment = NULL,
  owner = NULL,
  recipient_profile_str = NULL,
  ...
)
}
\arguments{
\item{name}{The name of the Provider.}

\item{comment}{Description about the provider.}

\item{owner}{Username of Provider owner.}

\item{recipient_profile_str}{This field is required when the \strong{authentication_type} is \strong{TOKEN} or not provided.}
}
\description{
Updates the information for an authentication provider, if the caller is a
metastore admin or is the owner of the provider. If the update changes the
provider name, the caller must be both a metastore admin and the owner of the
provider.
}
\keyword{internal}
