% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/providers.R
\name{providersDelete}
\alias{providersDelete}
\title{Delete a provider.}
\usage{
providersDelete(client, name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Name of the provider.}
}
\description{
Deletes an authentication provider, if the caller is a metastore admin or is
the owner of the provider.
}
