% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeview.R
\name{publish_lakeview}
\alias{publish_lakeview}
\alias{lakeviewPublish}
\title{Publish dashboard.}
\usage{
publish_lakeview(
  client,
  dashboard_id,
  embed_credentials = NULL,
  warehouse_id = NULL
)

lakeviewPublish(
  client,
  dashboard_id,
  embed_credentials = NULL,
  warehouse_id = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{dashboard_id}{Required. UUID identifying the dashboard to be published.}

\item{embed_credentials}{Flag to indicate if the publisher's credentials should be embedded in the published dashboard.}

\item{warehouse_id}{The ID of the warehouse that can be used to override the warehouse which was set in the draft.

See the \href{https://docs.databricks.com/api/workspace/lakeview/publish}{API documentation} for more
information on the request and response field structure.}
}
\description{
Publish the current draft dashboard.
}
