% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serving_endpoints.R
\name{put_serving_endpoint}
\alias{put_serving_endpoint}
\title{Update rate limits of a serving endpoint.}
\usage{
put_serving_endpoint(client, name, rate_limits = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. The name of the serving endpoint whose rate limits are being updated.}

\item{rate_limits}{The list of endpoint rate limits.}
}
\description{
Used to update the rate limits of a serving endpoint. NOTE: only external and
foundation model endpoints are supported as of now.
}
