% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\docType{data}
\name{queries}
\alias{queries}
\title{These endpoints are used for CRUD operations on query definitions. Query
definitions include the target SQL warehouse, query text, name, description,
tags, execution schedule, parameters, and visualizations.}
\format{
An object of class \code{list} of length 6.
}
\usage{
queries
}
\description{
\strong{Note}: Programmatic operations on refresh schedules via the Databricks SQL
API are deprecated. Query refresh schedules can be created, updated, fetched
and deleted using Jobs API, e.g. :method:jobs/create.
}
\section{Operations}{

\tabular{ll}{
\link[=queries_create]{create} \tab Create a new query definition.\cr
\link[=queries_delete]{delete} \tab Delete a query.\cr
\link[=queries_get]{get} \tab Get a query definition.\cr
\link[=queries_list]{list} \tab Get a list of queries.\cr
\link[=queries_restore]{restore} \tab Restore a query.\cr
\link[=queries_update]{update} \tab Change a query definition.\cr
}
}

\keyword{datasets}
