% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{queries_create}
\alias{queries_create}
\title{Create a new query definition.}
\usage{
queries_create(
  data_source_id = NULL,
  description = NULL,
  name = NULL,
  options = NULL,
  parent = NULL,
  query = NULL
)
}
\arguments{
\item{data_source_id}{The ID of the data source / SQL warehouse where this query will run.}

\item{description}{General description that can convey additional information about this query such as usage notes.}

\item{name}{The name or title of this query to display in list views.}

\item{options}{Exclusively used for storing a list parameter definitions.}

\item{parent}{The identifier of the workspace folder containing the query.}

\item{query}{The text of the query.}
}
\description{
Creates a new query definition. Queries created with this endpoint belong to
the authenticated user making the request.
}
\details{
The \code{data_source_id} field specifies the ID of the SQL warehouse to run this
query against. You can use the Data Sources API to see a complete list of
available SQL warehouses. Or you can copy the \code{data_source_id} from an
existing query.

\strong{Note}: You cannot add a visualization until you create the query.
}
\keyword{internal}
