% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{queries_delete}
\alias{queries_delete}
\title{Delete a query.}
\usage{
queries_delete(query_id, ...)
}
\arguments{
\item{query_id}{Required.}
}
\description{
Moves a query to the trash. Trashed queries immediately disappear from
searches and list views, and they cannot be used for alerts. The trash is
deleted after 30 days.
}
\keyword{internal}
