% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{queries_get}
\alias{queries_get}
\title{Get a query definition.}
\usage{
queries_get(query_id, ...)
}
\arguments{
\item{query_id}{Required.}
}
\description{
Retrieve a query object definition along with contextual permissions
information about the currently authenticated user.
}
\keyword{internal}
