% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{queries_list}
\alias{queries_list}
\title{Get a list of queries.}
\usage{
queries_list(order = NULL, page = NULL, page_size = NULL, q = NULL, ...)
}
\arguments{
\item{order}{Name of query attribute to order by.}

\item{page}{Page number to retrieve.}

\item{page_size}{Number of queries to return per page.}

\item{q}{Full text search term.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Gets a list of queries. Optionally, this list can be filtered by a search
term.
}
\keyword{internal}
