% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{queries_restore}
\alias{queries_restore}
\title{Restore a query.}
\usage{
queries_restore(query_id, ...)
}
\arguments{
\item{query_id}{Required.}
}
\description{
Restore a query that has been moved to the trash. A restored query appears in
list views and searches. You can use restored queries for alerts.
}
\keyword{internal}
