% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{queries_update}
\alias{queries_update}
\title{Change a query definition.}
\usage{
queries_update(
  query_id,
  data_source_id = NULL,
  description = NULL,
  name = NULL,
  options = NULL,
  query = NULL
)
}
\arguments{
\item{query_id}{Required.}

\item{data_source_id}{The ID of the data source / SQL warehouse where this query will run.}

\item{description}{General description that can convey additional information about this query such as usage notes.}

\item{name}{The name or title of this query to display in list views.}

\item{options}{Exclusively used for storing a list parameter definitions.}

\item{query}{The text of the query.}
}
\description{
Modify this query definition.
}
\details{
\strong{Note}: You cannot undo this operation.
}
\keyword{internal}
