% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{queriesList}
\alias{queriesList}
\title{Get a list of queries.}
\usage{
queriesList(client, order = NULL, page = NULL, page_size = NULL, q = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{order}{Name of query attribute to order by.}

\item{page}{Page number to retrieve.}

\item{page_size}{Number of queries to return per page.}

\item{q}{Full text search term.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Gets a list of queries. Optionally, this list can be filtered by a search
term.
}
