% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_history.R
\docType{data}
\name{query_history}
\alias{query_history}
\title{Access the history of queries through SQL warehouses.}
\format{
An object of class \code{list} of length 1.
}
\usage{
query_history
}
\description{
Access the history of queries through SQL warehouses.
}
\section{Operations}{

\tabular{ll}{
\link[=query_history_list]{list} \tab List Queries.\cr
}
}

\keyword{datasets}
