% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serving_endpoints.R
\name{query_serving_endpoint}
\alias{query_serving_endpoint}
\title{Query a serving endpoint.}
\usage{
query_serving_endpoint(
  client,
  name,
  dataframe_records = NULL,
  dataframe_split = NULL,
  extra_params = NULL,
  input = NULL,
  inputs = NULL,
  instances = NULL,
  max_tokens = NULL,
  messages = NULL,
  n = NULL,
  prompt = NULL,
  stop = NULL,
  stream = NULL,
  temperature = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. The name of the serving endpoint.}

\item{dataframe_records}{Pandas Dataframe input in the records orientation.}

\item{dataframe_split}{Pandas Dataframe input in the split orientation.}

\item{extra_params}{The extra parameters field used ONLY for \strong{completions, chat,} and \strong{embeddings external & foundation model} serving endpoints.}

\item{input}{The input string (or array of strings) field used ONLY for \strong{embeddings external & foundation model} serving endpoints and is the only field (along with extra_params if needed) used by embeddings queries.}

\item{inputs}{Tensor-based input in columnar format.}

\item{instances}{Tensor-based input in row format.}

\item{max_tokens}{The max tokens field used ONLY for \strong{completions} and \strong{chat external & foundation model} serving endpoints.}

\item{messages}{The messages field used ONLY for \strong{chat external & foundation model} serving endpoints.}

\item{n}{The n (number of candidates) field used ONLY for \strong{completions} and \strong{chat external & foundation model} serving endpoints.}

\item{prompt}{The prompt string (or array of strings) field used ONLY for \strong{completions external & foundation model} serving endpoints and should only be used with other completions query fields.}

\item{stop}{The stop sequences field used ONLY for \strong{completions} and \strong{chat external & foundation model} serving endpoints.}

\item{stream}{The stream field used ONLY for \strong{completions} and \strong{chat external & foundation model} serving endpoints.}

\item{temperature}{The temperature field used ONLY for \strong{completions} and \strong{chat external & foundation model} serving endpoints.}
}
\description{
Query a serving endpoint.
}
