% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_search_indexes.R
\name{query_vector_search_index}
\alias{query_vector_search_index}
\alias{vectorSearchIndexesQueryIndex}
\title{Query an index.}
\usage{
query_vector_search_index(
  client,
  index_name,
  columns,
  filters_json = NULL,
  num_results = NULL,
  query_text = NULL,
  query_vector = NULL,
  score_threshold = NULL
)

vectorSearchIndexesQueryIndex(
  client,
  index_name,
  columns,
  filters_json = NULL,
  num_results = NULL,
  query_text = NULL,
  query_vector = NULL,
  score_threshold = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{index_name}{Required. Name of the vector index to query.}

\item{columns}{Required. List of column names to include in the response.}

\item{filters_json}{JSON string representing query filters.}

\item{num_results}{Number of results to return.}

\item{query_text}{Query text.}

\item{query_vector}{Query vector.}

\item{score_threshold}{Threshold for the approximate nearest neighbor search.}
}
\description{
Query the specified vector index.
}
