% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_history.R
\name{queryHistoryList}
\alias{queryHistoryList}
\title{List Queries.}
\usage{
queryHistoryList(
  client,
  filter_by = NULL,
  include_metrics = NULL,
  max_results = NULL,
  page_token = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{filter_by}{A filter to limit query history results.}

\item{include_metrics}{Whether to include metrics about query.}

\item{max_results}{Limit the number of results returned in one page.}

\item{page_token}{A token that can be used to get the next page of results.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
List the history of queries through SQL warehouses.
}
\details{
You can filter by user ID, warehouse ID, status, and time range.
}
