% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_visualizations.R
\name{queryVisualizationsUpdate}
\alias{queryVisualizationsUpdate}
\title{Edit existing visualization.}
\usage{
queryVisualizationsUpdate(
  client,
  id,
  created_at = NULL,
  description = NULL,
  name = NULL,
  options = NULL,
  type = NULL,
  updated_at = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{id}{The UUID for this visualization.}

\item{created_at}{}

\item{description}{A short description of this visualization.}

\item{name}{The name of the visualization that appears on dashboards and the query screen.}

\item{options}{The options object varies widely from one visualization type to the next and is unsupported.}

\item{type}{The type of visualization: chart, table, pivot table, and so on.}

\item{updated_at}{}
}
\description{
Edit existing visualization.
}
