% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipient_activation.R
\docType{data}
\name{recipient_activation}
\alias{recipient_activation}
\title{Databricks Delta Sharing: Recipient Activation REST API}
\format{
An object of class \code{list} of length 2.
}
\usage{
recipient_activation
}
\description{
Databricks Delta Sharing: Recipient Activation REST API
}
\section{Operations}{

\tabular{ll}{
\link[=recipient_activation_get_activation_url_info]{get_activation_url_info} \tab Get a share activation URL.\cr
\link[=recipient_activation_retrieve_token]{retrieve_token} \tab Get an access token.\cr
}
}

\keyword{datasets}
