% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipient_activation.R
\name{recipient_activation_get_activation_url_info}
\alias{recipient_activation_get_activation_url_info}
\title{Get a share activation URL.}
\usage{
recipient_activation_get_activation_url_info(activation_url, ...)
}
\arguments{
\item{activation_url}{\link{required} The one time activation url.}
}
\description{
Gets an activation URL for a share.
}
\keyword{internal}
