% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipient_activation.R
\name{recipient_activation_retrieve_token}
\alias{recipient_activation_retrieve_token}
\title{Get an access token.}
\usage{
recipient_activation_retrieve_token(activation_url, ...)
}
\arguments{
\item{activation_url}{Required. The one time activation url.}
}
\description{
Retrieve access token with an activation url. This is a public API without
any authentication.
}
\keyword{internal}
