% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipients.R
\docType{data}
\name{recipients}
\alias{recipients}
\title{Databricks Delta Sharing: Recipients REST API}
\format{
An object of class \code{list} of length 7.
}
\usage{
recipients
}
\description{
Databricks Delta Sharing: Recipients REST API
}
\section{Operations}{

\tabular{ll}{
\link[=recipients_create]{create} \tab Create a share recipient.\cr
\link[=recipients_delete]{delete} \tab Delete a share recipient.\cr
\link[=recipients_get]{get} \tab Get a share recipient.\cr
\link[=recipients_list]{list} \tab List share recipients.\cr
\link[=recipients_rotate_token]{rotate_token} \tab Rotate a token.\cr
\link[=recipients_share_permissions]{share_permissions} \tab Get recipient share permissions.\cr
\link[=recipients_update]{update} \tab Update a share recipient.\cr
}
}

\keyword{datasets}
