% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipients.R
\name{recipients_create}
\alias{recipients_create}
\title{Create a share recipient.}
\usage{
recipients_create(
  name,
  authentication_type,
  comment = NULL,
  data_recipient_global_metastore_id = NULL,
  ip_access_list = NULL,
  owner = NULL,
  properties_kvpairs = NULL,
  sharing_code = NULL
)
}
\arguments{
\item{name}{Required. Name of Recipient.}

\item{authentication_type}{Required. The delta sharing authentication type.}

\item{comment}{Description about the recipient.}

\item{data_recipient_global_metastore_id}{The global Unity Catalog metastore id provided by the data recipient.\n This field is required when the \strong{authentication_type} is \strong{DATABRICKS}.\n The identifier is of format \strong{cloud}:\strong{region}:\strong{metastore-uuid}.}

\item{ip_access_list}{IP Access List.}

\item{owner}{Username of the recipient owner.}

\item{properties_kvpairs}{Recipient properties as map of string key-value pairs.\n.}

\item{sharing_code}{The one-time sharing code provided by the data recipient.}
}
\description{
Creates a new recipient with the delta sharing authentication type in the
metastore. The caller must be a metastore admin or has the
\strong{CREATE_RECIPIENT} privilege on the metastore.
}
\keyword{internal}
