% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipients.R
\name{recipients_delete}
\alias{recipients_delete}
\title{Delete a share recipient.}
\usage{
recipients_delete(name, ...)
}
\arguments{
\item{name}{\link{required} Name of the recipient.}
}
\description{
Deletes the specified recipient from the metastore. The caller must be the
owner of the recipient.
}
\keyword{internal}
