% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipients.R
\name{recipients_get}
\alias{recipients_get}
\title{Get a share recipient.}
\usage{
recipients_get(name, ...)
}
\arguments{
\item{name}{\link{required} Name of the recipient.}
}
\description{
Gets a share recipient from the metastore if:
}
\details{
\itemize{
\item the caller is the owner of the share recipient, or: * is a metastore admin
}
}
\keyword{internal}
