% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipients.R
\name{recipients_share_permissions}
\alias{recipients_share_permissions}
\title{Get recipient share permissions.}
\usage{
recipients_share_permissions(name)
}
\arguments{
\item{name}{Required. The name of the Recipient.}
}
\description{
Gets the share permissions for the specified Recipient. The caller must be a
metastore admin or the owner of the Recipient.
}
\keyword{internal}
