% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipients.R
\name{recipients_update}
\alias{recipients_update}
\title{Update a share recipient.}
\usage{
recipients_update(
  name,
  comment = NULL,
  ip_access_list = NULL,
  owner = NULL,
  properties_kvpairs = NULL
)
}
\arguments{
\item{name}{Name of Recipient.}

\item{comment}{Description about the recipient.}

\item{ip_access_list}{IP Access List.}

\item{owner}{Username of the recipient owner.}

\item{properties_kvpairs}{Recipient properties as map of string key-value pairs.}
}
\description{
Updates an existing recipient in the metastore. The caller must be a
metastore admin or the owner of the recipient. If the recipient name will be
updated, the user must be both a metastore admin and the owner of the
recipient.
}
\keyword{internal}
