% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipients.R
\name{recipientsList}
\alias{recipientsList}
\title{List share recipients.}
\usage{
recipientsList(client, data_recipient_global_metastore_id = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{data_recipient_global_metastore_id}{If not provided, all recipients will be returned.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Gets an array of all share recipients within the current metastore where:
}
\details{
\itemize{
\item the caller is a metastore admin, or * the caller is the owner. There is no
guarantee of a specific ordering of the elements in the array.
}
}
