% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipients.R
\name{recipientsRotateToken}
\alias{recipientsRotateToken}
\title{Rotate a token.}
\usage{
recipientsRotateToken(client, existing_token_expire_in_seconds, name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{existing_token_expire_in_seconds}{Required. The expiration time of the bearer token in ISO 8601 format.}

\item{name}{Required. The name of the recipient.}
}
\description{
Refreshes the specified recipient's delta sharing authentication token with
the provided token info. The caller must be the owner of the recipient.
}
