% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registered_models.R
\name{registered_models_create}
\alias{registered_models_create}
\title{Create a model.}
\usage{
registered_models_create(name, description = NULL, tags = NULL)
}
\arguments{
\item{name}{Required. Register models under this name.}

\item{description}{Optional description for registered model.}

\item{tags}{Additional metadata for registered model.}
}
\description{
Creates a new registered model with the name specified in the request body.
}
\details{
Throws \code{RESOURCE_ALREADY_EXISTS} if a registered model with the given name
exists.
}
\keyword{internal}
