% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registered_models.R
\name{registered_models_delete}
\alias{registered_models_delete}
\title{Delete a model.}
\usage{
registered_models_delete(name, ...)
}
\arguments{
\item{name}{\link{required} Registered model unique name identifier.}
}
\description{
Deletes a registered model.
}
\keyword{internal}
