% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registered_models.R
\name{registered_models_delete_tag}
\alias{registered_models_delete_tag}
\title{Delete a model tag.}
\usage{
registered_models_delete_tag(name, key, ...)
}
\arguments{
\item{name}{\link{required} Name of the registered model that the tag was logged under.}

\item{key}{\link{required} Name of the tag.}
}
\description{
Deletes the tag for a registered model.
}
\keyword{internal}
