% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registered_models.R
\name{registered_models_get}
\alias{registered_models_get}
\title{Get a model.}
\usage{
registered_models_get(name, ...)
}
\arguments{
\item{name}{Required. Registered model unique name identifier.}
}
\description{
Gets the registered model that matches the specified ID.
}
\keyword{internal}
