% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registered_models.R
\name{registered_models_get_latest_versions}
\alias{registered_models_get_latest_versions}
\title{Get the latest version.}
\usage{
registered_models_get_latest_versions(name, stages = NULL)
}
\arguments{
\item{name}{Required. Registered model unique name identifier.}

\item{stages}{List of stages.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Gets the latest version of a registered model.
}
\keyword{internal}
