% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registered_models.R
\name{registered_models_rename}
\alias{registered_models_rename}
\title{Rename a model.}
\usage{
registered_models_rename(name, new_name = NULL, ...)
}
\arguments{
\item{name}{\link{required} Registered model unique name identifier.}

\item{new_name}{If provided, updates the name for this \code{registered_model}.}
}
\description{
Renames a registered model.
}
\keyword{internal}
