% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registered_models.R
\name{registered_models_update}
\alias{registered_models_update}
\title{Update model.}
\usage{
registered_models_update(name, description = NULL, ...)
}
\arguments{
\item{name}{Required. Registered model unique name identifier.}

\item{description}{If provided, updates the description for this \code{registered_model}.}
}
\description{
Updates a registered model.
}
\keyword{internal}
