% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registered_models.R
\name{registeredModelsDelete}
\alias{registeredModelsDelete}
\title{Delete a Registered Model.}
\usage{
registeredModelsDelete(client, full_name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{full_name}{Required. The three-level (fully qualified) name of the registered model.}
}
\description{
Deletes a registered model and all its model versions from the specified
parent catalog and schema.
}
\details{
The caller must be a metastore admin or an owner of the registered model. For
the latter case, the caller must also be the owner or have the
\strong{USE_CATALOG} privilege on the parent catalog and the \strong{USE_SCHEMA}
privilege on the parent schema.
}
