% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registered_models.R
\name{registeredModelsUpdate}
\alias{registeredModelsUpdate}
\title{Update a Registered Model.}
\usage{
registeredModelsUpdate(
  client,
  full_name,
  comment = NULL,
  new_name = NULL,
  owner = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{full_name}{Required. The three-level (fully qualified) name of the registered model.}

\item{comment}{The comment attached to the registered model.}

\item{new_name}{New name for the registered model.}

\item{owner}{The identifier of the user who owns the registered model.}
}
\description{
Updates the specified registered model.
}
\details{
The caller must be a metastore admin or an owner of the registered model. For
the latter case, the caller must also be the owner or have the
\strong{USE_CATALOG} privilege on the parent catalog and the \strong{USE_SCHEMA}
privilege on the parent schema.

Currently only the name, the owner or the comment of the registered model can
be updated.
}
