% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry_webhooks.R
\name{registry_webhooks_create}
\alias{registry_webhooks_create}
\title{Create a webhook.}
\usage{
registry_webhooks_create(
  events,
  description = NULL,
  http_url_spec = NULL,
  job_spec = NULL,
  model_name = NULL,
  status = NULL,
  ...
)
}
\arguments{
\item{events}{\link{required} Events that can trigger a registry webhook: * \code{MODEL_VERSION_CREATED}: A new model version was created for the associated model.}

\item{description}{User-specified description for the webhook.}

\item{http_url_spec}{}

\item{job_spec}{}

\item{model_name}{Name of the model whose events would trigger this webhook.}

\item{status}{This describes an enum.}
}
\description{
\strong{NOTE}: This endpoint is in Public Preview.
}
\details{
Creates a registry webhook.
}
\keyword{internal}
