% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry_webhooks.R
\name{registry_webhooks_delete}
\alias{registry_webhooks_delete}
\title{Delete a webhook.}
\usage{
registry_webhooks_delete(id = NULL)
}
\arguments{
\item{id}{Webhook ID required to delete a registry webhook.}
}
\description{
\strong{NOTE:} This endpoint is in Public Preview.
}
\details{
Deletes a registry webhook.
}
\keyword{internal}
