% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry_webhooks.R
\name{registry_webhooks_list}
\alias{registry_webhooks_list}
\title{List registry webhooks.}
\usage{
registry_webhooks_list(
  events = NULL,
  model_name = NULL,
  page_token = NULL,
  ...
)
}
\arguments{
\item{events}{If \code{events} is specified, any webhook with one or more of the specified trigger events is included in the output.}

\item{model_name}{If not specified, all webhooks associated with the specified events are listed, regardless of their associated model.}

\item{page_token}{Token indicating the page of artifact results to fetch.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
\strong{NOTE:} This endpoint is in Public Preview.
}
\details{
Lists all registry webhooks.
}
\keyword{internal}
