% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry_webhooks.R
\name{registry_webhooks_update}
\alias{registry_webhooks_update}
\title{Update a webhook.}
\usage{
registry_webhooks_update(
  id,
  description = NULL,
  events = NULL,
  http_url_spec = NULL,
  job_spec = NULL,
  status = NULL,
  ...
)
}
\arguments{
\item{id}{Required. Webhook ID.}

\item{description}{User-specified description for the webhook.}

\item{events}{Events that can trigger a registry webhook: * \code{MODEL_VERSION_CREATED}: A new model version was created for the associated model.}

\item{http_url_spec}{}

\item{job_spec}{}

\item{status}{This describes an enum.}
}
\description{
\strong{NOTE:} This endpoint is in Public Preview.
}
\details{
Updates a registry webhook.
}
\keyword{internal}
