% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{reject_model_transition_request}
\alias{reject_model_transition_request}
\alias{modelRegistryRejectTransitionRequest}
\title{Reject a transition request.}
\usage{
reject_model_transition_request(client, name, version, stage, comment = NULL)

modelRegistryRejectTransitionRequest(
  client,
  name,
  version,
  stage,
  comment = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Name of the model.}

\item{version}{Required. Version of the model.

See the \href{https://docs.databricks.com/api/workspace/modelregistry/rejecttransitionrequest}{API documentation} for more
information on the request and response field structure.}

\item{stage}{Required. Target stage of the transition.}

\item{comment}{User-provided comment on the action.}
}
\description{
Rejects a model version stage transition request.
}
