% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance_profiles.R
\name{remove_instance_profile}
\alias{remove_instance_profile}
\alias{instanceProfilesRemove}
\title{Remove the instance profile.}
\usage{
remove_instance_profile(client, instance_profile_arn)

instanceProfilesRemove(client, instance_profile_arn)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{instance_profile_arn}{Required. The ARN of the instance profile to remove.

See the \href{https://docs.databricks.com/api/workspace/instanceprofiles/remove}{API documentation} for more
information on the request and response field structure.}
}
\description{
Remove the instance profile with the provided ARN. Existing clusters with
this instance profile will continue to function.
}
\details{
This API is only accessible to admin users.
}
