% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{rename_model}
\alias{rename_model}
\alias{modelRegistryRenameModel}
\title{Rename a model.}
\usage{
rename_model(client, name, new_name = NULL)

modelRegistryRenameModel(client, name, new_name = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Registered model unique name identifier.}

\item{new_name}{If provided, updates the name for this \code{registered_model}.

See the \href{https://docs.databricks.com/api/workspace/modelregistry/renamemodel}{API documentation} for more
information on the request and response field structure.}
}
\description{
Renames a registered model.
}
