% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\docType{data}
\name{repos}
\alias{repos}
\title{The Repos API allows users to manage their git repos. Users can use the API
to access all repos that they have manage permissions on.}
\format{
An object of class \code{list} of length 5.
}
\usage{
repos
}
\description{
Databricks Repos is a visual Git client in Databricks. It supports common Git
operations such a cloning a repository, committing and pushing, pulling,
branch management, and visual comparison of diffs when committing.
}
\details{
Within Repos you can develop code in notebooks or other files and follow data
science and engineering code development best practices using Git for version
control, collaboration, and CI/CD.
}
\section{Operations}{

\tabular{ll}{
\link[=repos_create]{create} \tab Create a repo.\cr
\link[=repos_delete]{delete} \tab Delete a repo.\cr
\link[=repos_get]{get} \tab Get a repo.\cr
\link[=repos_list]{list} \tab Get repos.\cr
\link[=repos_update]{update} \tab Update a repo.\cr
}
}

\keyword{datasets}
