% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{repos_create}
\alias{repos_create}
\title{Create a repo.}
\usage{
repos_create(url, provider, path = NULL, sparse_checkout = NULL)
}
\arguments{
\item{url}{Required. URL of the Git repository to be linked.}

\item{provider}{Required. Git provider.}

\item{path}{Desired path for the repo in the workspace.}

\item{sparse_checkout}{If specified, the repo will be created with sparse checkout enabled.}
}
\description{
Creates a repo in the workspace and links it to the remote Git repo
specified. Note that repos created programmatically must be linked to a
remote Git repo, unlike repos created in the browser.
}
\keyword{internal}
