% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{repos_delete}
\alias{repos_delete}
\title{Delete a repo.}
\usage{
repos_delete(repo_id, ...)
}
\arguments{
\item{repo_id}{Required. The ID for the corresponding repo to access.}
}
\description{
Deletes the specified repo.
}
\keyword{internal}
