% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{repos_get}
\alias{repos_get}
\title{Get a repo.}
\usage{
repos_get(repo_id)
}
\arguments{
\item{repo_id}{Required. The ID for the corresponding repo to access.}
}
\description{
Returns the repo with the given repo ID.
}
\keyword{internal}
