% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{repos_list}
\alias{repos_list}
\title{Get repos.}
\usage{
repos_list(next_page_token = NULL, path_prefix = NULL, ...)
}
\arguments{
\item{next_page_token}{Token used to get the next page of results.}

\item{path_prefix}{Filters repos that have paths starting with the given path prefix.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Returns repos that the calling user has Manage permissions on. Results are
paginated with each page containing twenty repos.
}
\keyword{internal}
