% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{repos_update}
\alias{repos_update}
\title{Update a repo.}
\usage{
repos_update(repo_id, branch = NULL, sparse_checkout = NULL, tag = NULL, ...)
}
\arguments{
\item{repo_id}{\link{required} The ID for the corresponding repo to access.}

\item{branch}{Branch that the local version of the repo is checked out to.}

\item{sparse_checkout}{If specified, update the sparse checkout settings.}

\item{tag}{Tag that the local version of the repo is checked out to.}
}
\description{
Updates the repo to a different branch or tag, or updates the repo to the
latest commit on the same branch.
}
\keyword{internal}
