% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{reposDelete}
\alias{reposDelete}
\title{Delete a repo.}
\usage{
reposDelete(client, repo_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{repo_id}{Required. The ID for the corresponding repo to access.}
}
\description{
Deletes the specified repo.
}
