% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{reposGetPermissions}
\alias{reposGetPermissions}
\title{Get repo permissions.}
\usage{
reposGetPermissions(client, repo_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{repo_id}{Required. The repo for which to get or manage permissions.}
}
\description{
Gets the permissions of a repo. Repos can inherit permissions from their root
object.
}
