% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{reposSetPermissions}
\alias{reposSetPermissions}
\title{Set repo permissions.}
\usage{
reposSetPermissions(client, repo_id, access_control_list = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{repo_id}{Required. The repo for which to get or manage permissions.}

\item{access_control_list}{This field has no description yet.}
}
\description{
Sets permissions on a repo. Repos can inherit permissions from their root
object.
}
