% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{reposSetRepoPermissions}
\alias{reposSetRepoPermissions}
\title{Set repo permissions.}
\usage{
reposSetRepoPermissions(client, repo_id, access_control_list = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{repo_id}{Required. The repo for which to get or manage permissions.}

\item{access_control_list}{}
}
\description{
Sets permissions on a repo. Repos can inherit permissions from their root
object.
}
