% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{resize_cluster}
\alias{resize_cluster}
\title{Resize cluster.}
\usage{
resize_cluster(client, cluster_id, autoscale = NULL, num_workers = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{cluster_id}{Required. The cluster to be resized.}

\item{autoscale}{Parameters needed in order to automatically scale clusters up and down based on load.}

\item{num_workers}{Number of worker nodes that this cluster should have.}
}
\description{
Resizes a cluster to have a desired number of workers. This will fail unless
the cluster is in a \code{RUNNING} state.
}
