% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{restart_cluster}
\alias{restart_cluster}
\alias{clustersRestart}
\title{Restart cluster.}
\usage{
restart_cluster(client, cluster_id, restart_user = NULL)

clustersRestart(client, cluster_id, restart_user = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{cluster_id}{Required. The cluster to be started.}

\item{restart_user}{\if{html}{\out{<needs content added>}}.

See the \href{https://docs.databricks.com/api/workspace/clusters/restart}{API documentation} for more
information on the request and response field structure.}
}
\description{
Restarts a Spark cluster with the supplied ID. If the cluster is not
currently in a \code{RUNNING} state, nothing will happen.
}
