% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboards.R
\name{restore_dashboard}
\alias{restore_dashboard}
\alias{dashboardsRestore}
\title{Restore a dashboard.}
\usage{
restore_dashboard(client, dashboard_id)

dashboardsRestore(client, dashboard_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{dashboard_id}{Required. This field has no description yet.

See the \href{https://docs.databricks.com/api/workspace/dashboards/restore}{API documentation} for more
information on the request and response field structure.}
}
\description{
A restored dashboard appears in list views and searches and can be shared.
}
