% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipient_activation.R
\name{retrieve_recipient_activation_token}
\alias{retrieve_recipient_activation_token}
\alias{recipientActivationRetrieveToken}
\title{Get an access token.}
\usage{
retrieve_recipient_activation_token(client, activation_url)

recipientActivationRetrieveToken(client, activation_url)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{activation_url}{Required. The one time activation url.}
}
\description{
Retrieve access token with an activation url. This is a public API without
any authentication.
}
