% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipients.R
\name{rotate_recipient_token}
\alias{rotate_recipient_token}
\alias{recipientsRotateToken}
\title{Rotate a token.}
\usage{
rotate_recipient_token(client, name, existing_token_expire_in_seconds)

recipientsRotateToken(client, name, existing_token_expire_in_seconds)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. The name of the recipient.

See the \href{https://docs.databricks.com/api/workspace/recipients/rotatetoken}{API documentation} for more
information on the request and response field structure.}

\item{existing_token_expire_in_seconds}{Required. The expiration time of the bearer token in ISO 8601 format.}
}
\description{
Refreshes the specified recipient's delta sharing authentication token with
the provided token info. The caller must be the owner of the recipient.
}
