% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\docType{data}
\name{schemas}
\alias{schemas}
\title{A schema (also called a database) is the second layer of Unity Catalog’s
three-level namespace. A schema organizes tables, views and functions. To
access (or list) a table or view in a schema, users must have the USE_SCHEMA
data permission on the schema and its parent catalog, and they must have the
SELECT permission on the table or view.}
\format{
An object of class \code{list} of length 5.
}
\usage{
schemas
}
\description{
A schema (also called a database) is the second layer of Unity Catalog’s
three-level namespace. A schema organizes tables, views and functions. To
access (or list) a table or view in a schema, users must have the USE_SCHEMA
data permission on the schema and its parent catalog, and they must have the
SELECT permission on the table or view.
}
\section{Operations}{

\tabular{ll}{
\link[=schemas_create]{create} \tab Create a schema.\cr
\link[=schemas_delete]{delete} \tab Delete a schema.\cr
\link[=schemas_get]{get} \tab Get a schema.\cr
\link[=schemas_list]{list} \tab List schemas.\cr
\link[=schemas_update]{update} \tab Update a schema.\cr
}
}

\keyword{datasets}
