% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\name{schemas_create}
\alias{schemas_create}
\title{Create a schema.}
\usage{
schemas_create(
  name,
  catalog_name,
  comment = NULL,
  properties = NULL,
  storage_root = NULL,
  ...
)
}
\arguments{
\item{name}{Required. Name of schema, relative to parent catalog.}

\item{catalog_name}{Required. Name of parent catalog.}

\item{comment}{User-provided free-form text description.}

\item{properties}{A map of key-value properties attached to the securable.}

\item{storage_root}{Storage root URL for managed tables within schema.}
}
\description{
Creates a new schema for catalog in the Metatastore. The caller must be a
metastore admin, or have the \strong{CREATE_SCHEMA} privilege in the parent
catalog.
}
\keyword{internal}
