% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\name{schemas_delete}
\alias{schemas_delete}
\title{Delete a schema.}
\usage{
schemas_delete(full_name, ...)
}
\arguments{
\item{full_name}{\link{required} Full name of the schema.}
}
\description{
Deletes the specified schema from the parent catalog. The caller must be the
owner of the schema or an owner of the parent catalog.
}
\keyword{internal}
