% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\name{schemas_get}
\alias{schemas_get}
\title{Get a schema.}
\usage{
schemas_get(full_name)
}
\arguments{
\item{full_name}{Required. Full name of the schema.}
}
\description{
Gets the specified schema within the metastore. The caller must be a
metastore admin, the owner of the schema, or a user that has the
\strong{USE_SCHEMA} privilege on the schema.
}
\keyword{internal}
