% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\name{schemas_list}
\alias{schemas_list}
\title{List schemas.}
\usage{
schemas_list(catalog_name, ...)
}
\arguments{
\item{catalog_name}{Required. Parent catalog for schemas of interest.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Gets an array of schemas for a catalog in the metastore. If the caller is the
metastore admin or the owner of the parent catalog, all schemas for the
catalog will be retrieved. Otherwise, only schemas owned by the caller (or
for which the caller has the \strong{USE_SCHEMA} privilege) will be retrieved.
There is no guarantee of a specific ordering of the elements in the array.
}
\keyword{internal}
