% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{search_experiment_experiments}
\alias{search_experiment_experiments}
\alias{experimentsSearchExperiments}
\title{Search experiments.}
\usage{
search_experiment_experiments(
  client,
  filter = NULL,
  max_results = NULL,
  order_by = NULL,
  page_token = NULL,
  view_type = NULL
)

experimentsSearchExperiments(
  client,
  filter = NULL,
  max_results = NULL,
  order_by = NULL,
  page_token = NULL,
  view_type = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{filter}{String representing a SQL filter condition (e.g.}

\item{max_results}{Maximum number of experiments desired.}

\item{order_by}{List of columns for ordering search results, which can include experiment name and last updated timestamp with an optional 'DESC' or 'ASC' annotation, where 'ASC' is the default.}

\item{page_token}{Token indicating the page of experiments to fetch.}

\item{view_type}{Qualifier for type of experiments to be returned.}
}
\value{
\code{data.frame} with all of the response pages.

See the \href{https://docs.databricks.com/api/workspace/experiments/searchexperiments}{API documentation} for more
information on the request and response field structure.
}
\description{
Searches for experiments that satisfy specified search criteria.
}
