% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{search_experiment_runs}
\alias{search_experiment_runs}
\alias{experimentsSearchRuns}
\title{Search for runs.}
\usage{
search_experiment_runs(
  client,
  experiment_ids = NULL,
  filter = NULL,
  max_results = NULL,
  order_by = NULL,
  page_token = NULL,
  run_view_type = NULL
)

experimentsSearchRuns(
  client,
  experiment_ids = NULL,
  filter = NULL,
  max_results = NULL,
  order_by = NULL,
  page_token = NULL,
  run_view_type = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{experiment_ids}{List of experiment IDs to search over.}

\item{filter}{A filter expression over params, metrics, and tags, that allows returning a subset of runs.}

\item{max_results}{Maximum number of runs desired.}

\item{order_by}{List of columns to be ordered by, including attributes, params, metrics, and tags with an optional 'DESC' or 'ASC' annotation, where 'ASC' is the default.}

\item{page_token}{Token for the current page of runs.}

\item{run_view_type}{Whether to display only active, only deleted, or all runs.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Searches for runs that satisfy expressions.
}
\details{
Search expressions can use \code{mlflowMetric} and \code{mlflowParam} keys.',
}
