% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\docType{data}
\name{secrets}
\alias{secrets}
\title{The Secrets API allows you to manage secrets, secret scopes, and access
permissions.}
\format{
An object of class \code{list} of length 10.
}
\usage{
secrets
}
\description{
Sometimes accessing data requires that you authenticate to external data
sources through JDBC. Instead of directly entering your credentials into a
notebook, use Databricks secrets to store your credentials and reference them
in notebooks and jobs.
}
\details{
Administrators, secret creators, and users granted permission can read
Databricks secrets. While Databricks makes an effort to redact secret values
that might be displayed in notebooks, it is not possible to prevent such
users from reading secrets.
}
\section{Operations}{

\tabular{ll}{
\link[=secrets_create_scope]{create_scope} \tab Create a new secret scope.\cr
\link[=secrets_delete_acl]{delete_acl} \tab Delete an ACL.\cr
\link[=secrets_delete_scope]{delete_scope} \tab Delete a secret scope.\cr
\link[=secrets_delete_secret]{delete_secret} \tab Delete a secret.\cr
\link[=secrets_get_acl]{get_acl} \tab Get secret ACL details.\cr
\link[=secrets_list_acls]{list_acls} \tab Lists ACLs.\cr
\link[=secrets_list_scopes]{list_scopes} \tab List all scopes.\cr
\link[=secrets_list_secrets]{list_secrets} \tab List secret keys.\cr
\link[=secrets_put_acl]{put_acl} \tab Create/update an ACL.\cr
\link[=secrets_put_secret]{put_secret} \tab Add a secret.\cr
}
}

\keyword{datasets}
