% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{secrets_create_scope}
\alias{secrets_create_scope}
\title{Create a new secret scope.}
\usage{
secrets_create_scope(
  scope,
  initial_manage_principal = NULL,
  keyvault_metadata = NULL,
  scope_backend_type = NULL,
  ...
)
}
\arguments{
\item{scope}{\link{required} Scope name requested by the user.}

\item{initial_manage_principal}{The principal that is initially granted \code{MANAGE} permission to the created scope.}

\item{keyvault_metadata}{The metadata for the secret scope if the type is \code{AZURE_KEYVAULT}.}

\item{scope_backend_type}{The backend type the scope will be created with.}
}
\description{
The scope name must consist of alphanumeric characters, dashes, underscores,
and periods, and may not exceed 128 characters. The maximum number of scopes
in a workspace is 100.
}
\keyword{internal}
