% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{secrets_delete_scope}
\alias{secrets_delete_scope}
\title{Delete a secret scope.}
\usage{
secrets_delete_scope(scope, ...)
}
\arguments{
\item{scope}{Required. Name of the scope to delete.}
}
\description{
Deletes a secret scope.
}
\details{
Throws \code{RESOURCE_DOES_NOT_EXIST} if the scope does not exist. Throws
\code{PERMISSION_DENIED} if the user does not have permission to make this API
call.
}
\keyword{internal}
