% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{secrets_delete_secret}
\alias{secrets_delete_secret}
\title{Delete a secret.}
\usage{
secrets_delete_secret(scope, key, ...)
}
\arguments{
\item{scope}{\link{required} The name of the scope that contains the secret to delete.}

\item{key}{\link{required} Name of the secret to delete.}
}
\description{
Deletes the secret stored in this secret scope. You must have \code{WRITE} or
\code{MANAGE} permission on the secret scope.
}
\details{
Throws \code{RESOURCE_DOES_NOT_EXIST} if no such secret scope or secret exists.
Throws \code{PERMISSION_DENIED} if the user does not have permission to make this
API call.
}
\keyword{internal}
