% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{secrets_list_acls}
\alias{secrets_list_acls}
\title{Lists ACLs.}
\usage{
secrets_list_acls(scope)
}
\arguments{
\item{scope}{Required. The name of the scope to fetch ACL information from.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
List the ACLs for a given secret scope. Users must have the \code{MANAGE}
permission to invoke this API.
}
\details{
Throws \code{RESOURCE_DOES_NOT_EXIST} if no such secret scope exists. Throws
\code{PERMISSION_DENIED} if the user does not have permission to make this API
call.
}
\keyword{internal}
