% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{secrets_list_scopes}
\alias{secrets_list_scopes}
\title{List all scopes.}
\usage{
secrets_list_scopes()
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Lists all secret scopes available in the workspace.
}
\details{
Throws \code{PERMISSION_DENIED} if the user does not have permission to make this
API call.#'
}
\keyword{internal}
