% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{secrets_list_secrets}
\alias{secrets_list_secrets}
\title{List secret keys.}
\usage{
secrets_list_secrets(scope)
}
\arguments{
\item{scope}{Required. The name of the scope to list secrets within.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Lists the secret keys that are stored at this scope. This is a metadata-only
operation; secret data cannot be retrieved using this API. Users need the
READ permission to make this call.
}
\details{
The lastUpdatedTimestamp returned is in milliseconds since epoch. Throws
\code{RESOURCE_DOES_NOT_EXIST} if no such secret scope exists. Throws
\code{PERMISSION_DENIED} if the user does not have permission to make this API
call.
}
\keyword{internal}
