% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{secretsGetSecret}
\alias{secretsGetSecret}
\title{Get a secret.}
\usage{
secretsGetSecret(client, scope, key)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{scope}{Required. The name of the scope to fetch secret information from.}

\item{key}{Required. The key to fetch secret for.}
}
\description{
Gets the bytes representation of a secret value for the specified scope and
key.
}
\details{
Users need the READ permission to make this call.

Note that the secret value returned is in bytes. The interpretation of the
bytes is determined by the caller in DBUtils and the type the data is decoded
into.

Throws \code{PERMISSION_DENIED} if the user does not have permission to make
this API call. Throws \code{RESOURCE_DOES_NOT_EXIST} if no such secret or secret
scope exists.
}
