% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{secretsListScopes}
\alias{secretsListScopes}
\title{List all scopes.}
\usage{
secretsListScopes(client)
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Lists all secret scopes available in the workspace.
}
\details{
Throws \code{PERMISSION_DENIED} if the user does not have permission to make this
API call.#'
}
