% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service_principals.R
\docType{data}
\name{service_principals}
\alias{service_principals}
\title{Identities for use with jobs, automated tools, and systems such as scripts,
apps, and CI/CD platforms. Databricks recommends creating service principals
to run production jobs or modify production data. If all processes that act
on production data run with service principals, interactive users do not need
any write, delete, or modify privileges in production. This eliminates the
risk of a user overwriting production data by accident.}
\format{
An object of class \code{list} of length 6.
}
\usage{
service_principals
}
\description{
Identities for use with jobs, automated tools, and systems such as scripts,
apps, and CI/CD platforms. Databricks recommends creating service principals
to run production jobs or modify production data. If all processes that act
on production data run with service principals, interactive users do not need
any write, delete, or modify privileges in production. This eliminates the
risk of a user overwriting production data by accident.
}
\section{Operations}{

\tabular{ll}{
\link[=service_principals_create]{create} \tab Create a service principal.\cr
\link[=service_principals_delete]{delete} \tab Delete a service principal.\cr
\link[=service_principals_get]{get} \tab Get service principal details.\cr
\link[=service_principals_list]{list} \tab List service principals.\cr
\link[=service_principals_patch]{patch} \tab Update service principal details.\cr
\link[=service_principals_update]{update} \tab Replace service principal.\cr
}
}

\keyword{datasets}
