% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service_principals.R
\name{service_principals_create}
\alias{service_principals_create}
\title{Create a service principal.}
\usage{
service_principals_create(
  active = NULL,
  application_id = NULL,
  display_name = NULL,
  entitlements = NULL,
  external_id = NULL,
  groups = NULL,
  id = NULL,
  roles = NULL
)
}
\arguments{
\item{active}{If this user is active.}

\item{application_id}{UUID relating to the service principal.}

\item{display_name}{String that represents a concatenation of given and family names.}

\item{entitlements}{}

\item{external_id}{}

\item{groups}{}

\item{id}{Databricks service principal ID.}

\item{roles}{}
}
\description{
Creates a new service principal in the Databricks workspace.
}
\keyword{internal}
