% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service_principals.R
\name{service_principals_delete}
\alias{service_principals_delete}
\title{Delete a service principal.}
\usage{
service_principals_delete(id)
}
\arguments{
\item{id}{Required. Unique ID for a service principal in the Databricks Workspace.}
}
\description{
Delete a single service principal in the Databricks Workspace.
}
\keyword{internal}
