% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service_principals.R
\name{service_principals_get}
\alias{service_principals_get}
\title{Get service principal details.}
\usage{
service_principals_get(id, ...)
}
\arguments{
\item{id}{\link{required} Unique ID for a service principal in the Databricks Workspace.}
}
\description{
Gets the details for a single service principal define in the Databricks
Workspace.
}
\keyword{internal}
