% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service_principals.R
\name{service_principals_list}
\alias{service_principals_list}
\title{List service principals.}
\usage{
service_principals_list(
  attributes = NULL,
  count = NULL,
  excluded_attributes = NULL,
  filter = NULL,
  sort_by = NULL,
  sort_order = NULL,
  start_index = NULL,
  ...
)
}
\arguments{
\item{attributes}{Comma-separated list of attributes to return in response.}

\item{count}{Desired number of results per page.}

\item{excluded_attributes}{Comma-separated list of attributes to exclude in response.}

\item{filter}{Query by which the results have to be filtered.}

\item{sort_by}{Attribute to sort the results.}

\item{sort_order}{The order to sort the results.}

\item{start_index}{Specifies the index of the first result.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Gets the set of service principals associated with a Databricks Workspace.
}
\keyword{internal}
