% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service_principals.R
\name{service_principals_patch}
\alias{service_principals_patch}
\title{Update service principal details.}
\usage{
service_principals_patch(id, operations = NULL)
}
\arguments{
\item{id}{Required. Unique ID for a group in the Databricks account.}

\item{operations}{}
}
\description{
Partially updates the details of a single service principal in the Databricks
workspace.
}
\keyword{internal}
