% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service_principals.R
\name{service_principals_update}
\alias{service_principals_update}
\title{Replace service principal.}
\usage{
service_principals_update(
  id,
  active = NULL,
  application_id = NULL,
  display_name = NULL,
  entitlements = NULL,
  external_id = NULL,
  groups = NULL,
  roles = NULL,
  ...
)
}
\arguments{
\item{id}{Databricks service principal ID.}

\item{active}{If this user is active.}

\item{application_id}{UUID relating to the service principal.}

\item{display_name}{String that represents a concatenation of given and family names.}

\item{entitlements}{}

\item{external_id}{}

\item{groups}{}

\item{roles}{}
}
\description{
Updates the details of a single service principal.
}
\details{
This action replaces the existing service principal with the same name.
}
\keyword{internal}
