% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service_principals.R
\name{servicePrincipalsCreate}
\alias{servicePrincipalsCreate}
\title{Create a service principal.}
\usage{
servicePrincipalsCreate(
  client,
  active = NULL,
  application_id = NULL,
  display_name = NULL,
  entitlements = NULL,
  external_id = NULL,
  groups = NULL,
  id = NULL,
  roles = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{active}{If this user is active.}

\item{application_id}{UUID relating to the service principal.}

\item{display_name}{String that represents a concatenation of given and family names.}

\item{entitlements}{}

\item{external_id}{}

\item{groups}{}

\item{id}{Databricks service principal ID.}

\item{roles}{}
}
\description{
Creates a new service principal in the Databricks workspace.
}
