% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service_principals.R
\name{servicePrincipalsPatch}
\alias{servicePrincipalsPatch}
\title{Update service principal details.}
\usage{
servicePrincipalsPatch(client, id, operations = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{id}{Required. Unique ID for a service principal in the Databricks workspace.}

\item{operations}{}
}
\description{
Partially updates the details of a single service principal in the Databricks
workspace.
}
