% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service_principals.R
\name{servicePrincipalsUpdate}
\alias{servicePrincipalsUpdate}
\title{Replace service principal.}
\usage{
servicePrincipalsUpdate(
  client,
  id,
  active = NULL,
  application_id = NULL,
  display_name = NULL,
  entitlements = NULL,
  external_id = NULL,
  groups = NULL,
  roles = NULL,
  schemas = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{id}{Databricks service principal ID.}

\item{active}{If this user is active.}

\item{application_id}{UUID relating to the service principal.}

\item{display_name}{String that represents a concatenation of given and family names.}

\item{entitlements}{Entitlements assigned to the service principal.}

\item{external_id}{This field has no description yet.}

\item{groups}{This field has no description yet.}

\item{roles}{Corresponds to AWS instance profile/arn role.}

\item{schemas}{The schema of the List response.}
}
\description{
Updates the details of a single service principal.
}
\details{
This action replaces the existing service principal with the same name.
}
