% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serving_endpoints.R
\docType{data}
\name{serving_endpoints}
\alias{serving_endpoints}
\title{The Serving Endpoints API allows you to create, update, and delete model
serving endpoints.}
\format{
An object of class \code{list} of length 9.
}
\usage{
serving_endpoints
}
\description{
You can use a serving endpoint to serve models from the Databricks Model
Registry. Endpoints expose the underlying models as scalable REST API
endpoints using serverless compute. This means the endpoints and associated
compute resources are fully managed by Databricks and will not appear in your
cloud account. A serving endpoint can consist of one or more MLflow models
from the Databricks Model Registry, called served models. A serving endpoint
can have at most ten served models. You can configure traffic settings to
define how requests should be routed to your served models behind an
endpoint. Additionally, you can configure the scale of resources that should
be applied to each served model.
}
\section{Operations}{

\tabular{ll}{
\link[=serving_endpoints_build_logs]{build_logs} \tab Retrieve the logs associated with building the model's environment for a given serving endpoint's served model.\cr
\link[=serving_endpoints_create]{create} \tab Create a new serving endpoint.\cr
\link[=serving_endpoints_delete]{delete} \tab Delete a serving endpoint.\cr
\link[=serving_endpoints_export_metrics]{export_metrics} \tab Retrieve the metrics corresponding to a serving endpoint for the current time in Prometheus or OpenMetrics exposition format.\cr
\link[=serving_endpoints_get]{get} \tab Get a single serving endpoint.\cr
\link[=serving_endpoints_list]{list} \tab Retrieve all serving endpoints.\cr
\link[=serving_endpoints_logs]{logs} \tab Retrieve the most recent log lines associated with a given serving endpoint's served model.\cr
\link[=serving_endpoints_query]{query} \tab Query a serving endpoint with provided model input.\cr
\link[=serving_endpoints_update_config]{update_config} \tab Update a serving endpoint with a new config.\cr
}
}

\keyword{datasets}
