% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serving_endpoints.R
\name{serving_endpoints_delete}
\alias{serving_endpoints_delete}
\title{Delete a serving endpoint.}
\usage{
serving_endpoints_delete(name, ...)
}
\arguments{
\item{name}{\link{required} The name of the serving endpoint.}
}
\description{
Delete a serving endpoint.
}
\keyword{internal}
