% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serving_endpoints.R
\name{serving_endpoints_export_metrics}
\alias{serving_endpoints_export_metrics}
\title{Retrieve the metrics corresponding to a serving endpoint for the current time
in Prometheus or OpenMetrics exposition format.}
\usage{
serving_endpoints_export_metrics(name)
}
\arguments{
\item{name}{Required. The name of the serving endpoint to retrieve metrics for.}
}
\description{
Retrieves the metrics associated with the provided serving endpoint in either
Prometheus or OpenMetrics exposition format. Please note that this API is in
preview and may change in the future.
}
\keyword{internal}
