% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serving_endpoints.R
\name{serving_endpoints_get}
\alias{serving_endpoints_get}
\title{Get a single serving endpoint.}
\usage{
serving_endpoints_get(name)
}
\arguments{
\item{name}{Required. The name of the serving endpoint.}
}
\description{
Retrieves the details for a single serving endpoint.
}
\keyword{internal}
