% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serving_endpoints.R
\name{serving_endpoints_logs}
\alias{serving_endpoints_logs}
\title{Retrieve the most recent log lines associated with a given serving endpoint's
served model.}
\usage{
serving_endpoints_logs(name, served_model_name)
}
\arguments{
\item{name}{Required. The name of the serving endpoint that the served model belongs to.}

\item{served_model_name}{Required. The name of the served model that logs will be retrieved for.}
}
\description{
Retrieves the service logs associated with the provided served model. Please
note that this API is in preview and may change in the future.
}
\keyword{internal}
