% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serving_endpoints.R
\name{servingEndpointsCreateAndWait}
\alias{servingEndpointsCreateAndWait}
\title{Create a new serving endpoint.}
\usage{
servingEndpointsCreateAndWait(
  client,
  name,
  config,
  rate_limits = NULL,
  tags = NULL,
  timeout = 20,
  callback = cli_reporter
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. The name of the serving endpoint.}

\item{config}{Required. The core config of the serving endpoint.}

\item{rate_limits}{Rate limits to be applied to the serving endpoint.}

\item{tags}{Tags to be attached to the serving endpoint and automatically propagated to billing logs.}

\item{timeout}{Time to wait for the operation to complete in minutes.}

\item{callback}{Function to report the status of the operation. By default, it reports to console.}
}
\description{
This is a long-running operation, which blocks until Serving Endpoints on Databricks reach
NOT_UPDATING state with the timeout of 20 minutes, that you can change via \code{timeout} parameter.
By default, the state of Databricks Serving Endpoints is reported to console. You can change this behavior
by changing the \code{callback} parameter.
}
