% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serving_endpoints.R
\name{servingEndpointsDelete}
\alias{servingEndpointsDelete}
\title{Delete a serving endpoint.}
\usage{
servingEndpointsDelete(client, name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. The name of the serving endpoint.}
}
\description{
Delete a serving endpoint.
}
