% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serving_endpoints.R
\name{servingEndpointsExportMetrics}
\alias{servingEndpointsExportMetrics}
\title{Get metrics of a serving endpoint.}
\usage{
servingEndpointsExportMetrics(client, name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. The name of the serving endpoint to retrieve metrics for.}
}
\description{
Retrieves the metrics associated with the provided serving endpoint in either
Prometheus or OpenMetrics exposition format.
}
