% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serving_endpoints.R
\name{servingEndpointsGet}
\alias{servingEndpointsGet}
\title{Get a single serving endpoint.}
\usage{
servingEndpointsGet(client, name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. The name of the serving endpoint.}
}
\description{
Retrieves the details for a single serving endpoint.
}
