% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serving_endpoints.R
\name{servingEndpointsList}
\alias{servingEndpointsList}
\title{Retrieve all serving endpoints.#'}
\usage{
servingEndpointsList(client)
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Retrieve all serving endpoints.#'
}
