% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serving_endpoints.R
\name{servingEndpointsQuery}
\alias{servingEndpointsQuery}
\title{Query a serving endpoint with provided model input.}
\usage{
servingEndpointsQuery(client, name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. The name of the serving endpoint.}
}
\description{
Query a serving endpoint with provided model input.
}
