% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serving_endpoints.R
\name{servingEndpointsUpdateServingEndpointPermissions}
\alias{servingEndpointsUpdateServingEndpointPermissions}
\title{Update serving endpoint permissions.}
\usage{
servingEndpointsUpdateServingEndpointPermissions(
  client,
  serving_endpoint_id,
  access_control_list = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{serving_endpoint_id}{Required. The serving endpoint for which to get or manage permissions.}

\item{access_control_list}{}
}
\description{
Updates the permissions on a serving endpoint. Serving endpoints can inherit
permissions from their root object.
}
