% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{set_cluster_permissions}
\alias{set_cluster_permissions}
\alias{clustersSetPermissions}
\title{Set cluster permissions.}
\usage{
set_cluster_permissions(client, cluster_id, access_control_list = NULL)

clustersSetPermissions(client, cluster_id, access_control_list = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{cluster_id}{Required. The cluster for which to get or manage permissions.

See the \href{https://docs.databricks.com/api/workspace/clusters/setpermissions}{API documentation} for more
information on the request and response field structure.}

\item{access_control_list}{This field has no description yet.}
}
\description{
Sets permissions on a cluster. Clusters can inherit permissions from their
root object.
}
