% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_policies.R
\name{set_cluster_policy_permissions}
\alias{set_cluster_policy_permissions}
\alias{clusterPoliciesSetPermissions}
\title{Set cluster policy permissions.}
\usage{
set_cluster_policy_permissions(
  client,
  cluster_policy_id,
  access_control_list = NULL
)

clusterPoliciesSetPermissions(
  client,
  cluster_policy_id,
  access_control_list = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{cluster_policy_id}{Required. The cluster policy for which to get or manage permissions.

See the \href{https://docs.databricks.com/api/workspace/clusterpolicies/setpermissions}{API documentation} for more
information on the request and response field structure.}

\item{access_control_list}{This field has no description yet.}
}
\description{
Sets permissions on a cluster policy. Cluster policies can inherit
permissions from their root object.
}
