% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance_pools.R
\name{set_instance_pool_permissions}
\alias{set_instance_pool_permissions}
\alias{instancePoolsSetPermissions}
\title{Set instance pool permissions.}
\usage{
set_instance_pool_permissions(
  client,
  instance_pool_id,
  access_control_list = NULL
)

instancePoolsSetPermissions(
  client,
  instance_pool_id,
  access_control_list = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{instance_pool_id}{Required. The instance pool for which to get or manage permissions.}

\item{access_control_list}{This field has no description yet.}
}
\description{
Sets permissions on an instance pool. Instance pools can inherit permissions
from their root object.
}
