% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{set_model_permissions}
\alias{set_model_permissions}
\alias{modelRegistrySetPermissions}
\title{Set registered model permissions.}
\usage{
set_model_permissions(client, registered_model_id, access_control_list = NULL)

modelRegistrySetPermissions(
  client,
  registered_model_id,
  access_control_list = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{registered_model_id}{Required. The registered model for which to get or manage permissions.

See the \href{https://docs.databricks.com/api/workspace/modelregistry/setpermissions}{API documentation} for more
information on the request and response field structure.}

\item{access_control_list}{This field has no description yet.}
}
\description{
Sets permissions on a registered model. Registered models can inherit
permissions from their root object.
}
