% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{set_model_version_tag}
\alias{set_model_version_tag}
\title{Set a version tag.}
\usage{
set_model_version_tag(client, name, version, key, value)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Unique name of the model.}

\item{version}{Required. Model version number.}

\item{key}{Required. Name of the tag.}

\item{value}{Required. String value of the tag being logged.}
}
\description{
Sets a model version tag.
}
