% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{set_permission}
\alias{set_permission}
\alias{permissionsSet}
\title{Set object permissions.}
\usage{
set_permission(
  client,
  request_object_type,
  request_object_id,
  access_control_list = NULL
)

permissionsSet(
  client,
  request_object_type,
  request_object_id,
  access_control_list = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{request_object_type}{Required. The type of the request object.}

\item{request_object_id}{Required. The id of the request object.}

\item{access_control_list}{This field has no description yet.}
}
\description{
Sets permissions on an object. Objects can inherit permissions from their
parent objects or root object.
}
