% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registered_models.R
\name{set_registered_model_alias}
\alias{set_registered_model_alias}
\alias{registeredModelsSetAlias}
\title{Set a Registered Model Alias.}
\usage{
set_registered_model_alias(client, full_name, alias, version_num)

registeredModelsSetAlias(client, full_name, alias, version_num)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{full_name}{Required. Full name of the registered model.}

\item{alias}{Required. The name of the alias.}

\item{version_num}{Required. The version number of the model version to which the alias points.}
}
\description{
Set an alias on the specified registered model.
}
\details{
The caller must be a metastore admin or an owner of the registered model. For
the latter case, the caller must also be the owner or have the
\strong{USE_CATALOG} privilege on the parent catalog and the \strong{USE_SCHEMA}
privilege on the parent schema.
}
