% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{set_repo_permissions}
\alias{set_repo_permissions}
\alias{reposSetPermissions}
\title{Set repo permissions.}
\usage{
set_repo_permissions(client, repo_id, access_control_list = NULL)

reposSetPermissions(client, repo_id, access_control_list = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{repo_id}{Required. The repo for which to get or manage permissions.

See the \href{https://docs.databricks.com/api/workspace/repos/setpermissions}{API documentation} for more
information on the request and response field structure.}

\item{access_control_list}{This field has no description yet.}
}
\description{
Sets permissions on a repo. Repos can inherit permissions from their root
object.
}
