% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_management.R
\name{set_token_management_permissions}
\alias{set_token_management_permissions}
\alias{tokenManagementSetPermissions}
\title{Set token permissions.}
\usage{
set_token_management_permissions(client, access_control_list = NULL)

tokenManagementSetPermissions(client, access_control_list = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{access_control_list}{This field has no description yet.

See the \href{https://docs.databricks.com/api/workspace/tokenmanagement/setpermissions}{API documentation} for more
information on the request and response field structure.}
}
\description{
Sets permissions on all tokens. Tokens can inherit permissions from their
root object.
}
