% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouses.R
\name{set_warehouse_permissions}
\alias{set_warehouse_permissions}
\title{Set SQL warehouse permissions.}
\usage{
set_warehouse_permissions(client, warehouse_id, access_control_list = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{warehouse_id}{Required. The SQL warehouse for which to get or manage permissions.}

\item{access_control_list}{This field has no description yet.}
}
\description{
Sets permissions on a SQL warehouse. SQL warehouses can inherit permissions
from their root object.
}
