% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouses.R
\name{set_warehouse_workspace_config}
\alias{set_warehouse_workspace_config}
\title{Set the workspace configuration.}
\usage{
set_warehouse_workspace_config(
  client,
  channel = NULL,
  config_param = NULL,
  data_access_config = NULL,
  enabled_warehouse_types = NULL,
  global_param = NULL,
  google_service_account = NULL,
  instance_profile_arn = NULL,
  security_policy = NULL,
  sql_configuration_parameters = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{channel}{Optional: Channel selection details.}

\item{config_param}{Deprecated: Use sql_configuration_parameters.}

\item{data_access_config}{Spark confs for external hive metastore configuration JSON serialized size must be less than <= 512K.}

\item{enabled_warehouse_types}{List of Warehouse Types allowed in this workspace (limits allowed value of the type field in CreateWarehouse and EditWarehouse).}

\item{global_param}{Deprecated: Use sql_configuration_parameters.}

\item{google_service_account}{GCP only: Google Service Account used to pass to cluster to access Google Cloud Storage.}

\item{instance_profile_arn}{AWS Only: Instance profile used to pass IAM role to the cluster.}

\item{security_policy}{Security policy for warehouses.}

\item{sql_configuration_parameters}{SQL configuration parameters.}
}
\description{
Sets the workspace level configuration that is shared by all SQL warehouses
in a workspace.
}
