% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace_conf.R
\name{set_workspace_conf_status}
\alias{set_workspace_conf_status}
\title{Enable/disable features.}
\usage{
set_workspace_conf_status(client)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}
}
\description{
Sets the configuration status for a workspace, including enabling or
disabling it.
}
