% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shares.R
\docType{data}
\name{shares}
\alias{shares}
\title{Databricks Delta Sharing: Shares REST API}
\format{
An object of class \code{list} of length 7.
}
\usage{
shares
}
\description{
Databricks Delta Sharing: Shares REST API
}
\section{Operations}{

\tabular{ll}{
\link[=shares_create]{create} \tab Create a share.\cr
\link[=shares_delete]{delete} \tab Delete a share.\cr
\link[=shares_get]{get} \tab Get a share.\cr
\link[=shares_list]{list} \tab List shares.\cr
\link[=shares_share_permissions]{share_permissions} \tab Get permissions.\cr
\link[=shares_update]{update} \tab Update a share.\cr
\link[=shares_update_permissions]{update_permissions} \tab Update permissions.\cr
}
}

\keyword{datasets}
