% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shares.R
\name{shares_delete}
\alias{shares_delete}
\title{Delete a share.}
\usage{
shares_delete(name, ...)
}
\arguments{
\item{name}{\link{required} The name of the share.}
}
\description{
Deletes a data object share from the metastore. The caller must be an owner
of the share.
}
\keyword{internal}
