% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shares.R
\name{shares_get}
\alias{shares_get}
\title{Get a share.}
\usage{
shares_get(name, include_shared_data = NULL, ...)
}
\arguments{
\item{name}{Required. The name of the share.}

\item{include_shared_data}{Query for data to include in the share.}
}
\description{
Gets a data object share from the metastore. The caller must be a metastore
admin or the owner of the share.
}
\keyword{internal}
