% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shares.R
\name{shares_share_permissions}
\alias{shares_share_permissions}
\title{Get permissions.}
\usage{
shares_share_permissions(name)
}
\arguments{
\item{name}{Required. The name of the share.}
}
\description{
Gets the permissions for a data share from the metastore. The caller must be
a metastore admin or the owner of the share.
}
\keyword{internal}
