% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shares.R
\name{sharesList}
\alias{sharesList}
\title{List shares.}
\usage{
sharesList(client)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Gets an array of data object shares from the metastore. The caller must be a
metastore admin or the owner of the share. There is no guarantee of a
specific ordering of the elements in the array.
}
