% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shares.R
\name{sharesSharePermissions}
\alias{sharesSharePermissions}
\title{Get permissions.}
\usage{
sharesSharePermissions(client, name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. The name of the share.}
}
\description{
Gets the permissions for a data share from the metastore. The caller must be
a metastore admin or the owner of the share.
}
