% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{start_cluster}
\alias{start_cluster}
\title{Start terminated cluster.}
\usage{
start_cluster(client, cluster_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{cluster_id}{Required. The cluster to be started.}
}
\description{
Starts a terminated Spark cluster with the supplied ID. This works similar to
\code{createCluster} except:
}
\details{
\itemize{
\item The previous cluster id and attributes are preserved. * The cluster starts
with the last specified cluster size. * If the previous cluster was an
autoscaling cluster, the current cluster starts with the minimum number of
nodes. * If the cluster is not currently in a \code{TERMINATED} state, nothing
will happen. * Clusters launched to run a job cannot be started.
}
}
