% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{start_pipeline_update}
\alias{start_pipeline_update}
\title{Start a pipeline.}
\usage{
start_pipeline_update(
  client,
  pipeline_id,
  cause = NULL,
  full_refresh = NULL,
  full_refresh_selection = NULL,
  refresh_selection = NULL,
  validate_only = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{pipeline_id}{Required. This field has no description yet.}

\item{cause}{This field has no description yet.}

\item{full_refresh}{If true, this update will reset all tables before running.}

\item{full_refresh_selection}{A list of tables to update with fullRefresh.}

\item{refresh_selection}{A list of tables to update without fullRefresh.}

\item{validate_only}{If true, this update only validates the correctness of pipeline source code but does not materialize or publish any datasets.}
}
\description{
Starts a new update for the pipeline. If there is already an active update
for the pipeline, the request will fail and the active update will remain
running.
}
