% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouses.R
\name{start_warehouse}
\alias{start_warehouse}
\alias{warehousesStart}
\title{Start a warehouse.}
\usage{
start_warehouse(client, id)

warehousesStart(client, id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{id}{Required. Required.

See the \href{https://docs.databricks.com/api/workspace/warehouses/start}{API documentation} for more
information on the request and response field structure.}
}
\description{
Starts a SQL warehouse.
}
