% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement_execution.R
\name{statement_execution_cancel_execution}
\alias{statement_execution_cancel_execution}
\title{Cancel statement execution.}
\usage{
statement_execution_cancel_execution(statement_id)
}
\arguments{
\item{statement_id}{Required.}
}
\description{
Requests that an executing statement be canceled. Callers must poll for
status to see the terminal state.
}
\keyword{internal}
