% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement_execution.R
\name{statement_execution_get_statement}
\alias{statement_execution_get_statement}
\title{Get status, manifest, and result first chunk.}
\usage{
statement_execution_get_statement(statement_id, ...)
}
\arguments{
\item{statement_id}{\link{required}}
}
\description{
Polls for statement status; when status.state=SUCCEEDED will also return the
result manifest, and the first chunk of result data.
}
\details{
\strong{NOTE} This call currently may take up to 5 seconds to get the latest
status and result.
}
\keyword{internal}
