% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement_execution.R
\name{statement_execution_get_statement_result_chunk_n}
\alias{statement_execution_get_statement_result_chunk_n}
\title{Get result chunk by index.}
\usage{
statement_execution_get_statement_result_chunk_n(statement_id, chunk_index)
}
\arguments{
\item{statement_id}{Required.}

\item{chunk_index}{Required.}
}
\description{
After statement execution has SUCCEEDED, result data can be fetched by
chunks.
}
\details{
The first chunk (\code{chunk_index=0}) is typically fetched through
\code{getStatementResult}, and subsequent chunks with this call. The response
structure is identical to the nested \code{result} element described in
getStatementResult, and similarly includes \code{next_chunk_index} and
\code{next_chunk_internal_link} for simple iteration through the result set.
}
\keyword{internal}
