% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_credentials.R
\name{storage_credentials_create}
\alias{storage_credentials_create}
\title{Create a storage credential.}
\usage{
storage_credentials_create(
  name,
  metastore_id,
  aws_iam_role = NULL,
  azure_service_principal = NULL,
  comment = NULL,
  gcp_service_account_key = NULL,
  read_only = NULL,
  skip_validation = NULL,
  ...
)
}
\arguments{
\item{name}{Required. The credential name.}

\item{metastore_id}{Required. Databricks Unity Catalog metastore ID.}

\item{aws_iam_role}{The AWS IAM role configuration.}

\item{azure_service_principal}{The Azure service principal configuration.}

\item{comment}{Comment associated with the credential.}

\item{gcp_service_account_key}{The GCP service account key configuration.}

\item{read_only}{Whether the storage credential is only usable for read operations.}

\item{skip_validation}{Supplying true to this argument skips validation of the created credential.}
}
\description{
Creates a new storage credential. The request object is specific to the
cloud:
}
\details{
\itemize{
\item \strong{AwsIamRole} for AWS credentials * \strong{AzureServicePrincipal} for Azure
credentials * \strong{GcpServiceAcountKey} for GCP credentials.
}

The caller must be a metastore admin and have the
\strong{CREATE_STORAGE_CREDENTIAL} privilege on the metastore.
}
\keyword{internal}
