% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_credentials.R
\name{storage_credentials_delete}
\alias{storage_credentials_delete}
\title{Delete a credential.}
\usage{
storage_credentials_delete(name, force = NULL, ...)
}
\arguments{
\item{name}{\link{required} Name of the storage credential.}

\item{force}{Force deletion even if there are dependent external locations or external tables.}
}
\description{
Deletes a storage credential from the metastore. The caller must be an owner
of the storage credential.
}
\keyword{internal}
