% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_credentials.R
\name{storage_credentials_get}
\alias{storage_credentials_get}
\title{Get a credential.}
\usage{
storage_credentials_get(name)
}
\arguments{
\item{name}{Required. Name of the storage credential.}
}
\description{
Gets a storage credential from the metastore. The caller must be a metastore
admin, the owner of the storage credential, or have some permission on the
storage credential.
}
\keyword{internal}
