% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_credentials.R
\name{storage_credentials_update}
\alias{storage_credentials_update}
\title{Update a credential.}
\usage{
storage_credentials_update(
  metastore_id,
  name,
  aws_iam_role = NULL,
  azure_service_principal = NULL,
  comment = NULL,
  force = NULL,
  gcp_service_account_key = NULL,
  owner = NULL,
  read_only = NULL,
  skip_validation = NULL
)
}
\arguments{
\item{metastore_id}{Required. Unity Catalog metastore ID.}

\item{name}{The credential name.}

\item{aws_iam_role}{The AWS IAM role configuration.}

\item{azure_service_principal}{The Azure service principal configuration.}

\item{comment}{Comment associated with the credential.}

\item{force}{Force update even if there are dependent external locations or external tables.}

\item{gcp_service_account_key}{The GCP service account key configuration.}

\item{owner}{Username of current owner of credential.}

\item{read_only}{Whether the storage credential is only usable for read operations.}

\item{skip_validation}{Supplying true to this argument skips validation of the updated credential.}
}
\description{
Updates a storage credential on the metastore. The caller must be the owner
of the storage credential or a metastore admin. If the caller is a metastore
admin, only the \strong{owner} credential can be changed.
}
\keyword{internal}
