% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_credentials.R
\name{storage_credentials_validate}
\alias{storage_credentials_validate}
\title{Validate a storage credential.}
\usage{
storage_credentials_validate(
  aws_iam_role = NULL,
  azure_service_principal = NULL,
  external_location_name = NULL,
  gcp_service_account_key = NULL,
  read_only = NULL,
  storage_credential_name = NULL,
  url = NULL,
  ...
)
}
\arguments{
\item{aws_iam_role}{The AWS IAM role configuration.}

\item{azure_service_principal}{The Azure service principal configuration.}

\item{external_location_name}{The name of an existing external location to validate.}

\item{gcp_service_account_key}{The GCP service account key configuration.}

\item{read_only}{Whether the storage credential is only usable for read operations.}

\item{storage_credential_name}{The name of the storage credential to validate.}

\item{url}{The external location url to validate.}
}
\description{
Validates a storage credential. At least one of \strong{external_location_name}
and \strong{url} need to be provided. If only one of them is provided, it will be
used for validation. And if both are provided, the \strong{url} will be used for
validation, and \strong{external_location_name} will be ignored when checking
overlapping urls.
}
\details{
Either the \strong{storage_credential_name} or the cloud-specific credential must
be provided.

The caller must be a metastore admin or the storage credential owner or have
the \strong{CREATE_EXTERNAL_LOCATION} privilege on the metastore and the storage
credential.
}
\keyword{internal}
