% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{submit_job}
\alias{submit_job}
\alias{jobsSubmit}
\title{Create and trigger a one-time run.}
\usage{
submit_job(
  client,
  access_control_list = NULL,
  email_notifications = NULL,
  git_source = NULL,
  health = NULL,
  idempotency_token = NULL,
  notification_settings = NULL,
  queue = NULL,
  run_name = NULL,
  tasks = NULL,
  timeout_seconds = NULL,
  webhook_notifications = NULL
)

jobsSubmit(
  client,
  access_control_list = NULL,
  email_notifications = NULL,
  git_source = NULL,
  health = NULL,
  idempotency_token = NULL,
  notification_settings = NULL,
  queue = NULL,
  run_name = NULL,
  tasks = NULL,
  timeout_seconds = NULL,
  webhook_notifications = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{access_control_list}{List of permissions to set on the job.}

\item{email_notifications}{An optional set of email addresses notified when the run begins or completes.}

\item{git_source}{An optional specification for a remote Git repository containing the source code used by tasks.}

\item{health}{An optional set of health rules that can be defined for this job.}

\item{idempotency_token}{An optional token that can be used to guarantee the idempotency of job run requests.}

\item{notification_settings}{Optional notification settings that are used when sending notifications to each of the \code{email_notifications} and \code{webhook_notifications} for this run.}

\item{queue}{The queue settings of the one-time run.}

\item{run_name}{An optional name for the run.}

\item{tasks}{This field has no description yet.}

\item{timeout_seconds}{An optional timeout applied to each run of this job.}

\item{webhook_notifications}{A collection of system notification IDs to notify when the run begins or completes.}
}
\description{
Submit a one-time run. This endpoint allows you to submit a workload directly
without creating a job. Runs submitted using this endpoint don’t display in
the UI. Use the \code{jobs/runs/get} API to check the run state after the job is
submitted.
}
