% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_schemas.R
\docType{data}
\name{system_schemas}
\alias{system_schemas}
\title{A system schema is a schema that lives within the system catalog. A system
schema may contain information about customer usage of Unity Catalog such as
audit-logs, billing-logs, lineage information, etc.}
\format{
An object of class \code{list} of length 3.
}
\usage{
system_schemas
}
\description{
A system schema is a schema that lives within the system catalog. A system
schema may contain information about customer usage of Unity Catalog such as
audit-logs, billing-logs, lineage information, etc.
}
\section{Operations}{

\tabular{ll}{
\link[=system_schemas_disable]{disable} \tab Disable a system schema.\cr
\link[=system_schemas_enable]{enable} \tab Enable a system schema.\cr
\link[=system_schemas_list]{list} \tab List system schemas.\cr
}
}

\keyword{datasets}
