% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_schemas.R
\name{system_schemas_disable}
\alias{system_schemas_disable}
\title{Disable a system schema.}
\usage{
system_schemas_disable(metastore_id, schema_name)
}
\arguments{
\item{metastore_id}{Required. The metastore ID under which the system schema lives.}

\item{schema_name}{Required. Full name of the system schema.}
}
\description{
Disables the system schema and removes it from the system catalog. The caller
must be an account admin or a metastore admin.
}
\keyword{internal}
