% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_schemas.R
\name{system_schemas_enable}
\alias{system_schemas_enable}
\title{Enable a system schema.}
\usage{
system_schemas_enable()
}
\description{
Enables the system schema and adds it to the system catalog. The caller must
be an account admin or a metastore admin.#'
}
\keyword{internal}
