% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_schemas.R
\name{system_schemas_list}
\alias{system_schemas_list}
\title{List system schemas.}
\usage{
system_schemas_list(metastore_id)
}
\arguments{
\item{metastore_id}{Required. The ID for the metastore in which the system schema resides.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Gets an array of system schemas for a metastore. The caller must be an
account admin or a metastore admin.
}
\keyword{internal}
