% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_schemas.R
\name{systemSchemasDisable}
\alias{systemSchemasDisable}
\title{Disable a system schema.}
\usage{
systemSchemasDisable(client, metastore_id, schema_name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{metastore_id}{Required. The metastore ID under which the system schema lives.}

\item{schema_name}{Required. Full name of the system schema.}
}
\description{
Disables the system schema and removes it from the system catalog. The caller
must be an account admin or a metastore admin.
}
