% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_constraints.R
\docType{data}
\name{table_constraints}
\alias{table_constraints}
\title{Primary key and foreign key constraints encode relationships between fields
in tables.}
\format{
An object of class \code{list} of length 2.
}
\usage{
table_constraints
}
\description{
Primary and foreign keys are informational only and are not enforced. Foreign
keys must reference a primary key in another table. This primary key is the
parent constraint of the foreign key and the table this primary key is on is
the parent table of the foreign key. Similarly, the foreign key is the child
constraint of its referenced primary key; the table of the foreign key is the
child table of the primary key.
}
\details{
You can declare primary keys and foreign keys as part of the table
specification during table creation. You can also add or drop constraints on
existing tables.
}
\section{Operations}{

\tabular{ll}{
\link[=table_constraints_create]{create} \tab Create a table constraint.\cr
\link[=table_constraints_delete]{delete} \tab Delete a table constraint.\cr
}
}

\keyword{datasets}
