% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\docType{data}
\name{tables}
\alias{tables}
\title{A table resides in the third layer of Unity Catalog’s three-level
namespace. It contains rows of data. To create a table, users must have
CREATE_TABLE and USE_SCHEMA permissions on the schema, and they must have the
USE_CATALOG permission on its parent catalog. To query a table, users must
have the SELECT permission on the table, and they must have the USE_CATALOG
permission on its parent catalog and the USE_SCHEMA permission on its parent
schema.}
\format{
An object of class \code{list} of length 4.
}
\usage{
tables
}
\description{
A table can be managed or external. From an API perspective, a \strong{VIEW} is a
particular kind of table (rather than a managed or external table).
}
\section{Operations}{

\tabular{ll}{
\link[=tables_delete]{delete} \tab Delete a table.\cr
\link[=tables_get]{get} \tab Get a table.\cr
\link[=tables_list]{list} \tab List tables.\cr
\link[=tables_list_summaries]{list_summaries} \tab List table summaries.\cr
}
}

\keyword{datasets}
