% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tables_delete}
\alias{tables_delete}
\title{Delete a table.}
\usage{
tables_delete(full_name, ...)
}
\arguments{
\item{full_name}{\link{required} Full name of the table.}
}
\description{
Deletes a table from the specified parent catalog and schema. The caller must
be the owner of the parent catalog, have the \strong{USE_CATALOG} privilege on the
parent catalog and be the owner of the parent schema, or be the owner of the
table and have the \strong{USE_CATALOG} privilege on the parent catalog and the
\strong{USE_SCHEMA} privilege on the parent schema.
}
\keyword{internal}
