% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tables_list_summaries}
\alias{tables_list_summaries}
\title{List table summaries.}
\usage{
tables_list_summaries(
  catalog_name,
  max_results = NULL,
  page_token = NULL,
  schema_name_pattern = NULL,
  table_name_pattern = NULL,
  ...
)
}
\arguments{
\item{catalog_name}{\link{required} Name of parent catalog for tables of interest.}

\item{max_results}{Maximum number of tables to return (page length).}

\item{page_token}{Opaque token to send for the next page of results (pagination).}

\item{schema_name_pattern}{A sql LIKE pattern (\% and _) for schema names.}

\item{table_name_pattern}{A sql LIKE pattern (\% and _) for table names.}
}
\description{
Gets an array of summaries for tables for a schema and catalog within the
metastore. The table summaries returned are either:
}
\details{
\itemize{
\item summaries for all tables (within the current metastore and parent catalog
and schema), when the user is a metastore admin, or: * summaries for all
tables and schemas (within the current metastore and parent catalog) for
which the user has ownership or the \strong{SELECT} privilege on the table and
ownership or \strong{USE_SCHEMA} privilege on the schema, provided that the user
also has ownership or the \strong{USE_CATALOG} privilege on the parent catalog.
}

There is no guarantee of a specific ordering of the elements in the array.
}
\keyword{internal}
