% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tablesGet}
\alias{tablesGet}
\title{Get a table.}
\usage{
tablesGet(client, full_name, include_delta_metadata = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{full_name}{Required. Full name of the table.}

\item{include_delta_metadata}{Whether delta metadata should be included in the response.}
}
\description{
Gets a table from the metastore for a specific catalog and schema. The caller
must be a metastore admin, be the owner of the table and have the
\strong{USE_CATALOG} privilege on the parent catalog and the \strong{USE_SCHEMA}
privilege on the parent schema, or be the owner of the table and have the
\strong{SELECT} privilege on it as well.
}
