% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tablesUpdate}
\alias{tablesUpdate}
\title{Update a table owner.}
\usage{
tablesUpdate(client, full_name, owner = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{full_name}{Required. Full name of the table.}

\item{owner}{}
}
\description{
Change the owner of the table. The caller must be the owner of the parent
catalog, have the \strong{USE_CATALOG} privilege on the parent catalog and be the
owner of the parent schema, or be the owner of the table and have the
\strong{USE_CATALOG} privilege on the parent catalog and the \strong{USE_SCHEMA}
privilege on the parent schema.
}
