% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_management.R
\docType{data}
\name{token_management}
\alias{token_management}
\title{Enables administrators to get all tokens and delete tokens for other users.
Admins can either get every token, get a specific token by ID, or get all
tokens for a particular user.}
\format{
An object of class \code{list} of length 4.
}
\usage{
token_management
}
\description{
Enables administrators to get all tokens and delete tokens for other users.
Admins can either get every token, get a specific token by ID, or get all
tokens for a particular user.
}
\section{Operations}{

\tabular{ll}{
\link[=token_management_create_obo_token]{create_obo_token} \tab Create on-behalf token.\cr
\link[=token_management_delete]{delete} \tab Delete a token.\cr
\link[=token_management_get]{get} \tab Get token info.\cr
\link[=token_management_list]{list} \tab List all tokens.\cr
}
}

\keyword{datasets}
