% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_management.R
\name{token_management_create_obo_token}
\alias{token_management_create_obo_token}
\title{Create on-behalf token.}
\usage{
token_management_create_obo_token(
  application_id,
  lifetime_seconds,
  comment = NULL,
  ...
)
}
\arguments{
\item{application_id}{Required. Application ID of the service principal.}

\item{lifetime_seconds}{Required. The number of seconds before the token expires.}

\item{comment}{Comment that describes the purpose of the token.}
}
\description{
Creates a token on behalf of a service principal.
}
\keyword{internal}
