% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_management.R
\name{tokenManagementDelete}
\alias{tokenManagementDelete}
\title{Delete a token.}
\usage{
tokenManagementDelete(client, token_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{token_id}{Required. The ID of the token to get.}
}
\description{
Deletes a token, specified by its ID.
}
