% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_management.R
\name{tokenManagementGet}
\alias{tokenManagementGet}
\title{Get token info.}
\usage{
tokenManagementGet(client, token_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{token_id}{Required. The ID of the token to get.}
}
\description{
Gets information about a token, specified by its ID.
}
