% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_management.R
\name{tokenManagementGetTokenPermissionLevels}
\alias{tokenManagementGetTokenPermissionLevels}
\title{Get token permission levels.}
\usage{
tokenManagementGetTokenPermissionLevels(client)
}
\description{
Gets the permission levels that a user can have on an object.#'
}
