% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_management.R
\name{tokenManagementSetPermissions}
\alias{tokenManagementSetPermissions}
\title{Set token permissions.}
\usage{
tokenManagementSetPermissions(client, access_control_list = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{access_control_list}{}
}
\description{
Sets permissions on all tokens. Tokens can inherit permissions from their
root object.
}
