% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\docType{data}
\name{tokens}
\alias{tokens}
\title{The Token API allows you to create, list, and revoke tokens that can be used
to authenticate and access Databricks REST APIs.}
\format{
An object of class \code{list} of length 3.
}
\usage{
tokens
}
\description{
The Token API allows you to create, list, and revoke tokens that can be used
to authenticate and access Databricks REST APIs.
}
\section{Operations}{

\tabular{ll}{
\link[=tokens_create]{create} \tab Create a user token.\cr
\link[=tokens_delete]{delete} \tab Revoke token.\cr
\link[=tokens_list]{list} \tab List tokens.\cr
}
}

\keyword{datasets}
