% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{tokens_delete}
\alias{tokens_delete}
\title{Revoke token.}
\usage{
tokens_delete(token_id)
}
\arguments{
\item{token_id}{Required. The ID of the token to be revoked.}
}
\description{
Revokes an access token.
}
\details{
If a token with the specified ID is not valid, this call returns an error
\strong{RESOURCE_DOES_NOT_EXIST}.
}
\keyword{internal}
