% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{transition_model_stage}
\alias{transition_model_stage}
\alias{modelRegistryTransitionStage}
\title{Transition a stage.}
\usage{
transition_model_stage(
  client,
  name,
  version,
  stage,
  archive_existing_versions,
  comment = NULL
)

modelRegistryTransitionStage(
  client,
  name,
  version,
  stage,
  archive_existing_versions,
  comment = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Name of the model.}

\item{version}{Required. Version of the model.

See the \href{https://docs.databricks.com/api/workspace/modelregistry/transitionstage}{API documentation} for more
information on the request and response field structure.}

\item{stage}{Required. Target stage of the transition.}

\item{archive_existing_versions}{Required. Specifies whether to archive all current model versions in the target stage.}

\item{comment}{User-provided comment on the action.}
}
\description{
Transition a model version's stage. This is a Databricks workspace version of
the MLflow endpoint that also accepts a comment associated with the
transition to be recorded.',
}
\details{
MLflow endpoint: https://www.mlflow.org/docs/latest/rest-api.html#transition-modelversion-stage
}
