% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_requests.R
\name{transition_requests_approve}
\alias{transition_requests_approve}
\title{Approve transition requests.}
\usage{
transition_requests_approve(
  name,
  version,
  stage,
  archive_existing_versions,
  comment = NULL,
  ...
)
}
\arguments{
\item{name}{\link{required} Name of the model.}

\item{version}{\link{required} Version of the model.}

\item{stage}{\link{required} Target stage of the transition.}

\item{archive_existing_versions}{\link{required} Specifies whether to archive all current model versions in the target stage.}

\item{comment}{User-provided comment on the action.}
}
\description{
Approves a model version stage transition request.
}
\keyword{internal}
