% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_requests.R
\name{transition_requests_create}
\alias{transition_requests_create}
\title{Make a transition request.}
\usage{
transition_requests_create(name, version, stage, comment = NULL)
}
\arguments{
\item{name}{Required. Name of the model.}

\item{version}{Required. Version of the model.}

\item{stage}{Required. Target stage of the transition.}

\item{comment}{User-provided comment on the action.}
}
\description{
Creates a model version stage transition request.
}
\keyword{internal}
