% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_requests.R
\name{transition_requests_delete}
\alias{transition_requests_delete}
\title{Delete a ransition request.}
\usage{
transition_requests_delete(name, version, stage, creator, comment = NULL, ...)
}
\arguments{
\item{name}{\link{required} Name of the model.}

\item{version}{\link{required} Version of the model.}

\item{stage}{\link{required} Target stage of the transition request.}

\item{creator}{\link{required} Username of the user who created this request.}

\item{comment}{User-provided comment on the action.}
}
\description{
Cancels a model version stage transition request.
}
\keyword{internal}
