% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_requests.R
\name{transition_requests_list}
\alias{transition_requests_list}
\title{List transition requests.}
\usage{
transition_requests_list(name, version, ...)
}
\arguments{
\item{name}{\link{required} Name of the model.}

\item{version}{\link{required} Version of the model.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Gets a list of all open stage transition requests for the model version.
}
\keyword{internal}
