% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metastores.R
\name{unassign_metastore}
\alias{unassign_metastore}
\title{Delete an assignment.}
\usage{
unassign_metastore(client, workspace_id, metastore_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{workspace_id}{Required. A workspace ID.}

\item{metastore_id}{Required. Query for the ID of the metastore to delete.}
}
\description{
Deletes a metastore assignment. The caller must be an account administrator.
}
