% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraries.R
\name{uninstall_cluster_library}
\alias{uninstall_cluster_library}
\alias{librariesUninstall}
\title{Uninstall libraries.}
\usage{
uninstall_cluster_library(client, cluster_id, libraries)

librariesUninstall(client, cluster_id, libraries)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{cluster_id}{Required. Unique identifier for the cluster on which to uninstall these libraries.}

\item{libraries}{Required. The libraries to uninstall.}
}
\description{
Set libraries to be uninstalled on a cluster. The libraries won't be
uninstalled until the cluster is restarted. Uninstalling libraries that are
not installed on the cluster will have no impact but is not an error.
}
