% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_access_control_proxy.R
\name{update_account_access_control_proxy_rule_set}
\alias{update_account_access_control_proxy_rule_set}
\alias{accountAccessControlProxyUpdateRuleSet}
\title{Update a rule set.}
\usage{
update_account_access_control_proxy_rule_set(client, name, rule_set)

accountAccessControlProxyUpdateRuleSet(client, name, rule_set)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Name of the rule set.}

\item{rule_set}{Required. This field has no description yet.

See the \href{https://docs.databricks.com/api/workspace/accountaccesscontrolproxy/updateruleset}{API documentation} for more
information on the request and response field structure.}
}
\description{
Replace the rules of a rule set. First, use a GET rule set request to read
the current version of the rule set before modifying it. This pattern helps
prevent conflicts between concurrent updates.
}
