% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{update_alert}
\alias{update_alert}
\alias{alertsUpdate}
\title{Update an alert.}
\usage{
update_alert(client, alert_id, name, options, query_id, rearm = NULL)

alertsUpdate(client, alert_id, name, options, query_id, rearm = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{alert_id}{Required. This field has no description yet.}

\item{name}{Required. Name of the alert.}

\item{options}{Required. Alert configuration options.}

\item{query_id}{Required. Query ID.}

\item{rearm}{Number of seconds after being triggered before the alert rearms itself and can be triggered again.

See the \href{https://docs.databricks.com/api/workspace/alerts/update}{API documentation} for more
information on the request and response field structure.}
}
\description{
Updates an alert.
}
