% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogs.R
\name{update_catalog}
\alias{update_catalog}
\title{Update a catalog.}
\usage{
update_catalog(
  client,
  name,
  comment = NULL,
  enable_predictive_optimization = NULL,
  isolation_mode = NULL,
  new_name = NULL,
  owner = NULL,
  properties = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. The name of the catalog.}

\item{comment}{User-provided free-form text description.}

\item{enable_predictive_optimization}{Whether predictive optimization should be enabled for this object and objects under it.}

\item{isolation_mode}{Whether the current securable is accessible from all workspaces or a specific set of workspaces.}

\item{new_name}{New name for the catalog.}

\item{owner}{Username of current owner of catalog.}

\item{properties}{A map of key-value properties attached to the securable.}
}
\description{
Updates the catalog that matches the supplied name. The caller must be either
the owner of the catalog, or a metastore admin (when changing the owner field
of the catalog).
}
