% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_rooms.R
\name{update_clean_room}
\alias{update_clean_room}
\alias{cleanRoomsUpdate}
\title{Update a clean room.}
\usage{
update_clean_room(
  client,
  name,
  catalog_updates = NULL,
  comment = NULL,
  owner = NULL
)

cleanRoomsUpdate(
  client,
  name,
  catalog_updates = NULL,
  comment = NULL,
  owner = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. The name of the clean room.}

\item{catalog_updates}{Array of shared data object updates.}

\item{comment}{User-provided free-form text description.}

\item{owner}{Username of current owner of clean room.}
}
\description{
Updates the clean room with the changes and data objects in the request. The
caller must be the owner of the clean room or a metastore admin.
}
\details{
When the caller is a metastore admin, only the \strong{owner} field can be
updated.

In the case that the clean room name is changed \strong{updateCleanRoom} requires
that the caller is both the clean room owner and a metastore admin.

For each table that is added through this method, the clean room owner must
also have \strong{SELECT} privilege on the table. The privilege must be maintained
indefinitely for recipients to be able to access the table. Typically, you
should use a group as the clean room owner.

Table removals through \strong{update} do not require additional privileges.
}
