% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{update_cluster_permissions}
\alias{update_cluster_permissions}
\alias{clustersUpdatePermissions}
\title{Update cluster permissions.}
\usage{
update_cluster_permissions(client, cluster_id, access_control_list = NULL)

clustersUpdatePermissions(client, cluster_id, access_control_list = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{cluster_id}{Required. The cluster for which to get or manage permissions.}

\item{access_control_list}{This field has no description yet.}
}
\description{
Updates the permissions on a cluster. Clusters can inherit permissions from
their root object.
}
