% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{update_connection}
\alias{update_connection}
\title{Update a connection.}
\usage{
update_connection(client, name, options, new_name = NULL, owner = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Name of the connection.}

\item{options}{Required. A map of key-value properties attached to the securable.}

\item{new_name}{New name for the connection.}

\item{owner}{Username of current owner of the connection.}
}
\description{
Updates the connection that matches the supplied name.
}
