% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboards.R
\name{update_dashboard}
\alias{update_dashboard}
\title{Change a dashboard definition.}
\usage{
update_dashboard(client, dashboard_id, name = NULL, run_as_role = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{dashboard_id}{Required. This field has no description yet.}

\item{name}{The title of this dashboard that appears in list views and at the top of the dashboard page.}

\item{run_as_role}{Sets the \strong{Run as} role for the object.}
}
\description{
Modify this dashboard definition. This operation only affects attributes of
the dashboard object. It does not add, modify, or remove widgets.
}
\details{
\strong{Note}: You cannot undo this operation.
}
