% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_namespace.R
\name{update_default_namespace}
\alias{update_default_namespace}
\alias{defaultNamespaceUpdate}
\title{Update the default namespace setting.}
\usage{
update_default_namespace(client, allow_missing, setting, field_mask)

defaultNamespaceUpdate(client, allow_missing, setting, field_mask)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{allow_missing}{Required. This should always be set to true for Settings API.}

\item{setting}{Required. This represents the setting configuration for the default namespace in the Databricks workspace.

See the \href{https://docs.databricks.com/api/workspace/defaultnamespace/update}{API documentation} for more
information on the request and response field structure.}

\item{field_mask}{Required. Field mask is required to be passed into the PATCH request.}
}
\description{
Updates the default namespace setting for the workspace. A fresh etag needs
to be provided in \code{PATCH} requests (as part of the setting field). The etag
can be retrieved by making a \code{GET} request before the \code{PATCH} request. Note
that if the setting does not exist, \code{GET} returns a NOT_FOUND error and the
etag is present in the error response, which should be set in the \code{PATCH}
request. If the setting is updated concurrently, \code{PATCH} fails with 409 and
the request must be retried by using the fresh etag in the 409 response.
}
