% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esm_enablement.R
\name{update_esm_enablement}
\alias{update_esm_enablement}
\alias{esmEnablementUpdate}
\title{Update the enhanced security monitoring setting.}
\usage{
update_esm_enablement(client, allow_missing, setting, field_mask)

esmEnablementUpdate(client, allow_missing, setting, field_mask)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{allow_missing}{Required. This should always be set to true for Settings API.}

\item{setting}{Required. This field has no description yet.

See the \href{https://docs.databricks.com/api/workspace/esmenablement/update}{API documentation} for more
information on the request and response field structure.}

\item{field_mask}{Required. Field mask is required to be passed into the PATCH request.}
}
\description{
Updates the enhanced security monitoring setting for the workspace. A fresh
etag needs to be provided in \code{PATCH} requests (as part of the setting field).
The etag can be retrieved by making a \code{GET} request before the \code{PATCH}
request. If the setting is updated concurrently, \code{PATCH} fails with 409 and
the request must be retried by using the fresh etag in the 409 response.
}
