% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{update_experiment_permissions}
\alias{update_experiment_permissions}
\alias{experimentsUpdatePermissions}
\title{Update experiment permissions.}
\usage{
update_experiment_permissions(
  client,
  experiment_id,
  access_control_list = NULL
)

experimentsUpdatePermissions(client, experiment_id, access_control_list = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{experiment_id}{Required. The experiment for which to get or manage permissions.

See the \href{https://docs.databricks.com/api/workspace/experiments/updatepermissions}{API documentation} for more
information on the request and response field structure.}

\item{access_control_list}{This field has no description yet.}
}
\description{
Updates the permissions on an experiment. Experiments can inherit permissions
from their root object.
}
