% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{update_experiment_run}
\alias{update_experiment_run}
\title{Update a run.}
\usage{
update_experiment_run(
  client,
  end_time = NULL,
  run_id = NULL,
  run_uuid = NULL,
  status = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{end_time}{Unix timestamp in milliseconds of when the run ended.}

\item{run_id}{ID of the run to update.}

\item{run_uuid}{Deprecated, use run_id instead. ID of the run to update.}

\item{status}{Updated status of the run.}
}
\description{
Updates run metadata.
}
