% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{update_function}
\alias{update_function}
\alias{functionsUpdate}
\title{Update a function.}
\usage{
update_function(client, name, owner = NULL)

functionsUpdate(client, name, owner = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. The fully-qualified name of the function (of the form \strong{catalog_name}.\strong{schema_name}.\strong{function__name}).}

\item{owner}{Username of current owner of function.}
}
\description{
Updates the function that matches the supplied name. Only the owner of the
function can be updated. If the user is not a metastore admin, the user must
be a member of the group that is the new function owner. - Is a metastore
admin - Is the owner of the function's parent catalog - Is the owner of the
function's parent schema and has the \strong{USE_CATALOG} privilege on its parent
catalog - Is the owner of the function itself and has the \strong{USE_CATALOG}
privilege on its parent catalog as well as the \strong{USE_SCHEMA} privilege on
the function's parent schema.
}
