% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_credentials.R
\name{update_git_credential}
\alias{update_git_credential}
\alias{gitCredentialsUpdate}
\title{Update a credential.}
\usage{
update_git_credential(
  client,
  credential_id,
  git_provider = NULL,
  git_username = NULL,
  personal_access_token = NULL
)

gitCredentialsUpdate(
  client,
  credential_id,
  git_provider = NULL,
  git_username = NULL,
  personal_access_token = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{credential_id}{Required. The ID for the corresponding credential to access.}

\item{git_provider}{Git provider.}

\item{git_username}{Git username.}

\item{personal_access_token}{The personal access token used to authenticate to the corresponding Git provider.

See the \href{https://docs.databricks.com/api/workspace/gitcredentials/update}{API documentation} for more
information on the request and response field structure.}
}
\description{
Updates the specified Git credential.
}
