% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_init_scripts.R
\name{update_global_init_script}
\alias{update_global_init_script}
\alias{globalInitScriptsUpdate}
\title{Update init script.}
\usage{
update_global_init_script(
  client,
  script_id,
  name,
  script,
  enabled = NULL,
  position = NULL
)

globalInitScriptsUpdate(
  client,
  script_id,
  name,
  script,
  enabled = NULL,
  position = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{script_id}{Required. The ID of the global init script.}

\item{name}{Required. The name of the script.}

\item{script}{Required. The Base64-encoded content of the script.}

\item{enabled}{Specifies whether the script is enabled.}

\item{position}{The position of a script, where 0 represents the first script to run, 1 is the second script to run, in ascending order.}
}
\description{
Updates a global init script, specifying only the fields to change. All
fields are optional. Unspecified fields retain their current value.
}
