% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance_pools.R
\name{update_instance_pool_permissions}
\alias{update_instance_pool_permissions}
\alias{instancePoolsUpdatePermissions}
\title{Update instance pool permissions.}
\usage{
update_instance_pool_permissions(
  client,
  instance_pool_id,
  access_control_list = NULL
)

instancePoolsUpdatePermissions(
  client,
  instance_pool_id,
  access_control_list = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{instance_pool_id}{Required. The instance pool for which to get or manage permissions.

See the \href{https://docs.databricks.com/api/workspace/instancepools/updatepermissions}{API documentation} for more
information on the request and response field structure.}

\item{access_control_list}{This field has no description yet.}
}
\description{
Updates the permissions on an instance pool. Instance pools can inherit
permissions from their root object.
}
