% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{update_job}
\alias{update_job}
\title{Update job settings partially.}
\usage{
update_job(client, job_id, fields_to_remove = NULL, new_settings = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{job_id}{Required. The canonical identifier of the job to update.}

\item{fields_to_remove}{Remove top-level fields in the job settings.}

\item{new_settings}{The new settings for the job.}
}
\description{
Add, update, or remove specific settings of an existing job. Use the \href{:method:jobs/reset}{\emph{Reset} endpoint} to overwrite all job settings.
}
