% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{update_job_permissions}
\alias{update_job_permissions}
\alias{jobsUpdatePermissions}
\title{Update job permissions.}
\usage{
update_job_permissions(client, job_id, access_control_list = NULL)

jobsUpdatePermissions(client, job_id, access_control_list = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{job_id}{Required. The job for which to get or manage permissions.}

\item{access_control_list}{This field has no description yet.}
}
\description{
Updates the permissions on a job. Jobs can inherit permissions from their
root object.
}
