% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeview.R
\name{update_lakeview}
\alias{update_lakeview}
\alias{lakeviewUpdate}
\title{Update dashboard.}
\usage{
update_lakeview(
  client,
  dashboard_id,
  display_name = NULL,
  etag = NULL,
  serialized_dashboard = NULL,
  warehouse_id = NULL
)

lakeviewUpdate(
  client,
  dashboard_id,
  display_name = NULL,
  etag = NULL,
  serialized_dashboard = NULL,
  warehouse_id = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{dashboard_id}{Required. UUID identifying the dashboard.}

\item{display_name}{The display name of the dashboard.}

\item{etag}{The etag for the dashboard.}

\item{serialized_dashboard}{The contents of the dashboard in serialized string form.}

\item{warehouse_id}{The warehouse ID used to run the dashboard.}
}
\description{
Update a draft dashboard.
}
