% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metastores.R
\name{update_metastore}
\alias{update_metastore}
\title{Update a metastore.}
\usage{
update_metastore(
  client,
  id,
  delta_sharing_organization_name = NULL,
  delta_sharing_recipient_token_lifetime_in_seconds = NULL,
  delta_sharing_scope = NULL,
  new_name = NULL,
  owner = NULL,
  privilege_model_version = NULL,
  storage_root_credential_id = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{id}{Required. Unique ID of the metastore.}

\item{delta_sharing_organization_name}{The organization name of a Delta Sharing entity, to be used in Databricks-to-Databricks Delta Sharing as the official name.}

\item{delta_sharing_recipient_token_lifetime_in_seconds}{The lifetime of delta sharing recipient token in seconds.}

\item{delta_sharing_scope}{The scope of Delta Sharing enabled for the metastore.}

\item{new_name}{New name for the metastore.}

\item{owner}{The owner of the metastore.}

\item{privilege_model_version}{Privilege model version of the metastore, of the form \code{major.minor} (e.g., \code{1.0}).}

\item{storage_root_credential_id}{UUID of storage credential to access the metastore storage_root.}
}
\description{
Updates information for a specific metastore. The caller must be a metastore
admin. If the \strong{owner} field is set to the empty string (\strong{''}), the
ownership is updated to the System User.
}
