% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metastores.R
\name{update_metastore_assignment}
\alias{update_metastore_assignment}
\alias{metastoresUpdateAssignment}
\title{Update an assignment.}
\usage{
update_metastore_assignment(
  client,
  workspace_id,
  default_catalog_name = NULL,
  metastore_id = NULL
)

metastoresUpdateAssignment(
  client,
  workspace_id,
  default_catalog_name = NULL,
  metastore_id = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{workspace_id}{Required. A workspace ID.}

\item{default_catalog_name}{The name of the default catalog for the metastore.}

\item{metastore_id}{The unique ID of the metastore.}
}
\description{
Updates a metastore assignment. This operation can be used to update
\strong{metastore_id} or \strong{default_catalog_name} for a specified Workspace, if
the Workspace is already assigned a metastore. The caller must be an account
admin to update \strong{metastore_id}; otherwise, the caller can be a Workspace
admin.
}
