% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{update_model}
\alias{update_model}
\alias{modelRegistryUpdateModel}
\title{Update model.}
\usage{
update_model(client, name, description = NULL)

modelRegistryUpdateModel(client, name, description = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Registered model unique name identifier.

See the \href{https://docs.databricks.com/api/workspace/modelregistry/updatemodel}{API documentation} for more
information on the request and response field structure.}

\item{description}{If provided, updates the description for this \code{registered_model}.}
}
\description{
Updates a registered model.
}
