% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{update_model_comment}
\alias{update_model_comment}
\alias{modelRegistryUpdateComment}
\title{Update a comment.}
\usage{
update_model_comment(client, id, comment)

modelRegistryUpdateComment(client, id, comment)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{id}{Required. Unique identifier of an activity.

See the \href{https://docs.databricks.com/api/workspace/modelregistry/updatecomment}{API documentation} for more
information on the request and response field structure.}

\item{comment}{Required. User-provided comment on the action.}
}
\description{
Post an edit to a comment on a model version.
}
