% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R, R/model_versions.R
\name{update_model_version}
\alias{update_model_version}
\alias{modelRegistryUpdateModelVersion}
\alias{modelVersionsUpdate}
\title{Update model version.}
\usage{
update_model_version(client, full_name, version, comment = NULL)

modelRegistryUpdateModelVersion(client, name, version, description = NULL)

update_model_version(client, full_name, version, comment = NULL)

modelVersionsUpdate(client, full_name, version, comment = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{full_name}{Required. The three-level (fully qualified) name of the model version.}

\item{version}{Required. The integer version number of the model version.}

\item{comment}{The comment attached to the model version.}

\item{name}{Required. Name of the registered model.}

\item{description}{If provided, updates the description for this \code{registered_model}.}
}
\description{
Updates the model version.

Updates the specified model version.
}
\details{
The caller must be a metastore admin or an owner of the parent registered
model. For the latter case, the caller must also be the owner or have the
\strong{USE_CATALOG} privilege on the parent catalog and the \strong{USE_SCHEMA}
privilege on the parent schema.

Currently only the comment of the model version can be updated.
}
