% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{update_model_webhook}
\alias{update_model_webhook}
\alias{modelRegistryUpdateWebhook}
\title{Update a webhook.}
\usage{
update_model_webhook(
  client,
  id,
  description = NULL,
  events = NULL,
  http_url_spec = NULL,
  job_spec = NULL,
  status = NULL
)

modelRegistryUpdateWebhook(
  client,
  id,
  description = NULL,
  events = NULL,
  http_url_spec = NULL,
  job_spec = NULL,
  status = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{id}{Required. Webhook ID.}

\item{description}{User-specified description for the webhook.}

\item{events}{Events that can trigger a registry webhook: * \code{MODEL_VERSION_CREATED}: A new model version was created for the associated model.}

\item{http_url_spec}{This field has no description yet.}

\item{job_spec}{This field has no description yet.}

\item{status}{Enable or disable triggering the webhook, or put the webhook into test mode.}
}
\description{
\strong{NOTE:} This endpoint is in Public Preview.
}
\details{
Updates a registry webhook.
}
