% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{update_pipeline_permissions}
\alias{update_pipeline_permissions}
\alias{pipelinesUpdatePermissions}
\title{Update pipeline permissions.}
\usage{
update_pipeline_permissions(client, pipeline_id, access_control_list = NULL)

pipelinesUpdatePermissions(client, pipeline_id, access_control_list = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{pipeline_id}{Required. The pipeline for which to get or manage permissions.

See the \href{https://docs.databricks.com/api/workspace/pipelines/updatepermissions}{API documentation} for more
information on the request and response field structure.}

\item{access_control_list}{This field has no description yet.}
}
\description{
Updates the permissions on a pipeline. Pipelines can inherit permissions from
their root object.
}
