% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{update_query}
\alias{update_query}
\alias{queriesUpdate}
\title{Change a query definition.}
\usage{
update_query(
  client,
  query_id,
  data_source_id = NULL,
  description = NULL,
  name = NULL,
  options = NULL,
  query = NULL,
  run_as_role = NULL
)

queriesUpdate(
  client,
  query_id,
  data_source_id = NULL,
  description = NULL,
  name = NULL,
  options = NULL,
  query = NULL,
  run_as_role = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{query_id}{Required. This field has no description yet.}

\item{data_source_id}{Data source ID maps to the ID of the data source used by the resource and is distinct from the warehouse ID.}

\item{description}{General description that conveys additional information about this query such as usage notes.}

\item{name}{The title of this query that appears in list views, widget headings, and on the query page.}

\item{options}{Exclusively used for storing a list parameter definitions.}

\item{query}{The text of the query to be run.}

\item{run_as_role}{Sets the \strong{Run as} role for the object.}
}
\description{
Modify this query definition.
}
\details{
\strong{Note}: You cannot undo this operation.
}
