% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipients.R
\name{update_recipient}
\alias{update_recipient}
\alias{recipientsUpdate}
\title{Update a share recipient.}
\usage{
update_recipient(
  client,
  name,
  comment = NULL,
  ip_access_list = NULL,
  new_name = NULL,
  owner = NULL,
  properties_kvpairs = NULL
)

recipientsUpdate(
  client,
  name,
  comment = NULL,
  ip_access_list = NULL,
  new_name = NULL,
  owner = NULL,
  properties_kvpairs = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Name of the recipient.}

\item{comment}{Description about the recipient.}

\item{ip_access_list}{IP Access List.}

\item{new_name}{New name for the recipient.}

\item{owner}{Username of the recipient owner.}

\item{properties_kvpairs}{Recipient properties as map of string key-value pairs.}
}
\description{
Updates an existing recipient in the metastore. The caller must be a
metastore admin or the owner of the recipient. If the recipient name will be
updated, the user must be both a metastore admin and the owner of the
recipient.
}
