% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serving_endpoints.R
\name{update_serving_endpoint_config}
\alias{update_serving_endpoint_config}
\alias{servingEndpointsUpdateConfig}
\title{Update config of a serving endpoint.}
\usage{
update_serving_endpoint_config(
  client,
  name,
  auto_capture_config = NULL,
  served_entities = NULL,
  served_models = NULL,
  traffic_config = NULL
)

servingEndpointsUpdateConfig(
  client,
  name,
  auto_capture_config = NULL,
  served_entities = NULL,
  served_models = NULL,
  traffic_config = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. The name of the serving endpoint to update.}

\item{auto_capture_config}{Configuration for Inference Tables which automatically logs requests and responses to Unity Catalog.}

\item{served_entities}{A list of served entities for the endpoint to serve.}

\item{served_models}{(Deprecated, use served_entities instead) A list of served models for the endpoint to serve.}

\item{traffic_config}{The traffic config defining how invocations to the serving endpoint should be routed.

See the \href{https://docs.databricks.com/api/workspace/servingendpoints/updateconfig}{API documentation} for more
information on the request and response field structure.}
}
\description{
Updates any combination of the serving endpoint's served entities, the
compute configuration of those served entities, and the endpoint's traffic
config. An endpoint that already has an update in progress can not be updated
until the current update completes or fails.
}
