% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_credentials.R
\name{update_storage_credential}
\alias{update_storage_credential}
\alias{storageCredentialsUpdate}
\title{Update a credential.}
\usage{
update_storage_credential(
  client,
  name,
  aws_iam_role = NULL,
  azure_managed_identity = NULL,
  azure_service_principal = NULL,
  cloudflare_api_token = NULL,
  comment = NULL,
  databricks_gcp_service_account = NULL,
  force = NULL,
  new_name = NULL,
  owner = NULL,
  read_only = NULL,
  skip_validation = NULL
)

storageCredentialsUpdate(
  client,
  name,
  aws_iam_role = NULL,
  azure_managed_identity = NULL,
  azure_service_principal = NULL,
  cloudflare_api_token = NULL,
  comment = NULL,
  databricks_gcp_service_account = NULL,
  force = NULL,
  new_name = NULL,
  owner = NULL,
  read_only = NULL,
  skip_validation = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Name of the storage credential.}

\item{aws_iam_role}{The AWS IAM role configuration.}

\item{azure_managed_identity}{The Azure managed identity configuration.}

\item{azure_service_principal}{The Azure service principal configuration.}

\item{cloudflare_api_token}{The Cloudflare API token configuration.}

\item{comment}{Comment associated with the credential.}

\item{databricks_gcp_service_account}{The \if{html}{\out{<Databricks>}} managed GCP service account configuration.}

\item{force}{Force update even if there are dependent external locations or external tables.}

\item{new_name}{New name for the storage credential.}

\item{owner}{Username of current owner of credential.}

\item{read_only}{Whether the storage credential is only usable for read operations.}

\item{skip_validation}{Supplying true to this argument skips validation of the updated credential.}
}
\description{
Updates a storage credential on the metastore.
}
