% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_management.R
\name{update_token_management_permissions}
\alias{update_token_management_permissions}
\title{Update token permissions.}
\usage{
update_token_management_permissions(client, access_control_list = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{access_control_list}{This field has no description yet.}
}
\description{
Updates the permissions on all tokens. Tokens can inherit permissions from
their root object.
}
