% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{update_user_permissions}
\alias{update_user_permissions}
\alias{usersUpdatePermissions}
\title{Update password permissions.}
\usage{
update_user_permissions(client, access_control_list = NULL)

usersUpdatePermissions(client, access_control_list = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{access_control_list}{This field has no description yet.}
}
\description{
Updates the permissions on all passwords. Passwords can inherit permissions
from their root object.
}
