% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{upload_file}
\alias{upload_file}
\alias{filesUpload}
\title{Upload a file.}
\usage{
upload_file(client, file_path, contents, overwrite = NULL)

filesUpload(client, file_path, contents, overwrite = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{file_path}{Required. The absolute path of the file.}

\item{contents}{This field has no description yet.}

\item{overwrite}{If true, an existing file will be overwritten.

See the \href{https://docs.databricks.com/api/workspace/files/upload}{API documentation} for more
information on the request and response field structure.}
}
\description{
Uploads a file of up to 5 GiB. The file contents should be sent as the
request body as raw bytes (an octet stream); do not encode or otherwise
modify the bytes before sending. The contents of the resulting file will be
exactly the bytes sent in the request body. If the request is successful,
there is no response body.
}
