% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_search_indexes.R
\name{upsert_vector_search_index_data}
\alias{upsert_vector_search_index_data}
\alias{vectorSearchIndexesUpsertDataVectorIndex}
\title{Upsert data into an index.}
\usage{
upsert_vector_search_index_data(client, index_name, inputs_json)

vectorSearchIndexesUpsertDataVectorIndex(client, index_name, inputs_json)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{index_name}{Required. Name of the vector index where data is to be upserted.}

\item{inputs_json}{Required. JSON string representing the data to be upserted.}
}
\description{
Handles the upserting of data into a specified vector index.
}
