% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\docType{data}
\name{users}
\alias{users}
\title{User identities recognized by Databricks and represented by email addresses.}
\format{
An object of class \code{list} of length 6.
}
\usage{
users
}
\description{
Databricks recommends using SCIM provisioning to sync users and groups
automatically from your identity provider to your Databricks workspace. SCIM
streamlines onboarding a new employee or team by using your identity provider
to create users and groups in Databricks workspace and give them the proper
level of access. When a user leaves your organization or no longer needs
access to Databricks workspace, admins can terminate the user in your
identity provider and that user’s account will also be removed from
Databricks workspace. This ensures a consistent offboarding process and
prevents unauthorized users from accessing sensitive data.
}
\section{Operations}{

\tabular{ll}{
\link[=users_create]{create} \tab Create a new user.\cr
\link[=users_delete]{delete} \tab Delete a user.\cr
\link[=users_get]{get} \tab Get user details.\cr
\link[=users_list]{list} \tab List users.\cr
\link[=users_patch]{patch} \tab Update user details.\cr
\link[=users_update]{update} \tab Replace a user.\cr
}
}

\keyword{datasets}
