% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{users_delete}
\alias{users_delete}
\title{Delete a user.}
\usage{
users_delete(id)
}
\arguments{
\item{id}{Required. Unique ID for a user in the Databricks Workspace.}
}
\description{
Deletes a user. Deleting a user from a Databricks Workspace also removes
objects associated with the user.
}
\keyword{internal}
