% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{users_get}
\alias{users_get}
\title{Get user details.}
\usage{
users_get(id)
}
\arguments{
\item{id}{Required. Unique ID for a user in the Databricks workspace.}
}
\description{
Gets information for a specific user in Databricks workspace.
}
\keyword{internal}
