% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{users_patch}
\alias{users_patch}
\title{Update user details.}
\usage{
users_patch(id, operations = NULL)
}
\arguments{
\item{id}{Required. Unique ID for a group in the Databricks account.}

\item{operations}{}
}
\description{
Partially updates a user resource by applying the supplied operations on
specific user attributes.
}
\keyword{internal}
