% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{users_update}
\alias{users_update}
\title{Replace a user.}
\usage{
users_update(
  id,
  active = NULL,
  display_name = NULL,
  emails = NULL,
  entitlements = NULL,
  external_id = NULL,
  groups = NULL,
  name = NULL,
  roles = NULL,
  user_name = NULL,
  ...
)
}
\arguments{
\item{id}{Databricks user ID.}

\item{active}{If this user is active.}

\item{display_name}{String that represents a concatenation of given and family names.}

\item{emails}{All the emails associated with the Databricks user.}

\item{entitlements}{}

\item{external_id}{}

\item{groups}{}

\item{name}{}

\item{roles}{}

\item{user_name}{Email address of the Databricks user.}
}
\description{
Replaces a user's information with the data supplied in request.
}
\keyword{internal}
