% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{usersCreate}
\alias{usersCreate}
\title{Create a new user.}
\usage{
usersCreate(
  client,
  active = NULL,
  display_name = NULL,
  emails = NULL,
  entitlements = NULL,
  external_id = NULL,
  groups = NULL,
  id = NULL,
  name = NULL,
  roles = NULL,
  user_name = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{active}{If this user is active.}

\item{display_name}{String that represents a concatenation of given and family names.}

\item{emails}{All the emails associated with the Databricks user.}

\item{entitlements}{}

\item{external_id}{}

\item{groups}{}

\item{id}{Databricks user ID.}

\item{name}{}

\item{roles}{}

\item{user_name}{Email address of the Databricks user.}
}
\description{
Creates a new user in the Databricks workspace. This new user will also be
added to the Databricks account.
}
