% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{usersDelete}
\alias{usersDelete}
\title{Delete a user.}
\usage{
usersDelete(client, id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{id}{Required. Unique ID for a user in the Databricks workspace.}
}
\description{
Deletes a user. Deleting a user from a Databricks workspace also removes
objects associated with the user.
}
