% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{usersGet}
\alias{usersGet}
\title{Get user details.}
\usage{
usersGet(client, id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{id}{Required. Unique ID for a user in the Databricks workspace.}
}
\description{
Gets information for a specific user in Databricks workspace.
}
