% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{usersGetPasswordPermissionLevels}
\alias{usersGetPasswordPermissionLevels}
\title{Get password permission levels.}
\usage{
usersGetPasswordPermissionLevels(client)
}
\description{
Gets the permission levels that a user can have on an object.#'
}
