% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{usersGetPasswordPermissions}
\alias{usersGetPasswordPermissions}
\title{Get password permissions.}
\usage{
usersGetPasswordPermissions(client)
}
\description{
Gets the permissions of all passwords. Passwords can inherit permissions from
their root object.#'
}
