% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{usersGetPermissionLevels}
\alias{usersGetPermissionLevels}
\title{Get password permission levels.}
\usage{
usersGetPermissionLevels(client)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}
}
\description{
Gets the permission levels that a user can have on an object.
}
