% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{usersGetPermissions}
\alias{usersGetPermissions}
\title{Get password permissions.}
\usage{
usersGetPermissions(client)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}
}
\description{
Gets the permissions of all passwords. Passwords can inherit permissions from
their root object.
}
