% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{usersPatch}
\alias{usersPatch}
\title{Update user details.}
\usage{
usersPatch(client, id, operations = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{id}{Required. Unique ID for a user in the Databricks workspace.}

\item{operations}{}
}
\description{
Partially updates a user resource by applying the supplied operations on
specific user attributes.
}
