% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{usersSetPasswordPermissions}
\alias{usersSetPasswordPermissions}
\title{Set password permissions.}
\usage{
usersSetPasswordPermissions(client, access_control_list = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{access_control_list}{}
}
\description{
Sets permissions on all passwords. Passwords can inherit permissions from
their root object.
}
