% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{usersUpdatePasswordPermissions}
\alias{usersUpdatePasswordPermissions}
\title{Update password permissions.}
\usage{
usersUpdatePasswordPermissions(client, access_control_list = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{access_control_list}{}
}
\description{
Updates the permissions on all passwords. Passwords can inherit permissions
from their root object.
}
