% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_search_endpoints.R
\name{vectorSearchEndpointsDeleteEndpoint}
\alias{vectorSearchEndpointsDeleteEndpoint}
\title{Delete an endpoint.}
\usage{
vectorSearchEndpointsDeleteEndpoint(client, endpoint_name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{endpoint_name}{Required. Name of the endpoint.}
}
\description{
Delete an endpoint.
}
