% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_search_endpoints.R
\name{vectorSearchEndpointsListEndpoints}
\alias{vectorSearchEndpointsListEndpoints}
\title{List all endpoints.}
\usage{
vectorSearchEndpointsListEndpoints(client, page_token = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{page_token}{Token for pagination.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
List all endpoints.
}
