% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_search_indexes.R
\name{vectorSearchIndexesCreateIndex}
\alias{vectorSearchIndexesCreateIndex}
\title{Create an index.}
\usage{
vectorSearchIndexesCreateIndex(
  client,
  name,
  endpoint_name,
  primary_key,
  index_type,
  delta_sync_index_spec = NULL,
  direct_access_index_spec = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Name of the index.}

\item{endpoint_name}{Required. Name of the endpoint to be used for serving the index.}

\item{primary_key}{Required. Primary key of the index.}

\item{index_type}{Required. There are 2 types of Vector Search indexes: - \code{DELTA_SYNC}: An index that automatically syncs with a source Delta Table, automatically and incrementally updating the index as the underlying data in the Delta Table changes.}

\item{delta_sync_index_spec}{Specification for Delta Sync Index.}

\item{direct_access_index_spec}{Specification for Direct Vector Access Index.}
}
\description{
Create a new index.
}
