% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_search_indexes.R
\name{vectorSearchIndexesDeleteIndex}
\alias{vectorSearchIndexesDeleteIndex}
\title{Delete an index.}
\usage{
vectorSearchIndexesDeleteIndex(client, index_name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{index_name}{Required. Name of the index.}
}
\description{
Delete an index.
}
