% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_search_indexes.R
\name{vectorSearchIndexesGetIndex}
\alias{vectorSearchIndexesGetIndex}
\title{Get an index.}
\usage{
vectorSearchIndexesGetIndex(client, index_name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{index_name}{Required. Name of the index.}
}
\description{
Get an index.
}
