% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_search_indexes.R
\name{vectorSearchIndexesSyncIndex}
\alias{vectorSearchIndexesSyncIndex}
\title{Synchronize an index.}
\usage{
vectorSearchIndexesSyncIndex(client, index_name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{index_name}{Required. Name of the vector index to synchronize.}
}
\description{
Triggers a synchronization process for a specified vector index.
}
