% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volumes.R
\docType{data}
\name{volumes}
\alias{volumes}
\title{Volumes are a Unity Catalog (UC) capability for accessing, storing,
governing, organizing and processing files. Use cases include running machine
learning on unstructured data such as image, audio, video, or PDF files,
organizing data sets during the data exploration stages in data science,
working with libraries that require access to the local file system on
cluster machines, storing library and config files of arbitrary formats such
as .whl or .txt centrally and providing secure access across workspaces to
it, or transforming and querying non-tabular data files in ETL.}
\format{
An object of class \code{list} of length 5.
}
\usage{
volumes
}
\description{
Volumes are a Unity Catalog (UC) capability for accessing, storing,
governing, organizing and processing files. Use cases include running machine
learning on unstructured data such as image, audio, video, or PDF files,
organizing data sets during the data exploration stages in data science,
working with libraries that require access to the local file system on
cluster machines, storing library and config files of arbitrary formats such
as .whl or .txt centrally and providing secure access across workspaces to
it, or transforming and querying non-tabular data files in ETL.
}
\section{Operations}{

\tabular{ll}{
\link[=volumes_create]{create} \tab Create a Volume.\cr
\link[=volumes_delete]{delete} \tab Delete a Volume.\cr
\link[=volumes_list]{list} \tab List Volumes.\cr
\link[=volumes_read]{read} \tab Get a Volume.\cr
\link[=volumes_update]{update} \tab Update a Volume.\cr
}
}

\keyword{datasets}
