% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volumes.R
\name{volumes_list}
\alias{volumes_list}
\title{List Volumes.}
\usage{
volumes_list(catalog_name, schema_name)
}
\arguments{
\item{catalog_name}{Required. The identifier of the catalog.}

\item{schema_name}{Required. The identifier of the schema.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Gets an array of all volumes for the current metastore under the parent
catalog and schema.
}
\details{
The returned volumes are filtered based on the privileges of the calling
user. For example, the metastore admin is able to list all the volumes. A
regular user needs to be the owner or have the \strong{READ VOLUME} privilege on
the volume to recieve the volumes in the response. For the latter case, the
caller must also be the owner or have the \strong{USE_CATALOG} privilege on the
parent catalog and the \strong{USE_SCHEMA} privilege on the parent schema.

There is no guarantee of a specific ordering of the elements in the array.
}
\keyword{internal}
