% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volumes.R
\name{volumes_update}
\alias{volumes_update}
\title{Update a Volume.}
\usage{
volumes_update(full_name_arg, comment = NULL, name = NULL, owner = NULL)
}
\arguments{
\item{full_name_arg}{Required. The three-level (fully qualified) name of the volume.}

\item{comment}{The comment attached to the volume.}

\item{name}{The name of the volume.}

\item{owner}{The identifier of the user who owns the volume.}
}
\description{
Updates the specified volume under the specified parent catalog and schema.
}
\details{
The caller must be a metastore admin or an owner of the volume. For the
latter case, the caller must also be the owner or have the \strong{USE_CATALOG}
privilege on the parent catalog and the \strong{USE_SCHEMA} privilege on the
parent schema.

Currently only the name, the owner or the comment of the volume could be
updated.
}
\keyword{internal}
