% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volumes.R
\name{volumesCreate}
\alias{volumesCreate}
\title{Create a Volume.}
\usage{
volumesCreate(
  client,
  catalog_name,
  name,
  schema_name,
  volume_type,
  comment = NULL,
  storage_location = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{catalog_name}{Required. The name of the catalog where the schema and the volume are.}

\item{name}{Required. The name of the volume.}

\item{schema_name}{Required. The name of the schema where the volume is.}

\item{volume_type}{Required.}

\item{comment}{The comment attached to the volume.}

\item{storage_location}{The storage location on the cloud.}
}
\description{
Creates a new volume.
}
\details{
The user could create either an external volume or a managed volume. An
external volume will be created in the specified external location, while a
managed volume will be located in the default location which is specified by
the parent schema, or the parent catalog, or the Metastore.

For the volume creation to succeed, the user must satisfy following
conditions: - The caller must be a metastore admin, or be the owner of the
parent catalog and schema, or have the \strong{USE_CATALOG} privilege on the
parent catalog and the \strong{USE_SCHEMA} privilege on the parent schema. - The
caller must have \strong{CREATE VOLUME} privilege on the parent schema.

For an external volume, following conditions also need to satisfy - The
caller must have \strong{CREATE EXTERNAL VOLUME} privilege on the external
location. - There are no other tables, nor volumes existing in the specified
storage location. - The specified storage location is not under the location
of other tables, nor volumes, or catalogs or schemas.
}
