% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volumes.R
\name{volumesRead}
\alias{volumesRead}
\title{Get a Volume.}
\usage{
volumesRead(client, full_name_arg)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{full_name_arg}{Required. The three-level (fully qualified) name of the volume.}
}
\description{
Gets a volume from the metastore for a specific catalog and schema.
}
\details{
The caller must be a metastore admin or an owner of (or have the \strong{READ
VOLUME} privilege on) the volume. For the latter case, the caller must also
be the owner or have the \strong{USE_CATALOG} privilege on the parent catalog and
the \strong{USE_SCHEMA} privilege on the parent schema.
}
