% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouses.R
\docType{data}
\name{warehouses}
\alias{warehouses}
\title{A SQL warehouse is a compute resource that lets you run SQL commands on data
objects within Databricks SQL. Compute resources are infrastructure resources
that provide processing capabilities in the cloud.}
\format{
An object of class \code{list} of length 9.
}
\usage{
warehouses
}
\description{
A SQL warehouse is a compute resource that lets you run SQL commands on data
objects within Databricks SQL. Compute resources are infrastructure resources
that provide processing capabilities in the cloud.
}
\section{Operations}{

\tabular{ll}{
\link[=warehouses_create]{create} \tab Create a warehouse.\cr
\link[=warehouses_delete]{delete} \tab Delete a warehouse.\cr
\link[=warehouses_edit]{edit} \tab Update a warehouse.\cr
\link[=warehouses_get]{get} \tab Get warehouse info.\cr
\link[=warehouses_get_workspace_warehouse_config]{get_workspace_warehouse_config} \tab Get the workspace configuration.\cr
\link[=warehouses_list]{list} \tab List warehouses.\cr
\link[=warehouses_set_workspace_warehouse_config]{set_workspace_warehouse_config} \tab Set the workspace configuration.\cr
\link[=warehouses_start]{start} \tab Start a warehouse.\cr
\link[=warehouses_stop]{stop} \tab Stop a warehouse.\cr
}
}

\keyword{datasets}
