% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouses.R
\name{warehouses_create}
\alias{warehouses_create}
\title{Create a warehouse.}
\usage{
warehouses_create(
  auto_stop_mins = NULL,
  channel = NULL,
  cluster_size = NULL,
  creator_name = NULL,
  enable_photon = NULL,
  enable_serverless_compute = NULL,
  instance_profile_arn = NULL,
  max_num_clusters = NULL,
  min_num_clusters = NULL,
  name = NULL,
  spot_instance_policy = NULL,
  tags = NULL,
  warehouse_type = NULL,
  timeout = 20,
  callback = cli_reporter
)
}
\arguments{
\item{auto_stop_mins}{The amount of time in minutes that a SQL Endpoint must be idle (i.e., no RUNNING queries) before it is automatically stopped.}

\item{channel}{Channel Details.}

\item{cluster_size}{Size of the clusters allocated for this endpoint.}

\item{creator_name}{endpoint creator name.}

\item{enable_photon}{Configures whether the endpoint should use Photon optimized clusters.}

\item{enable_serverless_compute}{Configures whether the endpoint should use Serverless Compute (aka Nephos) Defaults to value in global endpoint settings.}

\item{instance_profile_arn}{Deprecated.}

\item{max_num_clusters}{Maximum number of clusters that the autoscaler will create to handle concurrent queries.}

\item{min_num_clusters}{Minimum number of available clusters that will be maintained for this SQL Endpoint.}

\item{name}{Logical name for the cluster.}

\item{spot_instance_policy}{Configurations whether the warehouse should use spot instances.}

\item{tags}{A set of key-value pairs that will be tagged on all resources (e.g., AWS instances and EBS volumes) associated with this SQL Endpoints.}

\item{warehouse_type}{}
}
\description{
This is a long-running operation, which blocks until Warehouses on Databricks reach
RUNNING state with the timeout of 20 minutes, that you can change via \code{timeout} parameter.
By default, the state of Databricks Warehouses is reported to console. You can change this behavior
by changing the \code{callback} parameter.
}
\details{
Creates a new SQL warehouse.
}
\keyword{internal}
