% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouses.R
\name{warehouses_delete}
\alias{warehouses_delete}
\title{Delete a warehouse.}
\usage{
warehouses_delete(id)
}
\arguments{
\item{id}{Required. Required.}
}
\description{
Deletes a SQL warehouse.
}
\keyword{internal}
