% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouses.R
\name{warehouses_get}
\alias{warehouses_get}
\title{Get warehouse info.}
\usage{
warehouses_get(id, ...)
}
\arguments{
\item{id}{\link{required} Required.}
}
\description{
Gets the information for a single SQL warehouse.
}
\keyword{internal}
