% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouses.R
\name{warehouses_get_workspace_warehouse_config}
\alias{warehouses_get_workspace_warehouse_config}
\title{Get the workspace configuration.}
\usage{
warehouses_get_workspace_warehouse_config()
}
\description{
Gets the workspace level configuration that is shared by all SQL warehouses
in a workspace.#'
}
\keyword{internal}
