% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouses.R
\name{warehouses_list}
\alias{warehouses_list}
\title{List warehouses.}
\usage{
warehouses_list(run_as_user_id = NULL, ...)
}
\arguments{
\item{run_as_user_id}{Service Principal which will be used to fetch the list of endpoints.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Lists all SQL warehouses that a user has manager permissions on.
}
\keyword{internal}
