% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouses.R
\name{warehouses_stop}
\alias{warehouses_stop}
\title{Stop a warehouse.}
\usage{
warehouses_stop(id, timeout = 20, callback = cli_reporter)
}
\arguments{
\item{id}{Required. Required.}
}
\description{
This is a long-running operation, which blocks until Warehouses on Databricks reach
STOPPED state with the timeout of 20 minutes, that you can change via \code{timeout} parameter.
By default, the state of Databricks Warehouses is reported to console. You can change this behavior
by changing the \code{callback} parameter.
}
\details{
Stops a SQL warehouse.
}
\keyword{internal}
