% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouses.R
\name{warehousesGetPermissions}
\alias{warehousesGetPermissions}
\title{Get SQL warehouse permissions.}
\usage{
warehousesGetPermissions(client, warehouse_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{warehouse_id}{Required. The SQL warehouse for which to get or manage permissions.}
}
\description{
Gets the permissions of a SQL warehouse. SQL warehouses can inherit
permissions from their root object.
}
