% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouses.R
\name{warehousesGetWarehousePermissionLevels}
\alias{warehousesGetWarehousePermissionLevels}
\title{Get SQL warehouse permission levels.}
\usage{
warehousesGetWarehousePermissionLevels(client, warehouse_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{warehouse_id}{Required. The SQL warehouse for which to get or manage permissions.}
}
\description{
Gets the permission levels that a user can have on an object.
}
