% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouses.R
\name{warehousesStart}
\alias{warehousesStart}
\title{Start a warehouse.}
\usage{
warehousesStart(client, id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{id}{Required. Required.}
}
\description{
Starts a SQL warehouse.
}
