% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouses.R
\name{warehousesUpdateWarehousePermissions}
\alias{warehousesUpdateWarehousePermissions}
\title{Update SQL warehouse permissions.}
\usage{
warehousesUpdateWarehousePermissions(
  client,
  warehouse_id,
  access_control_list = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{warehouse_id}{Required. The SQL warehouse for which to get or manage permissions.}

\item{access_control_list}{}
}
\description{
Updates the permissions on a SQL warehouse. SQL warehouses can inherit
permissions from their root object.
}
