% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\docType{data}
\name{workspace}
\alias{workspace}
\title{The Workspace API allows you to list, import, export, and delete notebooks
and folders.}
\format{
An object of class \code{list} of length 6.
}
\usage{
workspace
}
\description{
A notebook is a web-based interface to a document that contains runnable
code, visualizations, and explanatory text.
}
\section{Operations}{

\tabular{ll}{
\link[=workspace_delete]{delete} \tab Delete a workspace object.\cr
\link[=workspace_export]{export} \tab Export a notebook.\cr
\link[=workspace_get_status]{get_status} \tab Get status.\cr
\link[=workspace_import]{import} \tab Import a notebook.\cr
\link[=workspace_list]{list} \tab List contents.\cr
\link[=workspace_mkdirs]{mkdirs} \tab Create a directory.\cr
}
}

\keyword{datasets}
