% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace_bindings.R
\docType{data}
\name{workspace_bindings}
\alias{workspace_bindings}
\title{A catalog in Databricks can be configured as \strong{OPEN} or \strong{ISOLATED}. An
\strong{OPEN} catalog can be accessed from any workspace, while an \strong{ISOLATED}
catalog can only be access from a configured list of workspaces.}
\format{
An object of class \code{list} of length 2.
}
\usage{
workspace_bindings
}
\description{
A catalog's workspace bindings can be configured by a metastore admin or the
owner of the catalog.
}
\section{Operations}{

\tabular{ll}{
\link[=workspace_bindings_get]{get} \tab Get catalog workspace bindings.\cr
\link[=workspace_bindings_update]{update} \tab Update catalog workspace bindings.\cr
}
}

\keyword{datasets}
