% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace_bindings.R
\name{workspace_bindings_get}
\alias{workspace_bindings_get}
\title{Get catalog workspace bindings.}
\usage{
workspace_bindings_get(name)
}
\arguments{
\item{name}{Required. The name of the catalog.}
}
\description{
Gets workspace bindings of the catalog. The caller must be a metastore admin
or an owner of the catalog.
}
\keyword{internal}
