% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace_bindings.R
\name{workspace_bindings_update}
\alias{workspace_bindings_update}
\title{Update catalog workspace bindings.}
\usage{
workspace_bindings_update(
  name,
  assign_workspaces = NULL,
  unassign_workspaces = NULL
)
}
\arguments{
\item{name}{Required. The name of the catalog.}

\item{assign_workspaces}{A list of workspace IDs.}

\item{unassign_workspaces}{A list of workspace IDs.}
}
\description{
Updates workspace bindings of the catalog. The caller must be a metastore
admin or an owner of the catalog.
}
\keyword{internal}
