% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace_conf.R
\docType{data}
\name{workspace_conf}
\alias{workspace_conf}
\title{This API allows updating known workspace settings for advanced users.}
\format{
An object of class \code{list} of length 2.
}
\usage{
workspace_conf
}
\description{
This API allows updating known workspace settings for advanced users.
}
\section{Operations}{

\tabular{ll}{
\link[=workspace_conf_get_status]{get_status} \tab Check configuration status.\cr
\link[=workspace_conf_set_status]{set_status} \tab Enable/disable features.\cr
}
}

\keyword{datasets}
