% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace_conf.R
\name{workspace_conf_get_status}
\alias{workspace_conf_get_status}
\title{Check configuration status.}
\usage{
workspace_conf_get_status(keys, ...)
}
\arguments{
\item{keys}{Required.}
}
\description{
Gets the configuration status for a workspace.
}
\keyword{internal}
