% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace_conf.R
\name{workspace_conf_set_status}
\alias{workspace_conf_set_status}
\title{Enable/disable features.}
\usage{
workspace_conf_set_status()
}
\description{
Sets the configuration status for a workspace, including enabling or
disabling it.
}
\keyword{internal}
