% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{workspace_get_status}
\alias{workspace_get_status}
\title{Get status.}
\usage{
workspace_get_status(path, ...)
}
\arguments{
\item{path}{Required. The absolute path of the notebook or directory.}
}
\description{
Gets the status of an object or a directory. If \code{path} does not exist, this
call returns an error \code{RESOURCE_DOES_NOT_EXIST}.
}
\keyword{internal}
