% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{workspace_import}
\alias{workspace_import}
\title{Import a notebook.}
\usage{
workspace_import(
  path,
  content = NULL,
  format = NULL,
  language = NULL,
  overwrite = NULL
)
}
\arguments{
\item{path}{Required. The absolute path of the notebook or directory.}

\item{content}{The base64-encoded content.}

\item{format}{This specifies the format of the file to be imported.}

\item{language}{The language of the object.}

\item{overwrite}{The flag that specifies whether to overwrite existing object.}
}
\description{
Imports a notebook or the contents of an entire directory. If \code{path} already
exists and \code{overwrite} is set to \code{false}, this call returns an error
\code{RESOURCE_ALREADY_EXISTS}. One can only use \code{DBC} format to import a
directory.
}
\keyword{internal}
