% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{workspace_list}
\alias{workspace_list}
\title{List contents.}
\usage{
workspace_list(path, notebooks_modified_after = NULL, ...)
}
\arguments{
\item{path}{Required. The absolute path of the notebook or directory.}

\item{notebooks_modified_after}{\if{html}{\out{<content needed>}}.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Lists the contents of a directory, or the object if it is not a directory.If
the input path does not exist, this call returns an error
\code{RESOURCE_DOES_NOT_EXIST}.
}
\keyword{internal}
