% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace_bindings.R
\name{workspaceBindingsUpdateBindings}
\alias{workspaceBindingsUpdateBindings}
\title{Update securable workspace bindings.}
\usage{
workspaceBindingsUpdateBindings(
  client,
  securable_type,
  securable_name,
  add = NULL,
  remove = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{securable_type}{Required. The type of the securable.}

\item{securable_name}{Required. The name of the securable.}

\item{add}{List of workspace bindings.}

\item{remove}{List of workspace bindings.}
}
\description{
Updates workspace bindings of the securable. The caller must be a metastore
admin or an owner of the securable.
}
