% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace_conf.R
\name{workspaceConfGetStatus}
\alias{workspaceConfGetStatus}
\title{Check configuration status.}
\usage{
workspaceConfGetStatus(client, keys)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{keys}{Required. This field has no description yet.}
}
\description{
Gets the configuration status for a workspace.
}
