% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace_conf.R
\name{workspaceConfSetStatus}
\alias{workspaceConfSetStatus}
\title{Enable/disable features.}
\usage{
workspaceConfSetStatus(client)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}
}
\description{
Sets the configuration status for a workspace, including enabling or
disabling it.
}
