% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{workspaceExport}
\alias{workspaceExport}
\title{Export a workspace object.}
\usage{
workspaceExport(client, path, format = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{path}{Required. The absolute path of the object or directory.}

\item{format}{This specifies the format of the exported file.}
}
\description{
Exports an object or the contents of an entire directory.
}
\details{
If \code{path} does not exist, this call returns an error
\code{RESOURCE_DOES_NOT_EXIST}.

If the exported data would exceed size limit, this call returns
\code{MAX_NOTEBOOK_SIZE_EXCEEDED}. Currently, this API does not support exporting
a library.
}
