% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{workspaceGetPermissions}
\alias{workspaceGetPermissions}
\title{Get workspace object permissions.}
\usage{
workspaceGetPermissions(client, workspace_object_type, workspace_object_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{workspace_object_type}{Required. The workspace object type for which to get or manage permissions.}

\item{workspace_object_id}{Required. The workspace object for which to get or manage permissions.}
}
\description{
Gets the permissions of a workspace object. Workspace objects can inherit
permissions from their parent objects or root object.
}
