% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{workspaceList}
\alias{workspaceList}
\title{List contents.}
\usage{
workspaceList(client, path, notebooks_modified_after = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{path}{Required. The absolute path of the notebook or directory.}

\item{notebooks_modified_after}{UTC timestamp in milliseconds.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Lists the contents of a directory, or the object if it is not a directory. If
the input path does not exist, this call returns an error
\code{RESOURCE_DOES_NOT_EXIST}.
}
