% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfs.R
\name{add_dbfs_block}
\alias{add_dbfs_block}
\title{Append data block.}
\usage{
add_dbfs_block(client, handle, data)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{handle}{Required. The handle on an open stream.}

\item{data}{Required. The base64-encoded data to append to the stream.}
}
\description{
Appends a block of data to the stream specified by the input handle. If the
handle does not exist, this call will throw an exception with
\code{RESOURCE_DOES_NOT_EXIST}.
}
\details{
If the block of data exceeds 1 MB, this call will throw an exception with
\code{MAX_BLOCK_SIZE_EXCEEDED}.
}
