% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{alertsUpdate}
\alias{alertsUpdate}
\title{Update an alert.}
\usage{
alertsUpdate(client, alert_id, name, options, query_id, rearm = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{alert_id}{Required. This field has no description yet.}

\item{name}{Required. Name of the alert.}

\item{options}{Required. Alert configuration options.}

\item{query_id}{Required. Query ID.}

\item{rearm}{Number of seconds after being triggered before the alert rearms itself and can be triggered again.}
}
\description{
Updates an alert.
}
