% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{approve_model_transition_request}
\alias{approve_model_transition_request}
\alias{modelRegistryApproveTransitionRequest}
\title{Approve transition request.}
\usage{
approve_model_transition_request(
  client,
  name,
  version,
  stage,
  archive_existing_versions,
  comment = NULL
)

modelRegistryApproveTransitionRequest(
  client,
  name,
  version,
  stage,
  archive_existing_versions,
  comment = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Name of the model.}

\item{version}{Required. Version of the model.

See the \href{https://docs.databricks.com/api/workspace/modelregistry/approvetransitionrequest}{API documentation} for more
information on the request and response field structure.}

\item{stage}{Required. Target stage of the transition.}

\item{archive_existing_versions}{Required. Specifies whether to archive all current model versions in the target stage.}

\item{comment}{User-provided comment on the action.}
}
\description{
Approves a model version stage transition request.
}
