% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{change_cluster_owner}
\alias{change_cluster_owner}
\title{Change cluster owner.}
\usage{
change_cluster_owner(client, cluster_id, owner_username)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{cluster_id}{Required. \if{html}{\out{<needs content added>}}.}

\item{owner_username}{Required. New owner of the cluster_id after this RPC.}
}
\description{
Change the owner of the cluster. You must be an admin and the cluster must be
terminated to perform this operation. The service principal application ID
can be supplied as an argument to \code{owner_username}.
}
