% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_rooms.R
\name{cleanRoomsDelete}
\alias{cleanRoomsDelete}
\title{Delete a clean room.}
\usage{
cleanRoomsDelete(client, name_arg)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name_arg}{Required. The name of the clean room.}
}
\description{
Deletes a data object clean room from the metastore. The caller must be an
owner of the clean room.
}
