% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{clusters_list}
\alias{clusters_list}
\title{List all clusters.}
\usage{
clusters_list(can_use_client = NULL, ...)
}
\arguments{
\item{can_use_client}{Filter clusters based on what type of client it can be used for.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Returns information about all pinned clusters, currently active clusters, up
to 70 of the most recently terminated interactive clusters in the past 7
days, and up to 30 of the most recently terminated job clusters in the past 7
days.
}
\details{
For example, if there is 1 pinned cluster, 4 active clusters, 45 terminated
interactive clusters in the past 7 days, and 50 terminated job clusters in
the past 7 days, then this API returns the 1 pinned cluster, 4 active
clusters, all 45 terminated interactive clusters, and the 30 most recently
terminated job clusters.
}
\keyword{internal}
