% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command_execution.R
\name{command_execution_command_status}
\alias{command_execution_command_status}
\title{Get command info.}
\usage{
command_execution_command_status(cluster_id, context_id, command_id)
}
\arguments{
\item{cluster_id}{Required.}

\item{context_id}{Required.}

\item{command_id}{Required.}
}
\description{
Gets the status of and, if available, the results from a currently executing
command.
}
\details{
The command ID is obtained from a prior successful call to \strong{execute}.
}
\keyword{internal}
