% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{configure}
\alias{configure}
\title{Override default authentication credentials for Databricks SDK for R}
\usage{
configure(...)
}
\description{
Override default authentication credentials for Databricks SDK for R
}
