% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{connectionsCreate}
\alias{connectionsCreate}
\title{Create a connection.}
\usage{
connectionsCreate(
  client,
  name,
  connection_type,
  options_kvpairs,
  comment = NULL,
  owner = NULL,
  properties_kvpairs = NULL,
  read_only = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Name of the connection.}

\item{connection_type}{Required. The type of connection.}

\item{options_kvpairs}{Required. Object properties as map of string key-value pairs.}

\item{comment}{User-provided free-form text description.}

\item{owner}{Username of current owner of the connection.}

\item{properties_kvpairs}{An object containing map of key-value properties attached to the connection.}

\item{read_only}{If the connection is read only.}
}
\description{
Creates a new connection
}
\details{
Creates a new connection to an external data source. It allows users to
specify connection details and configurations for interaction with the
external server.
}
