% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{connectionsDelete}
\alias{connectionsDelete}
\title{Delete a connection.}
\usage{
connectionsDelete(client, name_arg)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name_arg}{Required. The name of the connection to be deleted.}
}
\description{
Deletes the connection that matches the supplied name.
}
