% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{connectionsGet}
\alias{connectionsGet}
\title{Get a connection.}
\usage{
connectionsGet(client, name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Name of the connection.}
}
\description{
Gets a connection from it's name.
}
