% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{connectionsUpdate}
\alias{connectionsUpdate}
\title{Update a connection.}
\usage{
connectionsUpdate(client, name, options_kvpairs, name_arg)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Name of the connection.}

\item{options_kvpairs}{Required. A map of key-value properties attached to the securable.}

\item{name_arg}{Required. Name of the connection.}
}
\description{
Updates the connection that matches the supplied name.
}
