% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command_execution.R
\name{context_command_execution_status}
\alias{context_command_execution_status}
\title{Get status.}
\usage{
context_command_execution_status(client, cluster_id, context_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{cluster_id}{Required. This field has no description yet.}

\item{context_id}{Required. This field has no description yet.}
}
\description{
Gets the status for an execution context.
}
