% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_tables.R
\name{create_online_table}
\alias{create_online_table}
\alias{onlineTablesCreate}
\title{Create an Online Table.}
\usage{
create_online_table(client, name = NULL, spec = NULL)

onlineTablesCreate(client, name = NULL, spec = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Full three-part (catalog, schema, table) name of the table.}

\item{spec}{Specification of the online table.}
}
\description{
Create a new Online Table.
}
