% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shares.R
\name{create_share}
\alias{create_share}
\title{Create a share.}
\usage{
create_share(client, name, comment = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Name of the share.}

\item{comment}{User-provided free-form text description.}
}
\description{
Creates a new share for data objects. Data objects can be added after
creation with \strong{update}. The caller must be a metastore admin or have the
\strong{CREATE_SHARE} privilege on the metastore.
}
