% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metastores.R
\name{current_metastore}
\alias{current_metastore}
\title{Get metastore assignment for workspace.}
\usage{
current_metastore(client)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}
}
\description{
Gets the metastore assignment for the workspace being accessed.
}
