% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_policies.R
\name{delete_cluster_policy}
\alias{delete_cluster_policy}
\alias{clusterPoliciesDelete}
\title{Delete a cluster policy.}
\usage{
delete_cluster_policy(client, policy_id)

clusterPoliciesDelete(client, policy_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{policy_id}{Required. The ID of the policy to delete.}
}
\description{
Delete a policy for a cluster. Clusters governed by this policy can still
run, but cannot be edited.
}
