% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboards.R
\name{delete_dashboard}
\alias{delete_dashboard}
\title{Remove a dashboard.}
\usage{
delete_dashboard(client, dashboard_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{dashboard_id}{Required. This field has no description yet.}
}
\description{
Moves a dashboard to the trash. Trashed dashboards do not appear in list
views or searches, and cannot be shared.
}
