% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_widgets.R
\name{delete_dashboard_widget}
\alias{delete_dashboard_widget}
\title{Remove widget.}
\usage{
delete_dashboard_widget(client, id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{id}{Required. Widget ID returned by :method:dashboardwidgets/create.}
}
\description{
Remove widget.
}
