% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{delete_experiment_run}
\alias{delete_experiment_run}
\alias{experimentsDeleteRun}
\title{Delete a run.}
\usage{
delete_experiment_run(client, run_id)

experimentsDeleteRun(client, run_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{run_id}{Required. ID of the run to delete.

See the \href{https://docs.databricks.com/api/workspace/experiments/deleterun}{API documentation} for more
information on the request and response field structure.}
}
\description{
Marks a run for deletion.
}
