% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_locations.R
\name{delete_external_location}
\alias{delete_external_location}
\alias{externalLocationsDelete}
\title{Delete an external location.}
\usage{
delete_external_location(client, name, force = NULL)

externalLocationsDelete(client, name, force = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Name of the external location.

See the \href{https://docs.databricks.com/api/workspace/externallocations/delete}{API documentation} for more
information on the request and response field structure.}

\item{force}{Force deletion even if there are dependent external tables or mounts.}
}
\description{
Deletes the specified external location from the metastore. The caller must
be the owner of the external location.
}
