% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{delete_file_directory}
\alias{delete_file_directory}
\alias{filesDeleteDirectory}
\title{Delete a directory.}
\usage{
delete_file_directory(client, directory_path)

filesDeleteDirectory(client, directory_path)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{directory_path}{Required. The absolute path of a directory.}
}
\description{
Deletes an empty directory.
}
\details{
To delete a non-empty directory, first delete all of its contents. This can
be done by listing the directory contents and deleting each file and
subdirectory recursively.
}
