% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_credentials.R
\name{delete_git_credential}
\alias{delete_git_credential}
\title{Delete a credential.}
\usage{
delete_git_credential(client, credential_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{credential_id}{Required. The ID for the corresponding credential to access.}
}
\description{
Deletes the specified Git credential.
}
