% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{delete_model}
\alias{delete_model}
\alias{modelRegistryDeleteModel}
\title{Delete a model.}
\usage{
delete_model(client, name)

modelRegistryDeleteModel(client, name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Registered model unique name identifier.}
}
\description{
Deletes a registered model.
}
