% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{delete_model_comment}
\alias{delete_model_comment}
\alias{modelRegistryDeleteComment}
\title{Delete a comment.}
\usage{
delete_model_comment(client, id)

modelRegistryDeleteComment(client, id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{id}{Required. This field has no description yet.}
}
\description{
Deletes a comment on a model version.
}
