% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{delete_pipeline}
\alias{delete_pipeline}
\alias{pipelinesDelete}
\title{Delete a pipeline.}
\usage{
delete_pipeline(client, pipeline_id)

pipelinesDelete(client, pipeline_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{pipeline_id}{Required. This field has no description yet.

See the \href{https://docs.databricks.com/api/workspace/pipelines/delete}{API documentation} for more
information on the request and response field structure.}
}
\description{
Deletes a pipeline.
}
