% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{delete_repo}
\alias{delete_repo}
\alias{reposDelete}
\title{Delete a repo.}
\usage{
delete_repo(client, repo_id)

reposDelete(client, repo_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{repo_id}{Required. The ID for the corresponding repo to access.

See the \href{https://docs.databricks.com/api/workspace/repos/delete}{API documentation} for more
information on the request and response field structure.}
}
\description{
Deletes the specified repo.
}
