% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service_principals.R
\name{delete_service_principal}
\alias{delete_service_principal}
\title{Delete a service principal.}
\usage{
delete_service_principal(client, id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{id}{Required. Unique ID for a service principal in the Databricks workspace.}
}
\description{
Delete a single service principal in the Databricks workspace.
}
