% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serving_endpoints.R
\name{delete_serving_endpoint}
\alias{delete_serving_endpoint}
\alias{servingEndpointsDelete}
\title{Delete a serving endpoint.}
\usage{
delete_serving_endpoint(client, name)

servingEndpointsDelete(client, name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. The name of the serving endpoint.

See the \href{https://docs.databricks.com/api/workspace/servingendpoints/delete}{API documentation} for more
information on the request and response field structure.}
}
\description{
Delete a serving endpoint.
}
