% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_credentials.R
\name{delete_storage_credential}
\alias{delete_storage_credential}
\alias{storageCredentialsDelete}
\title{Delete a credential.}
\usage{
delete_storage_credential(client, name, force = NULL)

storageCredentialsDelete(client, name, force = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Name of the storage credential.}

\item{force}{Force deletion even if there are dependent external locations or external tables.}
}
\description{
Deletes a storage credential from the metastore. The caller must be an owner
of the storage credential.
}
