% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_search_endpoints.R
\name{delete_vector_search_endpoint}
\alias{delete_vector_search_endpoint}
\alias{vectorSearchEndpointsDeleteEndpoint}
\title{Delete an endpoint.}
\usage{
delete_vector_search_endpoint(client, endpoint_name)

vectorSearchEndpointsDeleteEndpoint(client, endpoint_name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{endpoint_name}{Required. Name of the endpoint.

See the \href{https://docs.databricks.com/api/workspace/vectorsearchendpoints/deleteendpoint}{API documentation} for more
information on the request and response field structure.}
}
\description{
Delete an endpoint.
}
