% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_search_indexes.R
\name{delete_vector_search_index}
\alias{delete_vector_search_index}
\title{Delete an index.}
\usage{
delete_vector_search_index(client, index_name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{index_name}{Required. Name of the index.}
}
\description{
Delete an index.
}
