% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouses.R
\name{delete_warehouse}
\alias{delete_warehouse}
\alias{warehousesDelete}
\title{Delete a warehouse.}
\usage{
delete_warehouse(client, id)

warehousesDelete(client, id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{id}{Required. Required.}
}
\description{
Deletes a SQL warehouse.
}
