% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command_execution.R
\name{destroy_command_execution}
\alias{destroy_command_execution}
\alias{commandExecutionDestroy}
\title{Delete an execution context.}
\usage{
destroy_command_execution(client, cluster_id, context_id)

commandExecutionDestroy(client, cluster_id, context_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{cluster_id}{Required. This field has no description yet.}

\item{context_id}{Required. This field has no description yet.

See the \href{https://docs.databricks.com/api/workspace/commandexecution/destroy}{API documentation} for more
information on the request and response field structure.}
}
\description{
Deletes an execution context.
}
