% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{get_alert}
\alias{get_alert}
\alias{alertsGet}
\title{Get an alert.}
\usage{
get_alert(client, alert_id)

alertsGet(client, alert_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{alert_id}{Required. This field has no description yet.}
}
\description{
Gets an alert.
}
