% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{get_app_deployment_status}
\alias{get_app_deployment_status}
\title{Get deployment status for an application.}
\usage{
get_app_deployment_status(client, deployment_id, include_app_log = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{deployment_id}{Required. The deployment id for an application.}

\item{include_app_log}{Boolean flag to include application logs.}
}
\description{
Get deployment status for an application
}
