% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{get_app_events}
\alias{get_app_events}
\alias{appsGetEvents}
\title{Get deployment events for an application.}
\usage{
get_app_events(client, name)

appsGetEvents(client, name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. The name of an application.

See the \href{https://docs.databricks.com/api/workspace/apps/getevents}{API documentation} for more
information on the request and response field structure.}
}
\description{
Get deployment events for an application
}
