% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatic_cluster_update.R
\name{get_automatic_cluster_update}
\alias{get_automatic_cluster_update}
\alias{automaticClusterUpdateGet}
\title{Get the automatic cluster update setting.}
\usage{
get_automatic_cluster_update(client, etag = NULL)

automaticClusterUpdateGet(client, etag = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{etag}{etag used for versioning.}
}
\description{
Gets the automatic cluster update setting.
}
