% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_policies.R
\name{get_cluster_policy_permissions}
\alias{get_cluster_policy_permissions}
\alias{clusterPoliciesGetPermissions}
\title{Get cluster policy permissions.}
\usage{
get_cluster_policy_permissions(client, cluster_policy_id)

clusterPoliciesGetPermissions(client, cluster_policy_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{cluster_policy_id}{Required. The cluster policy for which to get or manage permissions.

See the \href{https://docs.databricks.com/api/workspace/clusterpolicies/getpermissions}{API documentation} for more
information on the request and response field structure.}
}
\description{
Gets the permissions of a cluster policy. Cluster policies can inherit
permissions from their root object.
}
