% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_credentials.R
\name{get_git_credential}
\alias{get_git_credential}
\alias{gitCredentialsGet}
\title{Get a credential entry.}
\usage{
get_git_credential(client, credential_id)

gitCredentialsGet(client, credential_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{credential_id}{Required. The ID for the corresponding credential to access.

See the \href{https://docs.databricks.com/api/workspace/gitcredentials/get}{API documentation} for more
information on the request and response field structure.}
}
\description{
Gets the Git credential with the specified credential ID.
}
