% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{get_group}
\alias{get_group}
\alias{groupsGet}
\title{Get group details.}
\usage{
get_group(client, id)

groupsGet(client, id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{id}{Required. Unique ID for a group in the Databricks workspace.

See the \href{https://docs.databricks.com/api/workspace/groups/get}{API documentation} for more
information on the request and response field structure.}
}
\description{
Gets the information for a specific group in the Databricks workspace.
}
