% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip_access_lists.R
\name{get_ip_access_list}
\alias{get_ip_access_list}
\title{Get access list.}
\usage{
get_ip_access_list(client, ip_access_list_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{ip_access_list_id}{Required. The ID for the corresponding IP access list.}
}
\description{
Gets an IP access list, specified by its list ID.
}
