% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{get_job_permission_levels}
\alias{get_job_permission_levels}
\title{Get job permission levels.}
\usage{
get_job_permission_levels(client, job_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{job_id}{Required. The job for which to get or manage permissions.}
}
\description{
Gets the permission levels that a user can have on an object.
}
