% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{get_permission}
\alias{get_permission}
\alias{permissionsGet}
\title{Get object permissions.}
\usage{
get_permission(client, request_object_type, request_object_id)

permissionsGet(client, request_object_type, request_object_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{request_object_type}{Required. The type of the request object.}

\item{request_object_id}{Required. The id of the request object.}
}
\description{
Gets the permissions of an object. Objects can inherit permissions from their
parent objects or root object.
}
