% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipients.R
\name{get_recipient}
\alias{get_recipient}
\alias{recipientsGet}
\title{Get a share recipient.}
\usage{
get_recipient(client, name)

recipientsGet(client, name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Name of the recipient.}
}
\description{
Gets a share recipient from the metastore if:
}
\details{
\itemize{
\item the caller is the owner of the share recipient, or: * is a metastore admin
}
}
