% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipient_activation.R
\name{get_recipient_activation_url_info}
\alias{get_recipient_activation_url_info}
\title{Get a share activation URL.}
\usage{
get_recipient_activation_url_info(client, activation_url)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{activation_url}{Required. The one time activation url.}
}
\description{
Gets an activation URL for a share.
}
