% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{get_repo_permissions}
\alias{get_repo_permissions}
\alias{reposGetPermissions}
\title{Get repo permissions.}
\usage{
get_repo_permissions(client, repo_id)

reposGetPermissions(client, repo_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{repo_id}{Required. The repo for which to get or manage permissions.

See the \href{https://docs.databricks.com/api/workspace/repos/getpermissions}{API documentation} for more
information on the request and response field structure.}
}
\description{
Gets the permissions of a repo. Repos can inherit permissions from their root
object.
}
