% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrict_workspace_admins.R
\name{get_restrict_workspace_admin}
\alias{get_restrict_workspace_admin}
\alias{restrictWorkspaceAdminsGet}
\title{Get the restrict workspace admins setting.}
\usage{
get_restrict_workspace_admin(client, etag = NULL)

restrictWorkspaceAdminsGet(client, etag = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{etag}{etag used for versioning.

See the \href{https://docs.databricks.com/api/workspace/restrictworkspaceadmins/get}{API documentation} for more
information on the request and response field structure.}
}
\description{
Gets the restrict workspace admins setting.
}
