% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serving_endpoints.R
\name{get_serving_endpoint}
\alias{get_serving_endpoint}
\title{Get a single serving endpoint.}
\usage{
get_serving_endpoint(client, name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. The name of the serving endpoint.}
}
\description{
Retrieves the details for a single serving endpoint.
}
