% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shares.R
\name{get_share}
\alias{get_share}
\title{Get a share.}
\usage{
get_share(client, name, include_shared_data = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. The name of the share.}

\item{include_shared_data}{Query for data to include in the share.}
}
\description{
Gets a data object share from the metastore. The caller must be a metastore
admin or the owner of the share.
}
