% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_management.R
\name{get_token_management_permissions}
\alias{get_token_management_permissions}
\title{Get token permissions.}
\usage{
get_token_management_permissions(client)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}
}
\description{
Gets the permissions of all tokens. Tokens can inherit permissions from their
root object.
}
