% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{get_user_permission_levels}
\alias{get_user_permission_levels}
\title{Get password permission levels.}
\usage{
get_user_permission_levels(client)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}
}
\description{
Gets the permission levels that a user can have on an object.
}
