% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{get_user_permissions}
\alias{get_user_permissions}
\alias{usersGetPermissions}
\title{Get password permissions.}
\usage{
get_user_permissions(client)

usersGetPermissions(client)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()

See the \href{https://docs.databricks.com/api/workspace/users/getpermissions}{API documentation} for more
information on the request and response field structure.}
}
\description{
Gets the permissions of all passwords. Passwords can inherit permissions from
their root object.
}
