% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_search_indexes.R
\name{get_vector_search_index}
\alias{get_vector_search_index}
\alias{vectorSearchIndexesGetIndex}
\title{Get an index.}
\usage{
get_vector_search_index(client, index_name)

vectorSearchIndexesGetIndex(client, index_name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{index_name}{Required. Name of the index.

See the \href{https://docs.databricks.com/api/workspace/vectorsearchindexes/getindex}{API documentation} for more
information on the request and response field structure.}
}
\description{
Get an index.
}
