% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouses.R
\name{get_warehouse}
\alias{get_warehouse}
\title{Get warehouse info.}
\usage{
get_warehouse(client, id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{id}{Required. Required.}
}
\description{
Gets the information for a single SQL warehouse.
}
