% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouses.R
\name{get_warehouse_permission_levels}
\alias{get_warehouse_permission_levels}
\alias{warehousesGetPermissionLevels}
\title{Get SQL warehouse permission levels.}
\usage{
get_warehouse_permission_levels(client, warehouse_id)

warehousesGetPermissionLevels(client, warehouse_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{warehouse_id}{Required. The SQL warehouse for which to get or manage permissions.}
}
\description{
Gets the permission levels that a user can have on an object.
}
