% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace_bindings.R
\name{get_workspace_binding}
\alias{get_workspace_binding}
\title{Get catalog workspace bindings.}
\usage{
get_workspace_binding(client, name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. The name of the catalog.}
}
\description{
Gets workspace bindings of the catalog. The caller must be a metastore admin
or an owner of the catalog.
}
