% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{groupsCreate}
\alias{groupsCreate}
\title{Create a new group.}
\usage{
groupsCreate(
  client,
  display_name = NULL,
  entitlements = NULL,
  external_id = NULL,
  groups = NULL,
  id = NULL,
  members = NULL,
  meta = NULL,
  roles = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{display_name}{String that represents a human-readable group name.}

\item{entitlements}{}

\item{external_id}{}

\item{groups}{}

\item{id}{Databricks group ID.}

\item{members}{}

\item{meta}{Container for the group identifier.}

\item{roles}{}
}
\description{
Creates a group in the Databricks workspace with a unique name, using the
supplied group details.
}
