% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{groupsUpdate}
\alias{groupsUpdate}
\title{Replace a group.}
\usage{
groupsUpdate(
  client,
  id,
  display_name = NULL,
  entitlements = NULL,
  external_id = NULL,
  groups = NULL,
  members = NULL,
  meta = NULL,
  roles = NULL,
  schemas = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{id}{Databricks group ID.}

\item{display_name}{String that represents a human-readable group name.}

\item{entitlements}{Entitlements assigned to the group.}

\item{external_id}{This field has no description yet.}

\item{groups}{This field has no description yet.}

\item{members}{This field has no description yet.}

\item{meta}{Container for the group identifier.}

\item{roles}{Corresponds to AWS instance profile/arn role.}

\item{schemas}{The schema of the group.}
}
\description{
Updates the details of a group by replacing the entire group entity.
}
