% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip_access_lists.R
\name{ipAccessListsList}
\alias{ipAccessListsList}
\title{Get access lists.}
\usage{
ipAccessListsList(client)
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Gets all IP access lists for the specified workspace.#'
}
