% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobsGetRun}
\alias{jobsGetRun}
\title{Get a single job run.}
\usage{
jobsGetRun(
  client,
  run_id,
  include_history = NULL,
  include_resolved_values = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{run_id}{Required. The canonical identifier of the run for which to retrieve the metadata.}

\item{include_history}{Whether to include the repair history in the response.}

\item{include_resolved_values}{Whether to include resolved parameter values in the response.}
}
\description{
Retrieve the metadata of a run.
}
