% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobsReset}
\alias{jobsReset}
\title{Update all job settings (reset).}
\usage{
jobsReset(client, job_id, new_settings)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{job_id}{Required. The canonical identifier of the job to reset.}

\item{new_settings}{Required. The new settings of the job.}
}
\description{
Overwrite all settings for the given job. Use the \href{:method:jobs/update}{\emph{Update} endpoint} to update job settings partially.
}
