% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{list_alerts}
\alias{list_alerts}
\alias{alertsList}
\title{Get alerts.}
\usage{
list_alerts(client)

alertsList(client)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}
}
\description{
Gets a list of alerts.
}
