% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{list_cluster_node_types}
\alias{list_cluster_node_types}
\title{List node types.}
\usage{
list_cluster_node_types(client)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}
}
\description{
Returns a list of supported Spark node types. These node types can be used to
launch a cluster.
}
