% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{list_cluster_zones}
\alias{list_cluster_zones}
\alias{clustersListZones}
\title{List availability zones.}
\usage{
list_cluster_zones(client)

clustersListZones(client)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}
}
\description{
Returns a list of availability zones where clusters can be created in (For
example, us-west-2a). These zones can be used to launch a cluster.
}
