% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{list_connections}
\alias{list_connections}
\alias{connectionsList}
\title{List connections.}
\usage{
list_connections(client)

connectionsList(client)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
List all connections.
}
