% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sources.R
\name{list_data_sources}
\alias{list_data_sources}
\title{Get a list of SQL warehouses.}
\usage{
list_data_sources(client)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}
}
\description{
Retrieves a full list of SQL warehouses available in this workspace. All
fields that appear in this API response are enumerated for clarity. However,
you need only a SQL warehouse's \code{id} to create new queries against it.
}
