% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{list_repos}
\alias{list_repos}
\alias{reposList}
\title{Get repos.}
\usage{
list_repos(client, next_page_token = NULL, path_prefix = NULL)

reposList(client, next_page_token = NULL, path_prefix = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{next_page_token}{Token used to get the next page of results.}

\item{path_prefix}{Filters repos that have paths starting with the given path prefix.}
}
\value{
\code{data.frame} with all of the response pages.

See the \href{https://docs.databricks.com/api/workspace/repos/list}{API documentation} for more
information on the request and response field structure.
}
\description{
Returns repos that the calling user has Manage permissions on. Results are
paginated with each page containing twenty repos.
}
