% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{list_tokens}
\alias{list_tokens}
\alias{tokensList}
\title{List tokens.}
\usage{
list_tokens(client)

tokensList(client)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Lists all the valid tokens for a user-workspace pair.
}
