% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_lflow_databricks.R
\name{m_lflow_databricks_transition_stage}
\alias{m_lflow_databricks_transition_stage}
\title{Transition a stage.}
\usage{
m_lflow_databricks_transition_stage(
  name,
  version,
  stage,
  archive_existing_versions,
  comment = NULL
)
}
\arguments{
\item{name}{Required. Name of the model.}

\item{version}{Required. Version of the model.}

\item{stage}{Required. Target stage of the transition.}

\item{archive_existing_versions}{Required. Specifies whether to archive all current model versions in the target stage.}

\item{comment}{User-provided comment on the action.}
}
\description{
Transition a model version's stage. This is a Databricks Workspace version of
the \link{MLflow endpoint} that also accepts a comment associated with the
transition to be recorded.',
}
\details{

}
\keyword{internal}
