% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_user.R
\name{me}
\alias{me}
\alias{currentUserMe}
\title{Get current user info.}
\usage{
me(client)

currentUserMe(client)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()

See the \href{https://docs.databricks.com/api/workspace/currentuser/me}{API documentation} for more
information on the request and response field structure.}
}
\description{
Get details about the current method caller's identity.
}
