% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metastores.R
\name{metastores_unassign}
\alias{metastores_unassign}
\title{Delete an assignment.}
\usage{
metastores_unassign(workspace_id, metastore_id, ...)
}
\arguments{
\item{workspace_id}{\link{required} A workspace ID.}

\item{metastore_id}{\link{required} Query for the ID of the metastore to delete.}
}
\description{
Deletes a metastore assignment. The caller must be an account administrator.
}
\keyword{internal}
