% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_versions.R
\name{model_versions_delete}
\alias{model_versions_delete}
\title{Delete a model version.}
\usage{
model_versions_delete(name, version, ...)
}
\arguments{
\item{name}{Required. Name of the registered model.}

\item{version}{Required. Model version number.}
}
\description{
Deletes a model version.
}
\keyword{internal}
