% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_versions.R
\name{model_versions_set_tag}
\alias{model_versions_set_tag}
\title{Set a version tag.}
\usage{
model_versions_set_tag(name, version, key, value)
}
\arguments{
\item{name}{Required. Unique name of the model.}

\item{version}{Required. Model version number.}

\item{key}{Required. Name of the tag.}

\item{value}{Required. String value of the tag being logged.}
}
\description{
Sets a model version tag.
}
\keyword{internal}
