% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_versions.R
\name{model_versions_transition_stage}
\alias{model_versions_transition_stage}
\title{Transition a stage.}
\usage{
model_versions_transition_stage(
  name,
  version,
  stage,
  archive_existing_versions,
  ...
)
}
\arguments{
\item{name}{Required. Name of the registered model.}

\item{version}{Required. Model version number.}

\item{stage}{Required. Transition \code{model_version} to new stage.}

\item{archive_existing_versions}{Required. When transitioning a model version to a particular stage, this flag dictates whether all existing model versions in that stage should be atomically moved to the 'archived' stage.}
}
\description{
Transition to the next model stage.
}
\keyword{internal}
