% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_versions.R
\name{model_versions_update}
\alias{model_versions_update}
\title{Update model version.}
\usage{
model_versions_update(name, version, description = NULL, ...)
}
\arguments{
\item{name}{\link{required} Name of the registered model.}

\item{version}{\link{required} Model version number.}

\item{description}{If provided, updates the description for this \code{registered_model}.}
}
\description{
Updates the model version.
}
\keyword{internal}
