% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{permissions_update}
\alias{permissions_update}
\title{Update permission.}
\usage{
permissions_update(
  request_object_type,
  request_object_id,
  access_control_list = NULL,
  ...
)
}
\arguments{
\item{request_object_type}{\link{required} \if{html}{\out{<needs content>}}.}

\item{request_object_id}{\link{required}}

\item{access_control_list}{}
}
\description{
Updates the permissions on an object.
}
\keyword{internal}
