% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{queriesUpdate}
\alias{queriesUpdate}
\title{Change a query definition.}
\usage{
queriesUpdate(
  client,
  query_id,
  data_source_id = NULL,
  description = NULL,
  name = NULL,
  options = NULL,
  query = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{query_id}{Required.}

\item{data_source_id}{Data source ID.}

\item{description}{General description that conveys additional information about this query such as usage notes.}

\item{name}{The title of this query that appears in list views, widget headings, and on the query page.}

\item{options}{Exclusively used for storing a list parameter definitions.}

\item{query}{The text of the query to be run.}
}
\description{
Modify this query definition.
}
\details{
\strong{Note}: You cannot undo this operation.
}
