% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registered_models.R
\name{registered_models_set_tag}
\alias{registered_models_set_tag}
\title{Set a tag.}
\usage{
registered_models_set_tag(name, key, value)
}
\arguments{
\item{name}{Required. Unique name of the model.}

\item{key}{Required. Name of the tag.}

\item{value}{Required. String value of the tag being logged.}
}
\description{
Sets a tag on a registered model.
}
\keyword{internal}
