% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{rename_model}
\alias{rename_model}
\title{Rename a model.}
\usage{
rename_model(client, name, new_name = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Registered model unique name identifier.}

\item{new_name}{If provided, updates the name for this \code{registered_model}.}
}
\description{
Renames a registered model.
}
