% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\name{schemasGet}
\alias{schemasGet}
\title{Get a schema.}
\usage{
schemasGet(client, full_name, include_browse = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{full_name}{Required. Full name of the schema.}

\item{include_browse}{Whether to include schemas in the response for which the principal can only access selective metadata for.}
}
\description{
Gets the specified schema within the metastore. The caller must be a
metastore admin, the owner of the schema, or a user that has the
\strong{USE_SCHEMA} privilege on the schema.
}
