% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{search_model_versions}
\alias{search_model_versions}
\title{Searches model versions.}
\usage{
search_model_versions(
  client,
  filter = NULL,
  max_results = NULL,
  order_by = NULL,
  page_token = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{filter}{String filter condition, like 'name='my-model-name''.}

\item{max_results}{Maximum number of models desired.}

\item{order_by}{List of columns to be ordered by including model name, version, stage with an optional 'DESC' or 'ASC' annotation, where 'ASC' is the default.}

\item{page_token}{Pagination token to go to next page based on previous search query.}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Searches for specific model versions based on the supplied \strong{filter}.
}
