% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serving_endpoints.R
\name{serving_endpoints_list}
\alias{serving_endpoints_list}
\title{Retrieve all serving endpoints.#'}
\usage{
serving_endpoints_list(...)
}
\description{
Retrieve all serving endpoints.#'
}
\keyword{internal}
