% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serving_endpoints.R
\name{servingEndpointsBuildLogs}
\alias{servingEndpointsBuildLogs}
\title{Retrieve the logs associated with building the model's environment for a
given serving endpoint's served model.}
\usage{
servingEndpointsBuildLogs(client, name, served_model_name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. The name of the serving endpoint that the served model belongs to.}

\item{served_model_name}{Required. The name of the served model that build logs will be retrieved for.}
}
\description{
Retrieves the build logs associated with the provided served model.
}
