% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbsql_permissions.R
\name{set_dbsql_permission}
\alias{set_dbsql_permission}
\title{Set object ACL.}
\usage{
set_dbsql_permission(
  client,
  object_type,
  object_id,
  access_control_list = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{object_type}{Required. The type of object permission to set.}

\item{object_id}{Required. Object ID.}

\item{access_control_list}{This field has no description yet.}
}
\description{
Sets the access control list (ACL) for a specified object. This operation
will complete rewrite the ACL.
}
