% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{set_experiment_permissions}
\alias{set_experiment_permissions}
\alias{experimentsSetPermissions}
\title{Set experiment permissions.}
\usage{
set_experiment_permissions(client, experiment_id, access_control_list = NULL)

experimentsSetPermissions(client, experiment_id, access_control_list = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{experiment_id}{Required. The experiment for which to get or manage permissions.

See the \href{https://docs.databricks.com/api/workspace/experiments/setpermissions}{API documentation} for more
information on the request and response field structure.}

\item{access_control_list}{This field has no description yet.}
}
\description{
Sets permissions on an experiment. Experiments can inherit permissions from
their root object.
}
