% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{set_user_permissions}
\alias{set_user_permissions}
\alias{usersSetPermissions}
\title{Set password permissions.}
\usage{
set_user_permissions(client, access_control_list = NULL)

usersSetPermissions(client, access_control_list = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{access_control_list}{This field has no description yet.}
}
\description{
Sets permissions on all passwords. Passwords can inherit permissions from
their root object.
}
