% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace_conf.R
\name{set_workspace_conf_status}
\alias{set_workspace_conf_status}
\alias{workspaceConfSetStatus}
\title{Enable/disable features.}
\usage{
set_workspace_conf_status(client)

workspaceConfSetStatus(client)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()

See the \href{https://docs.databricks.com/api/workspace/workspaceconf/setstatus}{API documentation} for more
information on the request and response field structure.}
}
\description{
Sets the configuration status for a workspace, including enabling or
disabling it.
}
