% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shares.R
\name{share_permissions}
\alias{share_permissions}
\alias{sharesSharePermissions}
\title{Get permissions.}
\usage{
share_permissions(client, name)

sharesSharePermissions(client, name)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. The name of the share.

See the \href{https://docs.databricks.com/api/workspace/shares/sharepermissions}{API documentation} for more
information on the request and response field structure.}
}
\description{
Gets the permissions for a data share from the metastore. The caller must be
a metastore admin or the owner of the share.
}
