% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{spark_cluster_versions}
\alias{spark_cluster_versions}
\alias{clustersSparkVersions}
\title{List available Spark versions.}
\usage{
spark_cluster_versions(client)

clustersSparkVersions(client)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}
}
\description{
Returns the list of available Spark versions. These versions can be used to
launch a cluster.
}
