% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_credentials.R
\name{storage_credentials_update}
\alias{storage_credentials_update}
\title{Update a credential.}
\usage{
storage_credentials_update(
  name,
  aws_iam_role = NULL,
  azure_managed_identity = NULL,
  azure_service_principal = NULL,
  comment = NULL,
  databricks_gcp_service_account = NULL,
  force = NULL,
  owner = NULL,
  read_only = NULL,
  skip_validation = NULL
)
}
\arguments{
\item{name}{The credential name.}

\item{aws_iam_role}{The AWS IAM role configuration.}

\item{azure_managed_identity}{The Azure managed identity configuration.}

\item{azure_service_principal}{The Azure service principal configuration.}

\item{comment}{Comment associated with the credential.}

\item{databricks_gcp_service_account}{The \if{html}{\out{<Databricks>}} managed GCP service account configuration.}

\item{force}{Force update even if there are dependent external locations or external tables.}

\item{owner}{Username of current owner of credential.}

\item{read_only}{Whether the storage credential is only usable for read operations.}

\item{skip_validation}{Supplying true to this argument skips validation of the updated credential.}
}
\description{
Updates a storage credential on the metastore. The caller must be the owner
of the storage credential or a metastore admin. If the caller is a metastore
admin, only the \strong{owner} credential can be changed.
}
\keyword{internal}
