% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metastores.R
\name{summary_metastore}
\alias{summary_metastore}
\title{Get a metastore summary.}
\usage{
summary_metastore(client)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}
}
\description{
Gets information about a metastore. This summary includes the storage
credential, the cloud vendor, the cloud region, and the global metastore ID.
}
