% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_schemas.R
\name{systemSchemasEnable}
\alias{systemSchemasEnable}
\title{Enable a system schema.}
\usage{
systemSchemasEnable(client)
}
\description{
Enables the system schema and adds it to the system catalog. The caller must
be an account admin or a metastore admin.#'
}
