% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tablesList}
\alias{tablesList}
\title{List tables.}
\usage{
tablesList(
  client,
  catalog_name,
  schema_name,
  include_delta_metadata = NULL,
  max_results = NULL,
  page_token = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{catalog_name}{Required. Name of parent catalog for tables of interest.}

\item{schema_name}{Required. Parent schema of tables.}

\item{include_delta_metadata}{Whether delta metadata should be included in the response.}

\item{max_results}{Maximum number of tables to return (page length).}

\item{page_token}{Opaque token to send for the next page of results (pagination).}
}
\value{
\code{data.frame} with all of the response pages.
}
\description{
Gets an array of all tables for the current metastore under the parent
catalog and schema. The caller must be a metastore admin or an owner of (or
have the \strong{SELECT} privilege on) the table. For the latter case, the caller
must also be the owner or have the \strong{USE_CATALOG} privilege on the parent
catalog and the \strong{USE_SCHEMA} privilege on the parent schema. There is no
guarantee of a specific ordering of the elements in the array.
}
