% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_management.R
\name{tokenManagementGetPermissionLevels}
\alias{tokenManagementGetPermissionLevels}
\title{Get token permission levels.}
\usage{
tokenManagementGetPermissionLevels(client)
}
\description{
Gets the permission levels that a user can have on an object.#'
}
