% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_management.R
\name{tokenManagementGetPermissions}
\alias{tokenManagementGetPermissions}
\title{Get token permissions.}
\usage{
tokenManagementGetPermissions(client)
}
\description{
Gets the permissions of all tokens. Tokens can inherit permissions from their
root object.#'
}
