% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeview.R
\name{trash_lakeview}
\alias{trash_lakeview}
\alias{lakeviewTrash}
\title{Trash dashboard.}
\usage{
trash_lakeview(client, dashboard_id)

lakeviewTrash(client, dashboard_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{dashboard_id}{Required. UUID identifying the dashboard.}
}
\description{
Trash a dashboard.
}
