% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metastores.R
\name{unassign_metastore}
\alias{unassign_metastore}
\alias{metastoresUnassign}
\title{Delete an assignment.}
\usage{
unassign_metastore(client, workspace_id, metastore_id)

metastoresUnassign(client, workspace_id, metastore_id)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{workspace_id}{Required. A workspace ID.

See the \href{https://docs.databricks.com/api/workspace/metastores/unassign}{API documentation} for more
information on the request and response field structure.}

\item{metastore_id}{Required. Query for the ID of the metastore to delete.}
}
\description{
Deletes a metastore assignment. The caller must be an account administrator.
}
