% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{update_experiment}
\alias{update_experiment}
\alias{experimentsUpdateExperiment}
\title{Update an experiment.}
\usage{
update_experiment(client, experiment_id, new_name = NULL)

experimentsUpdateExperiment(client, experiment_id, new_name = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{experiment_id}{Required. ID of the associated experiment.}

\item{new_name}{If provided, the experiment's name is changed to the new name.}
}
\description{
Updates experiment metadata.
}
