% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{update_experiment_permissions}
\alias{update_experiment_permissions}
\title{Update experiment permissions.}
\usage{
update_experiment_permissions(
  client,
  experiment_id,
  access_control_list = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{experiment_id}{Required. The experiment for which to get or manage permissions.}

\item{access_control_list}{This field has no description yet.}
}
\description{
Updates the permissions on an experiment. Experiments can inherit permissions
from their root object.
}
