% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grants.R
\name{update_grant}
\alias{update_grant}
\alias{grantsUpdate}
\title{Update permissions.}
\usage{
update_grant(client, securable_type, full_name, changes = NULL)

grantsUpdate(client, securable_type, full_name, changes = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{securable_type}{Required. Type of securable.

See the \href{https://docs.databricks.com/api/workspace/grants/update}{API documentation} for more
information on the request and response field structure.}

\item{full_name}{Required. Full name of securable.}

\item{changes}{Array of permissions change objects.}
}
\description{
Updates the permissions for a securable.
}
