% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{update_notebook_permissions}
\alias{update_notebook_permissions}
\alias{workspaceUpdatePermissions}
\title{Update workspace object permissions.}
\usage{
update_notebook_permissions(
  client,
  workspace_object_type,
  workspace_object_id,
  access_control_list = NULL
)

workspaceUpdatePermissions(
  client,
  workspace_object_type,
  workspace_object_id,
  access_control_list = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{workspace_object_type}{Required. The workspace object type for which to get or manage permissions.

See the \href{https://docs.databricks.com/api/workspace/workspace/updatepermissions}{API documentation} for more
information on the request and response field structure.}

\item{workspace_object_id}{Required. The workspace object for which to get or manage permissions.}

\item{access_control_list}{This field has no description yet.}
}
\description{
Updates the permissions on a workspace object. Workspace objects can inherit
permissions from their parent objects or root object.
}
