% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{update_repo_permissions}
\alias{update_repo_permissions}
\alias{reposUpdatePermissions}
\title{Update repo permissions.}
\usage{
update_repo_permissions(client, repo_id, access_control_list = NULL)

reposUpdatePermissions(client, repo_id, access_control_list = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{repo_id}{Required. The repo for which to get or manage permissions.}

\item{access_control_list}{This field has no description yet.}
}
\description{
Updates the permissions on a repo. Repos can inherit permissions from their
root object.
}
