% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace_bindings.R
\name{update_workspace_binding}
\alias{update_workspace_binding}
\alias{workspaceBindingsUpdate}
\title{Update catalog workspace bindings.}
\usage{
update_workspace_binding(
  client,
  name,
  assign_workspaces = NULL,
  unassign_workspaces = NULL
)

workspaceBindingsUpdate(
  client,
  name,
  assign_workspaces = NULL,
  unassign_workspaces = NULL
)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. The name of the catalog.}

\item{assign_workspaces}{A list of workspace IDs.}

\item{unassign_workspaces}{A list of workspace IDs.

See the \href{https://docs.databricks.com/api/workspace/workspacebindings/update}{API documentation} for more
information on the request and response field structure.}
}
\description{
Updates workspace bindings of the catalog. The caller must be a metastore
admin or an owner of the catalog.
}
