% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_search_endpoints.R
\name{vectorSearchEndpointsCreateEndpoint}
\alias{vectorSearchEndpointsCreateEndpoint}
\title{Create an endpoint.}
\usage{
vectorSearchEndpointsCreateEndpoint(client, name, endpoint_type)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{name}{Required. Name of endpoint.}

\item{endpoint_type}{Required. Type of endpoint.}
}
\description{
Create a new endpoint.
}
