% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouses.R
\name{warehousesGetWorkspaceWarehouseConfig}
\alias{warehousesGetWorkspaceWarehouseConfig}
\title{Get the workspace configuration.}
\usage{
warehousesGetWorkspaceWarehouseConfig(client)
}
\description{
Gets the workspace level configuration that is shared by all SQL warehouses
in a workspace.#'
}
