% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouses.R
\name{warehousesStop}
\alias{warehousesStop}
\title{Stop a warehouse.}
\usage{
warehousesStop(client, id, timeout = 20, callback = cli_reporter)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{id}{Required. Required.}
}
\description{
This is a long-running operation, which blocks until Warehouses on Databricks reach
STOPPED state with the timeout of 20 minutes, that you can change via \code{timeout} parameter.
By default, the state of Databricks Warehouses is reported to console. You can change this behavior
by changing the \code{callback} parameter.
}
\details{
Stops a SQL warehouse.
}
